### 2020-09-08 #2648 @whedon
[PRE REVIEW]: dymos: A Python package for optimal control of multidisciplinary systems
**Submitting author:** @robfalck (<a href="http://orcid.org/0000-0001-9864-4928">Rob Falck</a>)
**Repository:** <a href="https://github.com/OpenMDAO/dymos" target ="_blank">https://github.com/OpenMDAO/dymos</a>
**Version:** v0.16.0
