### 2020-09-08 8c2f554
Modified refinement to work with master updates
