+++ dymos/grid_refinement/hp_adaptive/hp_adaptive.py
@@ -5,6 +5,7 @@ from ...utils.lagrange import lagrange_matrices
 from ...utils.lgr import lgr
 from ...utils.lgl import lgl
 from ...utils.interpolate import LagrangeBarycentricInterpolant
+from ..error_estimation import interpolation_lagrange_matrix, eval_ode_on_grid
 
 from scipy.linalg import block_diag
 
@@ -13,84 +14,6 @@ import copy
 
 import openmdao.api as om
 import dymos as dm
-
-
-def interpolation_lagrange_matrices(old_grid, new_grid):
-    """
-    Evaluate lagrange matrix to interpolate state and control values from the solved grid onto the new grid
-
-    Parameters
-    ----------
-    old_grid: GridData
-        The GridData object representing the grid on which the problem has been solved
-    new_grid: GridData
-        The GridData object representing the new, higher-order grid
-
-    Returns
-    -------
-    L: np.ndarray
-        The lagrange interpolation matrix
-
-    D: np.ndarray
-        The lagrange differentiation matrix
-
-    """
-    L_blocks = []
-    D_blocks = []
-
-    for iseg in range(old_grid.num_segments):
-        i1, i2 = old_grid.subset_segment_indices['all'][iseg, :]
-        indices = old_grid.subset_node_indices['all'][i1:i2]
-        nodes_given = old_grid.node_stau[indices]
-
-        i1, i2 = new_grid.subset_segment_indices['all'][iseg, :]
-        indices = new_grid.subset_node_indices['all'][i1:i2]
-        nodes_eval = new_grid.node_stau[indices]
-
-        L_block, D_block = lagrange_matrices(nodes_given, nodes_eval)
-
-        L_blocks.append(L_block)
-        D_blocks.append(D_block)
-
-    L = block_diag(*L_blocks)
-    D = block_diag(*D_blocks)
-
-    return L, D
-
-
-def integration_matrix(grid):
-    """
-    Evaluate the Integration matrix of the given grid.
-
-    Parameters
-    ----------
-    grid: GridData
-        The GridData object representing the grid on which the integration matrix is to be evaluated
-
-    Returns
-    -------
-    int_matrix: np.ndarray
-        The integration matrix used to propagate initial states over segments
-
-    """
-    I_blocks = []
-
-    for iseg in range(grid.num_segments):
-        i1, i2 = grid.subset_segment_indices['all'][iseg, :]
-        indices = grid.subset_node_indices['all'][i1:i2]
-        nodes_given = grid.node_stau[indices]
-
-        i1, i2 = grid.subset_segment_indices['all'][iseg, :]
-        indices = grid.subset_node_indices['all'][i1:i2]
-        nodes_eval = grid.node_stau[indices][1:]
-
-        _, D_block = lagrange_matrices(nodes_given, nodes_eval)
-        I_block = np.linalg.inv(D_block[:, 1:])
-        I_blocks.append(I_block)
-
-    int_matrix = block_diag(*I_blocks)
-
-    return int_matrix
 
 
 def split_segments(old_seg_ends, B):
@@ -177,112 +100,6 @@ class HPAdaptive:
         self.x_dd = {}
         self.parent_seg_map = {}
 
-    def check_error(self):
-        """
-        Compute the error in every solved segment
-
-        Returns
-        -------
-        need_refinement: dict
-            Indicator for which segments of the given phase require grid refinement
-
-        """
-        refine_results = {}
-
-        if self.iteration_number > 0:
-            self.previous_error = copy.deepcopy(self.error)
-
-        for phase_path, phase in self.phases.items():
-            refine_results[phase_path] = {}
-
-            # Save the original grid to the refine results
-            tx = phase.options['transcription']
[…]
