### 2020-09-08 cb4dfe0
Fixed number of segments on robot arm prob
