+++ dymos/examples/robot_arm/test/test_robot_arm.py
@@ -72,7 +72,7 @@ class TestRobotArm(unittest.TestCase):
             cpd = p.check_partials(method='fd', compact_print=True, out_stream=None)
 
     def test_robot_arm_radau(self):
-        p = self.make_problem(transcription=Radau, optimizer='IPOPT', numseg=20)
+        p = self.make_problem(transcription=Radau, optimizer='IPOPT', numseg=30)
         dm.run_problem(p, refine=False)
 
         show_plots = False
