### 2020-09-09 7248332
added automatic target shape and unit detection method to utils.misc.  currently used in Pseudospectral transcriptions.
