### 2020-09-09 d2513b9
SolveIVP timeseries now supports introspection of units and shape
