+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -110,23 +110,6 @@ class SolveIVP(TranscriptionBase):
             if targets:
                 phase.connect(name, [f'ode.{t}' for t in targets])
 
-        # if phase.time_options['targets']:
-        #     phase.connect('time', ['ode.{0}'.format(t) for t in time_options['targets']])
-        #
-        # if phase.time_options['time_phase_targets']:
-        #     time_phase_tgts = time_options['time_phase_targets']
-        #     phase.connect('time_phase',
-        #                   ['ode.{0}'.format(t) for t in time_phase_tgts])
-        #
-        # if phase.time_options['t_initial_targets']:
-        #     time_phase_tgts = time_options['t_initial_targets']
-        #     phase.connect('t_initial', ['ode.{0}'.format(t) for t in time_phase_tgts])
-        #
-        # if phase.time_options['t_duration_targets']:
-        #     time_phase_tgts = time_options['t_duration_targets']
-        #     phase.connect('t_duration',
-        #                   ['ode.{0}'.format(t) for t in time_phase_tgts])
-
     def setup_states(self, phase):
         """
         Add an IndepVarComp for the states and setup the states as design variables.
@@ -135,7 +118,7 @@ class SolveIVP(TranscriptionBase):
                                                promotes_outputs=['*'])
 
         for state_name, options in phase.state_options.items():
-            indep_states_ivc.add_output('initial_states:{0}'.format(state_name),
+            indep_states_ivc.add_output(f'initial_states:{state_name}',
                                         val=np.ones(((1,) + options['shape'])),
                                         units=options['units'])
 
@@ -147,18 +130,18 @@ class SolveIVP(TranscriptionBase):
             # Connect the initial state to the first segment
             src_idxs = get_src_indices_by_row([0], options['shape'])
 
-            phase.connect('initial_states:{0}'.format(state_name),
-                          'segment_0.initial_states:{0}'.format(state_name),
+            phase.connect(f'initial_states:{state_name}',
+                          f'segment_0.initial_states:{state_name}',
                           src_indices=src_idxs, flat_src_indices=True)
 
-            phase.connect('segment_0.states:{0}'.format(state_name),
-                          'state_mux_comp.segment_0_states:{0}'.format(state_name))
+            phase.connect(f'segment_0.states:{state_name}',
+                          f'state_mux_comp.segment_0_states:{state_name}')
 
             targets = get_targets(ode=phase.ode, name=state_name, user_targets=options['targets'])
 
             if targets:
-                phase.connect('state_mux_comp.states:{0}'.format(state_name),
-                              ['ode.{0}'.format(t) for t in targets])
+                phase.connect(f'state_mux_comp.states:{state_name}',
+                              [f'ode.{t}' for t in targets])
 
             # Connect the final state in segment n to the initial state in segment n + 1
             for i in range(1, num_seg):
@@ -168,12 +151,12 @@ class SolveIVP(TranscriptionBase):
                     nnps_i = self.options['output_nodes_per_seg']
 
                 src_idxs = get_src_indices_by_row([nnps_i-1], shape=options['shape'])
-                phase.connect('segment_{0}.states:{1}'.format(i-1, state_name),
-                              'segment_{0}.initial_states:{1}'.format(i, state_name),
+                phase.connect(f'segment_{i - 1}.states:{state_name}',
+                              f'segment_{i}.initial_states:{state_name}',
                               src_indices=src_idxs, flat_src_indices=True)
 
-                phase.connect('segment_{0}.states:{1}'.format(i, state_name),
-                              'state_mux_comp.segment_{0}_states:{1}'.format(i, state_name))
[…]
