### 2020-09-09 ed4c1f2
minor cleanup
