### 2020-09-09 63a1900
Added transcription dependendent shape factor for array sizing. GL transcription refinement works now
