+++ dymos/test/test_run_problem.py
@@ -127,7 +127,7 @@ class TestRunProblem(unittest.TestCase):
         p.set_val('traj.phase0.t_duration', tf)
         p.set_val('traj.phase0.controls:u', phase0.interpolate(ys=[-0.6, 2.4],
                                                                nodes='control_input'))
-        dm.run_problem(p, refine=True)
+        dm.run_problem(p, refine=True, refine_method='hp')
 
         sqrt_two = np.sqrt(2)
         val = sqrt_two * tf
