### 2020-09-09 bd3081d
converted many examples of add_timeseries_output so that they no longer explicitly specify units
