+++ dymos/examples/aircraft_steady_flight/test/test_aircraft_cruise.py
@@ -173,7 +173,7 @@ class TestAircraftCruise(unittest.TestCase):
 
         phase.add_path_constraint('propulsion.tau', lower=0.01, upper=1.0, shape=(1,))
 
-        phase.add_timeseries_output('tas_comp.TAS', units='m/s')
+        phase.add_timeseries_output('tas_comp.TAS')
 
         p.model.connect('assumptions.S', 'phase0.parameters:S')
         p.model.connect('assumptions.mass_empty', 'phase0.parameters:mass_empty')
