+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -3,7 +3,7 @@ import numpy as np
 from .pseudospectral_base import PseudospectralBase
 from .components import GaussLobattoInterleaveComp
 from ..common import PathConstraintComp, GaussLobattoContinuityComp
-from ...utils.misc import get_rate_units, get_targets, get_target_metadata
+from ...utils.misc import get_rate_units, get_targets, get_target_metadata, get_source_metadata
 from ...utils.indexing import get_src_indices_by_row
 from ..grid_data import GridData, make_subset_map
 
@@ -542,29 +542,26 @@ class GaussLobatto(PseudospectralBase):
                 if var_type != 'ode':
                     continue
 
-                ode_outputs = {opts['prom_name']: opts for (k, opts) in
-                               phase.rhs_disc.get_io_metadata(iotypes=('output',)).items()}
-
-                if var in ode_outputs:
-                    shape = (1,) if len(ode_outputs[var]['shape']) == 1 else ode_outputs[var]['shape'][1:]
-                    units = ode_outputs[var]['units'] if units is None else units
-
-                    timeseries_comp = phase._get_subsystem(timeseries_name)
-                    timeseries_comp._add_output_configure(output_name, units, shape)
-
-                    interleave_comp = phase._get_subsystem('interleave_comp')
-                    if interleave_comp.add_var(output_name, shape, units):
-                        phase.connect(src_name='rhs_disc.{0}'.format(var),
-                                      tgt_name='interleave_comp.disc_values:{0}'.format(output_name))
-                        phase.connect(src_name='rhs_col.{0}'.format(var),
-                                      tgt_name='interleave_comp.col_values:{0}'.format(output_name))
-
-                    phase.connect(src_name='interleave_comp.all_values:{0}'.format(output_name),
-                                  tgt_name='{0}.input_values:{1}'.format(timeseries_name, output_name))
-                else:
+                try:
+                    shape, units = get_source_metadata(phase.rhs_disc, src=var,
+                                                       user_units=options['units'],
+                                                       user_shape=options['shape'])
+                except ValueError:
                     raise ValueError(f'Timeseries output {var} is not a known variable in'
                                      f' the phase {phase.pathname} nor is it a known output of '
                                      f' the ODE.')
+
+                timeseries_comp._add_output_configure(output_name, units, shape)
+
+                interleave_comp = phase._get_subsystem('interleave_comp')
+                if interleave_comp.add_var(output_name, shape, units):
+                    phase.connect(src_name=f'rhs_disc.{var}',
+                                  tgt_name=f'interleave_comp.disc_values:{output_name}')
+                    phase.connect(src_name=f'rhs_col.{var}',
+                                  tgt_name=f'interleave_comp.col_values:{output_name}')
+
+                phase.connect(src_name=f'interleave_comp.all_values:{output_name}',
+                              tgt_name=f'{timeseries_name}.input_values:{output_name}')
 
     def get_rate_source_path(self, state_name, nodes, phase):
         gd = self.grid_data
