### 2020-09-09 d338540
Added polynomial controls and parameters to load case
