### 2020-09-09 9e8f2b7
removed declaration of partials that are always 0 in robot arm problem
