+++ dymos/examples/robot_arm/robot_arm_ode.py
@@ -35,64 +35,23 @@ class RobotArmODE(om.ExplicitComponent):
         # set up the partials
         r = c = np.arange(nn)
 
-        self.declare_partials(of='x0_dot', wrt='x0', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x0_dot', wrt='x1', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x0_dot', wrt='x2', rows=r, cols=c, val=0.0)
         self.declare_partials(of='x0_dot', wrt='x3', rows=r, cols=c, val=1.0)
-        self.declare_partials(of='x0_dot', wrt='x4', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x0_dot', wrt='x5', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x0_dot', wrt='u0', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x0_dot', wrt='u1', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x0_dot', wrt='u2', rows=r, cols=c, val=0.0)
 
-        self.declare_partials(of='x1_dot', wrt='x0', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x1_dot', wrt='x1', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x1_dot', wrt='x2', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x1_dot', wrt='x3', rows=r, cols=c, val=0.0)
         self.declare_partials(of='x1_dot', wrt='x4', rows=r, cols=c, val=1.0)
-        self.declare_partials(of='x1_dot', wrt='x5', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x1_dot', wrt='u0', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x1_dot', wrt='u1', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x1_dot', wrt='u2', rows=r, cols=c, val=0.0)
 
-        self.declare_partials(of='x2_dot', wrt='x0', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x2_dot', wrt='x1', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x2_dot', wrt='x2', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x2_dot', wrt='x3', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x2_dot', wrt='x4', rows=r, cols=c, val=0.0)
         self.declare_partials(of='x2_dot', wrt='x5', rows=r, cols=c, val=1.0)
-        self.declare_partials(of='x2_dot', wrt='u0', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x2_dot', wrt='u1', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x2_dot', wrt='u2', rows=r, cols=c, val=0.0)
 
-        self.declare_partials(of='x3_dot', wrt='x0', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x3_dot', wrt='x1', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x3_dot', wrt='x2', rows=r, cols=c, val=0.0)
         self.declare_partials(of='x3_dot', wrt='x3', rows=r, cols=c, val=1.0)
-        self.declare_partials(of='x3_dot', wrt='x4', rows=r, cols=c, val=0.0)
-        self.declare_partials(of='x3_dot', wrt='x5', rows=r, cols=c, val=0.0)
[…]
