### 2020-09-09 06cdad5
Changed simulated values to use exp data rather than timeseries
