### 2020-09-10 d452c27
added a statement of need to the JOSS paper
