### 2020-09-10 9998b96
Made run_problem refinement method agnostic
