### 2020-09-10 69bb3e9
Separated write_iteration from refinement methods and broke into two functions to write error and refinement. Also made hp-method default refinement scheme
