### 2020-09-10 5b02328
Fixed test robot arm problem to solve
