+++ dymos/examples/robot_arm/test/test_robot_arm.py
@@ -32,36 +32,41 @@ class TestRobotArm(unittest.TestCase):
             p.driver.opt_settings['nlp_scaling_method'] = 'gradient-based'
             p.driver.opt_settings['max_iter'] = 500
             p.driver.opt_settings['print_level'] = 5
-
         traj = p.model.add_subsystem('traj', Trajectory())
+
         phase = traj.add_phase('phase', Phase(ode_class=RobotArmODE,
                                               transcription=transcription(num_segments=numseg, order=3)))
         phase.set_time_options(fix_initial=True, fix_duration=False)
 
-        phase.add_state('x0', fix_initial=True, fix_final=True, rate_source='x0_dot', targets=['x0'], units='m')
-        phase.add_state('x1', fix_initial=True, fix_final=True, rate_source='x1_dot', targets=['x1'], units='rad')
-        phase.add_state('x2', fix_initial=True, fix_final=True, rate_source='x2_dot', targets=['x2'], units='rad')
-        phase.add_state('x3', fix_initial=True, fix_final=True, rate_source='x3_dot', targets=['x3'], units='m/s')
-        phase.add_state('x4', fix_initial=True, fix_final=True, rate_source='x4_dot', targets=['x4'], units='rad/s')
-        phase.add_state('x5', fix_initial=True, fix_final=True, rate_source='x5_dot', targets=['x5'], units='rad/s')
+        phase.add_state('x0', fix_initial=True, fix_final=True, rate_source='x0_dot', units='m')
+        phase.add_state('x1', fix_initial=True, fix_final=True, rate_source='x1_dot', units='rad')
+        phase.add_state('x2', fix_initial=True, fix_final=True, rate_source='x2_dot', units='rad')
+        phase.add_state('x3', fix_initial=True, fix_final=True, rate_source='x3_dot', units='m/s')
+        phase.add_state('x4', fix_initial=True, fix_final=True, rate_source='x4_dot', units='rad/s')
+        phase.add_state('x5', fix_initial=True, fix_final=True, rate_source='x5_dot', units='rad/s')
 
-        phase.add_control('u0', opt=True, lower=-1, upper=1, scaler=0.1, targets=['u0'], units='m**2/s**2',
-                          rate_continuity=False)
-        phase.add_control('u1', opt=True, lower=-1, upper=1, scaler=0.1, targets=['u1'], units='m**3*rad/s**2',
-                          rate_continuity=False)
-        phase.add_control('u2', opt=True, lower=-1, upper=1, scaler=0.1, targets=['u2'], units='m**3*rad/s**2',
-                          rate_continuity=False)
+        phase.add_control('u0', opt=True, lower=-1, upper=1, scaler=0.1, units='m**2/s**2',
+                          continuity=False, rate_continuity=False)
+        phase.add_control('u1', opt=True, lower=-1, upper=1, scaler=0.1, units='m**3*rad/s**2',
+                          continuity=False, rate_continuity=False)
+        phase.add_control('u2', opt=True, lower=-1, upper=1, scaler=0.1, units='m**3*rad/s**2',
+                          continuity=False, rate_continuity=False)
 
-        phase.add_objective('time', scaler=1.0)
+        phase.add_path_constraint('u0', lower=-1, upper=1, scaler=0.1, units='m**2/s**2')
+        phase.add_path_constraint('u1', lower=-1, upper=1, scaler=0.1, units='m**3*rad/s**2')
[…]
