### 2020-09-10 446e5a5
Ted's edits for the JOSS paper
