### 2020-09-10 cbf7f1a
Fixed a bug in curvature computation
