### 2020-09-11 ecf4a63
added a statement of need to the JOSS paper (#381)
* added a statement of need to the JOSS paper

* Ted's edits for the JOSS paper
