### 2020-09-11 #2648 @robfalck
> 👋 @robfalck - please add an explicit "Statement of need" section

I've added an explicit Statement of Need section.
