+++ dymos/phase/phase.py
@@ -932,191 +932,6 @@ class Phase(om.Group):
         if include_timeseries is not _unspecified:
             self.parameter_options[name]['include_timeseries'] = include_timeseries
 
-    def add_design_parameter(self, name, val=_unspecified, units=_unspecified, opt=_unspecified,
-                             desc=_unspecified, lower=_unspecified, upper=_unspecified, scaler=_unspecified,
-                             adder=_unspecified, ref0=_unspecified, ref=_unspecified, targets=_unspecified,
-                             shape=_unspecified, dynamic=_unspecified, include_timeseries=_unspecified):
-        """
-        Add a design parameter (static control variable) to the phase.
-
-        This method is deprecated. Use add_parameter instead.
-
-        Parameters
-        ----------
-        name : str
-            Name of the design parameter.
-        val : float or ndarray
-            Default value of the design parameter at all nodes.
-        units : str or None or 0
-            Units in which the design parameter is defined.  If 0, use the units declared
-            for the parameter in the ODE.
-        opt : bool
-            If True (default) the value(s) of this design parameter will be design variables in
-            the optimization problem, in the path 'phase_name.indep_controls.controls:control_name'.
-            If False, the this design parameter will still be owned by an IndepVarComp in the phase,
-            but it will not be a design variable in the optimization.
-        desc : str
-            A description of the design parameter.
-        lower : float or ndarray
-            The lower bound of the design parameter value.
-        upper : float or ndarray
-            The upper bound of the design parameter value.
-        scaler : float or ndarray
-            The scaler of the design parameter value for the optimizer.
-        adder : float or ndarray
-            The adder of the design parameter value for the optimizer.
-        ref0 : float or ndarray
-            The zero-reference value of the design parameter for the optimizer.
-        ref : float or ndarray
-            The unit-reference value of the design parameter for the optimizer.
-        targets : Sequence of str or None
-            Targets in the ODE to which this parameter is connected.
-            In the future, if left _unspecified (the default), the phase parameter will try to connect to an ODE input
-            of the same name. Currently _unspecified is the same as None, but a deprecation warning is issued.
-            Set targets to None to prevent this (the old default behavior).
-        shape : Sequence of int
-            The shape of the design parameter.
-        dynamic : bool
-            True if the targets in the ODE may be dynamic (if the inputs are sized to the number
-            of nodes) else False.
-        include_timeseries : bool
-            True if the static design parameters should be included in output timeseries, else False.
-        """
-        msg = "DesignParameters and InputParameters are being replaced by Parameters in  " + \
-            "Dymos 1.0.0. Please use add_parameter or set_parameter_options to remove this " + \
-            "deprecation warning."
-        warn_deprecation(msg)
-        self.add_parameter(name, val=val, units=units, opt=opt, desc=desc, lower=lower,
-                           upper=upper, scaler=scaler, adder=adder, ref0=ref0, ref=ref,
-                           targets=targets, shape=shape, dynamic=dynamic,
-                           include_timeseries=include_timeseries)
-
-    def set_design_parameter_options(self, name, val=_unspecified, units=_unspecified, opt=_unspecified,
-                                     desc=_unspecified, lower=_unspecified, upper=_unspecified,
-                                     scaler=_unspecified, adder=_unspecified, ref0=_unspecified,
-                                     ref=_unspecified, targets=_unspecified, shape=_unspecified,
-                                     dynamic=_unspecified, include_timeseries=_unspecified):
-        """
-        Set options for an existing design parameter (static control variable) in the phase.
-
-        Parameters
-        ----------
-        name : str
-            Name of the design parameter.
-        val : float or ndarray
-            Default value of the design parameter at all nodes.
-        units : str or None or 0
[…]
