+++ dymos/trajectory/trajectory.py
@@ -167,6 +167,95 @@ class Trajectory(om.Group):
         if dynamic is not _unspecified:
             self.parameter_options[name]['dynamic'] = dynamic
 
+    def add_input_parameter(self, name, units, val=_unspecified, desc=_unspecified,
+                            targets=_unspecified, custom_targets=_unspecified,
+                            shape=_unspecified, dynamic=_unspecified):
+        """
+        Add an input parameter to the trajectory.
+
+        Parameters
+        ----------
+        name : str
+            Name of the input parameter.
+        val : float or ndarray
+            Default value of the input parameter at all nodes.
+        desc : str
+            A description of the input parameter.
+        targets : dict or None
+            A dictionary mapping the name of each phase in the trajectory to a sequence of ODE
+            targets for this parameter in each phase.
+        units : str or None or 0
+            Units in which the input parameter is defined.  If 0, use the units declared
+            for the parameter in the ODE.
+        shape : Sequence of int
+            The shape of the input parameter.
+        dynamic : bool
+            True if the targets in the ODE may be dynamic (if the inputs are sized to the number
+            of nodes) else False.
+        """
+        msg = "DesignParameters and InputParameters are being replaced by Parameters in  " + \
+            "Dymos 1.0.0. Please use add_parameter or set_parameter_options to remove this " + \
+            "deprecation warning."
+        warn_deprecation(msg)
+        self.add_parameter(name, units, val, desc, targets, custom_targets, shape, dynamic)
+
+    def add_design_parameter(self, name, units, val=_unspecified, desc=_unspecified, opt=_unspecified,
+                             targets=_unspecified, custom_targets=_unspecified,
+                             lower=_unspecified, upper=_unspecified, scaler=_unspecified,
+                             adder=_unspecified, ref0=_unspecified, ref=_unspecified,
+                             shape=_unspecified, dynamic=_unspecified):
+        """
+        Add a design parameter (static control) to the trajectory.
+
+        Parameters
+        ----------
+        name : str
+            Name of the design parameter.
+        val : float or ndarray
+            Default value of the design parameter at all nodes.
+        desc : str
+            A description of the design parameter.
+        targets : dict or None
+            If None, then the design parameter will be connected to the controllable parameter
+            in the ODE of each phase.  For each phase where no such controllable parameter exists,
+            a warning will be issued.  If targets is given as a dict, the dict should provide
+            the relevant phase names as keys, each associated with the respective controllable
+            parameter as a value.
+        custom_targets : dict or None
+            By default, the input parameter will be connect to the parameter/targets of the given
+            name in each phase.  This argument can be used to override that behavior on a phase
+            by phase basis.
+        units : str or None or 0
+            Units in which the design parameter is defined.  If 0, use the units declared
+            for the parameter in the ODE.
+        opt : bool
+            If True (default) the value(s) of this parameter will be design variables in
+            the optimization problem.
+        lower : float or ndarray
+            The lower bound of the design parameter value.
+        upper : float or ndarray
+            The upper bound of the design parameter value.
+        scaler : float or ndarray
+            The scaler of the design parameter value for the optimizer.
+        adder : float or ndarray
+            The adder of the design parameter value for the optimizer.
+        ref0 : float or ndarray
+            The zero-reference value of the design parameter for the optimizer.
+        ref : float or ndarray
+            The unit-reference value of the design parameter for the optimizer.
+        shape : Sequence of int
+            The shape of the design parameter.
+        dynamic : bool
+            True if the targets in the ODE may be dynamic (if the inputs are sized to the number
+            of nodes) else False.
+        """
[…]
