### 2020-09-12 9c4ff8c
small cleanup of path constraint setup which turned into a larger refactor
