+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -265,14 +265,9 @@ class GaussLobatto(PseudospectralBase):
                           'collocation_constraint.f_computed:{0}'.format(name),
                           src_indices=src_idxs, flat_src_indices=True)
 
-    def setup_path_constraints(self, phase):
-        gd = self.grid_data
-
-        if phase._path_constraints:
-            path_comp = PathConstraintComp(num_nodes=gd.num_nodes)
-            phase.add_subsystem('path_constraints', subsys=path_comp)
-
     def configure_path_constraints(self, phase):
+        super(GaussLobatto, self).configure_path_constraints(phase)
+
         time_units = phase.time_options['units']
         path_comp = phase._get_subsystem('path_constraints')
 
@@ -287,101 +282,50 @@ class GaussLobatto(PseudospectralBase):
             var_type = phase.classify_var(var)
 
             if var_type == 'time':
-                phase.connect(src_name='time',
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                shape = (1,)
-                units = time_units if con_units is None else con_units
-                linear = True
+                src = 'time'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'time_phase':
-                phase.connect(src_name='time_phase',
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                shape = (1,)
-                units = time_units if con_units is None else con_units
-                linear = True
+                src = 'time_phase'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'state':
-                phase.connect(src_name='interleave_comp.all_values:states:{0}'.format(var),
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                state_shape = phase.state_options[var]['shape']
-                state_units = phase.state_options[var]['units']
-                shape = state_shape
-                units = state_units if con_units is None else con_units
-                linear = False
+                src = f'interleave_comp.all_values:states:{var}'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type in ('indep_control', 'input_control'):
-                constraint_path = 'control_values:{0}'.format(var)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.control_options[var]['shape']
-                control_units = phase.control_options[var]['units']
-                shape = control_shape
-                units = control_units if con_units is None else con_units
-                linear = True
+                src = f'control_values:{var}'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type in ('indep_polynomial_control', 'input_polynomial_control'):
-                constraint_path = 'polynomial_control_values:{0}'.format(var)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.polynomial_control_options[var]['shape']
-                control_units = phase.polynomial_control_options[var]['units']
-                shape = control_shape
-                units = control_units if con_units is None else con_units
-                linear = False
+                src = f'polynomial_control_values:{var}'
+                tgt = f'path_constraints.all_values:{con_name}'
 
             elif var_type == 'control_rate':
                 control_name = var[:-5]
-                constraint_path = 'control_rates:{0}_rate'.format(control_name)
-                phase.connect(src_name=constraint_path,
-                              tgt_name='path_constraints.all_values:{0}'.format(con_name))
-                control_shape = phase.control_options[control_name]['shape']
-                control_units = phase.control_options[control_name]['units']
-                shape = control_shape
[…]
