### 2020-09-12 2647f11
Timeseries outputs automatically detect shape and default units from ODE outputs (#380)
* Transcription methods now handle their timeseries objects during configure.

* Added get_targets_metadata and get_source_metadata to dymos.utils.misc for use in introspection during configure. These methods automatically determine the shape and units of targets or sources.

* converted many examples of add_timeseries_output so that they no longer explicitly specify units

* path constraints are now handled during configure, but automatic shape and unit detection remains to be done.
