+++ dymos/transcriptions/solve_ivp/components/ode_integration_interface_system.py
@@ -2,7 +2,7 @@ import numpy as np
 from .odeint_control_interpolation_comp import ODEIntControlInterpolationComp
 from .state_rate_collector_comp import StateRateCollectorComp
 from ....phase.options import TimeOptionsDictionary
-from ....utils.misc import get_targets
+from ....utils.misc import get_targets, get_target_metadata
 import openmdao.api as om
 
 
@@ -18,15 +18,9 @@ class ODEIntegrationInterfaceSystem(om.Group):
         self.options.declare('control_options', default=None, types=dict, allow_none=True,
                              desc='Dictionary of control names/options for the segments parent Phase.')
 
-        # self.options.declare('control_interpolants', default={}, types=dict,
-        #                      desc='Dictionary of control names/interpolants.')
-
         self.options.declare('polynomial_control_options', default=None, types=dict, allow_none=True,
                              desc='Dictionary of polynomial control names/options for the segments '
                                   'parent Phase.')
-
-        # self.options.declare('polynomial_control_interpolants', default={}, types=dict,
-        #                      desc='Dictionary of polynomial control names/interpolants.')
 
         self.options.declare('parameter_options', default=None, types=dict, allow_none=True,
                              desc='Dictionary of parameter names/options for the segments '
@@ -39,7 +33,6 @@ class ODEIntegrationInterfaceSystem(om.Group):
                              desc='Keyword arguments provided when initializing the ODE System')
 
     def setup(self):
-        # The time IVC
         ivc = om.IndepVarComp()
         time_options = self.options['time_options']
         time_units = time_options['units']
@@ -67,12 +60,6 @@ class ODEIntegrationInterfaceSystem(om.Group):
             self.add_subsystem('indep_controls', self._interp_comp, promotes_outputs=['*'])
             self.connect('time', ['indep_controls.time'])
 
-        if self.options['parameter_options']:
-            for name, options in self.options['parameter_options'].items():
-                ivc.add_output('parameters:{0}'.format(name),
-                               shape=options['shape'],
-                               units=options['units'])
-
         # The ODE System
         if self.options['ode_class'] is not None:
             self.add_subsystem('ode', subsys=self.options['ode_class'](num_nodes=1,
@@ -84,18 +71,20 @@ class ODEIntegrationInterfaceSystem(om.Group):
                                                   time_units=time_options['units']))
 
     def configure(self):
+        ivc = self._get_subsystem('ivc')
+        ode = self._get_subsystem('ode')
 
         # Configure states
         for name, options in self.options['state_options'].items():
-            self.ivc.add_output('states:{0}'.format(name),
-                                shape=(1, np.prod(options['shape'])),
-                                units=options['units'])
+            ivc.add_output(f'states:{name}',
+                           shape=(1, np.prod(options['shape'])),
+                           units=options['units'])
 
             rate_src = self._get_rate_source_path(name)
 
-            self.connect(rate_src, 'state_rate_collector.state_rates_in:{0}_rate'.format(name))
+            self.connect(rate_src, f'state_rate_collector.state_rates_in:{name}_rate')
 
-            targets = get_targets(ode=self.ode, name=name, user_targets=options['targets'])
+            targets = get_targets(ode=ode, name=name, user_targets=options['targets'])
 
             if targets:
                 self.connect(f'states:{name}', [f'ode.{tgt}' for tgt in targets])
@@ -103,11 +92,11 @@ class ODEIntegrationInterfaceSystem(om.Group):
         # Configure controls
         if self.options['control_options']:
             for name, options in self.options['control_options'].items():
-                targets = get_targets(ode=self.ode, name=name,
+                targets = get_targets(ode=ode, name=name,
                                       user_targets=options['targets'])
[…]
