### 2020-09-14 31bbe9e
wildcard match in add_timeseries_outputs
