### 2020-09-14 e2ef503
Added test case for loading polynomial controls
