### 2020-09-15 8185e6b
use fnmatch.filter for wildcard matching
