+++ dymos/transcriptions/runge_kutta/runge_kutta.py
@@ -12,7 +12,7 @@ from ...utils.misc import CoerceDesvar, get_rate_units, get_targets, get_target_
 from ...utils.constants import INF_BOUND
 from ...utils.indexing import get_src_indices_by_row
 from ..grid_data import GridData
-from fnmatch import fnmatch
+from fnmatch import filter
 
 
 class RungeKutta(TranscriptionBase):
@@ -795,7 +795,7 @@ class RungeKutta(TranscriptionBase):
                 if '*' in var:  # match outputs from the ODE
                     ode_outputs = {opts['prom_name']: opts for (k, opts) in
                                    phase.ode.get_io_metadata(iotypes=('output',)).items()}
-                    matches = [n for n in list(ode_outputs.keys()) if fnmatch(n, var)]
+                    matches = filter(list(ode_outputs.keys()), var)
                 else:
                     matches = [var]
 
