### 2020-09-15 86ab4f3
check timeseries outputs and units
