### 2020-09-15 01bfdb4
test and fix issue with wildcards and periods in names
