+++ dymos/transcriptions/pseudospectral/radau_pseudospectral.py
@@ -367,7 +367,7 @@ class Radau(PseudospectralBase):
 
                 for v in matches:
                     if '*' in var:
-                        output_name = v
+                        output_name = v.split('.')[-1]
 
                     # Determine the path to the variable which we will be constraining
                     # This is more complicated for path constraints since, for instance,
@@ -388,7 +388,14 @@ class Radau(PseudospectralBase):
                                          f' the phase {phase.pathname} nor is it a known output of '
                                          f' the ODE.')
 
-                    timeseries_comp._add_output_configure(output_name, units, shape, desc='')
+                    try:
+                        timeseries_comp._add_output_configure(output_name, units, shape, desc='')
+                    except ValueError as e:  # OK if it already exists
+                        if 'already exists' in str(e):
+                            continue
+                        else:
+                            raise e
+
                     phase.connect(src_name='rhs_all.{0}'.format(v),
                                   tgt_name='{0}.input_values:{1}'.format(timeseries_name,
                                                                          output_name))
