+++ dymos/transcriptions/runge_kutta/runge_kutta.py
@@ -801,7 +801,7 @@ class RungeKutta(TranscriptionBase):
 
                 for v in matches:
                     if '*' in var:
-                        output_name = v
+                        output_name = v.split('.')[-1]
 
                     # Determine the path to the variable which we will be constraining
                     # This is more complicated for path constraints since, for instance,
@@ -822,7 +822,13 @@ class RungeKutta(TranscriptionBase):
                                          f' the phase {phase.pathname} nor is it a known output of '
                                          f' the ODE.')
 
-                    timeseries_comp._add_output_configure(output_name, units, shape, desc='')
+                    try:
+                        timeseries_comp._add_output_configure(output_name, units, shape, desc='')
+                    except ValueError as e:  # OK if it already exists
+                        if 'already exists' in str(e):
+                            continue
+                        else:
+                            raise e
 
                     # Failed to find variable, assume it is in the RHS
                     phase.connect(src_name=f'ode.{v}',
