### 2020-09-16 #2648 @dpsanders
👋 @robfalck: Thanks for your submission to JOSS!

Reading through the paper I have a few initial comments:

- I don't feel that the paper is currently accessible to non-experts in the field. As an example, it's not clear how ODEs (which is never defined) are related to the rest of the story, and you seem to assume knowledge of OpenMDAO itself.

- It's not clear what is provided by OpenMDAO and what is new in Dymos.

- It's not clear what "phases" are.

- I would like to see a simple piece of example code and output in the paper.

In terms of the README in the repository:

- The code itself seems so verbose that it actually ends up being hard to read, even for a simple model.
Also there seems to be code within strings.
