+++ dymos/load_case.py
@@ -2,6 +2,7 @@ import numpy as np
 import openmdao.api as om
 from openmdao.recorders.case import Case
 from .phase.phase import Phase
+from .utils.lgl import lgl
 
 
 def _split_var_path(path):
@@ -112,6 +113,7 @@ def load_case(problem, previous_solution):
     phase_io = {'inputs': problem.model.list_inputs(out_stream=None, units=True, prom_name=True),
                 'outputs': problem.model.list_outputs(out_stream=None, units=True, prom_name=True)}
 
+    phase_inputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in phase_io['inputs']}
     phase_outputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in phase_io['outputs']}
 
     for phase_abs_path, phase in phase_paths.items():
@@ -164,22 +166,18 @@ def load_case(problem, previous_solution):
                                        s.endswith(f'{phase_name}.polynomial_controls:{polynomial_control_name}')][0]
             prev_polynomial_control_path = [s for s in prev_outputs if
                                             s.endswith(f'{phase_name}.'
-                                                       f'timeseries.polynomial_controls:{polynomial_control_name}')][0]
+                                                       f'polynomial_controls:{polynomial_control_name}')][0]
             prev_polynomial_control_val = prev_outputs[prev_polynomial_control_path]['value']
             prev_polynomial_control_units = prev_outputs[prev_polynomial_control_path]['units']
-            problem.set_val(polynomial_control_path,
-                            phase.interpolate(xs=prev_time, ys=prev_polynomial_control_val,
-                                              nodes='control_input', kind='slinear'),
+            problem.set_val(polynomial_control_path, prev_polynomial_control_val,
                             units=prev_polynomial_control_units)
 
         # Interpolate the timeseries parameter outputs from the previous solution onto the new grid
         for parameter_name, options in phase.parameter_options.items():
-            parameter_path = [s for s in phase_outputs if s.endswith(f'{phase_name}.parameters:{parameter_name}')][0]
+            parameter_path = [s for s in phase_inputs if s.endswith(f'{phase_name}.parameters:{parameter_name}')][0]
             prev_parameter_path = [s for s in prev_outputs if
                                    s.endswith(f'{phase_name}.timeseries.parameters:{parameter_name}')][0]
             prev_parameter_val = prev_outputs[prev_parameter_path]['value'][0, ...]
-            prev_parameter_val = np.repeat(prev_parameter_val, axis=0,
-                                           repeats=phase.options['transcriptions'].grid_data.num_nodes)
             prev_parameter_units = prev_outputs[prev_parameter_path]['units']
 
             problem.set_val(parameter_path, prev_parameter_val, units=prev_parameter_units)
