### 2020-09-16 f79faaa
Multiple variable input for add_timeseries_output (#387)
* allow list of names input to add_timeseries_output

* wildcard match in add_timeseries_outputs

* check timeseries outputs and units
