### 2020-09-17 4fd4928
Added an upgrade guide and a new doc macro for demonstrating old vs. new behavior
