+++ dymos/run_problem.py
@@ -78,6 +78,7 @@ def run_problem(problem, refine=False, refine_method='hp', refine_iteration_limi
     A Dymos-specific interface to execute an OpenMDAO problem containing Dymos Trajectories or
     Phases.  This function can iteratively call run_driver to perform grid refinement, and automatically
     call simulate following a run to check the validity of a result.
+
     Parameters
     ----------
     problem : om.Problem
