+++ dymos/test/test_upgrade_guide.py
@@ -1,17 +1,19 @@
 import unittest
 
-import matplotlib.pyplot as plt
-plt.switch_backend('Agg')
-plt.style.use('ggplot')
-
-from dymos.utils.doc_utils import save_for_docs
 
 
-class TestDocSSTOEarth(unittest.TestCase):
+class TestUpgrade_0_16_0(unittest.TestCase):
 
-    @save_for_docs
-    def test_doc_ssto_earth(self):
-        import matplotlib.pyplot as plt
+    def test_parameters(self):
+        """
+        # upgrade_doc: begin set_val
+        p.set_val('traj.phase0.design_parameters:thrust', 2.1, units='MN')
+        # upgrade_doc: end set_val
+        # upgrade_doc: begin parameter_timeseries
+        thrust = p.get_val('traj.phase0.timeseries.design_parameters:thrust')
+        # upgrade_doc: end parameter_timeseries
+        """
+        import numpy as np
         import openmdao.api as om
         from openmdao.utils.assert_utils import assert_near_equal
         import dymos as dm
@@ -80,7 +82,10 @@ class TestDocSSTOEarth(unittest.TestCase):
         p.set_val('traj.phase0.states:vy', phase.interpolate(ys=[1.0E-6, 0], nodes='state_input'))
         p.set_val('traj.phase0.states:m', phase.interpolate(ys=[117000, 1163], nodes='state_input'))
         p.set_val('traj.phase0.controls:theta', phase.interpolate(ys=[1.5, -0.76], nodes='control_input'))
+
+        # upgrade_doc: begin set_val
         p.set_val('traj.phase0.parameters:thrust', 2.1, units='MN')
+        # upgrade_doc: end set_val
 
         #
         # Solve the Problem
@@ -94,52 +99,65 @@ class TestDocSSTOEarth(unittest.TestCase):
         assert_near_equal(p.get_val('traj.phase0.timeseries.states:y')[-1], 1.85E5, 1e-4)
         assert_near_equal(p.get_val('traj.phase0.timeseries.states:vx')[-1], 7796.6961, 1e-4)
         assert_near_equal(p.get_val('traj.phase0.timeseries.states:vy')[-1], 0, 1e-4)
-        #
-        # Get the explitly simulated results
-        #
-        exp_out = traj.simulate()
 
-        #
-        # Plot the results
-        #
-        fig, axes = plt.subplots(nrows=2, ncols=1, figsize=(10, 8))
+        # upgrade_doc: begin parameter_timeseries
+        thrust = p.get_val('traj.phase0.timeseries.parameters:thrust')
+        # upgrade_doc: end parameter_timeseries
+        nn = phase.options['transcription'].grid_data.num_nodes
+        assert_near_equal(thrust, 2.1E6 * np.ones([nn, 1]))
 
-        axes[0].plot(p.get_val('traj.phase0.timeseries.states:x'),
-                     p.get_val('traj.phase0.timeseries.states:y'),
-                     marker='o',
-                     ms=4,
-                     linestyle='None',
-                     label='solution')
+    def test_parameter_no_include_timeseries(self):
+        import openmdao.api as om
+        import dymos as dm
+        from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
 
-        axes[0].plot(exp_out.get_val('traj.phase0.timeseries.states:x'),
-                     exp_out.get_val('traj.phase0.timeseries.states:y'),
-                     marker=None,
-                     linestyle='-',
-                     label='simulation')
+        p = om.Problem(model=om.Group())
 
-        axes[0].set_xlabel('range (m)')
-        axes[0].set_ylabel('altitude (m)')
-        axes[0].set_aspect('equal')
+        p.driver = om.ScipyOptimizeDriver()
+        p.driver.declare_coloring()
 
[…]
