### 2020-09-17 6817c05
fixed comments to be clear about where interpolation is performed
