### 2020-09-17 53d47d3
fix for removal of _loc_subsys_map
