+++ dymos/transcriptions/pseudospectral/components/state_interp_comp.py
@@ -41,7 +41,11 @@ class StateInterpComp(om.ExplicitComponent):
             'time_units', default=None, allow_none=True, types=str,
             desc='Units of the integration variable')
 
-    def setup(self):
+    def configure_io(self):
+        """
+        I/O creation is delayed until configure so that we can determine the shape and units for
+        the states.
+        """
         time_units = self.options['time_units']
 
         num_disc_nodes = self.options['grid_data'].subset_num_nodes['state_disc']
@@ -68,27 +72,27 @@ class StateInterpComp(om.ExplicitComponent):
 
             self.add_input(
                 name='state_disc:{0}'.format(state_name),
-                shape=(num_disc_nodes,) + shape,
+                shape=shape,
                 desc='Values of state {0} at discretization nodes'.format(state_name),
                 units=units)
 
             if transcription == 'gauss-lobatto':
                 self.add_input(
                     name='staterate_disc:{0}'.format(state_name),
-                    shape=(num_disc_nodes,) + shape,
+                    shape=shape,
                     units=rate_units,
                     desc='EOM time derivative of state {0} at '
                          'discretization nodes'.format(state_name))
 
                 self.add_output(
                     name='state_col:{0}'.format(state_name),
-                    shape=(num_col_nodes,) + shape, units=units,
+                    shape=shape, units=units,
                     desc='Interpolated values of state {0} at '
                          'collocation nodes'.format(state_name))
 
             self.add_output(
                 name='staterate_col:{0}'.format(state_name),
-                shape=(num_col_nodes,) + shape,
+                shape=shape,
                 units=rate_units,
                 desc='Interpolated rate of state {0} at collocation nodes'.format(state_name))
 
@@ -117,7 +121,7 @@ class StateInterpComp(om.ExplicitComponent):
         for name, options in state_options.items():
             shape = options['shape']
 
-            size = np.prod(shape)
+            size = np.prod(shape) / num_disc_nodes
             self.sizes[name] = size
 
             for key in self.jacs:
@@ -162,7 +166,6 @@ class StateInterpComp(om.ExplicitComponent):
 
     def _compute_radau(self, inputs, outputs):
         num_disc_nodes = self.options['grid_data'].subset_num_nodes['disc']
-        num_col_nodes = self.options['grid_data'].subset_num_nodes['col']
         state_options = self.options['state_options']
         dt_dstau = inputs['dt_dstau'][:, np.newaxis]
 
@@ -176,7 +179,7 @@ class StateInterpComp(om.ExplicitComponent):
                                  newshape=(num_disc_nodes, size))
 
             outputs[xdotc_str] = np.reshape(self.matrices['Ad'].dot(xd_flat) / dt_dstau,
-                                            newshape=(num_col_nodes,) + shape)
+                                            newshape=shape)
 
     def _compute_gauss_lobatto(self, inputs, outputs):
         state_options = self.options['state_options']
@@ -184,7 +187,6 @@ class StateInterpComp(om.ExplicitComponent):
         dt_dstau = inputs['dt_dstau'][:, np.newaxis]
 
         num_disc_nodes = self.options['grid_data'].subset_num_nodes['disc']
-        num_col_nodes = self.options['grid_data'].subset_num_nodes['col']
 
         Ai = self.matrices['Ai']
         Bi = self.matrices['Bi']
@@ -193,7 +195,7 @@ class StateInterpComp(om.ExplicitComponent):
 
         for name in state_options:
             shape = state_options[name]['shape']
-            size = np.prod(shape)
[…]
