+++ dymos/examples/brachistochrone/test/test_doc_brachistochrone_polynomial_controls.py
@@ -24,15 +24,12 @@ class TestBrachistochronePolynomialControl(unittest.TestCase):
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
         phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                        units=BrachistochroneODE.states['x']['units'],
                         fix_initial=True, fix_final=True, solve_segments=False)
 
         phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                        units=BrachistochroneODE.states['y']['units'],
                         fix_initial=True, fix_final=True, solve_segments=False)
 
         phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                        units=BrachistochroneODE.states['v']['units'],
                         fix_initial=True, fix_final=False, solve_segments=False)
 
         phase.add_polynomial_control('theta', order=1, units='deg', lower=0.01, upper=179.9)
@@ -116,15 +113,12 @@ class TestBrachistochronePolynomialControl(unittest.TestCase):
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
         phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                        units=BrachistochroneODE.states['x']['units'],
                         fix_initial=True, fix_final=True, solve_segments=False)
 
         phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                        units=BrachistochroneODE.states['y']['units'],
                         fix_initial=True, fix_final=True, solve_segments=False)
 
         phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                        units=BrachistochroneODE.states['v']['units'],
                         fix_initial=True, fix_final=False, solve_segments=False)
 
         phase.add_polynomial_control('theta', order=1, units='deg', lower=0.01, upper=179.9)
@@ -208,15 +202,12 @@ class TestBrachistochronePolynomialControl(unittest.TestCase):
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
         phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                        units=BrachistochroneODE.states['x']['units'],
                         fix_initial=True, fix_final=False, solve_segments=False)
 
         phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                        units=BrachistochroneODE.states['y']['units'],
                         fix_initial=True, fix_final=False, solve_segments=False)
 
         phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                        units=BrachistochroneODE.states['v']['units'],
                         fix_initial=True, fix_final=False, solve_segments=False)
 
         phase.add_polynomial_control('theta', order=1, units='deg', lower=0.01, upper=179.9)
@@ -305,15 +296,12 @@ class TestBrachistochronePolynomialControlBoundaryConstrained(unittest.TestCase)
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
         phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                        units=BrachistochroneODE.states['x']['units'],
                         fix_initial=True, fix_final=True, solve_segments=False)
 
         phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                        units=BrachistochroneODE.states['y']['units'],
                         fix_initial=True, fix_final=True, solve_segments=False)
 
[…]
