### 2020-09-19 1d1e18f
example docs now remove assert calls in the published docs.  added redbaron as a dependency needed for building the docs.
