+++ mkdocs/main.py
@@ -7,6 +7,7 @@ import re
 import sys
 import textwrap
 from pathlib import Path
+from redbaron import RedBaron
 
 try:
     from numpydoc.docscrape import FunctionDoc, ClassDoc
@@ -349,6 +350,20 @@ def get_parent_dir(env):
 
 
 def _get_test_source(reference):
+    """
+    Return the source code from the test specified by the gien reference.
+
+    Parameters
+    ----------
+    reference : str
+        A dotted path to the test method whose source is desired.
+
+    Returns
+    -------
+    str
+        The returned source, dedented and having had assert method calls removed.
+
+    """
     obj = get_object_from_reference(reference)
     try:
         method = obj._method
@@ -371,8 +386,44 @@ def _get_test_source(reference):
     source = textwrap.dedent(source)
     source = source.strip()
 
+    # Remove the assert method calls from documentation.
+    source = _strip_asserts(source)
+
     return source
 
+def _strip_asserts(source):
+    """
+    Remove assert method calls from source code.
+
+    Using RedBaron, replace some assert calls with print statements that print the actual
+    value given in the asserts. Depending on the calls, the actual value can be the first or second
+    argument.
+
+    Parameters
+    ----------
+    source : str
+        String containing source lines.
+
+    Returns
+    -------
+    str
+        Source with asserts removed.
+    """
+    rb = RedBaron(source)  # convert to RedBaron internal structure
+
+    # findAll is slow, so only check the ones that are present.
+    asserts = ['assertAlmostEqual', 'assertLess', 'assertGreater', 'assertEqual',
+               'assert_equal_arrays', 'assertTrue', 'assertFalse', 'assert_near_equal',
+               'assert_rel_error', 'assert_almost_equal', 'assert_allclose']
+
+    for assert_type in asserts:
+        assert_nodes = rb.findAll("NameNode", value=assert_type)
+        for i in reversed(range(len(assert_nodes))):
+            parent = assert_nodes[i].parent
+            for j in reversed(range(len(parent.value))):
+                assert_nodes[i].parent.remove(parent.value[j])
+
+    return rb.dumps()
 
 def _function_doc_markdown(func, reference, outstream=sys.stdout, indent='', method=False):
     """
@@ -498,3 +549,17 @@ def _options_dict_to_markdown(od):
     md = '\n'.join(lines[1:-1])
 
     return md
+
+
+if __name__ == '__main__':
+    # s = _get_test_source('dymos.examples.ssto.doc.test_doc_ssto_earth.TestDocSSTOEarth.test_doc_ssto_earth')
+    # print(s)
+
+    s = """
+        p = _get_problem()
+        J = p.compute_totals(of=['con.y1'], wrt=['des_vars.x'], return_format='array')
+        np.testing.assert_allclose(J, np.array([[1.0]]), rtol=1e-7, atol=0, equal_nan=True,
+                                   err_msg='', verbose=True)
+    """
+
+    print(_strip_asserts(s))
