+++ dymos/examples/ssto/doc/test_doc_ssto_earth.py
@@ -4,6 +4,7 @@ import matplotlib.pyplot as plt
 plt.switch_backend('Agg')
 plt.style.use('ggplot')
 
+from openmdao.utils.assert_utils import assert_near_equal
 from dymos.utils.doc_utils import save_for_docs
 
 
@@ -13,7 +14,6 @@ class TestDocSSTOEarth(unittest.TestCase):
     def test_doc_ssto_earth(self):
         import matplotlib.pyplot as plt
         import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
         import dymos as dm
 
         #
@@ -86,15 +86,12 @@ class TestDocSSTOEarth(unittest.TestCase):
         #
         dm.run_problem(p)
 
-        #
-        # Check the results.
-        #
         assert_near_equal(p.get_val('traj.phase0.timeseries.time')[-1], 143, tolerance=0.05)
         assert_near_equal(p.get_val('traj.phase0.timeseries.states:y')[-1], 1.85E5, 1e-4)
         assert_near_equal(p.get_val('traj.phase0.timeseries.states:vx')[-1], 7796.6961, 1e-4)
         assert_near_equal(p.get_val('traj.phase0.timeseries.states:vy')[-1], 0, 1e-4)
         #
-        # Get the explitly simulated results
+        # Get the explicitly simulated results
         #
         exp_out = traj.simulate()
 
