+++ dymos/test/test_upgrade_guide.py
@@ -1,5 +1,11 @@
 import unittest
 
+import openmdao.api as om
+
+import dymos as dm
+
+from openmdao.utils.assert_utils import assert_near_equal
+
 
 class TestUpgrade_0_16_0(unittest.TestCase):
 
@@ -14,7 +20,6 @@ class TestUpgrade_0_16_0(unittest.TestCase):
         """
         import numpy as np
         import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
         import dymos as dm
 
         #
@@ -160,3 +165,314 @@ class TestUpgrade_0_16_0(unittest.TestCase):
 
         with self.assertRaises(KeyError):
             p.get_val('phase0.timeseries.parameters:g}')
+
+    def test_simplified_ode_timeseries_output(self):
+        """
+        # upgrade_doc: begin simplified_ode_output_timeseries
+        phase.add_timeseries_output('tas_comp.TAS', shape=(1,), units='m/s')
+        # upgrade_doc: end simplified_ode_output_timeseries
+        """
+        import openmdao.api as om
+        import dymos as dm
+        from dymos.examples.aircraft_steady_flight.aircraft_ode import AircraftODE
+
+        p = om.Problem(model=om.Group())
+        p.driver = om.ScipyOptimizeDriver()
+        p.driver.declare_coloring()
+
+        transcription = dm.GaussLobatto(num_segments=1,
+                                        order=13,
+                                        compressed=False)
+        phase = dm.Phase(ode_class=AircraftODE, transcription=transcription)
+        p.model.add_subsystem('phase0', phase)
+
+        # Pass Reference Area from an external source
+        assumptions = p.model.add_subsystem('assumptions', om.IndepVarComp())
+        assumptions.add_output('S', val=427.8, units='m**2')
+        assumptions.add_output('mass_empty', val=1.0, units='kg')
+        assumptions.add_output('mass_payload', val=1.0, units='kg')
+
+        phase.set_time_options(initial_bounds=(0, 0),
+                               duration_bounds=(3600, 3600),
+                               duration_ref=3600)
+
+        phase.set_time_options(initial_bounds=(0, 0),
+                               duration_bounds=(3600, 3600),
+                               duration_ref=3600)
+
+        phase.add_state('range', units='km', fix_initial=True, fix_final=False, scaler=0.01,
+                        rate_source='range_rate_comp.dXdt:range',
+                        defect_scaler=0.01)
+        phase.add_state('mass_fuel', units='kg', fix_final=True, upper=20000.0, lower=0.0,
+                        rate_source='propulsion.dXdt:mass_fuel',
+                        scaler=1.0E-4, defect_scaler=1.0E-2)
+        phase.add_state('alt',
+                        rate_source='climb_rate',
+                        units='km', fix_initial=True)
+
+        phase.add_control('mach',  targets=['tas_comp.mach', 'aero.mach'], units=None, opt=False)
+
+        phase.add_control('climb_rate', targets=['gam_comp.climb_rate'], units='m/s', opt=False)
+
+        phase.add_parameter('S',
+                            targets=['aero.S', 'flight_equilibrium.S', 'propulsion.S'],
+                            units='m**2')
+
+        phase.add_parameter('mass_empty', targets=['mass_comp.mass_empty'], units='kg')
+        phase.add_parameter('mass_payload', targets=['mass_comp.mass_payload'], units='kg')
+
[…]
