### 2020-09-21 edc05dc
add missing units for add_timeseries_output with wildcards
