+++ dymos/transcriptions/pseudospectral/radau_pseudospectral.py
@@ -357,6 +357,7 @@ class Radau(PseudospectralBase):
 
             for var, options in phase._timeseries[timeseries_name]['outputs'].items():
                 output_name = options['output_name']
+                units = options.get('units', None)
 
                 if '*' in var:  # match outputs from the ODE
                     ode_outputs = {opts['prom_name']: opts for (k, opts) in
@@ -368,6 +369,7 @@ class Radau(PseudospectralBase):
                 for v in matches:
                     if '*' in var:
                         output_name = v.split('.')[-1]
+                        units = ode_outputs[v]['units']
 
                     # Determine the path to the variable which we will be constraining
                     # This is more complicated for path constraints since, for instance,
@@ -381,7 +383,7 @@ class Radau(PseudospectralBase):
 
                     try:
                         shape, units = get_source_metadata(phase.rhs_all, src=v,
-                                                           user_units=options['units'],
+                                                           user_units=units,
                                                            user_shape=options['shape'])
                     except ValueError:
                         raise ValueError(f'Timeseries output {v} is not a known variable in'
