### 2020-09-21 b94afb8
Changed interpolation of 2nd derivative to be done at segment level rather than phase level
