### 2020-09-21 cae7b11
reworded a docstring
