+++ dymos/phase/phase.py
@@ -120,7 +120,8 @@ class Phase(om.Group):
             specified here then the unit will be determined from the rate_source.
         shape : tuple of int
             The shape of the state variable.  For instance, a 3D cartesian position vector would have
-            a shape of (3,).  This only needs to be specified if the rate_source target is a control.
+            a shape of (3,).  This only needs to be specified if the rate_source target points to
+            a control or state whose shape isn't known in time.
         rate_source : str
             The path to the ODE output which provides the rate of this state variable.
         targets : str or Sequence of str
@@ -193,7 +194,8 @@ class Phase(om.Group):
             specified here then the unit will be determined from the rate_source.
         shape : tuple of int
             The shape of the state variable.  For instance, a 3D cartesian position vector would have
-            a shape of (3,).  This only needs to be specified if the rate_source target is a control.
+            a shape of (3,).  This only needs to be specified if the rate_source target points to
+            a control or state whose shape isn't known in time.
         rate_source : str
             The path to the ODE output which provides the rate of this state variable.
         targets : str or Sequence of str
