### 2020-09-22 95a3b14
add unit_dict to allow overriding units found in ODE
