+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -468,6 +468,7 @@ class GaussLobatto(PseudospectralBase):
             for var, options in phase._timeseries[timeseries_name]['outputs'].items():
                 output_name = options['output_name']
                 units = options.get('units', None)
+                timeseries_units = options.get('timeseries_units', None)
 
                 if '*' in var:  # match outputs from the ODE
                     ode_outputs = {opts['prom_name']: opts for (k, opts) in
@@ -480,6 +481,9 @@ class GaussLobatto(PseudospectralBase):
                     if '*' in var:
                         output_name = v.split('.')[-1]
                         units = ode_outputs[v]['units']
+                        # check for timeseries_units override of ODE units
+                        if v in timeseries_units:
+                            units = timeseries_units[v]
 
                     # Determine the path to the variable which we will be constraining
                     # This is more complicated for path constraints since, for instance,
