### 2020-09-22 f3739aa
Fixes to work with OpenMDAO PR #1693 (#396)
* fix for removal of _loc_subsys_map

* getting rid of direct access to model_ref
