### 2020-09-22 3902e85
Wildcard units (#399)
* add missing units for add_timeseries_output with wildcards

* add test for mixing wildcard ODE match with units dictionary

* add unit_dict to allow overriding units found in ODE
