+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -477,6 +477,8 @@ class SolveIVP(TranscriptionBase):
 
         for var, options in phase._timeseries['timeseries']['outputs'].items():
             output_name = options['output_name']
+            units = options.get('units', None)
+            timeseries_units = options.get('timeseries_units', None)
 
             if '*' in var:  # match outputs from the ODE
                 ode_outputs = {opts['prom_name']: opts for (k, opts) in
@@ -488,6 +490,10 @@ class SolveIVP(TranscriptionBase):
             for v in matches:
                 if '*' in var:
                     output_name = v.split('.')[-1]
+                    units = ode_outputs[v]['units']
+                    # check for timeseries_units override of ODE units
+                    if v in timeseries_units:
+                        units = timeseries_units[v]
 
                 # Determine the path to the variable which we will be constraining
                 # This is more complicated for path constraints since, for instance,
@@ -500,7 +506,7 @@ class SolveIVP(TranscriptionBase):
                     continue
 
                 shape, units = get_source_metadata(phase.ode, src=v, user_shape=options['shape'],
-                                                   user_units=options['units'])
+                                                   user_units=units)
 
                 try:
                     timeseries_comp._add_output_configure(output_name, shape=shape, units=units, desc='')
