+++ dymos/transcriptions/runge_kutta/runge_kutta.py
@@ -791,6 +791,8 @@ class RungeKutta(TranscriptionBase):
 
             for var, options in timeseries_options['outputs'].items():
                 output_name = options['output_name']
+                units = options.get('units', None)
+                timeseries_units = options.get('timeseries_units', None)
 
                 if '*' in var:  # match outputs from the ODE
                     ode_outputs = {opts['prom_name']: opts for (k, opts) in
@@ -802,6 +804,10 @@ class RungeKutta(TranscriptionBase):
                 for v in matches:
                     if '*' in var:
                         output_name = v.split('.')[-1]
+                        units = ode_outputs[v]['units']
+                        # check for timeseries_units override of ODE units
+                        if v in timeseries_units:
+                            units = timeseries_units[v]
 
                     # Determine the path to the variable which we will be constraining
                     # This is more complicated for path constraints since, for instance,
@@ -815,7 +821,7 @@ class RungeKutta(TranscriptionBase):
 
                     try:
                         shape, units = get_source_metadata(phase.ode, src=v,
-                                                           user_units=options['units'],
+                                                           user_units=units,
                                                            user_shape=options['shape'])
                     except ValueError:
                         raise ValueError(f'Timeseries output {v} is not a known variable in'
