### 2020-09-23 0a2f4bb
Fixed the analysis error raised in the event of a simulation failure.
