+++ dymos/examples/min_time_climb/test/test_min_time_climb_simulate_failure.py
@@ -1,0 +1,121 @@
+import unittest
+
+import openmdao.api as om
+from openmdao.utils.assert_utils import assert_near_equal
+import dymos as dm
+from dymos.examples.min_time_climb.min_time_climb_ode import MinTimeClimbODE
+from openmdao.utils.testing_utils import use_tempdirs
+
+
+def min_time_climb(optimizer='SLSQP', num_seg=3, transcription='gauss-lobatto',
+                   transcription_order=3, force_alloc_complex=False):
+
+    p = om.Problem(model=om.Group())
+
+    p.driver = om.pyOptSparseDriver()
+    p.driver.options['optimizer'] = optimizer
+    p.driver.declare_coloring()
+
+    if optimizer == 'SNOPT':
+        p.driver.opt_settings['Major iterations limit'] = 1000
+        p.driver.opt_settings['iSumm'] = 6
+        p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-6
+        p.driver.opt_settings['Major optimality tolerance'] = 1.0E-6
+        p.driver.opt_settings['Function precision'] = 1.0E-12
+        p.driver.opt_settings['Linesearch tolerance'] = 0.1
+        p.driver.opt_settings['Major step limit'] = 0.5
+
+    t = {'gauss-lobatto': dm.GaussLobatto(num_segments=num_seg, order=transcription_order),
+         'radau-ps': dm.Radau(num_segments=num_seg, order=transcription_order),
+         'runge-kutta': dm.RungeKutta(num_segments=num_seg)}
+
+    traj = dm.Trajectory()
+
+    phase = dm.Phase(ode_class=MinTimeClimbODE, transcription=t[transcription])
+    traj.add_phase('phase0', phase)
+
+    p.model.add_subsystem('traj', traj)
+
+    phase.set_time_options(fix_initial=True, duration_bounds=(50, 400),
+                           duration_ref=100.0)
+
+    phase.add_state('r', fix_initial=True, lower=0, upper=1.0E6,
+                    ref=1.0E3, defect_ref=1.0E3, units='m',
+                    rate_source='flight_dynamics.r_dot')
+
+    phase.add_state('h', fix_initial=True, lower=0, upper=20000.0,
+                    ref=1.0E2, defect_ref=1.0E2, units='m',
+                    rate_source='flight_dynamics.h_dot', targets=['h'])
+
+    phase.add_state('v', fix_initial=True, lower=10.0,
+                    ref=1.0E2, defect_ref=1.0E2, units='m/s',
+                    rate_source='flight_dynamics.v_dot', targets=['v'])
+
+    phase.add_state('gam', fix_initial=True, lower=-1.5, upper=1.5,
+                    ref=1.0, defect_ref=1.0, units='rad',
+                    rate_source='flight_dynamics.gam_dot', targets=['gam'])
+
+    phase.add_state('m', fix_initial=True, lower=10.0, upper=1.0E5,
+                    ref=1.0E3, defect_ref=1.0E3, units='kg',
+                    rate_source='prop.m_dot', targets=['m'])
+
+    phase.add_control('alpha', units='deg', lower=-8.0, upper=8.0, scaler=1.0,
+                      rate_continuity=True, rate_continuity_scaler=100.0,
+                      rate2_continuity=False, targets=['alpha'])
+
+    phase.add_parameter('S', val=49.2386, units='m**2', opt=False, targets=['S'])
+    phase.add_parameter('Isp', val=1600.0, units='s', opt=False, targets=['Isp'])
+    phase.add_parameter('throttle', val=1.0, opt=False, targets=['throttle'])
+
+    phase.add_boundary_constraint('h', loc='final', equals=20000, scaler=1.0E-3, units='m')
[…]
