### 2020-09-23 e1e859d
removed some unnecessary detail from the simulate failure test.
