+++ dymos/transcriptions/common/test/test_continuity_comp.py
@@ -63,15 +63,16 @@ class TestContinuityComp(unittest.TestCase):
         state_options['x']['units'] = 'm'
         state_options['y']['units'] = 'm/s'
 
-        # Need these for later in the test.
-        state_options['x']['shape'] = (1, )
-        state_options['y']['shape'] = (1, )
-
         control_options['u']['units'] = 'deg'
         control_options['u']['shape'] = (3,)
         control_options['u']['continuity'] = True
 
         control_options['v']['units'] = 'N'
+
+        # Need these for later in the test.
+        state_options['x']['shape'] = (1, )
+        state_options['y']['shape'] = (1, )
+        control_options['v']['shape'] = (1, )
 
         if transcription == 'gauss-lobatto':
             cnty_comp = GaussLobattoContinuityComp(grid_data=gd, time_units='s',
