+++ dymos/examples/hyper_sensitive/test/test_hyper_sensitive.py
@@ -87,7 +87,7 @@ class TestHyperSensitive(unittest.TestCase):
 
     def test_hyper_sensitive_radau(self):
         p = self.make_problem(transcription=Radau, optimizer='IPOPT')
-        dm.run_problem(p, refine=True)
+        dm.run_problem(p)
         ui, uf, J = self.solution()
 
         assert_near_equal(p.get_val('traj.phase0.timeseries.controls:u')[0],
@@ -105,7 +105,7 @@ class TestHyperSensitive(unittest.TestCase):
     def test_hyper_sensitive_gauss_lobatto(self):
         p = self.make_problem(transcription=GaussLobatto, optimizer='IPOPT')
         # p.run_driver()
-        dm.run_problem(p, refine=True)
+        dm.run_problem(p)
 
         ui, uf, J = self.solution()
 
