+++ dymos/utils/misc.py
@@ -89,7 +89,7 @@ def get_targets(ode, name, user_targets):
 
 
 def get_target_metadata(ode, name, user_targets=_unspecified, user_units=_unspecified,
-                        user_shape=_unspecified):
+                        user_shape=_unspecified, control_rate=False):
     """
     Return the targets of a state variable in a given ODE system.
     If the targets of the state is _unspecified, and the state name is a top level input name
@@ -111,6 +111,9 @@ def get_target_metadata(ode, name, user_targets=_unspecified, user_units=_unspec
         Units for the variable as given by the user.
     user_shape : str or None or Sequence or _unspecified
         Shape for the variable as given by the user.
+    control_rate : bool
+        When True, check for the control rate if the name is not in the ODE.
+
     Returns
     -------
     shape : tuple
@@ -123,12 +126,16 @@ def get_target_metadata(ode, name, user_targets=_unspecified, user_units=_unspec
     this method should be called from configure of some parent Group, and the ODE should
     be a system within that Group.
     """
+    rate_src = False
     ode_inputs = {opts['prom_name']: opts for (k, opts) in
                   ode.get_io_metadata(iotypes=('input', 'output')).items()}
 
     if user_targets is _unspecified:
         if name in ode_inputs:
             targets = [name]
+        elif control_rate and f'{name}_rate' in ode_inputs:
+            targets = [f'{name}_rate']
+            rate_src = True
         else:
             targets = []
     elif user_targets:
@@ -143,6 +150,8 @@ def get_target_metadata(ode, name, user_targets=_unspecified, user_units=_unspec
         target_units_set = {ode_inputs[tgt]['units'] for tgt in targets}
         if len(target_units_set) == 1:
             units = next(iter(target_units_set))
+            if rate_src:
+                units = f"{units}*s"
         else:
             raise ValueError(f'Unable to automatically assign units to {name}. '
                              f'Targets have multiple units: {target_units_set}. '
