### 2020-09-23 2243f2e
Configure Refactor 1 - States pull their units and shape from the rate_source. (#398)
* fix for removal of _loc_subsys_map

* getting rid of direct access to model_ref

* set of final fixes

* added a little docs

* reworded a docstring

* review

Co-authored-by: Bret Naylor <naylor.b@gmail.com>
Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
