+++ dymos/examples/brachistochrone/test/ex_brachistochrone_vector_states.py
@@ -50,13 +50,10 @@ def brachistochrone_min_time(transcription='gauss-lobatto', num_segments=8, tran
 
     phase.add_state('pos',
                     rate_source=BrachistochroneVectorStatesODE.states['pos']['rate_source'],
-                    units=BrachistochroneVectorStatesODE.states['pos']['units'],
-                    shape=BrachistochroneVectorStatesODE.states['pos']['shape'],
                     fix_initial=True, fix_final=fix_final, solve_segments=solve_segments)
     #
     phase.add_state('v',
                     rate_source=BrachistochroneVectorStatesODE.states['v']['rate_source'],
-                    units=BrachistochroneVectorStatesODE.states['v']['units'],
                     fix_initial=True, fix_final=False, solve_segments=solve_segments)
     #
     phase.add_control('theta',
