+++ dymos/transcriptions/common/continuity_comp.py
@@ -34,7 +34,7 @@ class ContinuityCompBase(om.ExplicitComponent):
             return
 
         for state_name, options in state_options.items():
-            shape = options['shape']
+            shape = options['shape'] if options['shape'] is not None else (1, )
             size = np.prod(shape)
             units = options['units']
 
@@ -201,7 +201,11 @@ class ContinuityCompBase(om.ExplicitComponent):
                 't_duration', dependent=True
             )
 
-    def setup(self):
+    def configure_io(self):
+        """
+        I/O creation is delayed until configure so that we can determine the shape and units for
+        the states.
+        """
         self.rate_jac_templates = {}
         self.name_maps = {}
 
