+++ dymos/transcriptions/pseudospectral/components/test/test_collocation_defect_sol_opt.py
@@ -9,6 +9,11 @@ from dymos.transcriptions.grid_data import GridData
 from dymos.transcriptions.pseudospectral.components import CollocationComp
 from dymos.transcriptions.pseudospectral.components.state_independents import StateIndependentsComp
 
+# Modify class so we can run it standalone.
+from dymos.utils.misc import CompWrapperConfig
+CollocationComp = CompWrapperConfig(CollocationComp)
+StateIndependentsComp = CompWrapperConfig(StateIndependentsComp)
+
 
 class TestCollocationCompSolOpt(unittest.TestCase):
 
