### 2020-09-24 92e36e8
revert travis pyoptsparse rework
