+++ .travis.yml
@@ -9,8 +9,8 @@ os:
   - linux
 
 env:
-  - PY=3.7 PETSc=3.10.0
-  - PY=3.6 PETSc=3.10.0 UPLOAD_DOCS=1
+  - PY=3.7 PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3.12
+  - PY=3.6 PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3.12 UPLOAD_DOCS=1
 
 language: generic
 
@@ -87,23 +87,52 @@ install:
     sudo apt-get install gfortran;
 
     python -m pip install --upgrade pip;
+
+    if [ "$PETSc" ]; then
+      echo " >> Installing PETSc";
+      conda install -c anaconda mpi4py --yes;
+      conda install -c conda-forge petsc=$PETSc petsc4py --yes;
+    fi
+
+    if [ "$PYOPTSPARSE" ]; then
+      echo " >> Installing pyOptSparse";
+      git clone -q https://github.com/OpenMDAO/build_pyoptsparse;
+
+      cd build_pyoptsparse;
+      chmod 755 ./build_pyoptsparse.sh;
+
+      if [ "$PETSc" ]; then
+        PAROPT=-a;
+      fi
+
+      if [ "$SNOPT" == "7.7" ] && [ "$SNOPT_LOCATION_77" ]; then
+        echo "  > Secure copying SNOPT 7.7 over SSH";
+        mkdir SNOPT;
+        scp -qr "$SNOPT_LOCATION_77" SNOPT;
+        ./build_pyoptsparse.sh -b "$PYOPTSPARSE" -s SNOPT/src $PAROPT;
+
+      elif [ "$SNOPT" == "7.2" ] && [ "$SNOPT_LOCATION_72" ]; then
+        echo "  > Secure copying SNOPT 7.2 over SSH";
+        mkdir SNOPT;
+        scp -qr "$SNOPT_LOCATION_72" SNOPT;
+        ./build_pyoptsparse.sh -b "$PYOPTSPARSE" -s SNOPT/source $PAROPT;
+
+      else
+        if [ "$SNOPT" ]; then
+          echo "SNOPT version $SNOPT was requested but source is not available";
+        fi
+        ./build_pyoptsparse.sh -b "$PYOPTSPARSE" $PAROPT;
+      fi
+
+      cd ..;
+    fi
+
     python -m pip install sqlitedict;
     python -m pip install nose;
     python -m pip install networkx;
     python -m pip install testflo==1.3.6;
     python -m pip install pyyaml;
     python -m pip install coveralls;
-
-    echo "Install pyoptsparse and along with IPOPT";
-    chmod 755 ./build_pyoptsparse_ipopt.sh; ./build_pyoptsparse_ipopt.sh;
-
-    if [ "$SNOPT_LOCATION" ] && [ "${PY:0:1}" = "3" ]; then
-      echo "Install SNOPT";
-      cd pyoptsparse/pyoptsparse/pySNOPT;
-      scp -r "$SNOPT_LOCATION";
-      cd ../../../;
-      pwd;
-    fi
 
     git clone https://github.com/OpenMDAO/MBI.git;
     cd MBI;
