### 2020-09-24 ca65eb7
added more upgrade docs for add_timeseries_output
