### 2020-09-24 0e25093
fixed an issue with plots in the docs.  fixed more upgrade guide inclusions.
