### 2020-09-24 d29aff1
Added restart arg and logic
