+++ dymos/examples/double_integrator/doc/test_doc_double_integrator.py
@@ -44,12 +44,11 @@ class TestDoubleIntegratorForDocs(unittest.TestCase):
         # Set the options for our variables.
         #
         phase.set_time_options(fix_initial=True, fix_duration=True, units='s')
-        phase.add_state('v', fix_initial=True, fix_final=True, rate_source='u', units='m/s',
-                        shape=(1, ))
+        phase.add_state('v', fix_initial=True, fix_final=True, rate_source='u', units='m/s')
         phase.add_state('x', fix_initial=True, rate_source='v', units='m')
 
         phase.add_control('u', units='m/s**2', scaler=0.01, continuity=False, rate_continuity=False,
-                          rate2_continuity=False, lower=-1.0, upper=1.0)
+                          rate2_continuity=False, shape=(1, ), lower=-1.0, upper=1.0)
 
         #
         # Maximize distance travelled.
