### 2020-09-24 dc86fab
added a doc section on using partial derivative coloring to improve performance
