+++ dymos/examples/min_time_climb/doc/test_doc_min_time_climb_fd.py
@@ -1,0 +1,174 @@
+import os.path
+from pprint import pprint
+import shutil
+import unittest
+
+import matplotlib
+matplotlib.use('Agg')
+import matplotlib.pyplot as plt
+plt.style.use('ggplot')
+
+from dymos.utils.doc_utils import save_for_docs
+from openmdao.utils.coloring import Coloring
+
+
+def _view_coloring(coloring_file, show_sparsity_text=False, show_sparsity=True,
+                   subjac_sparsity=False, color_var=None, show_meta=False):
+    coloring = Coloring.load(coloring_file)
+    if show_sparsity_text:
+        coloring.display_txt()
+
+    if show_sparsity:
+        coloring.display()
+
+    if subjac_sparsity:
+        print("\nSubjacobian sparsity:")
+        for tup in coloring._subjac_sparsity_iter():
+            print("(%s, %s)\n   rows=%s\n   cols=%s" % tup[:4])
+        print()
+
+    if color_var is not None:
+        fwd, rev = coloring.get_row_var_coloring(color_var)
+        print("\nVar: %s  (fwd solves: %d,  rev solves: %d)\n" % (color_var, fwd, rev))
+
+    if show_meta:
+        print("\nColoring metadata:")
+        pprint(coloring._meta)
+
+    coloring.summary()
+
+class TestMinTimeClimbForDocs(unittest.TestCase):
+
+    @save_for_docs
+    def test_min_time_climb_for_docs_partial_coloring(self):
+        import openmdao.api as om
+        from openmdao.utils.assert_utils import assert_near_equal
+
+        import dymos as dm
+        from dymos.examples.min_time_climb.doc.min_time_climb_ode_partial_coloring import MinTimeClimbODE
+
+        for fd in (False, True):
+            if fd:
+                pc_options = (False, True)
+            else:
+                pc_options = (False,)
+            for pc in pc_options:
+                with self.subTest(f'Finite Differencing: {fd}  Partial Coloring: {pc}'):
+                    print(f'Finite Differencing: {fd}  Partial Coloring: {pc}')
+
+                    #
+                    # Instantiate the problem and configure the optimization driver
+                    #
+                    p = om.Problem(model=om.Group())
+
+                    p.driver = om.pyOptSparseDriver()
+                    p.driver.options['optimizer'] = 'IPOPT'
+                    p.driver.declare_coloring(tol=1.0E-12)
+                    p.driver.opt_settings['max_iter'] = 500
+                    p.driver.opt_settings['alpha_for_y'] = 'safer-min-dual-infeas'
+
+                    #
+                    # Instantiate the trajectory and phase
+                    #
+                    traj = dm.Trajectory()
+
+                    phase = dm.Phase(ode_class=MinTimeClimbODE,
+                                     ode_init_kwargs={'fd': fd, 'partial_coloring': pc},
+                                     transcription=dm.GaussLobatto(num_segments=30))
+
+                    traj.add_phase('phase0', phase)
+
+                    p.model.add_subsystem('traj', traj)
+
+                    #
+                    # Set the options on the optimization variables
+                    #
+                    phase.set_time_options(fix_initial=True, duration_bounds=(50, 400),
+                                           duration_ref=100.0)
+
+                    phase.add_state('r', fix_initial=True, lower=0, upper=1.0E6,
+                                    ref=1.0E3, defect_ref=1.0E3, units='m',
+                                    rate_source='flight_dynamics.r_dot')
+
+                    phase.add_state('h', fix_initial=True, lower=0, upper=20000.0,
+                                    ref=1.0E2, defect_ref=1.0E2, units='m',
+                                    rate_source='flight_dynamics.h_dot')
+
+                    phase.add_state('v', fix_initial=True, lower=10.0,
[…]
