### 2020-09-24 d0de725
Removed solve_segments check, refine_iteration_limit logic and default
