### 2020-09-25 90dc1d2
tweaked figure size for sparsity plots in the partial coloring demo
