+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -5,7 +5,8 @@ from ..transcription_base import TranscriptionBase
 from .components import SegmentSimulationComp, SegmentStateMuxComp, \
     SolveIVPControlGroup, SolveIVPPolynomialControlGroup, SolveIVPTimeseriesOutputComp
 from ..common import TimeComp
-from ...utils.misc import get_rate_units, get_targets, get_target_metadata, get_source_metadata
+from ...utils.misc import get_rate_units, get_targets, get_target_metadata, get_source_metadata, \
+     _unspecified
 from ...utils.indexing import get_src_indices_by_row
 from fnmatch import filter
 
@@ -83,10 +84,12 @@ class SolveIVP(TranscriptionBase):
         phase.add_subsystem('time', time_comp, promotes=['*'])
 
     def configure_time(self, phase):
-        time_options = phase.time_options
+        super(SolveIVP, self).configure_time(phase)
         num_seg = self.grid_data.num_segments
         grid_data = self.grid_data
         output_nodes_per_seg = self.options['output_nodes_per_seg']
+
+        phase.time.configure_io()
 
         for i in range(num_seg):
             phase.connect('t_initial', f'segment_{i}.t_initial')
@@ -116,17 +119,16 @@ class SolveIVP(TranscriptionBase):
         """
         Add an IndepVarComp for the states and setup the states as design variables.
         """
-        indep_states_ivc = phase.add_subsystem('indep_states', om.IndepVarComp(),
-                                               promotes_outputs=['*'])
-
-        for state_name, options in phase.state_options.items():
-            indep_states_ivc.add_output(f'initial_states:{state_name}',
-                                        val=np.ones(((1,) + options['shape'])),
-                                        units=options['units'])
+        phase.add_subsystem('indep_states', om.IndepVarComp(),
+                            promotes_outputs=['*'])
 
     def configure_states(self, phase):
         num_seg = self.grid_data.num_segments
-        ode_inputs = phase.ode.list_inputs(out_stream=None, units=True, shape=True, values=False)
+
+        for state_name, options in phase.state_options.items():
+            phase.indep_states.add_output(f'initial_states:{state_name}',
+                                          val=np.ones(((1,) + options['shape'])),
+                                          units=options['units'])
 
         for state_name, options in phase.state_options.items():
             # Connect the initial state to the first segment
@@ -202,7 +204,12 @@ class SolveIVP(TranscriptionBase):
                                                               **phase.options['ode_init_kwargs']))
 
     def configure_ode(self, phase):
-        pass
+        gd = self.grid_data
+        num_seg = gd.num_segments
+
+        for i in range(num_seg):
+            seg_comp = phase.segments._get_subsystem(f'segment_{i}')
+            seg_comp.configure_io()
 
     def setup_controls(self, phase):
         output_nodes_per_seg = self.options['output_nodes_per_seg']
@@ -221,9 +228,31 @@ class SolveIVP(TranscriptionBase):
                                           'control_rates:*'])
 
     def configure_controls(self, phase):
+        ode = phase._get_subsystem(self._rhs_source)
+
+        # Interrogate shapes and units.
+        for name, options in phase.control_options.items():
+
+            full_shape, units = get_target_metadata(ode, name=name,
+                                                    user_targets=options['targets'],
+                                                    user_units=options['units'],
+                                                    user_shape=options['shape'],
+                                                    control_rate=True)
+
+            if options['units'] is None:
+                options['units'] = units
+
+            # Determine and store the pre-discretized state shape for use by other components.
+            if len(full_shape) < 2:
+                if options['shape'] in (_unspecified, None):
+                    options['shape'] = (1, )
+            else:
[…]
