+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -34,6 +34,7 @@ class GaussLobatto(PseudospectralBase):
         super(GaussLobatto, self).setup_time(phase)
 
     def configure_time(self, phase):
+        super(GaussLobatto, self).configure_time(phase)
         options = phase.time_options
 
         # The tuples here are (name, user_specified_targets, dynamic)

[…]
