### 2020-09-25 023859e
started debugging documentation guide
