+++ dymos/transcriptions/common/test/test_control_interp_comp.py
@@ -14,6 +14,7 @@ from dymos.transcriptions.grid_data import GridData
 
 # Modify class so we can run it standalone.
 from dymos.utils.misc import CompWrapperConfig
+TimeComp = CompWrapperConfig(TimeComp)
 ControlInterpComp = CompWrapperConfig(ControlInterpComp)
 
 
