+++ dymos/transcriptions/runge_kutta/runge_kutta.py
@@ -22,6 +22,10 @@ class RungeKutta(TranscriptionBase):
     RungeKutta transcription in Dymos uses the RungeKutta-based shooting method which propagates
     the states from the phase initial time to the phase final time.
     """
+    def __init__(self, grid_data=None, **kwargs):
+        super(RungeKutta, self).__init__(**kwargs)
+        self._rhs_source = 'ode'
+
     def initialize(self):
 
         self.options.declare('method', default='RK4', values=('RK4',),
@@ -239,10 +243,11 @@ class RungeKutta(TranscriptionBase):
                                                     user_units=options['units'],
                                                     user_shape=options['shape'])
 
-            if options['units'] is None:
+            if options['units'] is _unspecified:
                 # Units are from the rate source and should be converted.
-                units = f'{units}*{time_units}'
-                options['units'] = units
+                if units is not None:
+                    units = f'{units}*{time_units}'
+            options['units'] = units
 
             # Determine and store the pre-discretized state shape for use by other components.
             if len(full_shape) < 2:
@@ -816,14 +821,9 @@ class RungeKutta(TranscriptionBase):
                     prom_name = 'parameters:{0}'.format(param_name)
                     tgt_name = 'input_values:parameters:{0}'.format(param_name)
 
-                    shape, units = get_target_metadata(phase.ode, name=param_name,
-                                                       user_targets=options['targets'],
-                                                       user_shape=options['shape'],
-                                                       user_units=options['units'])
-
                     timeseries_comp._add_output_configure(f'parameters:{param_name}',
-                                                          shape=shape,
-                                                          units=units,
+                                                          shape=options['shape'],
+                                                          units=options['units'],
                                                           desc='')
 
                     phase.promotes(timeseries_name, inputs=[(tgt_name, prom_name)],
