### 2020-09-25 6f34f21
removed errant n2 call
