### 2020-09-26 1bb67b8
minor tweaks to the run_problem refinement warning. only warn if refine_iteration_limit > 0 and run_driver=False.
