+++ dymos/examples/hyper_sensitive/test/test_hyper_sensitive.py
@@ -123,12 +123,11 @@ class TestHyperSensitive(unittest.TestCase):
 
     def test_refinement_warning(self):
         p = self.make_problem(transcription=Radau, optimizer='IPOPT')
-        dm.run_problem(p)
 
         msg = "Refinement not performed. Set run_driver to True to perform refinement."
 
         with warnings.catch_warnings(record=True) as ctx:
             warnings.simplefilter('always')
-            dm.run_problem(p, run_driver=False)
+            dm.run_problem(p, run_driver=False, refine_iteration_limit=10)
 
         self.assertIn(msg, [str(w.message) for w in ctx])
