### 2020-09-26 1795fc2
accounting for change in format of AnalysisError raise
