### 2020-09-28 bee71d8
fixed min time climb analysis error
