### 2020-09-28 99386c3
change assert to shorten error message
