### 2020-09-28 b91ad39
change AnalysisError text and check for expected result
