+++ dymos/phase/options.py
@@ -17,7 +17,7 @@ class ControlOptionsDictionary(om.OptionsDictionary):
         self.declare(name='name', types=str,
                      desc='The name of ODE system parameter to be controlled.')
 
-        self.declare(name='units', types=str, default=None,
+        self.declare(name='units', default=_unspecified,
                      allow_none=True, desc='The units in which the control variable is defined.')
 
         self.declare(name='desc', types=str, default='',
@@ -126,7 +126,7 @@ class PolynomialControlOptionsDictionary(om.OptionsDictionary):
         self.declare(name='name', types=str,
                      desc='The name of ODE system parameter to be controlled.')
 
-        self.declare(name='units', types=str, default=None,
+        self.declare(name='units', default=_unspecified,
                      allow_none=True, desc='The units in which the control variable is defined.')
 
         self.declare(name='desc', types=str, default='',
