### 2020-09-28 300083c
fixed min time climb analysis error (#409)
* change assert to shorten error message

* change AnalysisError text and check for expected result
