### 2020-09-28 a044da2
addressing review of JOSS paper and cleaning up the readme a bit
