+++ joss/paper.md
@@ -31,19 +31,28 @@ bibliography: paper.bib
 Dymos is an library for solving optimal-control type optimization problems. 
 It contains capabilities typical of optimal control software and can handle a wide range of typical optimal control problems.
 Its software design (built on top of NASA's OpenMDAO Framework [@Gray2019a] also allows users to tackle problems where the trajectory is not necessarily central to the optimization and where the system dynamics may include expensive, implicit calculations.
-Implicit calculations may be present due to some calculation in the ODE, or more generally because the ODE is posed as a set of differential algebraic equations or differential inclusions [@Seywald1994].
-
+Building dymos as a library on top of OpenMDAO extends its capabilities due to the efficiency with which OpenMDAO can compute analytic derivatives, even in the presence of iterative behavior.
+Optimal control software generally require that the dynamics of the system be defined as a set of explicit ordinary differential equations (ODE) which compute the rates of the state variables to be integrated.
+Sometimes the dynamics are instead posed as a set of differential algebraic equations, where some equality constraint needs to be satisfied at the solution in addition to the ODE.
+One application of this approach is the method of differential inclusions, in which the state time-history is posed as a dynamic control, and the traditional control variables needed to achieve that trajectory are found using a nonlinear solver within the ODE [@Seywald1994].
 Support for implicit calculations gives users more freedom to pose dynamics in more natural ways, but typically causes numerical and computational cost challenges in an optimization context, especially when finite-differences are used to compute derivatives for the optimizer.
-Dymos employs analytic derivatives to overcome these challenges and makes such methods computationally feasible.
-To achieve this, dymos makes extensive use of OpenMDAO's built in analytic derivative features, its non-linear solver library, and its support for gradient-based optimization.
+The ability to efficiently solve problems with nested implicit behavior without completely reformulating the problem sets dymos apart from many other optimal control software tools.
 
 Dymos can be used stand-alone, or as a building block in a larger model where a significant portion of the optimization is focused on some non-controls aspect with a trajectory added to enforce some constraint upon the design.
 In some contexts, you may hear this kind of problem referred to as co-design, controls-co-design, or multidisciplinary design optimization.
 
 ## The dymos perspective on optimal control
 
-In dymos trajectories are subdivided into "phases," and within each phase, boundary or path constraints may be applied to any output of the system dynamics.
-In addition to typical collocated dynamic control variables, dymos offers the ability to use variable order polynomial controls that are useful in forcing a lower-order control solution upon a phase of the trajectory.
+Dymos is similar to some other trajectory optimization tools in that the trajectory of a system is subdivided into chunks of time called _phases_.
+Breaking the trajectory into phases provides several capabilities.
+Intermediate constraints along a trajectory can be enforced by applying boundary constraint to a phase that begins or ends at the time of interest.
+For instance, the optimal trajectory of a launch vehicle may be required to ascend vertically to clear a launch tower before it pitches over on its way to orbit.
+Path constraints can be applied within each phase to bound some performance parameter within that phase.
[…]
