### 2020-09-28 aa5fab4
removed verbose plot creation code from JOSS paper
