### 2020-09-28 9978df8
Default unit for param/state is now _unspecified. You can now specify None (or inherit it from a unitless target.) (#407)
* fix for removal of _loc_subsys_map

* getting rid of direct access to model_ref

* Default unit for param/state is now _unspecified. You can now specify None.

Co-authored-by: Bret Naylor <naylor.b@gmail.com>
Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
