+++ joss/paper.md
@@ -32,7 +32,7 @@ Dymos is an library for solving optimal-control type optimization problems.
 It contains capabilities typical of optimal control software and can handle a wide range of typical optimal control problems.
 Its software design (built on top of NASA's OpenMDAO Framework [@Gray2019a] also allows users to tackle problems where the trajectory is not necessarily central to the optimization and where the system dynamics may include expensive, implicit calculations.
 Building dymos as a library on top of OpenMDAO extends its capabilities due to the efficiency with which OpenMDAO can compute analytic derivatives, even in the presence of iterative behavior.
-Optimal control software generally require that the dynamics of the system be defined as a set of explicit ordinary differential equations (ODE) which compute the rates of the state variables to be integrated.
+Optimal control software generally requires that the dynamics of the system be defined as a set of explicit ordinary differential equations (ODE) which compute the rates of the state variables to be integrated.
 Sometimes the dynamics are instead posed as a set of differential algebraic equations, where some equality constraint needs to be satisfied at the solution in addition to the ODE.
 One application of this approach is the method of differential inclusions, in which the state time-history is posed as a dynamic control, and the traditional control variables needed to achieve that trajectory are found using a nonlinear solver within the ODE [@Seywald1994].
 Support for implicit calculations gives users more freedom to pose dynamics in more natural ways, but typically causes numerical and computational cost challenges in an optimization context, especially when finite-differences are used to compute derivatives for the optimizer.
