### 2020-09-29 6bcb289
upgrade_doc printed message about missing feature in a doc string is now a warning.  more work on the debugging N2 diagram docs, waiting for the latest AutoIVC changes to the N2
