### 2020-09-29 d4b3c41
Configure refactor for controls and time. (#412)
* fix for removal of _loc_subsys_map

* getting rid of direct access to model_ref

* a few fixes

* finished time and solv_ivp

* Default unit for param/state is now _unspecified. You can now specify None.

Co-authored-by: Bret Naylor <naylor.b@gmail.com>
Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
