### 2020-09-29 3729f71
Cleaned up skitests for IPOPT
