### 2020-09-29 880bb31
made a few new sections in debugging not yet filled with content
