### 2020-09-29 08cd34a
First set of bugs related to configure refactor. (#414)
* Several fixes and working on more.

* Cleaned up skiptests for IPOPT

Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
