### 2020-09-29 6a60a81
[WIP] Doc update for 0.16.0 (#392)
* Added an upgrade guide and a new doc macro for demonstrating old vs. new behavior

* example docs now remove assert calls in the published docs.  added redbaron as a dependency needed for building the docs.

* added more upgrade docs for add_timeseries_output

* fixed an issue with plots in the docs.  fixed more upgrade guide inclusions.

* Reversed the order of the "new behavior" and "deprecated behavior" tabs in the upgrade docs, so that the new way is displayed by default.

* added a doc section on using partial derivative coloring to improve performance
[…]
