+++ dymos/examples/min_time_climb/doc/min_time_climb_ode_partial_coloring.py
@@ -1,0 +1,44 @@
+import openmdao.api as om
+from ....models.atmosphere import USatm1976Comp
+from .aero_partial_coloring import AeroGroup
+from ..prop import PropGroup
+from ....models.eom import FlightPathEOM2D
+
+
+class MinTimeClimbODE(om.Group):
+
+    def initialize(self):
+        self.options.declare('num_nodes', types=int)
+        self.options.declare('fd', types=bool, default=False, desc='If True, use fd for partials')
+        self.options.declare('partial_coloring', types=bool, default=False,
+                             desc='If True and fd is True, color the approximated partials')
+
+    def setup(self):
+        nn = self.options['num_nodes']
+
+        self.add_subsystem(name='atmos',
+                           subsys=USatm1976Comp(num_nodes=nn),
+                           promotes_inputs=['h'])
+
+        self.add_subsystem(name='aero',
+                           subsys=AeroGroup(num_nodes=nn,
+                                            fd=self.options['fd'],
+                                            partial_coloring=self.options['partial_coloring']),
+                           promotes_inputs=['v', 'alpha', 'S'])
+
+        self.connect('atmos.sos', 'aero.sos')
+        self.connect('atmos.rho', 'aero.rho')
+
+        self.add_subsystem(name='prop',
+                           subsys=PropGroup(num_nodes=nn),
+                           promotes_inputs=['h', 'Isp', 'throttle'])
+
+        self.connect('aero.mach', 'prop.mach')
+
+        self.add_subsystem(name='flight_dynamics',
+                           subsys=FlightPathEOM2D(num_nodes=nn),
+                           promotes_inputs=['m', 'v', 'gam', 'alpha'])
+
+        self.connect('aero.f_drag', 'flight_dynamics.D')
+        self.connect('aero.f_lift', 'flight_dynamics.L')
+        self.connect('prop.thrust', 'flight_dynamics.T')
