+++ dymos/examples/min_time_climb/aero/test/test_kappa_comp.py
@@ -4,7 +4,6 @@ import numpy as np
 
 import openmdao.api as om
 from openmdao.utils.assert_utils import assert_check_partials, assert_near_equal
-
 from dymos.examples.min_time_climb.aero.kappa_comp import KappaComp
 
 
@@ -12,26 +11,13 @@ class TestKappaComp(unittest.TestCase):
 
     def test_value(self):
         n = 500
-
-        p = om.Problem(model=om.Group())
-
-        ivc = om.IndepVarComp()
-
-        ivc.add_output(name='mach', units=None, val=np.zeros(n))
-
-        p.model.add_subsystem(name='ivc', subsys=ivc, promotes_outputs=['*'])
-
-        p.model.add_subsystem(name='kappa_comp',
-                              subsys=KappaComp(num_nodes=n))
-
-        p.model.connect('mach', 'kappa_comp.mach')
-
+        p = om.Problem()
+        p.model.add_subsystem(name='kappa_comp', subsys=KappaComp(num_nodes=n))
         p.setup()
-
-        p['mach'] = np.linspace(0, 1.8, n)
+        p.set_val('kappa_comp.mach', np.linspace(0, 1.8, n))
         p.run_model()
 
-        M = p.get_val('mach')
+        M = p.get_val('kappa_comp.mach')
         kappa = p.get_val('kappa_comp.kappa')
 
         idxs_0 = np.where(M <= 1.15)[0]
@@ -44,28 +30,12 @@ class TestKappaComp(unittest.TestCase):
         assert_near_equal(kappa[idxs_1], kappa_analtic_1)
 
     def test_partials(self):
-
         n = 10
-
         p = om.Problem(model=om.Group())
-
-        ivc = om.IndepVarComp()
-
-        ivc.add_output(name='mach', units=None, val=np.zeros(n))
-
-        p.model.add_subsystem(name='ivc', subsys=ivc, promotes_outputs=['*'])
-
-        p.model.add_subsystem(name='kappa_comp',
-                              subsys=KappaComp(num_nodes=n))
-
-        p.model.connect('mach', 'kappa_comp.mach')
-
+        p.model.add_subsystem(name='kappa_comp', subsys=KappaComp(num_nodes=n))
         p.setup()
-
-        p['mach'] = np.linspace(0, 1.8, n)
-
+        p.set_val('kappa_comp.mach', np.linspace(0, 1.8, n))
         p.run_model()
-
         cpd = p.check_partials(compact_print=False, out_stream=None)
         assert_check_partials(cpd, atol=1.0E-5, rtol=1.0E-4)
 
