+++ mkdocs/docs/api/trajectory_api.md
@@ -1,3 +1,3 @@
 # The Trajectory API
 
-{{ api_doc('dymos.Trajectory', members=['add_phase', 'link_phases']) }}
+{{ api_doc('dymos.Trajectory', members=['add_phase','link_phases','add_parameter', 'simulate']) }}

[…]
