### 2020-09-30 0655517
reworked intro dymos perspective sections
