+++ joss/paper.md
@@ -28,41 +28,57 @@ bibliography: paper.bib
 
 # Summary
 
-Dymos is an library for solving optimal-control type optimization problems. 
-It contains capabilities typical of optimal control software and can handle a wide range of typical optimal control problems.
-Its software design (built on top of NASA's OpenMDAO Framework [@Gray2019a] also allows users to tackle problems where the trajectory is not necessarily central to the optimization and where the system dynamics may include expensive, implicit calculations.
-Building dymos as a library on top of OpenMDAO extends its capabilities due to the efficiency with which OpenMDAO can compute analytic derivatives, even in the presence of iterative behavior.
-Optimal control software generally requires that the dynamics of the system be defined as a set of explicit ordinary differential equations (ODE) which compute the rates of the state variables to be integrated.
-Sometimes the dynamics are instead posed as a set of differential algebraic equations, where some equality constraint needs to be satisfied at the solution in addition to the ODE.
+Dymos is an library for optimizing control schedules for dynamic systems --- sometimes referred to as trajectory optimization. 
+There are a number of other optimal control libraries that tackle a similar kind of problem, such as OTIS4, GPOPS, CASADI [CITATIONS!!]. 
+Dymos can operate in a similar manner to these codes where it is focused purely on the optimal control problem, 
+but it is unique in that it can also support broader cases where trajectory optimization is just one component in a larger physical model. 
+This allows Dymos to play a critical role in solving problems where the physical design of a system is optimized simultaneously with control schedule governing its transient performance. 
+This kind of problem is often referred to as co-design, controls-co-design, or multidisciplinary design optimization.
+Clear examples of Dymos used in this context are demonstrated in the coupled trajectory-thermal design of electric aircraft [CITATIONS]. 
+
+Optimal control software typically requires that the dynamics of the system be defined as a set of ordinary differential equations (ODE) that use explicit functions to compute the rates of the state variables to be time-integrated.
+Sometimes the dynamics are instead posed as a set of differential algebraic equations, where some residual equations need to be satisfied implicitly in order to solve ODE.
 One application of this approach is the method of differential inclusions, in which the state time-history is posed as a dynamic control, and the traditional control variables needed to achieve that trajectory are found using a nonlinear solver within the ODE [@Seywald1994].
 Support for implicit calculations gives users more freedom to pose dynamics in more natural ways, but typically causes numerical and computational cost challenges in an optimization context, especially when finite-differences are used to compute derivatives for the optimizer.
-The ability to efficiently solve problems with nested implicit behavior without completely reformulating the problem sets dymos apart from many other optimal control software tools.
+Some optimal control libraries tackle this numerical challenge with a monolithic algorithmic differentiation[CITE Greiwank paper?] approach. 
+While effective, the monolithic nature is both less efficient[CITE MARTINS ADjoint paper] and less flexible. 
 
-Dymos can be used stand-alone, or as a building block in a larger model where a significant portion of the optimization is focused on some non-controls aspect with a trajectory added to enforce some constraint upon the design.
-In some contexts, you may hear this kind of problem referred to as co-design, controls-co-design, or multidisciplinary design optimization.
+Instead, Dymos uses a modular approach that allows users to select any combination of finite-difference, complex-step[CITE Martins CS paper], algorithmic differentiation, and hand differentiation that suits their modeling needs. 
+This flexibility is achieved by leveraging the data passing and efficient differentiation fearures of NASA's OpenMDAO framework[@Gray2019a]. 
+Dymos extends OpenMDAO by adding differentiated time-integration schemes along with the necessary APIs for users to plug their ODEs and DAEs into those schemes.
[…]
