### 2020-09-30 c3e6863
some tweaks to justins edits, and some gitignore cleanup
