+++ joss/paper.md
@@ -29,21 +29,21 @@ bibliography: paper.bib
 # Summary
 
 Dymos is an library for optimizing control schedules for dynamic systems --- sometimes referred to as trajectory optimization. 
-There are a number of other optimal control libraries that tackle a similar kind of problem, such as OTIS4, GPOPS, CASADI [CITATIONS!!]. 
+There are a number of other optimal control libraries that tackle a similar kind of problem, such as OTIS4 [@Paris2006], GPOPS-II [@Patterson2014GPOPSII],and CASADI [@Andersson2018].
 Dymos can operate in a similar manner to these codes where it is focused purely on the optimal control problem, 
 but it is unique in that it can also support broader cases where trajectory optimization is just one component in a larger physical model. 
 This allows Dymos to play a critical role in solving problems where the physical design of a system is optimized simultaneously with control schedule governing its transient performance. 
 This kind of problem is often referred to as co-design, controls-co-design, or multidisciplinary design optimization.
-Clear examples of Dymos used in this context are demonstrated in the coupled trajectory-thermal design of electric aircraft [CITATIONS]. 
+Clear examples of Dymos used in this context are demonstrated in the coupled trajectory-thermal design of electric aircraft [@Hariton2020a].
 
 Optimal control software typically requires that the dynamics of the system be defined as a set of ordinary differential equations (ODE) that use explicit functions to compute the rates of the state variables to be time-integrated.
 Sometimes the dynamics are instead posed as a set of differential algebraic equations, where some residual equations need to be satisfied implicitly in order to solve ODE.
 One application of this approach is the method of differential inclusions, in which the state time-history is posed as a dynamic control, and the traditional control variables needed to achieve that trajectory are found using a nonlinear solver within the ODE [@Seywald1994].
 Support for implicit calculations gives users more freedom to pose dynamics in more natural ways, but typically causes numerical and computational cost challenges in an optimization context, especially when finite-differences are used to compute derivatives for the optimizer.
-Some optimal control libraries tackle this numerical challenge with a monolithic algorithmic differentiation[CITE Greiwank paper?] approach. 
-While effective, the monolithic nature is both less efficient[CITE MARTINS ADjoint paper] and less flexible. 
+Some optimal control libraries tackle this numerical challenge with a monolithic algorithmic differentiation[@griewank2003mathematical] approach.
+While effective, the monolithic nature is both less efficient[@martins2005coupled] and less flexible.
 
-Instead, Dymos uses a modular approach that allows users to select any combination of finite-difference, complex-step[CITE Martins CS paper], algorithmic differentiation, and hand differentiation that suits their modeling needs. 
+Instead, Dymos uses a modular approach that allows users to select any combination of finite-difference, complex-step [@Martins2003CS], algorithmic differentiation, and hand differentiation that suits their modeling needs.
 This flexibility is achieved by leveraging the data passing and efficient differentiation fearures of NASA's OpenMDAO framework[@Gray2019a]. 
 Dymos extends OpenMDAO by adding differentiated time-integration schemes along with the necessary APIs for users to plug their ODEs and DAEs into those schemes.
 It also provides further APIs to enable in the input and output of information from the optimal control problem, in order to couple with the physical design models. 
@@ -58,18 +58,17 @@ For instance, the optimal trajectory of a launch vehicle may be required to asce
 Path constraints can be applied within each phase to bound some performance parameter within that phase.
 For example, reentry vehicles may need to adjust their trajectory to limit aerodynamic heating.
 
-Each phase in a trajectory can use its own separate ODE. 
-For instance an aircraft with vertical takeoff and landing capability may use different ODEs for vertical flight and horizontal flight. 
+Each phase in a trajectory can also use its own separate ODE.
[…]
