+++ joss/paper.bib
@@ -181,3 +181,106 @@ keywords = {49M37, Line search, Nonlinear programming, 90C30, Barrier method, 65
       adsurl = {https://rdcu.be/b08Wh},
       doi = {https://doi.org/10.1038/s41592-019-0686-2},
 }
+
+@Article{Andersson2018,
+  Author = {Joel A E Andersson and Joris Gillis and Greg Horn
+            and James B Rawlings and Moritz Diehl},
+  Title = {{CasADi} -- {A} software framework for nonlinear optimization
+           and optimal control},
+  Journal = {Mathematical Programming Computation},
+  volume={11},
+  pages={1--36},
+  Year = {2019},
+}
+
+@article{Patterson2014GPOPSII,
+author = {Patterson, Michael A. and Rao, Anil V.},
+title = {GPOPS-II: A MATLAB Software for Solving Multiple-Phase Optimal Control Problems Using Hp-Adaptive Gaussian Quadrature Collocation Methods and Sparse Nonlinear Programming},
+year = {2014},
+issue_date = {October 2014},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {41},
+number = {1},
+issn = {0098-3500},
+url = {https://doi.org/10.1145/2558904},
+doi = {10.1145/2558904},
+abstract = {A general-purpose MATLAB software program called GPOPS--II is described for solving multiple-phase optimal control problems using variable-order Gaussian quadrature collocation methods. The software employs a Legendre-Gauss-Radau quadrature orthogonal collocation method where the continuous-time optimal control problem is transcribed to a large sparse nonlinear programming problem (NLP). An adaptive mesh refinement method is implemented that determines the number of mesh intervals and the degree of the approximating polynomial within each mesh interval to achieve a specified accuracy. The software can be interfaced with either quasi-Newton (first derivative) or Newton (second derivative) NLP solvers, and all derivatives required by the NLP solver are approximated using sparse finite-differencing of the optimal control problem functions. The key components of the software are described in detail and the utility of the software is demonstrated on five optimal control problems of varying complexity. The software described in this article provides researchers a useful platform upon which to solve a wide variety of complex constrained optimal control problems.},
+journal = {ACM Trans. Math. Softw.},
+month = oct,
+articleno = {1},
+numpages = {37},
+keywords = {Optimal control, direct collocation, applied mathematics, hp--adaptive methods, scientific computation, numerical methods, Gaussian quadrature, MATLAB}
+}
+
+@inbook{Paris2006,
+author = {Stephen Paris and John Riehl and Waldy Sjauw},
+title = {Enhanced Procedures for Direct Trajectory Optimization Using Nonlinear Programming and Implicit Integration},
+booktitle = {AIAA/AAS Astrodynamics Specialist Conference and Exhibit},
+doi = {10.2514/6.2006-6309},
+URL = {https://arc.aiaa.org/doi/abs/10.2514/6.2006-6309},
+eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2006-6309}
+}
+
+@article{Martins2003CS,
+author = {Martins, Joaquim R. R. A. and Sturdza, Peter and Alonso, Juan J.},
+title = {The Complex-Step Derivative Approximation},
+year = {2003},
+issue_date = {September 2003},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {29},
+number = {3},
+issn = {0098-3500},
+url = {https://doi.org/10.1145/838250.838251},
+doi = {10.1145/838250.838251},
[…]
