+++ joss/paper.bib
@@ -284,3 +284,13 @@ eprint = {
   year={2005},
   publisher={Springer}
 }
+
+@article{mader2008adjoint,
+  title={ADjoint: An approach for the rapid development of discrete adjoint solvers},
+  author={Mader, Charles A and Martins, Joaquim RRA and Alonso, Juan J and Van Der Weide, Edwin},
+  journal={AIAA journal},
+  volume={46},
+  number={4},
+  pages={863--873},
+  year={2008}
+}
