### 2020-09-30 ac10896
another citation for adjoints
