+++ joss/paper.bib
@@ -294,3 +294,13 @@ eprint = {
   pages={863--873},
   year={2008}
 }
+
+@article{kenway2019effective,
+  title={Effective adjoint approaches for computational fluid dynamics},
+  author={Kenway, Gaetan KW and Mader, Charles A and He, Ping and Martins, Joaquim RRA},
+  journal={Progress in Aerospace Sciences},
+  volume={110},
+  pages={100542},
+  year={2019},
+  publisher={Elsevier}
+}
