+++ joss/paper.bib
@@ -304,3 +304,26 @@ eprint = {
   year={2019},
   publisher={Elsevier}
 }
+
+@article{garciasans2019,
+author = {Garcia-Sanz, Mario},
+title = {Control Co-Design: An engineering game changer},
+journal = {Advanced Control for Applications},
+volume = {1},
+number = {1},
+pages = {e18},
+keywords = {concurrent engineering, control, control co-design, design, systems engineering},
+doi = {10.1002/adc2.18},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/adc2.18},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/adc2.18},
+note = {e18 ADC2-19-0039},
+abstract = {Abstract Over the last few decades, control engineers have focused on developing innovative control theories and algorithms to regulate systems. These control efforts are usually at the last stage of a sequential strategy that allows engineering departments to work independently and consecutively toward the design of new products and systems. Control algorithms are usually developed at the end of that sequential process, once the mechanical, electrical, and other subsystems are completely defined. This paper discusses a different approach, named Control Co-Design (CCD). Following a concurrent engineering strategy that considers multidisciplinary subsystem interactions from the beginning of the design process, CCD applies control concepts to design the entire system and reach optimal solutions that are not achievable otherwise. This approach is a game changer for the control engineer, who will be not only the designer of advanced control algorithms but also the natural leader of the design of new products and systems. This paper describes some historic engineering breakthroughs achieved by applying CCD and explores some relevant application areas. It also presents three complementary CCD methodologies that include control-inspired paradigms, formal mathematical co-optimization techniques, and co-simulation campaigns to enhance engineering creativity and achieve radically new optimal designs.},
+year = {2019}
+}
+
+@misc{allison2004complex,
+  title={Complex system optimization: A review of analytical target cascading, collaborative optimization, and other formulations},
+  author={Allison, J},
+  year={2004},
+  publisher={A master’s thesis submitted in partial fulfillment of the requirements for~…}
+}
