+++ joss/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-Dymos is an library for optimizing control schedules for dynamic systems --- sometimes referred to as trajectory optimization. 
+Dymos is a library for optimizing control schedules for dynamic systems --- sometimes referred to as trajectory optimization.
 There are a number of other optimal control libraries that tackle a similar kind of problem, such as OTIS4 [@Paris2006], GPOPS-II [@Patterson2014GPOPSII],and CASADI [@Andersson2018].
 Dymos can operate in a similar manner to these codes where it is focused purely on the optimal control problem, 
 but it is unique in that it can also support broader cases where trajectory optimization is just one component in a larger physical model. 
@@ -37,7 +37,7 @@ This kind of problem is often referred to as co-design, controls-co-design, or m
 Clear examples of Dymos used in this context are demonstrated in the coupled trajectory-thermal design of electric aircraft [@Hariton2020a].
 
 Optimal control software typically requires that the dynamics of the system be defined as a set of ordinary differential equations (ODE) that use explicit functions to compute the rates of the state variables to be time-integrated.
-Sometimes the dynamics are instead posed as a set of differential algebraic equations, where some residual equations need to be satisfied implicitly in order to solve ODE.
+Sometimes the dynamics are instead posed as a set of differential algebraic equations (DAE), where some residual equations need to be satisfied implicitly in order to solve ODE.
 One application of this approach is the method of differential inclusions, in which the state time-history is posed as a dynamic control, and the traditional control variables needed to achieve that trajectory are found using a nonlinear solver within the ODE [@Seywald1994].
 Support for implicit calculations gives users more freedom to pose dynamics in more natural ways, but typically causes numerical and computational cost challenges in an optimization context, especially when finite-differences are used to compute derivatives for the optimizer.
 Some optimal control libraries tackle this numerical challenge with a monolithic algorithmic differentiation[@griewank2003mathematical] approach.
@@ -46,12 +46,12 @@ While effective, the monolithic nature is both less efficient[@mader2008adjoint;
 Instead, Dymos uses a modular approach that allows users to select any combination of finite-difference, complex-step [@Martins2003CS], algorithmic differentiation, and hand differentiation that suits their modeling needs.
 This flexibility is achieved by leveraging the data passing and efficient differentiation fearures of NASA's OpenMDAO framework[@Gray2019a]. 
 Dymos extends OpenMDAO by adding differentiated time-integration schemes along with the necessary APIs for users to plug their ODEs and DAEs into those schemes.
-It also provides further APIs to enable in the input and output of information from the optimal control problem, in order to couple with the physical design models. 
+It also provides further APIs to enable the input and output of information from the optimal control problem, in order to couple with the physical design models.
 
 
 ## The dymos perspective on optimal control
 
-Dymos breaks the trajectory of a system is into chunks of time called _phases_.
+Dymos breaks the trajectory of a system into chunks of time called _phases_.
 Breaking the trajectory into phases provides several capabilities.
 Intermediate constraints along a trajectory can be enforced by applying boundary constraint to a phase that begins or ends at the time of interest.
 For instance, the optimal trajectory of a launch vehicle may be required to ascend vertically to clear a launch tower before it pitches over on its way to orbit.
@@ -67,9 +67,9 @@ Every phase uses its own specific time discretization tailored to the dynamics i
 If one part of a trajectory has fast dynamics and another has slow dynamics, the time history can be broken into two phases with separate time discritizations.
 The available time discretization schemes in Dymos are based on pseudospectral methods, using two common direct collocation transcriptions:
 the high-order Gauss-Lobatto transcription [@Herman1996] and the Radau pseudospectral method [@Garg2009].
[…]
