### 2020-09-30 1778c50
addressing review of JOSS paper and cleaning up the readme a bit (#411)
* addressing review of JOSS paper and cleaning up the readme a bit

* removed verbose plot creation code from JOSS paper

* change per review from ken

* reworked intro dymos perspective sections

* some tweaks to justins edits

* different adjoint citation

* another citation for adjoints

Co-authored-by: Justin Gray <justin.s.gray@gmail.com>
