### 2020-09-30 #2648 @robfalck
@dpsanders

We've made some edits that hopefully will address your concerns here.

I would like to know the intent of demonstrating it's use within the paper itself.  Developing dymos models and optimizing them is not easily demonstrated in a compact format.

I've included an example but I wonder if we should instead link to examples in the repo.  Including this example significantly increases the length of the paper.
