### 2020-10-02 7390928
updated AnalysisError formatting
