+++ dymos/examples/min_time_climb/test/test_min_time_climb_simulate_failure.py
@@ -93,8 +93,9 @@ class TestMinTimeClimbSimulateFailure(unittest.TestCase):
 
         with self.assertRaises(om.AnalysisError) as e:
             dm.run_problem(p, run_driver=False, simulate=True)
-        expected = 'SegmentSimulationComp (traj.phases.phase0.segments.segment_2): Error calling compute(),' \
-                   ' solve_ivp failed: Required step size is less than spacing between numbers.'
+        expected = '\'traj.phases.phase0.segments.segment_2\' <class SegmentSimulationComp>: ' \
+                   'Error calling compute(), solve_ivp failed: Required step size is less than ' \
+                   'spacing between numbers.'
         self.assertEqual(str(e.exception), expected)
 
 
