### 2020-10-03 eff0764
clean up the readme a bit
