### 2020-10-04 c37a962
Fixed some doc typos I saw
