### 2020-10-05 eeacf59
Fixed some doc typos I saw (#419)
