### 2020-10-06 fd95804
added a new add_linkage_constraint method to Trajectory to generalize inter-phase constriants.
