### 2020-10-08 b1e0510
add_linkage_constraint is now working as the backend of link_phases. Warnings occur because initial and final values are often taken from a single source into separate inputs of the linkage constraint. This should be changed by always passing in the initial and final values and indexing accordingly. docs are still to be updated.
