+++ dymos/trajectory/phase_linkage_comp.py
@@ -1,0 +1,97 @@
+import numpy as np
+import openmdao.api as om
+from .options import LinkageOptionsDictionary
+
+
+class PhaseLinkageComp(om.ExplicitComponent):
+    """
+    Provides a 'linkage' capability between two phases to provide
+    continuity in states, time, and other variables between two
+    phases.
+
+    Conceptually, each linkage can be thought of as a set of compatibility constraints involving
+    one or more variables.
+    """
+    def initialize(self):
+        self.options.declare('linkages', default=[])
+        self._io_names = {}
+
+    def add_linkage_configure(self, lnk):
+        """
+        Add a linkage constraint to be managed by this component.
+
+        Each constraint equation consists of a variable in the first phase (var_a) at a given
+        location in the first phase (loc_a - either 'initial' or 'final') and a variable in the
+        second phase (var_b), with corresponding location (loc_b).
+
+        The resulting linkage equation of constraint is:
+
+        C_linkage = sign_a * vars_a + sign_b * vars_b
+
+        Constraining this linkage to a value of zero, with the default signs
+        (sign_a = 1, sign_b = -1) will result in the two variables having the same value at the
+        given locations.
+
+
+        Parameters
+        ----------
+        linkage_options : LinkageOptionsDictionary
+            The linkage options dictionary defining the given linkage constraint.
+
+        """
+        if lnk['connected']:
+            return
+
+        self.options['linkages'].append(lnk)
+
+        phase_a = lnk['phase_a']
+        phase_b = lnk['phase_b']
+        shape = lnk['shape']
+        units = lnk['units']
+
+        var_a = lnk['constraint_name'] if lnk['constraint_name'] else lnk['var_a']
+        var_b = lnk['constraint_name'] if lnk['constraint_name'] else lnk['var_b']
+
+        loc_a = 'initial' if lnk['loc_a'] in {'initial', '--', '-+'} else 'final'
+        loc_b = 'initial' if lnk['loc_b'] in {'initial', '--', '-+'} else 'final'
+
+        input_a = f'{phase_a}:{var_a}_{loc_a}'
+        input_b = f'{phase_b}:{var_b}_{loc_b}'
+        output = f'{phase_a}:{var_a}_{loc_a}|{phase_b}:{var_b}_{loc_b}'
+
+        lnk['input_a'] = input_a
+        lnk['input_b'] = input_b
+        lnk['output'] = output
+
+        try:
+            self.add_input(name=input_a, shape=shape, val=np.zeros(shape), units=units)
+        except ValueError:
+            pass
+
+        try:
+            self.add_input(name=input_b, shape=shape, val=np.zeros(shape), units=units)
+        except ValueError:
+            pass
+
+        self.add_output(name=output, shape=shape, val=np.zeros(shape), units=units)
+
+        if lnk['equals'] is None and lnk['lower'] is None and lnk['upper'] is None:
+            lnk['equals'] = 0.0
+
+        self.add_constraint(name=output, equals=lnk['equals'], lower=lnk['lower'],
+                            upper=lnk['upper'], ref=lnk['ref'], ref0=lnk['ref0'],
+                            scaler=lnk['scaler'], adder=lnk['adder'], linear=lnk['linear'])
+
+        ar = np.arange(np.prod(lnk['shape']))
+
+        self.declare_partials(of=output, wrt=input_a, rows=ar, cols=ar, val=lnk['sign_a'])
+        self.declare_partials(of=output, wrt=input_b, rows=ar, cols=ar, val=lnk['sign_b'])
+
+    def compute(self, inputs, outputs):
+
+        for lnk in self.options['linkages']:
+            input_a = lnk['input_a']
[…]

[…]
