### 2020-10-09 d199d45
added some argument checking to add_linkage_constraint
