+++ dymos/trajectory/test/test_phase_linkage_comp.py
@@ -1,0 +1,122 @@
+import unittest
+
+import numpy as np
+from numpy.testing import assert_almost_equal
+
+import openmdao.api as om
+from openmdao.utils.assert_utils import assert_check_partials
+
+from dymos.trajectory.phase_linkage_comp import PhaseLinkageComp
+from dymos.trajectory.options import LinkageOptionsDictionary
+
+
+class TestPhaseLinkageComp(unittest.TestCase):
+
+    def setUp(self):
+
+        self.p = om.Problem(model=om.Group())
+
+        ivp = self.p.model.add_subsystem('ivc', subsys=om.IndepVarComp(), promotes_outputs=['*'])
+
+        ndn = 20
+        nn = 30
+
+        ivp.add_output('phase0:x', val=np.zeros((ndn, 1)), units='m')
+        ivp.add_output('phase0:u', val=np.zeros((nn, 3)), units='deg')
+        ivp.add_output('phase0:v', val=np.zeros((nn, 3, 3)), units='N')
+
+        ivp.add_output('phase1:x', val=np.zeros((ndn, 1)), units='m')
+        ivp.add_output('phase1:u', val=np.zeros((nn, 3)), units='deg')
+        ivp.add_output('phase1:v', val=np.zeros((nn, 3, 3)), units='N')
+
+        linkage_comp = PhaseLinkageComp()
+
+        x_lnk = LinkageOptionsDictionary()
+        x_lnk['phase_a'] = 'phase0'
+        x_lnk['phase_b'] = 'phase1'
+        x_lnk['var_a'] = 'x'
+        x_lnk['var_b'] = 'x'
+        x_lnk['loc_a'] = 'final'
+        x_lnk['loc_b'] = 'initial'
+        x_lnk['units'] = 'm'
+        x_lnk['shape'] = (1,)
+        x_lnk['equals'] = 0.0
+
+        u_lnk = LinkageOptionsDictionary()
+        u_lnk['phase_a'] = 'phase0'
+        u_lnk['phase_b'] = 'phase1'
+        u_lnk['var_a'] = 'u'
+        u_lnk['var_b'] = 'u'
+        u_lnk['loc_a'] = 'final'
+        u_lnk['loc_b'] = 'initial'
+        u_lnk['units'] = 'deg'
+        u_lnk['shape'] = (3,)
+        u_lnk['equals'] = 0.0
+
+        v_lnk = LinkageOptionsDictionary()
+        v_lnk['phase_a'] = 'phase0'
+        v_lnk['phase_b'] = 'phase1'
+        v_lnk['var_a'] = 'v'
+        v_lnk['var_b'] = 'v'
+        v_lnk['loc_a'] = 'final'
+        v_lnk['loc_b'] = 'initial'
+        v_lnk['units'] = 'N'
+        v_lnk['shape'] = (3, 3)
+        v_lnk['equals'] = 0.0
+
+        linkage_comp.add_linkage_configure(x_lnk)
+        linkage_comp.add_linkage_configure(u_lnk)
+        linkage_comp.add_linkage_configure(v_lnk)
+
+        self.p.model.add_subsystem('linkage_comp', subsys=linkage_comp)
+
+        self.p.model.connect('phase0:x', 'linkage_comp.phase0:x_final',
+                             src_indices=[-1],
+                             flat_src_indices=True)
+
+        self.p.model.connect('phase1:x', 'linkage_comp.phase1:x_initial',
+                             src_indices=[0],
+                             flat_src_indices=True)
+
+        self.p.model.connect('phase0:u', 'linkage_comp.phase0:u_final',
+                             src_indices=om.slicer[-1, ...])
+
+        self.p.model.connect('phase1:u', 'linkage_comp.phase1:u_initial',
[…]
