+++ dymos/trajectory/trajectory.py
@@ -315,14 +315,13 @@ class Trajectory(om.Group):
     def _setup_linkages(self):
         has_linkage_constraints = False
         for pair, var_dict in self._linkages.items():
-            phase_name1, phase_name2 = pair
-
-            phase2 = self._phases[phase_name2]
 
             for name in pair:
                 if name not in self._phases:
                     raise ValueError(f'Invalid linkage.  Phase \'{name}\' does not exist in '
                                      f'trajectory \'{self.pathname}\'.')
+
+            phase2 = self._phases[pair[1]]
 
             for var_pair, options in var_dict.items():
                 var1, var2 = var_pair
@@ -433,7 +432,7 @@ class Trajectory(om.Group):
 
         phase_a = self._get_subsystem(f'phases.{phase_name_a}')
         phase_b = self._get_subsystem(f'phases.{phase_name_b}')
-        
+
         phases = {'a': self._get_subsystem(f'phases.{phase_name_a}'),
                   'b': self._get_subsystem(f'phases.{phase_name_b}')}
 
@@ -489,9 +488,13 @@ class Trajectory(om.Group):
             else:
                 rhs_source = phases[i].options['transcription']._rhs_source
                 sources[i] = f'{rhs_source}:{vars[i]}'
-                shapes[i], units[i] = get_source_metadata(phases[i]._get_subsystem(rhs_source),
-                                                          sources[i], user_units=units[i],
-                                                          user_shape=_unspecified)
+                try:
+                    shapes[i], units[i] = get_source_metadata(phases[i]._get_subsystem(rhs_source),
+                                                              sources[i], user_units=units[i],
+                                                              user_shape=_unspecified)
+                except ValueError as e:
+                    raise ValueError(f'Invalid linkage:  Unable to find variable \'{vars[i]}\' in '
+                                     f'phase \'{phases[i].pathname}\' or its ODE.')
 
         linkage_options['src_a'] = sources['a']
         linkage_options['src_b'] = sources['b']
@@ -514,9 +517,9 @@ class Trajectory(om.Group):
 
             phase_b = self._get_subsystem(f'phases.{phase_name_b}')
 
-            for var_pair, options in var_dict.items():
+            for var_pair in var_dict.keys():
                 if tuple(var_pair) == ('*', '*'):
-                    expanded = True
+                    options = var_dict[var_pair]
                     self.add_linkage_constraint(phase_name_a, phase_name_b, var_a='time',
                                                 var_b='time', loc_a=options['loc_a'],
                                                 loc_b=options['loc_b'], sign_a=options['sign_a'],
@@ -530,12 +533,13 @@ class Trajectory(om.Group):
                     self._linkages[phase_pair].pop(var_pair)
 
     def _configure_linkages(self):
+        connected_linkage_inputs = []
 
         # First, if the user requested all states and time be continuous ('*', '*'), then
         # expand it out.
         self._expand_star_linkage_configure()
 
-        print('--- Linkage Report [{0}] ---'.format(self.pathname))
+        print(f'--- Linkage Report [{self.pathname}] ---')
 
         indent = '    '
 
@@ -543,8 +547,14 @@ class Trajectory(om.Group):
 
         for phase_pair, var_dict in self._linkages.items():
             phase_name_a, phase_name_b = phase_pair
-            phase_a = self._get_subsystem(f'phases.{phase_name_a}')
+            print(f'{indent}--- {phase_name_a} - {phase_name_b} ---')
+
             phase_b = self._get_subsystem(f'phases.{phase_name_b}')
+
+            # Pull out the maximum variable name length of all variables to make the print nicer.
+            var_len = [(len(var_pair[0]), len(var_pair[1])) for var_pair in var_dict]
+            max_varname_length = max(itertools.chain(*var_len))
+            padding = max_varname_length + 1
 
[…]

[…]
