### 2020-10-09 4d70286
added docs for add_linkage_constraint
