+++ dymos/examples/finite_burn_orbit_raise/test/test_two_burn_orbit_raise_linkages.py
@@ -1,10 +1,9 @@
 import unittest
+import warnings
+
 from openmdao.utils.testing_utils import use_tempdirs
 from openmdao.utils.general_utils import set_pyoptsparse_opt
 _, optimizer = set_pyoptsparse_opt('IPOPT', fallback=True)
-
-import matplotlib
-matplotlib.use('Agg')
 
 
 @unittest.skipIf(optimizer is not 'IPOPT', 'IPOPT not available')
@@ -248,6 +247,256 @@ class TestTwoBurnOrbitRaiseLinkages(unittest.TestCase):
             ax_xy.plot(x_exp[phs], y_exp[phs], 'b-', label='explicit')
 
         plt.show()
+
+    @use_tempdirs
+    def test_two_burn_orbit_raise_link_control_to_param(self):
+        import numpy as np
+
+        import matplotlib.pyplot as plt
+
+        import openmdao.api as om
+        from openmdao.utils.assert_utils import assert_near_equal
+        from openmdao.utils.general_utils import set_pyoptsparse_opt
+
+        import dymos as dm
+        from dymos.examples.finite_burn_orbit_raise.finite_burn_eom import FiniteBurnODE
+
+        traj = dm.Trajectory()
+        p = om.Problem(model=om.Group())
+        p.model.add_subsystem('traj', traj)
+
+        p.driver = om.pyOptSparseDriver()
+        _, optimizer = set_pyoptsparse_opt('IPOPT', fallback=True)
+        p.driver.options['optimizer'] = optimizer
+
+        p.driver.declare_coloring()
+
+        traj.add_parameter('c', opt=False, val=1.5, units='DU/TU',
+                           targets={'burn1': ['c'], 'coast': ['c'], 'burn2': ['c']})
+
+        # First Phase (burn)
+
+        burn1 = dm.Phase(ode_class=FiniteBurnODE,
+                         transcription=dm.GaussLobatto(num_segments=10, order=3, compressed=True))
+
+        burn1 = traj.add_phase('burn1', burn1)
+
+        burn1.set_time_options(fix_initial=True, duration_bounds=(.5, 10), units='TU')
+        burn1.add_state('r', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                        rate_source='r_dot', targets=['r'], units='DU')
+        burn1.add_state('theta', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                        rate_source='theta_dot', targets=['theta'], units='rad')
+        burn1.add_state('vr', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                        rate_source='vr_dot', targets=['vr'], units='DU/TU')
+        burn1.add_state('vt', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                        rate_source='vt_dot', targets=['vt'], units='DU/TU')
+        burn1.add_state('accel', fix_initial=True, fix_final=False,
+                        rate_source='at_dot', targets=['accel'], units='DU/TU**2')
+        burn1.add_state('deltav', fix_initial=True, fix_final=False,
+                        rate_source='deltav_dot', units='DU/TU')
+        burn1.add_control('u1', rate_continuity=True, rate2_continuity=True, units='deg',
+                          scaler=0.01,
+                          rate_continuity_scaler=0.001, rate2_continuity_scaler=0.001,
+                          lower=-30, upper=30, targets=['u1'])
+
+        # Second Phase (Coast)
+        coast = dm.Phase(ode_class=FiniteBurnODE,
[…]
