### 2020-10-09 c37c026
possible MPI issue with linkage introspection, but will need to fix pyoptsparse on travis first.
