+++ .travis.yml
@@ -97,15 +97,17 @@ install:
     echo "Install pyoptsparse and along with IPOPT";
     git clone https://github.com/OpenMDAO/build_pyoptsparse
     cd build_pyoptsparse
-    chmod 755 ./build_pyoptsparse_ipopt.sh; ./build_pyoptsparse_ipopt.sh;
+    chmod 755 ./build_pyoptsparse_ipopt.sh;
 
-    if [ "$SNOPT_LOCATION" ] && [ "${PY:0:1}" = "3" ]; then
+    if [ "$SNOPT_LOCATION" ]; then
       echo "Install SNOPT";
       cd pyoptsparse/pyoptsparse/pySNOPT;
       scp -r "$SNOPT_LOCATION";
       cd ../../../;
       pwd;
     fi
+
+    ./build_pyoptsparse_ipopt.sh v2.1.5 SNOPT=7.7
 
     git clone https://github.com/OpenMDAO/MBI.git;
     cd MBI;
