+++ .travis.yml
@@ -96,24 +96,26 @@ install:
 
     echo "Install pyoptsparse and along with IPOPT";
     git clone https://github.com/OpenMDAO/build_pyoptsparse;
-    chmod 755 ./build_pyoptsparse_ipopt.sh;
 
     if [ "$SNOPT_LOCATION" ]; then
       echo "Install SNOPT";
       cd build_pyoptsparse/pyoptsparse/pyoptsparse/pySNOPT;
       scp -r "$SNOPT_LOCATION";
       cd ../../../../;
-      pwd;
+      echo $PWD;
     fi
 
     cd build_pyoptsparse;
+    chmod 755 ./build_pyoptsparse_ipopt.sh;
     ./build_pyoptsparse_ipopt.sh v2.1.5 SNOPT=7.7
     cd ..
+    echo $PWD
 
     git clone https://github.com/OpenMDAO/MBI.git;
     cd MBI;
     python setup.py build install;
     cd ..;
+    echo $PWD
 
     if [ "$PETSc" ]; then
       echo " >> Installing parallel processing dependencies";
