+++ .travis.yml
@@ -1,46 +1,54 @@
 branches:
   only:
-    - master
-    - develop
+  - master
+  - develop
 
-sudo: false
+filter_secrets: false
 
 os:
-  - linux
-
-env:
-  - PY=3.7 PETSc=3.10.0
-  - PY=3.6 PETSc=3.10.0 UPLOAD_DOCS=1
+- linux
 
 language: generic
 
+env:
+  matrix:
+    # latest versions
+    - PY=3   NUMPY=1    SCIPY=1   PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3
+    # baseline versions
+    - PY=3.8 NUMPY=1.18 SCIPY=1.4 PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3.12 UPLOAD_DOCS=1
+    # older versions
+    - PY=3.7 NUMPY=1.17 SCIPY=1.3 PYOPTSPARSE=v2.1.5 SNOPT=7.7
+    - PY=3.6 NUMPY=1.16 SCIPY=1.2 PYOPTSPARSE=v1.2   SNOPT=7.2 PETSc=3.10.2
+
+git:
+  depth: 99999
+
 addons:
   apt:
+    update: true
     sources:
     - ubuntu-toolchain-r-test
     packages:
     - gfortran
-    - libblas-dev
+    - gcc
+    - g++
     - liblapack-dev
-    - libopenmpi-dev
-    - openmpi-bin
     ssh_known_hosts:
     - web543.webfaction.com
 
-cache:
-  apt: true
-  timeout: 300
-  directories:
-    - $HOME/.cache/pip
-    - $HOME/miniconda
-
 before_install:
-# Check for existence of files to determine if cache exists
+# When building master (e.g. after a merge), we will create a fresh env as a thorough test.
+# For other builds (e.g. pull requests), try to use cached env for faster testing.
+# Need to check for existence of files to determine if cache exists.
 # If the dir doesn't exist, but is slated to be cached later,
 # Travis unhelpfully creates it, which then causes "dir already exists"
 # errors when you go to actually install the thing, so we must non-intuitively
 # delete the file before re-creating it later.
-- if [ -f $HOME/miniconda/envs/PY$PY ]; then
+- if  [ "$TRAVIS_REPO_SLUG" = "OpenMDAO/OpenMDAO" ] && [ "$TRAVIS_BRANCH" = "master" ] && [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
+    echo "building master";
+    MASTER_BUILD=1;
+    rm -rf $HOME/miniconda;
+  elif [ -d $HOME/miniconda/envs/PY$PY ]; then
     echo "cached miniconda environment found";
     CACHED_ENV=1;
   else
@@ -48,14 +56,10 @@ before_install:
     rm -rf $HOME/miniconda;
   fi
 
-- if  [ "$TRAVIS_REPO_SLUG" = "OpenMDAO/dymos" ] && [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-    MASTER_BUILD=1;
-  fi
-
 install:
 # get key decrypted, placed, chmodded, and added for passwordless access to WebFaction
-- if [ "$encrypted_4548f0b84c7d_key" ]; then
-    openssl aes-256-cbc -K $encrypted_4548f0b84c7d_key -iv $encrypted_4548f0b84c7d_iv -in travis_deploy_rsa.enc -out /tmp/travis_deploy_rsa -d;
+- if [ "$encrypted_74d70a284b7d_key" ]; then
[…]
