+++ .travis.yml
@@ -1,53 +1,47 @@
 branches:
   only:
-  - master
-  - develop
+    - master
+    - develop
 
-filter_secrets: false
+sudo: false
 
 os:
-- linux
+  - linux
+
+env:
+  - PY=3.8 NUMPY=1.18.1 SCIPY=1.4.1 PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3.12 UPLOAD_DOCS=1
+  - PY=3.7 NUMPY=1.18.1 SCIPY=1.4.1 PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3.12
+  - PY=3.6 NUMPY=1.18.1 SCIPY=1.4.1 PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3.12
 
 language: generic
 
-env:
-  matrix:
-    # latest versions
-    - PY=3   NUMPY=1    SCIPY=1   PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3
-    # baseline versions
-    - PY=3.8 NUMPY=1.18 SCIPY=1.4 PYOPTSPARSE=v2.1.5 SNOPT=7.7 PETSc=3.12 UPLOAD_DOCS=1
-    # older versions
-    - PY=3.6 NUMPY=1.16 SCIPY=1.2 PYOPTSPARSE=v1.2   SNOPT=7.2 PETSc=3.10.2
-
-git:
-  depth: 99999
-
 addons:
   apt:
-    update: true
     sources:
     - ubuntu-toolchain-r-test
     packages:
     - gfortran
-    - gcc
-    - g++
+    - libblas-dev
     - liblapack-dev
+    - libopenmpi-dev
+    - openmpi-bin
     ssh_known_hosts:
     - web543.webfaction.com
 
+cache:
+  apt: true
+  timeout: 300
+  directories:
+    - $HOME/.cache/pip
+    - $HOME/miniconda
+
 before_install:
-# When building master (e.g. after a merge), we will create a fresh env as a thorough test.
-# For other builds (e.g. pull requests), try to use cached env for faster testing.
-# Need to check for existence of files to determine if cache exists.
+# Check for existence of files to determine if cache exists
 # If the dir doesn't exist, but is slated to be cached later,
 # Travis unhelpfully creates it, which then causes "dir already exists"
 # errors when you go to actually install the thing, so we must non-intuitively
 # delete the file before re-creating it later.
-- if  [ "$TRAVIS_REPO_SLUG" = "OpenMDAO/OpenMDAO" ] && [ "$TRAVIS_BRANCH" = "master" ] && [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-    echo "building master";
-    MASTER_BUILD=1;
-    rm -rf $HOME/miniconda;
-  elif [ -d $HOME/miniconda/envs/PY$PY ]; then
+- if [ -f $HOME/miniconda/envs/PY$PY ]; then
     echo "cached miniconda environment found";
     CACHED_ENV=1;
   else
@@ -55,10 +49,14 @@ before_install:
     rm -rf $HOME/miniconda;
   fi
 
+- if  [ "$TRAVIS_REPO_SLUG" = "OpenMDAO/dymos" ] && [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
+    MASTER_BUILD=1;
+  fi
+
 install:
 # get key decrypted, placed, chmodded, and added for passwordless access to WebFaction
-- if [ "$encrypted_74d70a284b7d_key" ]; then
[…]
