### 2020-10-13 #2648 @dpsanders
👋 @robfalck: Thanks for the edits to the paper -- it's much clearer (and more inviting) now.
I'll proceed to look for reviewers.

Personally I like seeing the code a simple example in the paper and I don't see a particular reason for needing to keep the paper shorter but that is part of an ongoing editorial discussion.
