+++ .travis.yml
@@ -109,8 +109,8 @@ install:
 
     echo "Build pyoptsparse"
     cd build_pyoptsparse;
-    chmod 755 ./build_pyoptsparse_ipopt.sh;
-    ./build_pyoptsparse_ipopt.sh v2.1.5 SNOPT=7.7
+    chmod 755 ./build_pyoptsparse.sh;
+    ./build_pyoptsparse.sh v2.1.5 SNOPT=7.7
     cd ..
     echo $PWD
 
