+++ .travis.yml
@@ -145,7 +145,7 @@ script:
 - export LD_LIBRARY_PATH=/home/travis/ipopt/lib
 - dymos_path=$(python -c "import dymos as dm; import inspect; import pathlib; p=pathlib.PurePath(inspect.getfile(dm)).parent; print(p)" | tail -1);
 - export OMPI_MCA_rmaps_base_oversubscribe=1
-- testflo -n 1 $dymos_path --pre_announce --coverage --coverpkg dymos;
+- testflo $dymos_path --pre_announce --show_skipped --coverage --coverpkg dymos;
 - testflo -b -n 1 benchmark --pre_announce
 # change to mkdocs path and build the documentation
 - cd mkdocs;
