### 2020-10-14 97a08d3
remove pre-announce from travis
