### 2020-10-14 d4e7d11
added ODE linkage example. both initial and final values of all linkage input variales are now sent to the linkage comp to remove duplicate input errors.
