+++ dymos/trajectory/trajectory.py
@@ -1,5 +1,6 @@
 from collections import OrderedDict
 from collections.abc import Sequence
+from copy import deepcopy
 import itertools
 import warnings
 try:
@@ -17,10 +18,8 @@ from ..utils.constants import INF_BOUND
 from .options import LinkageOptionsDictionary
 from .phase_linkage_comp import PhaseLinkageComp
 from ..phase.options import TrajParameterOptionsDictionary
-from ..utils.misc import get_rate_units, get_source_metadata
-
-
-_unspecified = object()
+from ..utils.misc import get_rate_units, get_source_metadata, _unspecified
+from ..utils.indexing import get_src_indices_by_row
 
 
 class Trajectory(om.Group):
@@ -429,6 +428,8 @@ class Trajectory(om.Group):
         phase_name_b = linkage_options['phase_b']
         var_a = linkage_options['var_a']
         var_b = linkage_options['var_b']
+        loc_a = linkage_options['loc_a']
+        loc_b = linkage_options['loc_b']
 
         info_str = f'Error in linking {var_a} from {phase_name_a} to {var_b} in {phase_name_b}'
 
@@ -444,6 +445,9 @@ class Trajectory(om.Group):
         vars = {'a': var_a, 'b': var_b}
         units = {'a': _unspecified, 'b': _unspecified}
         shapes = {'a': _unspecified, 'b': _unspecified}
+        src_idxs = {'a': None, 'b': None}
+        flat_src_idxs = {'a': False, 'b': False}
+        locs = {'a': loc_a, 'b': loc_b}
 
         for i in ('a', 'b'):
             if classes[i] == 'time':
@@ -487,17 +491,26 @@ class Trajectory(om.Group):
                 shapes[i] = phases[i].parameter_options[vars[i]]['shape']
             else:
                 rhs_source = phases[i].options['transcription']._rhs_source
-                sources[i] = f'{rhs_source}:{vars[i]}'
+                sources[i] = f'{rhs_source}.{vars[i]}'
                 try:
                     shapes[i], units[i] = get_source_metadata(phases[i]._get_subsystem(rhs_source),
-                                                              sources[i], user_units=units[i],
+                                                              vars[i], user_units=units[i],
                                                               user_shape=_unspecified)
                 except ValueError as e:
                     raise ValueError(f'{info_str}: Unable to find variable \'{vars[i]}\' in '
                                      f'phase \'{phases[i].pathname}\' or its ODE.')
 
+                num_ode_nodes = phases[i]._get_subsystem(rhs_source).options['num_nodes']
+                src_idxs[i] = get_src_indices_by_row([0], shapes[i]) if locs[i] == 'initial' else \
+                    get_src_indices_by_row([num_ode_nodes-1], shapes[i])
+                flat_src_idxs[i] = True
+
         linkage_options._src_a = sources['a']
         linkage_options._src_b = sources['b']
+        linkage_options._src_idxs_a = src_idxs['a']
+        linkage_options._src_idxs_b = src_idxs['b']
+        linkage_options._flat_src_idxs_a = flat_src_idxs['a']
+        linkage_options._flat_src_idxs_b = flat_src_idxs['b']
         linkage_options['shape'] = shapes['b']
 
         if linkage_options['units'] is _unspecified:
@@ -519,7 +532,8 @@ class Trajectory(om.Group):
             a phase boundary.
 
         """
-        for phase_pair, var_dict in self._linkages.items():
+        linkages_copy = deepcopy(self._linkages)
+        for phase_pair, var_dict in linkages_copy.items():
             phase_name_a, phase_name_b = phase_pair
 
             phase_b = self._get_subsystem(f'phases.{phase_name_b}')
[…]
