+++ dymos/trajectory/options.py
@@ -1,5 +1,5 @@
 from numbers import Number
-from collections import Iterable
+from collections.abc import Iterable
 
 import openmdao.api as om
 from ..utils.misc import _unspecified
@@ -82,6 +82,8 @@ class LinkageOptionsDictionary(om.OptionsDictionary):
                      desc='optional alternative constraint name to override name conflicts.')
 
         self._src_a = None
+        self._src_idxs_a = None
+        self._src_idxs_b = None
         self._src_b = None
         self._input_a = None
         self._input_b = None
