+++ dymos/examples/cannonball/test/test_two_phase_cannonball_ode_output_linkage.py
@@ -1,0 +1,231 @@
+import unittest
+
+import matplotlib.pyplot as plt
+plt.switch_backend('Agg')
+
+
+class TestTwoPhaseCannonballODEOutputLinkage(unittest.TestCase):
+
+    def test_two_phase_cannonball_ode_output_linkage(self):
+        import openmdao.api as om
+        from openmdao.utils.assert_utils import assert_near_equal
+
+        import dymos as dm
+        from dymos.examples.cannonball.size_comp import CannonballSizeComp
+        from dymos.examples.cannonball.cannonball_phase import CannonballPhase
+
+        p = om.Problem(model=om.Group())
+
+        p.driver = om.pyOptSparseDriver()
+        p.driver.options['optimizer'] = 'SLSQP'
+        p.driver.declare_coloring()
+
+        external_params = p.model.add_subsystem('external_params', om.IndepVarComp())
+
+        external_params.add_output('radius', val=0.10, units='m')
+        external_params.add_output('dens', val=7.87, units='g/cm**3')
+
+        external_params.add_design_var('radius', lower=0.01, upper=0.10, ref0=0.01, ref=0.10)
+
+        p.model.add_subsystem('size_comp', CannonballSizeComp())
+
+        traj = p.model.add_subsystem('traj', dm.Trajectory())
+
+        transcription = dm.Radau(num_segments=5, order=3, compressed=True)
+        ascent = CannonballPhase(transcription=transcription)
+
+        ascent = traj.add_phase('ascent', ascent)
+
+        # All initial states except flight path angle are fixed
+        # Final flight path angle is fixed (we will set it to zero so that the phase ends at apogee)
+        ascent.set_time_options(fix_initial=True, duration_bounds=(1, 100), duration_ref=100, units='s')
+        ascent.set_state_options('r', fix_initial=True, fix_final=False)
+        ascent.set_state_options('h', fix_initial=True, fix_final=False)
+        ascent.set_state_options('gam', fix_initial=False, fix_final=True)
+        ascent.set_state_options('v', fix_initial=False, fix_final=False)
+
+        ascent.add_parameter('S', targets=['aero.S'], units='m**2')
+        ascent.add_parameter('mass', targets=['eom.m', 'kinetic_energy.m'], units='kg')
+
+        # Limit the muzzle energy
+        ascent.add_boundary_constraint('kinetic_energy.ke', loc='initial', units='J',
+                                       upper=400000, lower=0, ref=100000, shape=(1,))
+
+        # Second Phase (descent)
+        transcription = dm.GaussLobatto(num_segments=5, order=3, compressed=True)
+        descent = CannonballPhase(transcription=transcription)
+
+        traj.add_phase('descent', descent)
+
+        # All initial states and time are free (they will be linked to the final states of ascent.
+        # Final altitude is fixed (we will set it to zero so that the phase ends at ground impact)
+        descent.set_time_options(initial_bounds=(.5, 100), duration_bounds=(.5, 100),
+                                 duration_ref=100, units='s')
+        descent.add_state('r', )
+        descent.add_state('h', fix_initial=False, fix_final=True)
+        descent.add_state('gam', fix_initial=False, fix_final=False)
+        descent.add_state('v', fix_initial=False, fix_final=False)
+
+        descent.add_parameter('S', targets=['aero.S'], units='m**2')
+        descent.add_parameter('mass', targets=['eom.m', 'kinetic_energy.m'], units='kg')
+
+        descent.add_objective('r', loc='final', scaler=-1.0)
+
[…]

[…]
