### 2020-10-15 e6c253b
Some fixes for a couple of issues Rob found.
