### 2020-10-15 3bdb079
Some fixes for a couple of issues Rob found.
