### 2020-10-15 a5a573a
attempt to use get_src_indices_by_row instead of a slicer to work around openmdao issues
