+++ dymos/trajectory/trajectory.py
@@ -266,7 +266,6 @@ class Trajectory(om.Group):
         """
         if self.parameter_options:
             for name, options in self.parameter_options.items():
-                src_name = f'parameters:{name}'
 
                 if options['opt']:
                     lb = -INF_BOUND if options['lower'] is None else options['lower']
@@ -476,25 +475,30 @@ class Trajectory(om.Group):
         shapes = {'a': _unspecified, 'b': _unspecified}
         src_idxs = {'a': None, 'b': None}
         flat_src_idxs = {'a': False, 'b': False}
-        locs = {'a': loc_a, 'b': loc_b}
 
         for i in ('a', 'b'):
+            num_nodes = phases[i].options['transcription'].grid_data.num_nodes
+
             if classes[i] == 'time':
                 sources[i] = 'timeseries.time'
                 shapes[i] = (1,)
                 units[i] = phases[i].time_options['units']
+                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] == 'time_phase':
                 sources[i] = 'timeseries.time_phase'
                 units[i] = phases[i].time_options['units']
                 shapes[i] = (1,)
+                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] == 'state':
                 sources[i] = f'timeseries.states:{vars[i]}'
                 units[i] = phases[i].state_options[vars[i]]['units']
                 shapes[i] = phases[i].state_options[vars[i]]['shape']
+                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] in {'indep_control', 'input_control'}:
                 sources[i] = f'timeseries.controls:{vars[i]}'
                 units[i] = phases[i].control_options[vars[i]]['units']
                 shapes[i] = phases[i].control_options[vars[i]]['shape']
+                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] in {'control_rate', 'control_rate2'}:
                 sources[i] = f'timeseries.control_rates:{vars[i]}'
                 control_name = vars[i][:-5] if classes[i] == 'control_rate' else vars[i][:-6]
@@ -502,10 +506,12 @@ class Trajectory(om.Group):
                 deriv = 1 if classes[i] == 'control_rate' else 2
                 units[i] = get_rate_units(units[i], phases[i].time_options['units'], deriv=deriv)
                 shapes[i] = phases[i].control_options[control_name]['shape']
+                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] == 'polynomial_control':
                 sources[i] = f'timeseries.polynomial_controls:{vars[i]}'
                 units[i] = phases[i].polynomial_control_options[vars[i]]['units']
                 shapes[i] = phases[i].polynomial_control_options[vars[i]]['shape']
+                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] in {'polynomial_control_rate', 'polynomial_control_rate2'}:
                 sources[i] = f'timeseries.polynomial_control_rates:{vars[i]}'
                 control_name = vars[i][:-5] if classes[i] == 'polynomial_control_rate' else vars[i][:-6]
@@ -514,12 +520,15 @@ class Trajectory(om.Group):
                 deriv = 1 if classes[i] == 'control_rate' else 2
                 units[i] = get_rate_units(control_units, time_units, deriv=deriv)
                 shapes[i] = phases[i].polynomial_control_options[control_name]['shape']
[…]
