### 2020-10-16 f87fdab
reverted linkages to use slicer for src_indices pending OpenMDAO update.
