+++ dymos/trajectory/trajectory.py
@@ -473,8 +473,6 @@ class Trajectory(om.Group):
         vars = {'a': var_a, 'b': var_b}
         units = {'a': _unspecified, 'b': _unspecified}
         shapes = {'a': _unspecified, 'b': _unspecified}
-        src_idxs = {'a': None, 'b': None}
-        flat_src_idxs = {'a': False, 'b': False}
 
         for i in ('a', 'b'):
             num_nodes = phases[i].options['transcription'].grid_data.num_nodes
@@ -483,22 +481,18 @@ class Trajectory(om.Group):
                 sources[i] = 'timeseries.time'
                 shapes[i] = (1,)
                 units[i] = phases[i].time_options['units']
-                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] == 'time_phase':
                 sources[i] = 'timeseries.time_phase'
                 units[i] = phases[i].time_options['units']
                 shapes[i] = (1,)
-                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] == 'state':
                 sources[i] = f'timeseries.states:{vars[i]}'
                 units[i] = phases[i].state_options[vars[i]]['units']
                 shapes[i] = phases[i].state_options[vars[i]]['shape']
-                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] in {'indep_control', 'input_control'}:
                 sources[i] = f'timeseries.controls:{vars[i]}'
                 units[i] = phases[i].control_options[vars[i]]['units']
                 shapes[i] = phases[i].control_options[vars[i]]['shape']
-                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] in {'control_rate', 'control_rate2'}:
                 sources[i] = f'timeseries.control_rates:{vars[i]}'
                 control_name = vars[i][:-5] if classes[i] == 'control_rate' else vars[i][:-6]
@@ -506,12 +500,10 @@ class Trajectory(om.Group):
                 deriv = 1 if classes[i] == 'control_rate' else 2
                 units[i] = get_rate_units(units[i], phases[i].time_options['units'], deriv=deriv)
                 shapes[i] = phases[i].control_options[control_name]['shape']
-                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] == 'polynomial_control':
                 sources[i] = f'timeseries.polynomial_controls:{vars[i]}'
                 units[i] = phases[i].polynomial_control_options[vars[i]]['units']
                 shapes[i] = phases[i].polynomial_control_options[vars[i]]['shape']
-                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] in {'polynomial_control_rate', 'polynomial_control_rate2'}:
                 sources[i] = f'timeseries.polynomial_control_rates:{vars[i]}'
                 control_name = vars[i][:-5] if classes[i] == 'polynomial_control_rate' else vars[i][:-6]
@@ -520,12 +512,10 @@ class Trajectory(om.Group):
                 deriv = 1 if classes[i] == 'control_rate' else 2
                 units[i] = get_rate_units(control_units, time_units, deriv=deriv)
                 shapes[i] = phases[i].polynomial_control_options[control_name]['shape']
-                src_idxs[i] = get_src_indices_by_row([0, num_nodes-1], shape=shapes[i])
             elif classes[i] == 'parameter':
                 sources[i] = f'timeseries.parameters:{vars[i]}'
[…]
