### 2020-10-16 ee25676
cleaning up a few sensitive tests.
