### 2020-10-16 c3a20cf
pep8 and a test concurrency issue
