### 2020-10-16 83e0d40
Refactored phase linkages to provide more capability. Added add_linkage_constraint method. (#422)
* added a new add_linkage_constraint method to Trajectory to generalize inter-phase constriants.

* add_linkage_constraint is now working as the backend of link_phases.
Warnings occur because initial and final values are often taken from a single source into separate inputs of the linkage constraint.
This should be changed by always passing in the initial and final values and indexing accordingly.
docs are still to be updated.

* added some argument checking to add_linkage_constraint

* added docs
[…]
