+++ dymos/trajectory/test/test_trajectory.py
@@ -155,8 +155,8 @@ class TestTrajectory(unittest.TestCase):
     def test_linked_phases(self):
         burn1_accel = self.p.get_val('burn1.states:accel')
         burn2_accel = self.p.get_val('burn2.states:accel')
-        accel_link_error = self.p.get_val('linkages.burn1|burn2_accel')
-        assert_near_equal(accel_link_error, burn2_accel[0]-burn1_accel[-1])
+        accel_link_error = self.p.get_val('linkages.burn1:accel_final|burn2:accel_initial')
+        assert_near_equal(accel_link_error, burn1_accel[-1]-burn2_accel[0])
 
 
 class TestInvalidLinkages(unittest.TestCase):
@@ -252,9 +252,9 @@ class TestInvalidLinkages(unittest.TestCase):
         with self.assertRaises(ValueError) as e:
             p.setup(check=True)
 
-        self.assertEqual(str(e.exception), 'Cannot find linkage variable \'bar\' in '
-                                           'phase \'burn1\'.  Only states, time, controls, '
-                                           'or parameters may be linked via link_phases.')
+        self.assertEqual(str(e.exception), 'Error in linking bar from burn1 to bar in burn2: '
+                                           'Unable to find variable \'bar\' in phase '
+                                           '\'phases.burn1\' or its ODE.')
 
     def test_invalid_linkage_phase(self):
         p = om.Problem(model=om.Group())
@@ -350,6 +350,98 @@ class TestInvalidLinkages(unittest.TestCase):
 
         self.assertEqual(str(e.exception), 'Invalid linkage.  Phase \'foo\' does not exist in '
                                            'trajectory \'traj\'.')
+
+    def test_invalid_linkage_args(self):
+        p = om.Problem(model=om.Group())
+
+        traj = dm.Trajectory()
+        p.model.add_subsystem('traj', subsys=traj)
+
+        # Since we're only testing features like get_values that don't rely on a converged
+        # solution, no driver is attached.  We'll just invoke run_model.
+
+        # First Phase (burn)
+        burn1 = dm.Phase(ode_class=FiniteBurnODE, transcription=dm.GaussLobatto(num_segments=4, order=3))
+
+        traj.add_phase('burn1', burn1)
+
+        burn1.set_time_options(fix_initial=True, duration_bounds=(.5, 10), units='TU')
+        burn1.add_state('r', fix_initial=True, fix_final=False,
+                        rate_source='r_dot', targets=['r'], units='DU')
+        burn1.add_state('theta', fix_initial=True, fix_final=False,
+                        rate_source='theta_dot', targets=['theta'], units='rad')
+        burn1.add_state('vr', fix_initial=True, fix_final=False,
+                        rate_source='vr_dot', targets=['vr'], units='DU/TU')
+        burn1.add_state('vt', fix_initial=True, fix_final=False,
+                        rate_source='vt_dot', targets=['vt'], units='DU/TU')
+        burn1.add_state('accel', fix_initial=True, fix_final=False,
+                        rate_source='at_dot', targets=['accel'], units='DU/TU**2')
+        burn1.add_state('deltav', fix_initial=True, fix_final=False,
+                        rate_source='deltav_dot', units='DU/TU')
+        burn1.add_control('u1',  targets=['u1'], rate_continuity=True, rate2_continuity=True,
+                          units='deg', scaler=0.01, lower=-30, upper=30)
+        burn1.add_parameter('c', opt=False, val=1.5, targets=['c'], units='DU/TU')
+
+        # Second Phase (Coast)
+
[…]
