+++ dymos/transcriptions/common/__init__.py
@@ -5,5 +5,4 @@ from .polynomial_control_group import PolynomialControlGroup
 from .endpoint_conditions_comp import EndpointConditionsComp
 from .path_constraint_comp import PathConstraintComp
 from .timeseries_output_comp import PseudospectralTimeseriesOutputComp
-from .phase_linkage_comp import PhaseLinkageComp
 from .time_comp import TimeComp
