### 2020-10-16 9e78bf0
RungeKutta now raises DeprecationWarning on init
