### 2020-10-16 d962c68
removed endpoint conditions from transcriptions.  Use of two-character location flag is deprecated in favor of "initial" and "final"
