### 2020-10-16 1082be8
removed EndpointConditionsComp, setup_endpoint_conditions, and configure_endpoint_conditions
