+++ dymos/transcriptions/runge_kutta/runge_kutta.py
@@ -495,106 +495,6 @@ class RungeKutta(TranscriptionBase):
                               'continuity_comp.control_rates:{}_rate2'.format(name),
                               src_indices=src_idxs, flat_src_indices=True)
 
-    def setup_endpoint_conditions(self, phase):
-        jump_comp = phase.add_subsystem('indep_jumps', subsys=om.IndepVarComp(),
-                                        promotes_outputs=['*'])
-
-        jump_comp.add_output('initial_jump:time', val=0.0, units=phase.time_options['units'],
-                             desc='discontinuity in time at the start of the phase')
-
-        jump_comp.add_output('final_jump:time', val=0.0, units=phase.time_options['units'],
-                             desc='discontinuity in time at the end of the phase')
-
-        ic_comp = EndpointConditionsComp(loc='initial',
-                                         time_options=phase.time_options,
-                                         state_options=phase.state_options,
-                                         control_options=phase.control_options)
-
-        phase.add_subsystem(name='initial_conditions', subsys=ic_comp, promotes_outputs=['*'])
-
-        fc_comp = EndpointConditionsComp(loc='final',
-                                         time_options=phase.time_options,
-                                         state_options=phase.state_options,
-                                         control_options=phase.control_options)
-
-        phase.add_subsystem(name='final_conditions', subsys=fc_comp, promotes_outputs=['*'])
-
-    def configure_endpoint_conditions(self, phase):
-        phase.initial_conditions.configure_io()
-        phase.final_conditions.configure_io()
-
-        jump_comp = phase.indep_jumps
-        for state_name, options in phase.state_options.items():
-
-            jump_comp.add_output('initial_jump:{0}'.format(state_name),
-                                 val=np.zeros(options['shape']),
-                                 units=options['units'],
-                                 desc='discontinuity in {0} at the '
-                                      'start of the phase'.format(state_name))
-
-            jump_comp.add_output('final_jump:{0}'.format(state_name),
-                                 val=np.zeros(options['shape']),
-                                 units=options['units'],
-                                 desc='discontinuity in {0} at the '
-                                      'end of the phase'.format(state_name))
-
-        for control_name, options in phase.control_options.items():
-            size = np.prod(options['shape'])
-            ar = np.arange(size)
-
-            jump_comp.add_output('initial_jump:{0}'.format(control_name),
-                                 val=np.zeros(options['shape']),
-                                 units=options['units'],
-                                 desc='discontinuity in {0} at the '
-                                      'start of the phase'.format(control_name))
-
-            jump_comp.add_output('final_jump:{0}'.format(control_name),
-                                 val=np.zeros(options['shape']),
-                                 units=options['units'],
-                                 desc='discontinuity in {0} at the '
-                                      'end of the phase'.format(control_name))
-
-        phase.connect('time', 'initial_conditions.initial_value:time')
-        phase.connect('time', 'final_conditions.final_value:time')
-        phase.connect('initial_jump:time', 'initial_conditions.initial_jump:time')
-        phase.connect('final_jump:time', 'final_conditions.final_jump:time')
-
-        for state_name, options in phase.state_options.items():
-            size = np.prod(options['shape'])
-            ar = np.arange(size)
-
-            phase.connect('states:{0}'.format(state_name),
-                          'initial_conditions.initial_value:{0}'.format(state_name))
-
-            phase.connect('states:{0}'.format(state_name),
-                          'final_conditions.final_value:{0}'.format(state_name))
-
-            phase.connect('initial_jump:{0}'.format(state_name),
-                          'initial_conditions.initial_jump:{0}'.format(state_name),
-                          src_indices=ar, flat_src_indices=True)
-
[…]
