### 2020-10-17 ec1d01a
removed references to endpoint conditions in vanderpol tests
