### 2020-10-17 a031575
doc updates for the release notes
