### 2020-10-17 90a9223
test for expected deprecation warning if two-character location flags are used
