### 2020-10-17 b578ae5
RungeKutta now raises DeprecationWarning on init (#426)
