### 2020-10-17 90730ab
Remove endpoint conditions (#427)
* removed endpoint conditions from transcriptions.  Use of two-character location flag is deprecated in favor of "initial" and "final"

* removed EndpointConditionsComp, setup_endpoint_conditions, and configure_endpoint_conditions

* removed references to endpoint conditions in vanderpol tests

* doc updates for the release notes

* coverage improvements

* test for expected deprecation warning if two-character location flags are used
