### 2020-10-20 127fdd9
start of balanced field length example
