### 2020-10-21 3818e01
fixed issues where some parameters were promoted twice (once by *, once explicitly).  pending updates to OpenMDAO caught these.
