### 2020-10-21 4299680
added release notes for 0.16.0
