### 2020-10-22 4df1204
added release notes for version 0.16.0.  Updated the Issue Template.
