### 2020-10-26 #2648 @dpsanders
👋 @bjack205 and @thowell: Would either of you be able to review this submission for JOSS?
