### 2020-10-26 #2648 @dpsanders
👋 @moorepants and @goerz: Would you be able to review this submission for JOSS?
