### 2020-10-27 083b33c
Fixed a bug in SolveIVP where super was not calling the same method in its parent.
