### 2020-10-27 38d7e36
updated JOSS paper citation for pyoptsparse (#432)
