### 2020-10-29 #2648 @dpsanders
👋 @Jgoldfar: Would you be able to review this submission for JOSS?
