### 2020-10-29 188b49b
run_problem, and not modify_problem, is now responsible for adding the recorder to the solution problem.
