### 2020-10-30 8a641af
run_problem will now add a problem recorder and save by default (this fails under MPI, use problem.run_driver() in that case. changed lots of test that saved to the same file to use temp directories when running. the command line interface still adds a problem recorder, but run_problem will check for its presence before adding its own.
