+++ dymos/examples/ssto/doc/test_doc_ssto_polynomial_control.py
@@ -5,9 +5,11 @@ matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 plt.style.use('ggplot')
 
+from openmdao.utils.testing_utils import use_tempdirs
 from dymos.utils.doc_utils import save_for_docs
 
 
+@use_tempdirs
 class TestDocSSTOPolynomialControl(unittest.TestCase):
 
     @save_for_docs
