+++ dymos/examples/double_integrator/doc/test_doc_double_integrator.py
@@ -6,9 +6,11 @@ matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 plt.style.use('ggplot')
 
+from openmdao.utils.testing_utils import use_tempdirs
 from dymos.utils.doc_utils import save_for_docs
 
 
+@use_tempdirs
 class TestDoubleIntegratorForDocs(unittest.TestCase):
 
     @classmethod
