### 2020-10-30 c22fa40
load_case test is more rigorous now
