### 2020-10-30 89b601b
switching examples over to use tempdirs
