+++ dymos/examples/vanderpol/test/test_vanderpol.py
@@ -64,7 +64,7 @@ class TestVanderpolExampleMPI(unittest.TestCase):
         """
         p = vanderpol(transcription='gauss-lobatto', num_segments=75, delay=True,
                       use_pyoptsparse=True, optimizer='IPOPT')
-        dm.run_problem(p)  # find optimal control solution to stop oscillation
+        p.run_driver()  # find optimal control solution to stop oscillation
 
         if SHOW_PLOTS:
             vanderpol_dymos_plots(p)
