+++ dymos/transcriptions/runge_kutta/components/runge_kutta_stepsize_comp.py
@@ -2,6 +2,7 @@ from collections.abc import Sequence
 import numpy as np
 
 import openmdao.api as om
+from ....options import options as dymos_options
 
 
 class RungeKuttaStepsizeComp(om.ExplicitComponent):
@@ -9,6 +10,9 @@ class RungeKuttaStepsizeComp(om.ExplicitComponent):
     Given the duration of the phase and the segment relative lengths, compute the duration of
     each segment (the step size) for each segment (step).
     """
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._no_check_partials = not dymos_options['include_check_partials']
 
     def initialize(self):
         self.options.declare('num_segments', types=int,
