### 2020-11-01 6d5a1e9
set option include_check_partials to True on component unit test, and added some behavior to assert_check_partials to make sure there is some check_partials_data being tested
