+++ dymos/examples/finite_burn_orbit_raise/test/test_finite_burn_eom.py
@@ -4,7 +4,8 @@ import numpy as np
 
 import openmdao.api as om
 
-from openmdao.utils.assert_utils import assert_near_equal, assert_check_partials
+from openmdao.utils.assert_utils import assert_near_equal
+from dymos.utils.testing_utils import assert_check_partials
 
 from dymos.examples.finite_burn_orbit_raise.finite_burn_eom import FiniteBurnODE
 
