+++ dymos/transcriptions/pseudospectral/components/test/test_control_endpoint_defect_comp.py
@@ -3,7 +3,8 @@ import unittest
 import numpy as np
 
 import openmdao.api as om
-from openmdao.utils.assert_utils import assert_check_partials, assert_near_equal
+from openmdao.utils.assert_utils import assert_near_equal
+from dymos.utils.testing_utils import assert_check_partials
 
 from dymos.transcriptions.pseudospectral.components import ControlEndpointDefectComp
 from dymos.transcriptions.grid_data import GridData
