+++ dymos/examples/aircraft_steady_flight/propulsion/test/test_propulsion_group.py
@@ -2,7 +2,8 @@ import unittest
 
 import numpy as np
 import openmdao.api as om
-from openmdao.utils.assert_utils import assert_near_equal, assert_check_partials
+from openmdao.utils.assert_utils import assert_near_equal
+from dymos.utils.testing_utils import assert_check_partials
 
 from dymos.examples.aircraft_steady_flight.propulsion.propulsion_group import PropulsionGroup
 
