+++ dymos/examples/brachistochrone/doc/test_doc_brachistochrone_static_gravity.py
@@ -19,7 +19,7 @@ class TestBrachistochroneStaticGravity(unittest.TestCase):
     def test_brachistochrone_partials(self):
         import numpy as np
         import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_check_partials
+        from dymos.utils.testing_utils import assert_check_partials
         from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
 
         num_nodes = 5
