### 2020-11-01 99bbe0c
pep8 and some attention to test ordering
