### 2020-11-01 #2809 @whedon
[REVIEW]: dymos: A Python package for optimal control of multidisciplinary systems
**Submitting author:** @robfalck (<a href="http://orcid.org/0000-0001-9864-4928">Rob Falck</a>)
**Repository:** <a href="https://github.com/OpenMDAO/dymos" target ="_blank">https://github.com/OpenMDAO/dymos</a>
**Version:** 1.0.0
