### 2020-11-02 8d9c80b
fixed issues where some parameters were promoted twice (once by *, once explicitly).  pending updates to OpenMDAO caught these. (#429)
