### 2020-11-05 70fe3a8
run_problem now adds a recorder if not already present (#437)
* switching examples over to use tempdirs

* missed files
