+++ dymos/examples/min_time_climb/doc/test_doc_min_time_climb_fd.py
@@ -9,6 +9,7 @@ import matplotlib.pyplot as plt
 plt.style.use('ggplot')
 
 from dymos.utils.doc_utils import save_for_docs
+from openmdao.utils.testing_utils import use_tempdirs
 from openmdao.utils.coloring import Coloring
 
 
@@ -41,6 +42,7 @@ def _view_coloring(coloring_file, show_sparsity_text=False, show_sparsity=True,
     coloring.summary()
 
 
+@use_tempdirs
 class TestMinTimeClimbForDocs(unittest.TestCase):
 
     @save_for_docs
