### 2020-11-05 134ec5c
added test_cannonball_matrix_state to test that dymos works with matrix-shaped state variables
