### 2020-11-05 b109868
gauss lobatto and radau now work with matrix states, working on simulation
