+++ dymos/transcriptions/pseudospectral/radau_pseudospectral.py
@@ -145,7 +145,7 @@ class Radau(PseudospectralBase):
 
             phase.connect(rate_src,
                           'collocation_constraint.f_computed:{0}'.format(name),
-                          src_indices=src_idxs, flat_src_indices=True)
+                          src_indices=src_idxs)
 
     def configure_path_constraints(self, phase):
         super(Radau, self).configure_path_constraints(phase)
@@ -255,15 +255,14 @@ class Radau(PseudospectralBase):
                                                       desc=f'rate of state {state_name}')
 
                 src_rows = gd.input_maps['state_input_to_disc']
-                src_idxs = get_src_indices_by_row(src_rows, options['shape'])
                 phase.connect(src_name=f'states:{state_name}',
                               tgt_name=f'{timeseries_name}.input_values:states:{state_name}',
-                              src_indices=src_idxs, flat_src_indices=True)
+                              src_indices=om.slicer[src_rows, ...])
 
                 rate_src, src_idxs = self.get_rate_source_path(state_name, 'all', phase)
                 phase.connect(src_name=rate_src,
                               tgt_name=f'{timeseries_name}.input_values:state_rates:{state_name}',
-                              src_indices=src_idxs, flat_src_indices=True)
+                              src_indices=src_idxs)
 
             for control_name, options in phase.control_options.items():
                 control_units = options['units']
@@ -480,7 +479,7 @@ class Radau(PseudospectralBase):
             rate_path = 'rhs_all.{0}'.format(var)
             node_idxs = gd.subset_node_indices[nodes]
 
-        src_idxs = get_src_indices_by_row(node_idxs, shape=shape)
+        src_idxs = om.slicer[node_idxs, ...]
 
         return rate_path, src_idxs
 
