### 2020-11-05 2600d0d
simulation of matrix states now works
