+++ dymos/examples/cannonball/test/test_cannonball_matrix_state.py
@@ -38,13 +38,14 @@ class TestCannonballMatrixState(unittest.TestCase):
         traj.add_phase('phase', phase)
         p.model.add_subsystem('traj', traj)
 
-        phase.set_time_options(fix_initial=True, duration_bounds=(1, 5))
+        phase.set_time_options(fix_initial=True, duration_bounds=(1, 5), units=None)
         phase.add_state('z', rate_source='zdot', fix_initial=True, units=None)
 
         phase.add_boundary_constraint('z', loc='final', lower=0, upper=0, indices=[1])
         phase.add_objective('time', loc='final')
 
         p.driver = om.pyOptSparseDriver()
+        p.driver.declare_coloring(tol=1.0E-12)
 
         p.setup()
 
@@ -56,7 +57,7 @@ class TestCannonballMatrixState(unittest.TestCase):
 
     def test_cannonball_matrix_state_radau(self):
 
-        tx = dm.Radau(num_segments=10, order=3, solve_segments=True)
+        tx = dm.Radau(num_segments=10, order=3, solve_segments=False)
 
         p = self._make_problem(tx)
 
@@ -66,9 +67,48 @@ class TestCannonballMatrixState(unittest.TestCase):
         assert_near_equal(p.get_val('traj.phase.timeseries.states:z')[-1, 0, 1], 0.0, tolerance=1E-5)
         assert_near_equal(p.get_val('traj.phase.timeseries.states:z')[-1, 0, 0], 20.3873598, tolerance=1E-5)
 
-        p.model.traj.simulate()
+        c = om.CaseReader('dymos_simulation.db').get_case('final')
+
+        assert_near_equal(c.get_val('traj.phase.timeseries.time')[-1], 2.03873598, tolerance=1E-5)
+        assert_near_equal(c.get_val('traj.phase.timeseries.states:z')[-1, 0, 1], 0.0, tolerance=1E-5)
+        assert_near_equal(c.get_val('traj.phase.timeseries.states:z')[-1, 0, 0], 20.3873598, tolerance=1E-5)
 
     def test_cannonball_matrix_state_gl(self):
+        tx = dm.GaussLobatto(num_segments=10, order=3, solve_segments=False)
+
+        p = self._make_problem(tx)
+
+        p.run_driver()
+
+        assert_near_equal(p.get_val('traj.phase.timeseries.time')[-1], 2.03873598, tolerance=1E-5)
+        assert_near_equal(p.get_val('traj.phase.timeseries.states:z')[-1, 0, 1], 0.0, tolerance=1E-5)
+        assert_near_equal(p.get_val('traj.phase.timeseries.states:z')[-1, 0, 0], 20.3873598, tolerance=1E-5)
+
+        c = om.CaseReader('dymos_simulation.db').get_case('final')
+
+        assert_near_equal(c.get_val('traj.phase.timeseries.time')[-1], 2.03873598, tolerance=1E-5)
+        assert_near_equal(c.get_val('traj.phase.timeseries.states:z')[-1, 0, 1], 0.0, tolerance=1E-5)
+        assert_near_equal(c.get_val('traj.phase.timeseries.states:z')[-1, 0, 0], 20.3873598, tolerance=1E-5)
+
+    def test_cannonball_matrix_state_radau_solve_segments(self):
+
+        tx = dm.Radau(num_segments=10, order=3, solve_segments=True)
+
+        p = self._make_problem(tx)
+
+        dm.run_problem(p, simulate=True)
+
+        assert_near_equal(p.get_val('traj.phase.timeseries.time')[-1], 2.03873598, tolerance=1E-5)
[…]
