+++ dymos/examples/cannonball/test/test_cannonball_matrix_state.py
@@ -61,8 +61,8 @@ class TestCannonballMatrixState(unittest.TestCase):
         tx = dm.Radau(num_segments=10, order=3, solve_segments=False)
 
         p = self._make_problem(tx)
-
-        p.run_driver()
+        
+        dm.run_problem(p, simulate=True)
 
         assert_near_equal(p.get_val('traj.phase.timeseries.time')[-1], 2.03873598, tolerance=1E-5)
         assert_near_equal(p.get_val('traj.phase.timeseries.states:z')[-1, 0, 1], 0.0, tolerance=1E-5)
@@ -79,7 +79,7 @@ class TestCannonballMatrixState(unittest.TestCase):
 
         p = self._make_problem(tx)
 
-        p.run_driver()
+        dm.run_problem(p, simulate=True)
 
         assert_near_equal(p.get_val('traj.phase.timeseries.time')[-1], 2.03873598, tolerance=1E-5)
         assert_near_equal(p.get_val('traj.phase.timeseries.states:z')[-1, 0, 1], 0.0, tolerance=1E-5)
