### 2020-11-06 fcd0533
Fixed a bug that was causing matrix-shaped states to have connection errors. (#441)
* added test_cannonball_matrix_state to test that dymos works with matrix-shaped state variables

* gauss lobatto and radau now work with matrix states, working on simulation

* simulation of matrix states now works
