### 2020-11-06 83112a0
updated 0.16.0 release notes
