### 2020-11-06 c1f9e3a
added check partials PR to the release notes
