+++ dymos/transcriptions/common/time_comp.py
@@ -1,8 +1,13 @@
 import numpy as np
 import openmdao.api as om
+from ...options import options as dymos_options
 
 
 class TimeComp(om.ExplicitComponent):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._no_check_partials = not dymos_options['include_check_partials']
 
     def initialize(self):
         # Required
