+++ dymos/examples/brachistochrone/doc/test_doc_brachistochrone.py
@@ -25,7 +25,7 @@ class TestBrachistochroneForDocs(unittest.TestCase):
     def test_brachistochrone_partials(self):
         import numpy as np
         import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_check_partials
+        from dymos.utils.testing_utils import assert_check_partials
         from dymos.examples.brachistochrone.doc.brachistochrone_ode import BrachistochroneODE
 
         num_nodes = 5
