+++ dymos/__init__.py
@@ -5,3 +5,4 @@ from .transcriptions import GaussLobatto, Radau, RungeKutta
 from .trajectory.trajectory import Trajectory
 from .run_problem import run_problem
 from .load_case import load_case
+from .options import options
