+++ dymos/transcriptions/pseudospectral/components/control_endpoint_defect_comp.py
@@ -1,6 +1,7 @@
 import numpy as np
 import openmdao.api as om
 from ...grid_data import GridData
+from ....options import options as dymos_options
 
 
 class ControlEndpointDefectComp(om.ExplicitComponent):
@@ -14,6 +15,10 @@ class ControlEndpointDefectComp(om.ExplicitComponent):
     .. math:: {u}_{f} = \left[ L_{LGR} \right] vec{u}_{col}
 
     """
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._no_check_partials = not dymos_options['include_check_partials']
+
     def initialize(self):
 
         self.options.declare(
