+++ dymos/examples/min_time_climb/aero/test/test_cd0_comp.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 
 import openmdao.api as om
-from openmdao.utils.assert_utils import assert_check_partials
+from dymos.utils.testing_utils import assert_check_partials
 
 from dymos.examples.min_time_climb.aero.cd0_comp import CD0Comp
 
