+++ dymos/trajectory/test/test_phase_linkage_comp.py
@@ -4,7 +4,7 @@ import numpy as np
 from numpy.testing import assert_almost_equal
 
 import openmdao.api as om
-from openmdao.utils.assert_utils import assert_check_partials
+from dymos.utils.testing_utils import assert_check_partials
 
 from dymos.trajectory.phase_linkage_comp import PhaseLinkageComp
 from dymos.trajectory.options import LinkageOptionsDictionary
