### 2020-11-06 5887009
Release notes 0.16.0 (#431)
* added release notes for 0.16.0

* added release notes for version 0.16.0.  Updated the Issue Template.

* updated 0.16.0 release notes

* added check partials PR to the release notes
