+++ release_notes.md
@@ -1,4 +1,79 @@
-# Dymos Release Notes
+********************************
+# Release Notes for Dymos 0.16.0
+
+October 23, 2020
+
+This is a long-overdue release of Dymos that features several useful new features.
+By taking advantage of recent changes to the OpenMDAO setup process, we can use introspection to determine a lot of things automatically now.
+For instance, the user will no longer be required to specify the shape or default units of timeseries outputs or constraints (although they can override units if they choose).
+
+Also on the topic of timeseries, dymos now supports glob patterns when adding ODE outputs to the timeseries.
+Adding all 100 outputs of a large ODE to the timeseries output can now be accomplished in a single line:
+
+```
+phase.add_timeseries_output('*')
+```
+
+Introspection is also used for states, time, and controls.
+States will get their default units and shape from the associated state rate variable now.
+Controls, Polynomial Controls, and Parameters will now get their default units and shapes from their targets.
+
+With the advent of automatic IndepVarComps in OpenMDAO, there no longer needs to be a distinction between InputParameters and DesignParameters.
+From here on out, users just add `parameters` to a Trajectory or Phase.
+A user can choose to make them design variables (in which case they perform as DesignParameters), or not.
+Connecting an external variable to the parameter makes it function as a InputParameter.
+
+From here on, we plan on making releases of Dymos roughly once per month.
+
+## Backwards Incompatible API Changes & Deprecations:
+
+- `add_input_parameter` and `add_design_parameter` are **deprecated** and replaced with `add_parameter` [#365](https://github.com/OpenMDAO/dymos/pull/365)
+- The use of two-character location strings ('--' and '++') are replaced by 'initial' and 'final' when linking phases. [#427](https://github.com/OpenMDAO/dymos/pull/427)
+- The endpoint conditions component, used to pull out initial or final values of a variable in a phase, have been removed.  Values should instead by pulled from the timeseries. [#427](https://github.com/OpenMDAO/dymos/pull/427)
+- The results of simulate are now automatically recorded to a database recorder 'dymos_simulation.db' and stored in a case named 'final'.  Use `CaseReader('dymos_simulation.db').get_case('final')` to access the case. instead of the previous behavior `CaseReader('dymos_simulation.db').get_case(-1)` [#437](https://github.com/OpenMDAO/dymos/pull/437)
+- The RungeKutta transcription is **deprecated**. It is functionally equivalent to the GaussLobatto transcription with an order of 3, but since it's signficantly different under the hood it was being a drag on development.
+
+One note on another backwards incompatibility.
+Since we're using some newer OpenMDAO capabilities for the way in which parameters are handled, it is currently not possible to connect an upstream output to a parameter in dymos _when src_indices are used_.
+We plan on fixing this issue with an upcoming OpenMDAO release.
+If you rely on this capability, we recommend waiting for that update before using this version of Dymos, or using an intermediate "pass-through" component to pull the correct index from the original output and then passing _that_ as the value of a dymos parameter.
+
+## Enhancements:
+
+* Dymos components are no longer listed in check_partials output by default.  Use `dymos.options['include_check_partials'] = True` to override this for debugging.  [#438](https://github.com/OpenMDAO/dymos/pull/438)
[…]
