+++ .bumpversion.cfg
@@ -15,9 +15,9 @@ replace = version='{new_version}',
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
 
-[bumpversion:file:dymos/docs/conf.py]
-search = version = u'{current_version}'
-replace = version = u'{new_version}'
+[bumpversion:file:dymos/mkdocs/mkdocs.yml]
+search = version = u'site_name: Dymos {current_version}'
+replace = version = u'site_name: Dymos {new_version}'
 
 [bumpversion:part:release]
 optional_value = rel
