### 2020-11-06 18924a7
bumpversion now works on mkdocs
