+++ .bumpversion.cfg
@@ -15,7 +15,7 @@ replace = version='{new_version}',
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
 
-[bumpversion:file:dymos/mkdocs/mkdocs.yml]
+[bumpversion:file:mkdocs/mkdocs.yml]
 search = version = u'site_name: Dymos {current_version}'
 replace = version = u'site_name: Dymos {new_version}'
 
