### 2020-11-06 abba06c
missed an "include_check_partials" option in the phase linkage test
