+++ .bumpversion.cfg
@@ -16,8 +16,8 @@ search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
 
 [bumpversion:file:mkdocs/mkdocs.yml]
-search = version = u'site_name: Dymos {current_version}'
-replace = version = u'site_name: Dymos {new_version}'
+search = version = u'{current_version}'
+replace = version = u'{new_version}'
 
 [bumpversion:part:release]
 optional_value = rel
