### 2020-11-06 a62b543
Version 0.16.0 (#442)
* updates for new version

* bumpversion now works on mkdocs

* Bump version: 0.16.0-dev → 0.16.0

* missed an "include_check_partials" option in the phase linkage test
