+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.16.0-dev
+current_version = 0.16.0
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+))?
@@ -15,7 +15,7 @@ replace = version='{new_version}',
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
 
-[bumpversion:file:dymos/docs/conf.py]
+[bumpversion:file:mkdocs/mkdocs.yml]
 search = version = u'{current_version}'
 replace = version = u'{new_version}'
 
