### 2020-11-07 a38e1fa
add support for mike on travis and remove version number from docs for now
