+++ .bumpversion.cfg
@@ -15,10 +15,6 @@ replace = version='{new_version}',
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
 
-[bumpversion:file:mkdocs/mkdocs.yml]
-search = version = u'{current_version}'
-replace = version = u'{new_version}'
-
 [bumpversion:part:release]
 optional_value = rel
 values = 
