### 2020-11-09 257f0c5
Initial code to add timeseries plots
