+++ dymos/utils/command_line.py
@@ -63,6 +63,15 @@ def dymos_cmd(argv=None):
     parser.add_argument('-l', '--refine_limit', default=0,
                         help='The number of passes through the grid refinement algorithm'
                              ' to use. (default: 0)')
+
+    # NEW
+    parser.add_argument('-p', '--make_plots', action='store_true',
+                        help='If given, automatically generate plots of all timeseries outputs.')
+    parser.add_argument('-o', '--solution_record_file', default='dymos_solution.db',
+                        help='Set the name of the case recorder file for solution results. (default: dymos_solution.db)')
+    parser.add_argument('-i', '--simulation_record_file', default='dymos_simulation.db',
+                        help='Set the name of the case recorder file for simulation results. (default: dymos_simulation.db)')
+
     args = parser.parse_args(argv)  # sys.argv is used if argv parameter is None
 
     if args.solution == 'dymos_solution.db':  # make sure the loaded db is not being overwritten by the new db
@@ -75,7 +84,10 @@ def dymos_cmd(argv=None):
         'restart': args.solution,
         'simulate': args.simulate,
         'no_solve': args.no_solve,
-        'reset_grid': args.reset_grid
+        'reset_grid': args.reset_grid,
+        'make_plots': args.make_plots,
+        'solution_record_file': args.solution_record_file,
+        'simulation_record_file': args.simulation_record_file,
     }
 
     class DymosHooks:
@@ -108,7 +120,12 @@ def dymos_cmd(argv=None):
 
             refine_iterations = opts.get('refine_iteration_limit')
             run_problem(prob, refine_iteration_limit=refine_iterations,
-                        run_driver=not opts['no_solve'], simulate=opts['simulate'])
+                        run_driver=not opts['no_solve'],
+                        make_plots = opts['make_plots'],
+                        solution_record_file=opts['solution_record_file'],
+                        simulate=opts['simulate'],
+                        simulation_record_file = opts['simulation_record_file']
+                        )
 
     dymos_hooks = DymosHooks()
 
