### 2020-11-09 183f9e5
Fixed location and spacing for legend
