### 2020-11-10 d800e28
Bump version to 0.16.1-dev (#443)
* added support for mike to version the documentation

* Bump version: 0.16.0 → 0.16.1-dev

* add support for mike on travis and remove version number from docs for now
