+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.16.0
+current_version = 0.16.1-dev
 commit = True
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+))?
@@ -14,10 +14,6 @@ replace = version='{new_version}',
 [bumpversion:file:dymos/__init__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
-
-[bumpversion:file:mkdocs/mkdocs.yml]
-search = version = u'{current_version}'
-replace = version = u'{new_version}'
 
 [bumpversion:part:release]
 optional_value = rel
