### 2020-11-10 42e2a4f
Phase and trajectory parameters no longer have their shape prepended by (1,).
Also made an effort to further isolate a lot of check_partials tests where side effects from `dm.options['include_check_partials']` in other tests were causing unintended behavior.
