### 2020-11-10 9ecbd60
All code and tests for basic plotting from run_problem and timeseries_plots work
