+++ dymos/transcriptions/runge_kutta/components/test/test_rk_stepsize_comp.py
@@ -6,6 +6,7 @@ import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 from dymos.utils.testing_utils import assert_check_partials
 
+import dymos as dm
 from dymos.transcriptions.runge_kutta.components.runge_kutta_stepsize_comp import RungeKuttaStepsizeComp
 
 from dymos.utils.misc import CompWrapperConfig
@@ -13,6 +14,12 @@ RungeKuttaStepsizeComp = CompWrapperConfig(RungeKuttaStepsizeComp)
 
 
 class TestRKStepsizeComp(unittest.TestCase):
+
+    def setUp(self):
+        dm.options['include_check_partials'] = True
+
+    def tearDown(self):
+        dm.options['include_check_partials'] = False
 
     def test_rk_stepsize_comp(self):
         p = om.Problem(model=om.Group())
