### 2020-11-10 c4c7b8a
Parameter shapes are now as expected, not prepended by (1,) (#444)
* Phase and trajectory parameters no longer have their shape prepended by (1,).

Also made an effort to further isolate a lot of check_partials tests where side effects from `dm.options['include_check_partials']` in other tests were causing unintended behavior.

* added a fix for #439

* more include_check_partials cleanup
