### 2020-11-12 32179f4
state introspection pulls units and shape from rate_source unless given by user
