+++ dymos/transcriptions/pseudospectral/pseudospectral_base.py
@@ -6,7 +6,7 @@ import openmdao.api as om
 from ..transcription_base import TranscriptionBase
 from ..common import TimeComp, PseudospectralTimeseriesOutputComp
 from .components import StateIndependentsComp, StateInterpComp, CollocationComp
-from ...utils.misc import CoerceDesvar, get_rate_units, get_target_metadata, _unspecified
+from ...utils.misc import CoerceDesvar, get_rate_units, get_target_metadata, get_source_metadata, _unspecified
 from ...utils.constants import INF_BOUND
 from ...utils.indexing import get_src_indices_by_row
 
@@ -63,56 +63,87 @@ class PseudospectralBase(TranscriptionBase):
         phase.add_subsystem('indep_states', indep, promotes_inputs=prom_inputs,
                             promotes_outputs=['*'])
 
+    def _configure_state_introspection(self, state_name, phase):
+        options = phase.state_options[state_name]
+        time_units = phase.time_options['units']
+        user_units = options['units']
+        user_shape = options['shape']
+
+        need_units = user_units is _unspecified
+        need_shape = user_shape in {None, _unspecified}
+
+        if not(need_shape or need_units):
+            return user_shape, user_units
+
+        if options['targets']:
+
+
+        rate_src = options['rate_source']
+        rate_src_type = phase.classify_var(rate_src)
+        rate_src_path, _ = self.get_rate_source_path(state_name, 'col', phase)
+
+        if rate_src_type in ['time', 'time_phase']:
+            rate_src_units = phase.time_options['units']
+            shape = (1,)
+        elif rate_src_type == 'state':
+            rate_src_units = phase.state_options[rate_src]['units']
+            shape = phase.state_options[rate_src]['shape']
+        elif rate_src_type in ['input_control', 'indep_control']:
+            rate_src_units = phase.control_options[rate_src]['units']
+            shape = phase.control_options[rate_src]['shape']
+        elif rate_src_type in ['input_polynomial_control', 'indep_polynomial_control']:
+            rate_src_units = phase.polynomial_control_options[rate_src]['units']
+            shape = phase.polynomial_control_options[rate_src]['shape']
+        elif rate_src_type == 'parameter':
+            rate_src_units = phase.parameter_options[rate_src]['units']
+            shape = phase.parameter_options[rate_src]['shape']
+        elif rate_src_type == 'control_rate':
+            control_name = rate_src[:-5]
+            rate_src_units = phase.control_options[control_name]['units']
+            shape = phase.control_options[control_name]['shape']
+        elif rate_src_type == 'control_rate2':
+            control_name = rate_src[:-6]
+            rate_src_units = phase.control_options[control_name]['units']
+            shape = phase.control_options[control_name]['shape']
+        elif rate_src_type == 'polynomial_control_rate':
+            control_name = rate_src[:-5]
+            rate_src_units = phase.polynomial_control_options[control_name]['units']
+            shape = phase.polynomial_control_options[control_name]['shape']
+        elif rate_src_type == 'polynomial_control_rate2':
+            control_name = rate_src[:-6]
+            rate_src_units = phase.polynomial_control_options[control_name]['units']
+            shape = phase.polynomial_control_options[control_name]['shape']
+        elif rate_src_type == 'ode':
+            ode = phase._get_subsystem(self._rhs_source)
+            shape, rate_src_units = get_source_metadata(ode,
+                                                        src=rate_src,
+                                                        user_units=options['units'],
+                                                        user_shape=options['shape'])
+
+        ret_shape = shape if need_shape else user_shape
+        if ret_shape is None:
+            ret_shape = (1,)
+
+        if need_units:
+            if rate_src_units is None:
+                ret_units = None
[…]
