### 2020-11-12 9d1bfa2
added a new module in utils.introspection for introspection utility functions.  changed state metadata introspection so that user defined shape and units are preferred, then the units/shape of a target if there is a single target in the ODE, and finally shape and units based on the rate source if necessary.
