+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -8,8 +8,9 @@ from ..transcription_base import TranscriptionBase
 from .components import SegmentSimulationComp, SegmentStateMuxComp, \
     SolveIVPControlGroup, SolveIVPPolynomialControlGroup, SolveIVPTimeseriesOutputComp
 from ..common import TimeComp
-from ...utils.misc import get_rate_units, get_targets, get_target_metadata, get_source_metadata, \
+from ...utils.misc import get_rate_units, get_target_metadata, get_source_metadata, \
     _unspecified
+from ...utils.introspection import get_targets
 from ...utils.indexing import get_src_indices_by_row
 
 
