### 2020-11-12 a02b155
allow Numpy array as a boundary constraint
