### 2020-11-12 703339a
added a test to explicitly test connecting the final value of a variablre from one phase to a parameter of another.  requires OpenMDAO 3.4.1 (#445)
