+++ dymos/examples/cannonball/test/test_matrix_states.py
@@ -1,119 +1,0 @@
-import numpy as np
-
-import openmdao.api as om
-
-import dymos as dm
-
-
-class BrachistochroneVectorStatesODE(om.ExplicitComponent):
-
-    def initialize(self):
-        self.options.declare('num_nodes', types=int)
-
-    def setup(self):
-        nn = self.options['num_nodes']
-
-        # Inputs
-        self.add_input('v', val=np.zeros(nn), desc='velocity', units='m/s')
-
-        self.add_input('g', val=9.80665 * np.ones(nn), desc='grav. acceleration', units='m/s/s')
-
-        self.add_input('theta', val=np.zeros(nn), desc='angle of wire', units='rad')
-
-        self.add_output('pos_dot', val=np.zeros((nn, 2, 2)), desc='velocity components', units='m/s')
-
-        self.add_output('vdot', val=np.zeros(nn), desc='acceleration magnitude', units='m/s**2')
-
-        self.add_output('check', val=np.zeros(nn), desc='check solution: v/sin(theta) = constant',
-                        units='m/s')
-
-        # Setup partials
-        arange = np.arange(self.options['num_nodes'])
-
-        self.declare_partials(of='vdot', wrt='g', rows=arange, cols=arange)
-        self.declare_partials(of='vdot', wrt='theta', rows=arange, cols=arange)
-
-        rs = np.arange(2*nn, dtype=int)
-        cs = np.repeat(np.arange(nn, dtype=int), 2)
-
-        self.declare_partials('*', '*', method='cs')
-        self.declare_coloring(wrt='*', method='cs', show_sparsity=True)
-
-    def compute(self, inputs, outputs):
-        theta = inputs['theta']
-        cos_theta = np.cos(theta)
-        sin_theta = np.sin(theta)
-        g = inputs['g']
-        v = inputs['v']
-
-        outputs['vdot'] = g * cos_theta
-        outputs['pos_dot'][:, 0, 0] = v * sin_theta
-        outputs['pos_dot'][:, 0, 1] = -v * cos_theta
-
-        outputs['pos_dot'][:, 1, 0] = v * sin_theta
-        outputs['pos_dot'][:, 1, 1] = -v * cos_theta
-        outputs['check'] = v / sin_theta
-
-
-p = om.Problem(model=om.Group())
-
-p.driver = om.pyOptSparseDriver()
-p.driver.options['optimizer'] = 'SNOPT'
-p.driver.opt_settings['Major iterations limit'] = 100
-p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-6
-p.driver.opt_settings['Major optimality tolerance'] = 1.0E-6
-p.driver.opt_settings['iSumm'] = 6
-
-
-p.driver.declare_coloring()
-
-transcription = dm.Radau(num_segments=1,
-                         order=3,
-                         compressed=False)
-fix_final = True
-
-traj = dm.Trajectory()
-phase = dm.Phase(ode_class=BrachistochroneVectorStatesODE,
-                 transcription=transcription)
-traj.add_phase('phase0', phase)
-
-p.model.add_subsystem('traj0', traj)
-
-phase.set_time_options(fix_initial=True, duration_bounds=(.5, 10), units='s')
-
-# can't fix final position if you're solving the segments
-
-phase.add_state('pos',
-                rate_source='pos_dot', units='m',
-                fix_initial=True, fix_final=True, shape=(2, 2))
-#
-phase.add_state('v',
-                rate_source='vdot', units='m/s',
-                fix_initial=True, fix_final=False)
-#
-phase.add_control('theta',
[…]
