### 2020-11-12 083d06e
Fixes a bug where user-defined shapes of states were colliding with those found during introspection, and other state introspection updates. (#449)
* state introspection pulls units and shape from rate_source unless given by user

* added a new module in utils.introspection for introspection utility functions.  changed state metadata introspection so that user defined shape and units are preferred, then the units/shape of a target if there is a single target in the ODE, and finally shape and units based on the rate source if necessary.
