+++ dymos/transcriptions/pseudospectral/pseudospectral_base.py
@@ -6,7 +6,9 @@ import openmdao.api as om
 from ..transcription_base import TranscriptionBase
 from ..common import TimeComp, PseudospectralTimeseriesOutputComp
 from .components import StateIndependentsComp, StateInterpComp, CollocationComp
-from ...utils.misc import CoerceDesvar, get_rate_units, get_target_metadata, _unspecified
+from ...utils.misc import CoerceDesvar, get_rate_units, _unspecified, \
+    get_target_metadata, get_source_metadata
+from ...utils.introspection import get_targets, get_state_target_metadata
 from ...utils.constants import INF_BOUND
 from ...utils.indexing import get_src_indices_by_row
 
@@ -67,52 +69,12 @@ class PseudospectralBase(TranscriptionBase):
         grid_data = self.grid_data
         num_state_input_nodes = grid_data.subset_num_nodes['state_input']
         indep = phase.indep_states
-        time_units = phase.time_options['units']
-        ode = phase._get_subsystem(self._rhs_source)
 
         # add all the des-vars (either from the IndepVarComp or from the indep-var-like
         # outputs of the collocation comp)
         for name, options in phase.state_options.items():
 
-            rate_src = options['rate_source']
-
-            # Handle states that point to a control.
-            # This feels a little hackish
-            if rate_src in phase.control_options:
-                targets = phase.control_options[rate_src]['targets']
-                shape = phase.control_options[rate_src]['shape']
-                units = phase.control_options[rate_src]['units']
-                if targets is not _unspecified:
-                    rate_src = targets[0]
-                else:
-                    if units is not _unspecified and options['units'] in (_unspecified, None):
-                        options['units'] = f'{units}*{time_units}'
-                    if shape is not _unspecified and options['shape'] in (_unspecified, None):
-                        options['shape'] = shape
-
-            # Handle states that point to another state. States must be declared in the right
-            # order.
-            # This feels a little hackish
-            elif rate_src in phase.state_options and options['shape'] in (_unspecified, None):
-                options['shape'] = phase.state_options[rate_src]['shape']
-
-            full_shape, units = get_target_metadata(ode, name=name,
-                                                    user_targets=rate_src,
-                                                    user_units=options['units'],
-                                                    user_shape=options['shape'])
-
-            if options['units'] is _unspecified:
-                # Units are from the rate source and should be converted.
-                if units is not None:
-                    units = f'{units}*{time_units}'
-            options['units'] = units
-
-            # Determine and store the pre-discretized state shape for use by other components.
-            if len(full_shape) < 2:
-                if options['shape'] in (_unspecified, None):
-                    options['shape'] = (1, )
-            else:
-                options['shape'] = full_shape[1:]
+            self._configure_state_introspection(name, options, phase)
 
             size = np.prod(options['shape'])
             # In certain cases, we put an output on the IVC.
@@ -120,7 +82,7 @@ class PseudospectralBase(TranscriptionBase):
                 if not options['solve_segments'] and not options['connected_initial']:
                     indep.add_output(name='states:{0}'.format(name),
                                      shape=(num_state_input_nodes, size),
-                                     units=units)
+                                     units=options['units'])
 
             if options['opt']:
                 if options['solve_segments']:
