### 2020-11-13 ae656b8
move array checks to transcription_base
