+++ dymos/transcriptions/transcription_base.py
@@ -445,22 +445,22 @@ class TranscriptionBase(object):
                                      'compatible the provided indices. Provide them as a '
                                      'flat array with the same size as indices.'.format(var))
 
-            elif options['lower'] or options['upper'] or options['equals']:
+            elif 'lower' in options or 'upper' in options or 'equals' in options:
                 # Indices not provided, make sure lower/upper/equals have shape of source.
-                if options['lower'] and not np.isscalar(options['lower']) and \
-                        np.asarray(options['lower']).shape != shape:
+                if 'lower' in options and options['lower'] is not None and \
+                        not np.isscalar(options['lower']) and np.asarray(options['lower']).shape != shape:
                     raise ValueError('The lower bounds of boundary constraint on {0} are not '
                                      'compatible with its shape, and no indices were '
                                      'provided.'.format(var))
 
-                if options['upper'] and not np.isscalar(options['upper']) and \
-                        np.asarray(options['upper']).shape != shape:
+                if 'upper' in options and options['upper'] is not None and \
+                        not np.isscalar(options['upper']) and np.asarray(options['upper']).shape != shape:
                     raise ValueError('The upper bounds of boundary constraint on {0} are not '
                                      'compatible with its shape, and no indices were '
                                      'provided.'.format(var))
 
-                if options['equals'] and not np.isscalar(options['equals']) \
-                        and np.asarray(options['equals']).shape != shape:
+                if 'equals' in options and options['equals'] is not None and \
+                        not np.isscalar(options['equals']) and np.asarray(options['equals']).shape != shape:
                     raise ValueError('The equality boundary constraint value on {0} is not '
                                      'compatible with its shape, and no indices were '
                                      'provided.'.format(var))
