### 2020-11-13 b325d8d
release notes for 0.16.1 added.  setup.py now requires openmdao >= 3.4.1.
