### 2020-11-13 fb6e269
Added handling of openmdao verion to phase.py so that version 3.3.0 still works, aside from those features enabled by 3.4.1
