### 2020-11-15 #2809 @goerz
I noticed that two people, @hwangjt and @kmarsteller seem to have [contributed significantly](https://github.com/OpenMDAO/dymos/graphs/contributors) to the software repo, but do not appear as authors on the paper. I don't have a fundamental problem with this, but @robfalck may want to comment on what these two people contributed to the project and why their contribution is not relevant for authorship on the paper, just so that I can mark that point in the review checklist with confidence.

@dpsanders: What exactly are JOSS' guidelines on this, beyond the "Does the full list of paper authors seem appropriate and complete?"
