### 2020-11-15 OpenMDAO/dymos#453 @goerz
Release on PyPI
Is there a reason the package has not been released on PyPI? Usually, I would consider releases as a major factor in the maturity of a package (for the "Substantiality" requirement of https://github.com/openjournals/joss-reviews/issues/2809).

I noticed that you have tagged released on Github, but PyPI would certainly be the expected best practice for any Python package. Unless you have a good reason not to do so, I would ask you to publish on PyPI.
