### 2020-11-15 OpenMDAO/dymos#453 @goerz
Somewhat related: you probably want to revise https://github.com/OpenMDAO/dymos/blob/083d06e641ea18a4731eeae2118e9e563603ebe2/setup.py#L36

"Pre-Alpha" wouldn't be something publishable.
