### 2020-11-15 #2809 @goerz
@robfalck  The [documentation of dymos](https://openmdao.github.io/dymos/index.html#citation) points to https://arc.aiaa.org/doi/abs/10.2514/6.2019-0976, and that paper clearly already introduces the dymos package. That makes me somewhat inclined to reject the submission to JOSS as "not original". Can you comment on why you think the submission to JOSS is justified?

@dpsanders What are your guidelines on this? I would potentially consider a new major version of a software (or a "1.0" in this case) as warranting a second publication.
