### 2020-11-16 OpenMDAO/dymos#453 @robfalck
We've been intending to release dymos on PyPI, and hopefully will get that done this week.  Thank you for finding the oversight in setup.py.
