### 2020-11-16 #2809 @robfalck
> I noticed that two people, @hwangjt and @kmarsteller seem to have [contributed significantly](https://github.com/OpenMDAO/dymos/graphs/contributors) to the software repo, but do not appear as authors on the paper. I don't have a fundamental problem with this, but @robfalck may want to comment on what these two people contributed to the project and why their contribution is not relevant for authorship on the paper, just so that I can mark that point in the review checklist with confidence.

Originally @hwangt helped with the transition from the original "alpha" stage of dymos (we called it pointer back then) into it's current form.  At that point there were a few of us in the room coding together and he happened to be the one pushing the commits.  In addition, he's not been on the project for several years and, and we opted to not list him as a co author.

@kmarstellar is recently retired.  He's role in dymos was in aiding within interfacing to travis and helping getting the documentation working, but not with developing the core code.  We've since moved on from using some of the documentation systems he implemented, and we opted to not include him as a coauthor.
