### 2020-11-16 7a8591c
fix setup.py to recognize dymos is now production/stable status.
