### 2020-11-16 8555944
Release notes 0.16.1 (#451)
* release notes for 0.16.1 added.  setup.py now requires openmdao >= 3.4.1.

* Bump version: 0.16.1-dev → 0.16.1

* Added handling of openmdao verion to phase.py so that version 3.3.0 still works, aside from those features enabled by 3.4.1

* fix setup.py to recognize dymos is now production/stable status.
