### 2020-11-16 6e5a310
update bug fixes in release notes
