+++ release_notes.md
@@ -21,6 +21,8 @@ improved handling of parameters.
 
 * Fixes a bug where user-defined shapes of states were colliding with those found during introspection, and other state introspection updates. [#449](https://github.com/OpenMDAO/dymos/pull/449)
 
+* Fixes a bug that prevented the use of numpy arrays as a boundary constraints [#450](https://github.com/OpenMDAO/dymos/pull/450)
+
 ## Miscellaneous
 
 * Added test to verify functionality of OpenMDAO 3.4.1 that allows the final value of a control (or a partial portion of any output) to be connected to a parameter. [#445](https://github.com/OpenMDAO/dymos/pull/445)
@@ -131,4 +133,3 @@ If you rely on this capability, we recommend waiting for that update before usin
 * [__Enhancement__] ODE options can be specified at the phase level rather than in the ODE. This feature is experimental, but it allows one way of programmatically defining an ODE.
 * [__Enhancement__] Changed the use of OpenMDAO to do `import openmdao.api as om` for consistency with the OpenMDAO documentation.
 * [__Enhancement__] Dymos is now imported in the examples as import `dymos as dm`
-
