### 2020-11-16 4f76684
change optimizer for Travis compatibility
