### 2020-11-16 94e87be
allow Numpy array as a boundary constraint (#450)
* allow Numpy array as a boundary constraint

* move array checks to transcription_base

* add unit test

* rebase and update unit test for compatibility

* update bug fixes in release notes

* change optimizer for Travis compatibility
