### 2020-11-16 d13a6b3
minor fix to the release notes for 0.16.1 (#455)
