### 2020-11-17 5a93be6
fix for apache license string in setup.py classifiers (#458)
