### 2020-11-17 OpenMDAO/dymos#453 @robfalck
Release on PyPI.
Adding a long_description to the setup.py for PyPI is added as an issue.

https://pypi.org/project/dymos/
