### 2020-11-19 dce7f8e
balanced field coming along, but exposed two bugs in dymos that need to be fixed.
