### 2020-11-19 eb5fdf6
corrected issue where polynomial controls were not able to be linked across phases, and some other similar tests. (#462)
