### 2020-11-19 bcb8af4
Many changes to get to the point of being able to plot multiple phases
