+++ dymos/visualization/timeseries_plots.py
@@ -1,63 +1,119 @@
 import os
 import warnings
 
-import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.lines as mlines
+import matplotlib.patches as mpatches
 
 import openmdao.api as om
 
-from matplotlib.figure import Figure
-from matplotlib.backends.backend_agg import FigureCanvasAgg
+def _get_phases_node_in_problem_metadata(metadata,path=""):
+    """
+    Get the units for a timeseries variable and also the time associated with that variable.
 
-import matplotlib.pyplot as plt
+    Parameters
+    ----------
+    cr : CaseRecorder
+        The CaseRecorder that contains the problem metadata
+    varname : str
+        The CaseRecorder that contains the problem metadata
 
+    Returns
+    -------
+    units : tuple of two strings
+        The units for the requested timeseries variable and also the time associated with that variable
+    """
+    for item in metadata:
+        if item['name'] == 'phases':
+            return item, f'{path}'
+        else:
+            if 'children' in item:
+                children = item['children']
+                name = item['name']
+                if path:
+                    new_path = f'{path}:{name}'
+                else:
+                    new_path = name
+                phases_node, phase_node_path  = _get_phases_node_in_problem_metadata(children,new_path)
+                if phases_node:
+                    return phases_node, phase_node_path
+    return None, None
 
-def get_timeseries_units(cr, varname):
+def _get_timeseries_units(cr, varname):
+    """
+    Get the units for a timeseries variable and also the time associated with that variable.
+
+    Parameters
+    ----------
+    cr : CaseRecorder
+        The CaseRecorder that contains the problem metadata
+    varname : str
+        The CaseRecorder that contains the problem metadata
+
+    Returns
+    -------
+    units : tuple of two strings
+        The units for the requested timeseries variable and also the time associated with that variable
+    """
     root_children = cr.problem_metadata['tree']['children']
 
-    traj_name = varname.split('.')[0]
-    for item in root_children:
-        if item['name'] == traj_name:
-            traj_dict = item
-            break
+    # The `phases` node can be anywhere in the hierarchy so need to recurse
+    phases_node = _get_phases_node_in_problem_metadata(root_children)
 
-    # get phases
-    for item in traj_dict['children']:
-        if item['name'] == 'phases':
-            phases_dict = item
-            break
+    time_units, var_units = None, None
 
-    # get the phase we want
-    phase_name = varname.split('.')[1]
-    for item in phases_dict['children']:
+    # get the phase child
+    phase_name = varname.split('.')[0]
+    for item in phases_node['children']:
         if item['name'] == phase_name:
             phase_dict = item
             break
 
-    # get the timeseries
+    # get the timeseries child
     for item in phase_dict['children']:
         if item['name'] == 'timeseries':
             timeseries_dict = item
             break
 
-    # get the variable
+    # get the variable child
     for item in timeseries_dict['children']:
         if item['name'] == 'states:' + varname.split(':')[-1] :
             var_units = item['units']
         if item['name'] == 'time' :
             time_units = item['units']
 
+    if var_units is None:
+        raise RuntimeError(f'Units not found for variable: {varname}')
+    if time_units is None:
+        raise RuntimeError(f'Time units not found for variable: {varname}')
 
     return time_units, var_units
 
 def timeseries_plots(recorder_filename, plot_simulation=False, simulation_record_file=None, plot_dir="plots"):
+    """
[…]
