### 2020-11-19 b0f4aa0
All three example codes work now when plotting using the function
