### 2020-11-22 5992344
radau and solve_ivp working with parameters as state rates, working on refinement.
