### 2020-11-23 0c3f87a
error estimation for grid refinement will now pull rate source value converted to the correct units even if it is not an ODE output
