+++ dymos/grid_refinement/error_estimation.py
@@ -117,7 +117,12 @@ def eval_ode_on_grid(phase, transcription):
     """
     x = {}
     u = {}
+    u_rate = {}
+    u_rate2 = {}
     p = {}
+    p_rate = {}
+    p_rate2 = {}
+    param = {}
     f = {}
 
     # Build the interpolation matrix which interpolates from all nodes on the old grid to
@@ -144,83 +149,133 @@ def eval_ode_on_grid(phase, transcription):
     # Set the values in the refinement problem using the outputs from the first
     ode = p_refine.model.grid_refinement_system.ode
 
+    t_prev = phase.get_val('timeseries.time', units=phase.time_options['units'])
+    t_phase_prev = phase.get_val('timeseries.time_phase', units=phase.time_options['units'])
+    t_initial = np.repeat(t_prev[0, 0], repeats=transcription.grid_data.num_nodes, axis=0)
+    t_duration = np.repeat(t_prev[-1, 0], repeats=transcription.grid_data.num_nodes, axis=0)
+    t = np.dot(L, t_prev)
+    t_phase = np.dot(L, t_phase_prev)
+    targets = get_targets(ode, 'time', phase.time_options['targets'])
+    time_phase_targets = get_targets(ode, 'time_phase', phase.time_options['time_phase_targets'])
+    t_initial_targets = get_targets(ode, 't_initial', phase.time_options['t_initial_targets'])
+    t_duration_targets = get_targets(ode, 't_duration', phase.time_options['t_duration_targets'])
+    if targets:
+        p_refine.set_val(f'time', t)
+    if time_phase_targets:
+        p_refine.set_val(f'time_phase', t_phase)
+    if t_initial_targets:
+        p_refine.set_val(f't_initial', t_initial)
+    if t_duration_targets:
+        p_refine.set_val(f't_duration', t_duration)
+
     for name, options in phase.state_options.items():
         x_prev = phase.get_val(f'timeseries.states:{name}', units=options['units'])
         x[name] = np.dot(L, x_prev)
         targets = get_targets(ode, name, options['targets'])
-        if not targets:
-            continue
-        p_refine.set_val(f'states:{name}', x[name])
+        if targets:
+            p_refine.set_val(f'states:{name}', x[name])
 
     for name, options in phase.control_options.items():
         targets = get_targets(ode, name, options['targets'])
         rate_targets = get_targets(ode, f'{name}_rate', options['rate_targets'])
         rate2_targets = get_targets(ode, f'{name}_rate12', options['rate2_targets'])
 
+        u_prev = phase.get_val(f'timeseries.controls:{name}', units=options['units'])
+        u[name] = np.dot(L, u_prev)
         if targets:
-            u_prev = phase.get_val(f'timeseries.controls:{name}', units=options['units'])
-            u[name] = np.dot(L, u_prev)
             p_refine.set_val(f'controls:{name}', u[name])
 
+        u_rate_prev = phase.get_val(f'timeseries.control_rates:{name}_rate')
+        u_rate[name] = np.dot(L, u_rate_prev)
         if rate_targets:
-            u_rate_prev = phase.get_val(f'timeseries.control_rates:{name}_rate')
-            u_rate_hat = np.dot(L, u_rate_prev)
-            p_refine.set_val(f'control_rates:{name}_rate', u_rate_hat)
+            p_refine.set_val(f'control_rates:{name}_rate', u_rate[name])
 
+        u_rate2_prev = phase.get_val(f'timeseries.control_rates:{name}_rate2')
+        u_rate2[name] = np.dot(L, u_rate2_prev)
         if rate2_targets:
[…]
