### 2020-11-23 7d591e1
GaussLobatto now working with states that integrate parameter values. Tweaked the timeseries comparison to use linear interpolation since the nonlinear variety was causing a few outlying interpolated points.
