+++ dymos/examples/brachistochrone/test/test_brachistochrone_integrated_parameter.py
@@ -95,20 +95,24 @@ class TestBrachistochroneIntegratedControl(unittest.TestCase):
         import dymos as dm
 
         p = om.Problem(model=om.Group())
-        p.driver = om.ScipyOptimizeDriver()
+        p.driver = om.pyOptSparseDriver()
+        p.driver.options['optimizer'] = 'SNOPT'
+        p.driver.opt_settings['iSumm'] = 6
+        p.driver.declare_coloring()
 
-        phase = dm.Phase(ode_class=BrachistochroneODE, transcription=dm.GaussLobatto(num_segments=10))
+        phase = dm.Phase(ode_class=BrachistochroneODE,
+                         transcription=dm.GaussLobatto(num_segments=10))
 
         p.model.add_subsystem('phase0', phase)
 
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10), units='s')
 
-        phase.add_state('x', fix_initial=True, fix_final=False, rate_source='xdot', units='m')
-        phase.add_state('y', fix_initial=True, fix_final=False, rate_source='ydot', units='m')
+        phase.add_state('x', fix_initial=True, fix_final=True, rate_source='xdot', units='m')
+        phase.add_state('y', fix_initial=True, fix_final=True, rate_source='ydot', units='m')
         phase.add_state('v', fix_initial=True, rate_source='vdot', units='m/s')
-        phase.add_state('theta', targets='theta', fix_initial=False, rate_source='theta_dot', units='rad')
+        phase.add_state('theta', fix_initial=False, rate_source='theta_dot', lower=1E-3)
 
-        phase.add_parameter('theta_dot', units='rad/s', opt=True, lower=0, upper=1)
+        phase.add_parameter('theta_dot', units='deg/s', opt=True, lower=0, upper=100)
 
         phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
 
@@ -125,43 +129,35 @@ class TestBrachistochroneIntegratedControl(unittest.TestCase):
         p['phase0.states:x'] = phase.interpolate(ys=[0, 10], nodes='state_input')
         p['phase0.states:y'] = phase.interpolate(ys=[10, 5], nodes='state_input')
         p['phase0.states:v'] = phase.interpolate(ys=[0, 9.9], nodes='state_input')
-        p['phase0.states:theta'] = np.radians(phase.interpolate(ys=[0.05, 100.0],
-                                                                nodes='state_input'))
-        p['phase0.parameters:theta_dot'] = 1.0
+        p['phase0.states:theta'] = np.radians(phase.interpolate(ys=[0.05, 100.0], nodes='state_input'))
+        p['phase0.parameters:theta_dot'] = 60.0
 
         # Solve for the optimal trajectory
-        p.run_driver()
+        dm.run_problem(p, refine_iteration_limit=5)
 
         # Test the results
         assert_near_equal(p.get_val('phase0.timeseries.time')[-1], 1.8016, tolerance=1.0E-3)
 
         sim_out = phase.simulate(times_per_seg=20)
 
+        t_sol = p.get_val('phase0.timeseries.time')
         x_sol = p.get_val('phase0.timeseries.states:x')
         y_sol = p.get_val('phase0.timeseries.states:y')
         v_sol = p.get_val('phase0.timeseries.states:v')
         theta_sol = p.get_val('phase0.timeseries.states:theta')
[…]
