### 2020-11-23 a53a34f
SNOPT -> SLSQP
