+++ dymos/examples/brachistochrone/test/test_brachistochrone_integrated_parameter.py
@@ -96,8 +96,7 @@ class TestBrachistochroneIntegratedControl(unittest.TestCase):
 
         p = om.Problem(model=om.Group())
         p.driver = om.pyOptSparseDriver()
-        p.driver.options['optimizer'] = 'SNOPT'
-        p.driver.opt_settings['iSumm'] = 6
+        p.driver.options['optimizer'] = 'SLSQP'
         p.driver.declare_coloring()
 
         phase = dm.Phase(ode_class=BrachistochroneODE,
@@ -167,8 +166,7 @@ class TestBrachistochroneIntegratedControl(unittest.TestCase):
 
         p = om.Problem(model=om.Group())
         p.driver = om.pyOptSparseDriver()
-        p.driver.options['optimizer'] = 'SNOPT'
-        p.driver.opt_settings['iSumm'] = 6
+        p.driver.options['optimizer'] = 'SLSQP'
         p.driver.declare_coloring()
 
         phase = dm.Phase(ode_class=BrachistochroneODE,
