### 2020-11-23 de8829d
don't assume shape when unspecified
