### 2020-11-23 745d6d8
don't include src_indices in promotes when dynamic=False
