### 2020-11-23 d374ad4
add check_valid function for shape parameter instead of specifying types
