### 2020-11-24 73f5882
get_target_metadata now accepts argument 'dynamic' and returns the shape ignoring the number of nodes. moved setup_objective to configure_objective, since we may not know all of the objective metadata during setup. Fixed a bug in double integrator test that had the shape of a control as (0, 1) rather than (1,).
