### 2020-11-24 9d253a1
Parameter shapes (#467)
* handle different add_parameter shapes

* don't assume shape when unspecified

* pep8 fixes

* don't include src_indices in promotes when dynamic=False

* add check_valid function for shape parameter instead of specifying types

* get_target_metadata now accepts argument 'dynamic' and returns the shape ignoring the number of nodes.
moved setup_objective to configure_objective, since we may not know all of the objective metadata during setup.
Fixed a bug in double integrator test that had the shape of a control as (0, 1) rather than (1,).

Co-authored-by: Rob Falck <rfa
[…]
