### 2020-11-24 a542136
Test for integrated parameter requires explicit shape and units for the parameter, since theta_dot has no targets itself. Modified error message when units cannot be found via introspection for parameters or controls.
