### 2020-11-24 e14fe44
Fixed a bug that was preventing parameters from being state rate sources. (#466)
* radau and solve_ivp working with parameters as state rates, working on refinement.

* error estimation for grid refinement will now pull rate source value converted to the correct units even if it is not an ODE output

* GaussLobatto now working with states that integrate parameter values.
Tweaked the timeseries comparison to use linear interpolation since the nonlinear variety was causing a few outlying interpolated points.

* Test for integrated parameter requires explicit shape and units for the parameter, si
[…]
