### 2020-11-25 5727ad9
change in comparison to eliminate a warning message
