### 2020-11-25 23f36af
Large amount of code cleanup and PEP8 fixes
