+++ dymos/visualization/test/test_timeseries_plots.py
@@ -1,11 +1,9 @@
 import os
 import unittest
-import warnings
 
 import numpy as np
 
 import openmdao.api as om
-from openmdao.utils.assert_utils import assert_near_equal
 from openmdao.utils.testing_utils import use_tempdirs
 
 import dymos as dm
@@ -14,7 +12,7 @@ from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneOD
 from dymos.visualization.timeseries_plots import timeseries_plots
 
 
-# @use_tempdirs
+@use_tempdirs
 class TestSimpleTimeSeriesPlots(unittest.TestCase):
     def setUp(self):
         optimizer = 'SLSQP'
@@ -102,8 +100,8 @@ class TestSimpleTimeSeriesPlots(unittest.TestCase):
         self.assertTrue(os.path.exists('plots/test_v.png'))
 
     def test_brachistochrone_timeseries_plots_solution_only_set_solution_record_file(self):
-        dm.run_problem(self.p,make_plots=False, solution_record_file='solution_record_file.db' # will plot with timeseries_plots function
-                       ) # records to the hardcode file 'dymos_simulation.db'
+        # records to the default file 'dymos_simulation.db'
+        dm.run_problem(self.p, make_plots=False, solution_record_file='solution_record_file.db')
 
         timeseries_plots('solution_record_file.db')
         self.assertTrue(os.path.exists('plots/test_x.png'))
@@ -111,44 +109,23 @@ class TestSimpleTimeSeriesPlots(unittest.TestCase):
         self.assertTrue(os.path.exists('plots/test_v.png'))
 
     def test_brachistochrone_timeseries_plots_solution_and_simulation(self):
-        dm.run_problem(self.p, simulate=True, make_plots=False, # will plot with timeseries_plots function
-                       simulation_record_file='simulation_record_file.db') # records to the hardcode file 'dymos_simulation.db'
+        dm.run_problem(self.p, simulate=True, make_plots=False,
+                       simulation_record_file='simulation_record_file.db')
 
-        timeseries_plots('dymos_solution.db', plot_simulation=True,
-                         simulation_record_file='simulation_record_file.db')
-
-
-
-    def test_brachistochrone_timeseries_plots_plot_simulation_true_but_no_path_given(self):
-        dm.run_problem(self.p, make_plots=False)
-        with self.assertRaises(ValueError) as e:
-            timeseries_plots('dymos_solution.db', plot_simulation=True)
-        expected = 'If plot_simulation is True, simulation_record_file must be path to simulation case recorder file, not None'
-        self.assertEqual(str(e.exception), expected)
-
-    def test_brachistochrone_timeseries_plots_plot_simulation_false_but_path_given(self):
-        dm.run_problem(self.p, make_plots=False)
-
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter('always')
-            timeseries_plots('dymos_solution.db', simulation_record_file='simulation_record_file.db')
-
-        expected = 'Setting simulation_record_file but not setting plot_simulation will not result in plotting simulation data'
-
-        self.assertIn(expected, [str(ww.message) for ww in w])
+        timeseries_plots('dymos_solution.db', simulation_record_file='simulation_record_file.db')
 
     def test_brachistochrone_timeseries_plots_set_plot_dir(self):
         dm.run_problem(self.p, make_plots=False)
 
         plot_dir = "test_plot_dir"
-        timeseries_plots('dymos_solution.db',plot_dir=plot_dir)
[…]

[…]
