### 2020-11-25 f0e81d9
Added doc for timeseries_plot function
