+++ dymos/test/test_run_problem.py
@@ -349,9 +349,7 @@ class TestRunProblemPlotting(unittest.TestCase):
         os.remove(solution_record_file)
 
     def test_run_brachistochrone_problem_plot_simulation(self):
-        simulation_record_file = 'simulation_record_file.db'
-        dm.run_problem(self.p, make_plots=True, simulate=True,
-                       simulation_record_file=simulation_record_file)
+        dm.run_problem(self.p, make_plots=True, simulate=True)
 
         self.assertTrue(os.path.exists('plots/test_x.png'))
         self.assertTrue(os.path.exists('plots/test_y.png'))

[…]
