### 2020-11-25 1000805
if parameter value is given, use that value as the source of its shape (assuming float, int, or complex value).
