### 2020-11-25 ade5307
if parameter value is given, use that value as the source of its shape (assuming float, int, or complex value).
