+++ dymos/test/test_run_problem.py
@@ -3,16 +3,17 @@ import os
 import unittest
 
 import numpy as np
+from numpy.testing import assert_almost_equal
 
 import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 from openmdao.utils.testing_utils import use_tempdirs
-from openmdao.utils.general_utils import set_pyoptsparse_opt
-_, optimizer = set_pyoptsparse_opt('IPOPT', fallback=True)
 
 import dymos as dm
 from dymos.examples.hyper_sensitive.hyper_sensitive_ode import HyperSensitiveODE
 from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
+from openmdao.utils.general_utils import set_pyoptsparse_opt
+_, optimizer = set_pyoptsparse_opt('IPOPT', fallback=True)
 
 
 @use_tempdirs
@@ -166,7 +167,7 @@ class TestRunProblem(unittest.TestCase):
         phase0 = traj.add_phase('phase0', dm.Phase(ode_class=BrachistochroneODE,
                                                    transcription=dm.Radau(num_segments=10,
                                                                           order=3)))
-        phase0.set_time_options(fix_initial=True, fix_duration=True)
+        phase0.set_time_options(fix_initial=True, fix_duration=False)
         phase0.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
                          units=BrachistochroneODE.states['x']['units'],
                          fix_initial=True, fix_final=False, solve_segments=False)
@@ -201,6 +202,64 @@ class TestRunProblem(unittest.TestCase):
         p.set_val('traj.phase0.parameters:g', 9.80665)
 
         dm.run_problem(p)
+
+        self.assertTrue(os.path.exists('dymos_solution.db'))
+        # Assert the results are what we expect.
+        cr = om.CaseReader('dymos_solution.db')
+        case = cr.get_case('final')
+        assert_almost_equal(case.outputs['traj.phase0.timeseries.time'].max(), 1.8016, decimal=4)
+
+    def test_run_brachistochrone_problem_with_simulate(self):
+        p = om.Problem(model=om.Group())
+        p.driver = om.pyOptSparseDriver()
+        p.driver.declare_coloring()
+        p.driver.options['optimizer'] = 'SLSQP'
+
+        traj = p.model.add_subsystem('traj', dm.Trajectory())
+        phase0 = traj.add_phase('phase0', dm.Phase(ode_class=BrachistochroneODE,
+                                                   transcription=dm.Radau(num_segments=10,
+                                                                          order=3)))
+        phase0.set_time_options(fix_initial=True, fix_duration=False)
+        phase0.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
+                         units=BrachistochroneODE.states['x']['units'],
+                         fix_initial=True, fix_final=False, solve_segments=False)
+        phase0.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
+                         units=BrachistochroneODE.states['y']['units'],
+                         fix_initial=True, fix_final=False, solve_segments=False)
+        phase0.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
+                         units=BrachistochroneODE.states['v']['units'],
+                         fix_initial=True, fix_final=False, solve_segments=False)
+        phase0.add_control('theta', continuity=True, rate_continuity=True,
+                           units='deg', lower=0.01, upper=179.9)
+        phase0.add_parameter('g', units='m/s**2', val=9.80665)
+
[…]
