+++ dymos/visualization/test/test_timeseries_plots.py
@@ -13,7 +13,7 @@ from dymos.visualization.timeseries_plots import timeseries_plots
 
 
 @use_tempdirs
-class TestSimpleTimeSeriesPlots(unittest.TestCase):
+class TestTimeSeriesPlotsBasics(unittest.TestCase):
     def setUp(self):
         optimizer = 'SLSQP'
         num_segments = 8
@@ -75,13 +75,6 @@ class TestSimpleTimeSeriesPlots(unittest.TestCase):
         p['traj0.phase0.controls:theta'] = phase.interpolate(ys=[5, 100], nodes='control_input')
         p['traj0.phase0.parameters:g'] = 9.80665
 
-        # solution_record_file = 'test_timeseries_plots.db'
-        # simulation_record_file = 'dymos_solution.db'
-
-        # self.recorder = om.SqliteRecorder(solution_record_file)
-
-        # p.model.add_recorder(self.recorder)
-
         p.setup()
 
         self.p = p
@@ -126,312 +119,7 @@ class TestSimpleTimeSeriesPlots(unittest.TestCase):
 
 
 @use_tempdirs
-class TestTimeSeriesPlots(unittest.TestCase):
-
-    def tearDown(self):
-        for filename in ['test_time_series_plots.db']:
-            if os.path.exists(filename):
-                os.remove(filename)
-
-    def setUp(self):
-
-        self.traj = dm.Trajectory()
-        p = self.p = om.Problem(model=self.traj)
-
-        # Since we're only testing features like get_values that don't rely on a converged
-        # solution, no driver is attached.  We'll just invoke run_model.
-
-        # First Phase (burn)
-
-        burn1 = dm.Phase(ode_class=FiniteBurnODE, transcription=dm.GaussLobatto(num_segments=4,
-                                                                                order=3))
-
-        self.traj.add_phase('burn1', burn1)
-
-        burn1.set_time_options(fix_initial=True, duration_bounds=(.5, 10), units='TU')
-        burn1.add_state('r', fix_initial=True, fix_final=False,
-                        rate_source='r_dot', targets=['r'], units='DU')
-        burn1.add_state('theta', fix_initial=True, fix_final=False,
-                        rate_source='theta_dot', targets=['theta'], units='rad')
-        burn1.add_state('vr', fix_initial=True, fix_final=False,
-                        rate_source='vr_dot', targets=['vr'], units='DU/TU')
-        burn1.add_state('vt', fix_initial=True, fix_final=False,
-                        rate_source='vt_dot', targets=['vt'], units='DU/TU')
-        burn1.add_state('accel', fix_initial=True, fix_final=False,
-                        rate_source='at_dot', targets=['accel'], units='DU/TU**2')
-        burn1.add_state('deltav', fix_initial=True, fix_final=False,
-                        rate_source='deltav_dot', units='DU/TU')
-        burn1.add_control('u1',  targets=['u1'], rate_continuity=True, rate2_continuity=True,
-                          units='deg', scaler=0.01, lower=-30, upper=30)
-        burn1.add_parameter('c', opt=False, val=1.5, targets=['c'], units='DU/TU')
-
-        # Second Phase (Coast)
-
-        coast = dm.Phase(ode_class=FiniteBurnODE, transcription=dm.GaussLobatto(num_segments=10,
-                                                                                order=3))
-
-        self.traj.add_phase('coast', coast)
-
-        coast.set_time_options(initial_bounds=(0.5, 20), duration_bounds=(.5, 10), duration_ref=10,
-                               units='TU')
-        coast.add_state('r', fix_initial=False, fix_final=False, defect_scaler=100.0,
[…]
