### 2020-11-25 4280f6d
Removed commented out code
