### 2020-11-26 #2809 @goerz
Hmm.. ok with me, although it still seems a bit borderline, considering that the earlier paper announces Dymos in its abstract and extensively discusses benchmarks. In any case, this is clearly an editor's decision, so if @dpsanders has no objections to this manuscript being published in JOSS, I'll finish the review.
