### 2020-11-26 OpenMDAO/dymos#453 @goerz
You may also want to update the "default" installation instructions in the README and the documentation to just `python -m pip install dymos`
