### 2020-11-26 OpenMDAO/dymos#471 @goerz
Instructions for running tests locally
Can you add instructions to https://openmdao.github.io/dymos/contributing/contributing.html on how to run the automated tests manually?

From a fresh checkout of the `dymos` repo, what commands to do I have to type in to run the test locally? I had a look at `.travis.yml`, but the commands there seem surprisingly complex.

Personally, I tend to add a Makefile to my projects so that I can just to `make test` (on Unix-based systems, at least), but that's totally up to you.

This is a blocker for https://github.com/openjournals/joss-reviews/issues/2809
