### 2020-11-26 OpenMDAO/dymos#472 @goerz
Statement of Need
If I read the requirements for https://github.com/openjournals/joss-reviews/issues/2809 correctly, it seems that JOSS wants to have a "Statement of Need" *both* in the JOSS paper and in the software documentation. While you have the statement in the paper, I didn't find it in the documentation.

To be honest, "Statement of Need" as an explicit boilerplate section in the documentation seems a bit unnecessary to me (@dpsanders: How strict are you on this?), although what I would consider very useful a "related software" section in the documentation that gives an overview of existing/related software packages and discusses how Dymos fits into that ecosystem. I'd consider anything like that (even if not explicitly titled "Statement of Need", or any duplication of the statement from the JOSS manuscript sufficient to check of that point.
