### 2020-11-26 #2809 @goerz
The JOSS manuscript will definitely have to cite the [earlier publication](https://arc.aiaa.org/doi/abs/10.2514/6.2019-0976), though. Even if it's just "The methods implemented in Dymos have previously been described in *Ref*"
