### 2020-11-28 16bcafd
more work on takeoff portion of balanced field
