### 2020-11-28 9f72c49
more work on takeoff portion of balanced field
