+++ dymos/examples/balanced_field/doc/test_doc_balanced_field_length.py
@@ -19,6 +19,7 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
 
         import dymos as dm
         from dymos.examples.balanced_field.ground_roll_ode import GroundRollODE
+        from dymos.examples.balanced_field.takeoff_ode import TakeoffODE
 
         #
         # Instantiate the problem and configure the optimization driver
@@ -31,7 +32,8 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
         p.driver.opt_settings['Verify level'] = 3
         p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-6
         p.driver.opt_settings['Major optimality tolerance'] = 1.0E-6
-        p.driver.opt_settings['Major iterations limit'] = 20
+        p.driver.opt_settings['Major iterations limit'] = 50
+        p.driver.opt_settings['Minor iterations limit'] = 100000
         p.driver.declare_coloring()
 
         #
@@ -54,7 +56,7 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
         #
         # Set the options on the optimization variables
         #
-        p1.set_time_options(fix_initial=True, duration_bounds=(1, 200), duration_ref=10.0)
+        p1.set_time_options(fix_initial=True, duration_bounds=(1, 1000), duration_ref=10.0)
 
         p1.add_state('r', fix_initial=True, lower=0, ref=1000.0, defect_ref=1000.0,
                      rate_source='r_dot')
@@ -63,7 +65,9 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
 
         p1.add_parameter('h', opt=False, units='m')
         p1.add_parameter('T', val=27000 * 2, opt=False, units='lbf')
+        p1.add_parameter('m', val=174200, opt=False, units='lbm')
         p1.add_parameter('alpha', val=0.0, opt=False, units='deg')
+        p1.add_parameter('mu_r', val=0.03, opt=False, units=None)
 
         p1.add_timeseries_output('*')
         
@@ -90,15 +94,17 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
 
         p2.add_parameter('h', val=0.0, opt=False, units='m')
         p2.add_parameter('T', val=27000, opt=False, units='lbf')
+        p2.add_parameter('m', val=174200, opt=False, units='lbm')
+        p2.add_parameter('mu_r', val=0.03, opt=False, units=None)
         p2.add_parameter('alpha', val=0.0, opt=False, units='deg')
 
         p2.add_timeseries_output('*')
 
-        # p2.add_boundary_constraint('v_over_v_stall', loc='final', lower=1.2)
+        # p2.add_boundary_constraint('v_over_v_stall', loc='final', equals=1.2)
 
         # Minimize time at the end of the phase
         # p2.add_objective('time', loc='final', ref=1.0)
-        
+
         # Third Phase - Rejected Takeoff
         # V1 to Zero speed with no propulsion and braking.
         #
@@ -120,16 +126,17 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
 
         p3.add_parameter('h', val=0.0, opt=False, units='m')
         p3.add_parameter('T', val=0.0, opt=False, units='N')
+        p3.add_parameter('m', val=174200, opt=False, units='lbm')
         p3.add_parameter('mu_r', val=0.3, opt=False, units=None)
         p3.add_parameter('alpha', val=0.0, opt=False, units='deg')
 
         p3.add_timeseries_output('*')
 
-        p3.add_boundary_constraint('v', loc='final', equals=0, ref=100)
[…]
