### 2020-11-30 3c94146
adding detail on how to install some of the more complex prerequisites
