+++ mkdocs/docs/contributing/contributing.md
@@ -72,3 +72,77 @@ index=1) }}" }}
 ```
 
 Here `alt_text` is the text displayed when cursor is over the image, and `index` is the index of the plot figure from the test (default is 1, this index starts at 1).
+
+## Installing Dependencies
+
+Many of the simple dependencies for Dymos are installed via pip using
+
+```
+python -m pip install dymos
+```
+
+However, Dymos relies on a few more advanced dependencies:
+
+### mpi4py and petsc4py
+
+Together, `mpi4py` and `petsd4py` enable parallel processing in Dymos via MPI.
+The easiest way to install `mpi4py` is to use the [anaconda](https://www.anaconda.com/products/individual) python environment.
+
+```
+    conda create --yes -n PY$PY python=$PY;
+    conda activate PY$PY;
+    conda config --add channels conda-forge;
+
+    conda install mpi4py
+    conda install petsc4py
+```
+
+### pyoptsparse
+
+Dymos will work "out-of-the-box" using OpenMDAO's `ScipyOptimizeDriver` for many of the more simple problems.
+For more complex problems, the implicit optimization techniques used by Dymos can easily generate problems with hundreds or even thousands of design variables and constraints.
+When the nonlinear optimization problems generated Dymos become that complicated, the free optimizers available via `ScipyOptimizeDriver` can struggle to converge.
+
+For this reason, the authors tend to use the open-source [IPOPT](https://coin-or.github.io/Ipopt/) optimizer or the proprietary [SNOPT](https://ccom.ucsd.edu/~optimizers/solvers/snopt/) optimizer.
+Both of these optimizers are able to capitalize on the _sparse_ nature of the optimal control problems generated by Dymos, significantly reducing the time and memory required to solve problems.
+
+For uniform access to a variety of optimizers, including `SLSQP`, `IPOPT`, and `SNOPT`, Dymos uses OpenMDAO's `pyOptSparseDriver`, which interfaces to these external optimizers via [pyoptsparse](https://github.com/mdolab/pyoptsparse), produced by MDOLab at the University of Michigan.
+User's on OS X and Linux systems can use our [script](https://github.com/OpenMDAO/build_pyoptsparse) to build and install pyoptsparse with support for IPOPT.
+The script will also handle support for SNOPT if the user has access to the SNOPT source code.
+Installing these dependencies on Windows-based systems requires [Intel Fortran](https://software.intel.com/content/www/us/en/develop/tools/oneapi/components/fortran-compiler.html).
+Free Fortran compilers on Windows, at the time of this writing, are not compatible with the Microsoft ABI used by Python on Windows.
+
+## Running Tests
+
+Dymos tests can be run with any test runner such as [nosetests](https://nose.readthedocs.io/en/latest/) or [pytest](https://docs.pytest.org/en/stable/).
+However, due to some MPI-specific tests in our examples, we prefer our [testflo](https://github.com/OpenMDAO/testflo) package.
+The testflo utility can be installed using
+
+```
+python -m pip install testflo
+```
+
+Testflo can be invoked from the top-level Dymos directory with:
+
+```
+testflo .
+```
+
+With pyoptsparse correctly installed and things working correctly, the tests should conclude after several minutes with a message like the following:
+The lack of MPI capability or pyoptsparse will cause additional tests to be skipped.
+
+```
+The following tests were skipped:
+test_command_line.py:TestCommandLine.test_ex_brachistochrone_reset_grid
+
+
+OK
+
+
+Passed:  450
+Failed:  0
+Skipped: 1
+
+Ran 451 tests using 2 processes
+```
+
