### 2020-11-30 196b13b
Added more thorough installation instructions to documentation. Added a "Why Dymos?" section to the documentation to cover the statement of need. Added phase interpolate method to the documentation.
