+++ mkdocs/docs/api/phase_api.md
@@ -1,3 +1,3 @@
 # The Phase API
 
-{{ api_doc('dymos.Phase', members=['set_time_options', 'add_state', 'set_state_options', 'add_control', 'set_control_options', 'add_polynomial_control', 'set_polynomial_control_options', 'add_parameter', 'set_parameter_options', 'add_timeseries_output', 'simulate', 'set_refine_options']) }}
+{{ api_doc('dymos.Phase', members=['set_time_options', 'add_state', 'set_state_options', 'add_control', 'set_control_options', 'add_polynomial_control', 'set_polynomial_control_options', 'add_parameter', 'set_parameter_options', 'add_timeseries_output', 'simulate', 'set_refine_options', 'interpolate']) }}
