### 2020-11-30 fefe8fe
added add_boundary_constraint and add_path_constraint to the phase api reference
