### 2020-12-01 46eb1ad
State Discovery: basic operation works, tests added, needs docs.
