### 2020-12-02 24151f8
Now plotting all variables in timeseries not including input_values
