### 2020-12-02 324a793
made the conda instructions a bit more clear (hopefully)
