### 2020-12-02 53cc645
I402 Automatically generate plots of all timeseries outputs (#469)
* Initial code to add timeseries plots

* remove visualization from the ignores

* visualization folder

* removed files that should not have been committed

* Fixed location and spacing for legend

* All code and tests for basic plotting from run_problem and timeseries_plots work

* Many changes to get to the point of being able to plot multiple phases

* All three example codes work now when plotting using the function

* change in comparison to eliminate a warning message

* Large amount of code cleanup and PEP8 fixes

* Add
[…]
