### 2020-12-02 ac00d61
Some documentation cleanup for the JOSS review (#474)
* adding detail on how to install some of the more complex prerequisites

* Added more thorough installation instructions to documentation.
Added a "Why Dymos?" section to the documentation to cover the statement of need.
Added phase interpolate method to the documentation.

* added add_boundary_constraint and add_path_constraint to the phase api reference

* Review feedback on documentation updates.
Fixed errant argument in link_phases documentation.

* made the conda instructions a bit more clear (hopefully)
