+++ mkdocs/references.bib
@@ -68,7 +68,7 @@ url = {https://doi.org/10.2514/6.2018-1646},
 year = {2018}
 }
 
-@article {hwang:2018b,
+@article {hwang2018b,
 	title = {A computational architecture for coupling heterogeneous numerical models and computing coupled derivatives},
 	journal = {ACM TOMS},
 	year = {2018},
@@ -497,4 +497,23 @@ eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9780898718577}
   pages={4081--4106},
   year={2015},
   publisher={Elsevier}
+}
+
+@Article{Andersson2018,
+  Author = {Joel A E Andersson and Joris Gillis and Greg Horn
+            and James B Rawlings and Moritz Diehl},
+  Title = {{CasADi} -- {A} software framework for nonlinear optimization
+           and optimal control},
+  Journal = {Mathematical Programming Computation},
+  volume={11},
+  pages={1--36},
+  Year = {2019},
+}
+
+@inproceedings{gray2019coloring,
+  title={Using Graph Coloring To Compute Total Derivatives More Efficiently in OpenMDAO},
+  author={Gray, Justin S and Hearn, Tristan A and Naylor, Bret A},
+  booktitle={AIAA Aviation 2019 Forum},
+  pages={3108},
+  year={2019}
 }
