+++ mkdocs/mkdocs.yml
@@ -40,6 +40,7 @@ markdown_extensions:
 
 nav:
     - 'Home': 'index.md'
+    - 'Installation': 'installation.md'
     - 'Getting Started':
         - 'getting_started/collocation.md'
         - 'getting_started/optimal_control.md'
