+++ dymos/examples/flying_robot/test/test_flying_robot.py
@@ -2,8 +2,6 @@ import itertools
 import os
 import unittest
 
-from parameterized import parameterized
-
 import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 from openmdao.utils.testing_utils import use_tempdirs
