+++ dymos/transcriptions/common/test/test_control_interp_comp.py
@@ -1,8 +1,6 @@
 import itertools
 import unittest
 
-from parameterized import parameterized
-
 import numpy as np
 from numpy.testing import assert_almost_equal
 import openmdao.api as om
@@ -77,496 +75,488 @@ class TestControlRateComp(unittest.TestCase):
     def tearDown(self):
         dm.options['include_check_partials'] = False
 
-    @parameterized.expand(
-        itertools.product(['gauss-lobatto', 'radau-ps'],  # transcription
-                          [True, False],  # compressed
-                          ), name_func=lambda f, n, p: '_'.join(
-            ['test_control_interp_scalar', p.args[0], str(p.args[1])])
-    )
-    def test_control_interp_scalar(self, transcription='gauss-lobatto', compressed=True):
-        segends = np.array([0.0, 3.0, 10.0])
+    def test_control_interp_scalar(self):
+        param_list = itertools.product(['gauss-lobatto', 'radau-ps'],  # transcription
+                                        [True, False],  # compressed
+                                       )
+        for transcription, compressed in param_list:
+            with self.subTest():
+                segends = np.array([0.0, 3.0, 10.0])
 
-        gd = GridData(num_segments=2,
-                      transcription_order=5,
-                      segment_ends=segends,
-                      transcription=transcription,
-                      compressed=compressed)
+                gd = GridData(num_segments=2,
+                              transcription_order=5,
+                              segment_ends=segends,
+                              transcription=transcription,
+                              compressed=compressed)
 
-        p = om.Problem(model=om.Group())
+                p = om.Problem(model=om.Group())
 
-        controls = {'a': {'units': 'm', 'shape': (1,), 'dynamic': True},
-                    'b': {'units': 'm', 'shape': (1,), 'dynamic': True}}
+                controls = {'a': {'units': 'm', 'shape': (1,), 'dynamic': True},
+                            'b': {'units': 'm', 'shape': (1,), 'dynamic': True}}
 
-        ivc = om.IndepVarComp()
-        p.model.add_subsystem('ivc', ivc, promotes_outputs=['*'])
+                ivc = om.IndepVarComp()
+                p.model.add_subsystem('ivc', ivc, promotes_outputs=['*'])
 
-        ivc.add_output('controls:a',
-                       val=np.zeros((gd.subset_num_nodes['control_input'], 1)),
-                       units='m')
+                ivc.add_output('controls:a',
+                               val=np.zeros((gd.subset_num_nodes['control_input'], 1)),
+                               units='m')
 
-        ivc.add_output('controls:b',
-                       val=np.zeros((gd.subset_num_nodes['control_input'], 1)),
-                       units='m')
+                ivc.add_output('controls:b',
+                               val=np.zeros((gd.subset_num_nodes['control_input'], 1)),
+                               units='m')
 
-        ivc.add_output('t_initial', val=0.0, units='s')
-        ivc.add_output('t_duration', val=10.0, units='s')
+                ivc.add_output('t_initial', val=0.0, units='s')
+                ivc.add_output('t_duration', val=10.0, units='s')
 
-        p.model.add_subsystem('time_comp',
-                              subsys=TimeComp(num_nodes=gd.num_nodes, node_ptau=gd.node_ptau,
-                                              node_dptau_dstau=gd.node_dptau_dstau, units='s'),
-                              promotes_inputs=['t_initial', 't_duration'],
-                              promotes_outputs=['time', 'dt_dstau'])
+                p.model.add_subsystem('time_comp',
+                                      subsys=TimeComp(num_nodes=gd.num_nodes, node_ptau=gd.node_ptau,
+                                                      node_dptau_dstau=gd.node_dptau_dstau, units='s'),
[…]
