+++ dymos/transcriptions/common/test/test_continuity_comp.py
@@ -1,7 +1,6 @@
 import unittest
 
 import itertools
-from parameterized import parameterized
 
 import numpy as np
 
@@ -19,152 +18,151 @@ from dymos.utils.misc import CompWrapperConfig
 GaussLobattoContinuityComp = CompWrapperConfig(GaussLobattoContinuityComp)
 RadauPSContinuityComp = CompWrapperConfig(RadauPSContinuityComp)
 
+params_list = itertools.product(['gauss-lobatto', 'radau-ps'],  # transcription
+                                ['compressed', 'uncompressed']  # compressed
+                                )
+
 
 class TestContinuityComp(unittest.TestCase):
 
-    @parameterized.expand(
-        itertools.product(['gauss-lobatto', 'radau-ps'],  # transcription
-                          ['compressed', 'uncompressed'],  # compressed
-                          ), name_func=lambda f, n, p: '_'.join(['test_continuity_comp',
-                                                                 p.args[0],
-                                                                 p.args[1]])
-    )
-    def test_continuity_comp(self, transcription='gauss-lobatto', compressed='compressed'):
+    def test_continuity_comp(self):
+        for transcription, compressed in params_list:
+            with self.subTest():
 
-        dm.options['include_check_partials'] = True
-        num_seg = 3
+                dm.options['include_check_partials'] = True
+                num_seg = 3
 
-        gd = GridData(num_segments=num_seg,
-                      transcription_order=[5, 3, 3],
-                      segment_ends=[0.0, 3.0, 10.0, 20],
-                      transcription=transcription,
-                      compressed=compressed == 'compressed')
+                gd = GridData(num_segments=num_seg,
+                              transcription_order=[5, 3, 3],
+                              segment_ends=[0.0, 3.0, 10.0, 20],
+                              transcription=transcription,
+                              compressed=compressed == 'compressed')
 
-        self.p = om.Problem(model=om.Group())
+                self.p = om.Problem(model=om.Group())
 
-        ivp = self.p.model.add_subsystem('ivc', subsys=om.IndepVarComp(), promotes_outputs=['*'])
+                ivp = self.p.model.add_subsystem('ivc', subsys=om.IndepVarComp(), promotes_outputs=['*'])
 
-        nn = gd.subset_num_nodes['all']
+                nn = gd.subset_num_nodes['all']
 
-        ivp.add_output('x', val=np.arange(nn), units='m')
-        ivp.add_output('y', val=np.arange(nn), units='m/s')
-        ivp.add_output('u', val=np.zeros((nn, 3)), units='deg')
-        ivp.add_output('v', val=np.arange(nn), units='N')
-        ivp.add_output('u_rate', val=np.zeros((nn, 3)), units='deg/s')
-        ivp.add_output('v_rate', val=np.arange(nn), units='N/s')
-        ivp.add_output('u_rate2', val=np.zeros((nn, 3)), units='deg/s**2')
-        ivp.add_output('v_rate2', val=np.arange(nn), units='N/s**2')
-        ivp.add_output('t_duration', val=121.0, units='s')
+                ivp.add_output('x', val=np.arange(nn), units='m')
+                ivp.add_output('y', val=np.arange(nn), units='m/s')
+                ivp.add_output('u', val=np.zeros((nn, 3)), units='deg')
+                ivp.add_output('v', val=np.arange(nn), units='N')
+                ivp.add_output('u_rate', val=np.zeros((nn, 3)), units='deg/s')
+                ivp.add_output('v_rate', val=np.arange(nn), units='N/s')
+                ivp.add_output('u_rate2', val=np.zeros((nn, 3)), units='deg/s**2')
+                ivp.add_output('v_rate2', val=np.arange(nn), units='N/s**2')
+                ivp.add_output('t_duration', val=121.0, units='s')
 
[…]
