### 2020-12-02 847c622
State discovery docs + fixes from review
