+++ dymos/phase/test/test_state_discovery.py
@@ -30,13 +30,13 @@ class BrachistochroneODE(om.ExplicitComponent):
         self.add_input('theta', val=np.ones(nn), desc='angle of wire', units='rad')
 
         self.add_output('xdot', val=np.zeros(nn), desc='velocity component in x', units='m/s',
-                                    tags=['state_rate_source:x', 'state_units:m'])
+                        tags=['state_rate_source:x', 'state_units:m'])
 
         self.add_output('ydot', val=np.zeros(nn), desc='velocity component in y', units='m/s',
-                                    tags=['state_rate_source:y', 'state_units:m'])
+                        tags=['state_rate_source:y', 'state_units:m'])
 
         self.add_output('vdot', val=np.zeros(nn), desc='acceleration magnitude', units='m/s**2',
-                                    tags=['state_rate_source:v', 'state_units:m/s'])
+                        tags=['state_rate_source:v', 'state_units:m/s'])
 
         self.add_output('check', val=np.zeros(nn), desc='check solution: v/sin(theta) = constant',
                         units='m/s')
@@ -217,7 +217,6 @@ class TestStateDiscovery(unittest.TestCase):
 
         msg = ("State 'x' is missing a rate_source.")
         self.assertEqual(str(cm.exception), msg)
-
 
 
 if __name__ == "__main__":
