### 2020-12-03 0b1e8e0
moved solve_segments validation to _configure_solve_segments in PseudospectralBase from StateIndependentsComp.  Valid values for solve_segments are now True, False, None, "forward", and "backward".
