+++ dymos/examples/brachistochrone/test/test_brachistochrone_solve_segments.py
@@ -1,13 +1,94 @@
 import unittest
 from numpy.testing import assert_almost_equal
+import openmdao.api as om
+import dymos as dm
 
 import dymos.examples.brachistochrone.test.ex_brachistochrone_vector_states as ex_brachistochrone_vs
 import dymos.examples.brachistochrone.test.ex_brachistochrone as ex_brachistochrone
+from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
 from openmdao.utils.testing_utils import use_tempdirs
 
 from openmdao.utils.general_utils import set_pyoptsparse_opt
 
 OPT, OPTIMIZER = set_pyoptsparse_opt('SLSQP')
+
+
+def _make_problem(transcription='gauss-lobatto', num_segments=8, transcription_order=3,
+                  compressed=True, optimizer='SLSQP', run_driver=True, force_alloc_complex=False,
+                  solve_segments=False):
+    p = om.Problem(model=om.Group())
+
+    p.driver = om.pyOptSparseDriver()
+    p.driver.options['optimizer'] = optimizer
+    if optimizer == 'SNOPT':
+        p.driver.opt_settings['iSumm'] = 6
+        p.driver.opt_settings['Verify level'] = 3
+    p.driver.declare_coloring(tol=1.0E-12)
+
+    if transcription == 'gauss-lobatto':
+        t = dm.GaussLobatto(num_segments=num_segments,
+                            order=transcription_order,
+                            compressed=compressed)
+    elif transcription == 'radau-ps':
+        t = dm.Radau(num_segments=num_segments,
+                     order=transcription_order,
+                     compressed=compressed)
+    elif transcription == 'runge-kutta':
+        t = dm.RungeKutta(num_segments=num_segments,
+                          order=transcription_order,
+                          compressed=compressed)
+    traj = dm.Trajectory()
+    phase = dm.Phase(ode_class=BrachistochroneODE, transcription=t)
+    p.model.add_subsystem('traj0', traj)
+    traj.add_phase('phase0', phase)
+
+    phase.set_time_options(fix_initial=True, duration_bounds=(.5, 10))
+
+    phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
+                    fix_initial=False, fix_final=False, solve_segments=solve_segments)
+    phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
+                    fix_initial=False, fix_final=False, solve_segments=solve_segments)
+
+    # Note that by omitting the targets here Dymos will automatically attempt to connect
+    # to a top-level input named 'v' in the ODE, and connect to nothing if it's not found.
+    phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
+                    fix_initial=False, fix_final=False, solve_segments=solve_segments)
+
+    phase.add_control('theta',
+                      continuity=True, rate_continuity=True,
+                      units='deg', lower=0.01, upper=179.9)
+
+    phase.add_parameter('g', targets=['g'], units='m/s**2')
+
+    phase.add_timeseries('timeseries2',
+                         transcription=dm.Radau(num_segments=num_segments*5,
+                                                order=transcription_order,
+                                                compressed=compressed),
+                         subset='control_input')
+
+    phase.add_boundary_constraint('x', loc='initial', equals=0)
+    phase.add_boundary_constraint('y', loc='initial', equals=10)
+    phase.add_boundary_constraint('v', loc='initial', equals=0)
+
[…]
