### 2020-12-03 fe147cb
Drop Parameterized Dependency (#479)
* Replaced parameterized-based iteration with subTests

* formatting fix

Co-authored-by: Kaushik Ponnapalli <kaushik.s.ponnapalli@nasa.gov>
