### 2020-12-03 ca6b6fa
Discovery of states via tags. (#477)
* State Discovery: basic operation works, tests added, needs docs.

* State discovery docs + fixes from review

Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
