+++ dymos/phase/options.py
@@ -352,7 +352,7 @@ class StateOptionsDictionary(om.OptionsDictionary):
         self.declare(name='shape', types=Iterable, allow_none=True, default=None,
                      desc='shape of the state variable, as determined by introspection')
 
-        self.declare(name='rate_source', types=str,
+        self.declare(name='rate_source', types=str, allow_none=True, default=None,
                      desc='ODE-path to the derivative of the state variable')
 
         self.declare(name='targets', allow_none=True, default=_unspecified,
