### 2020-12-04 dd89bee
reworking things to allow solve_segments="backward"
