+++ dymos/transcriptions/pseudospectral/pseudospectral_base.py
@@ -84,82 +84,77 @@ class PseudospectralBase(TranscriptionBase):
                                      units=options['units'])
 
             if options['opt']:
-                if options['solve_segments']:
-                    # If we are using a solver on the defects, then our design variables
-                    # are the first nodes in each segment.
-                    # For instance, for two 5th order radau segments, the indep indices are [0, 6].
-                    # If we have vectorized states of size n, then there are n design variables
-                    # at each node.  For instance, with n=2, the desvar indices are [0, 1, 12, 13]
-                    num_seg = grid_data.num_segments
-                    # Get the desvar node indices
-                    desvar_node_idxs = np.asarray(self.state_idx_map[name]['indep'])
-                    # In compressed transcription, the desvar indices are just the first
-                    # index for each element in the state shape
-                    if self.options['compressed']:
-                        desvar_indices = np.arange(size, dtype=int)
-                    else:
-                        # In uncompressed transcription, we need desvar_indices repeated
-                        # once for each segment, with the number of nodes in all but the
-                        # last segments added to it
-                        desvar_indices = size * np.repeat(desvar_node_idxs, size) + \
-                            np.tile(np.arange(size, dtype=int), num_seg)
-                    desvar_indices = list(desvar_indices)
-                else:
-                    desvar_indices = list(range(size * num_state_input_nodes))
+                # Add the states as design variables.
+                #
+                # In the case of optimizer-driven collocation, this includes the values at all
+                # nodes (excluding initial and/or final if fix_initial and/or fix_final is specified).
+                #
+                # In the case of solve_segments == 'forward', the design var nodes are the first
+                # node in the  phase (when compressed) or the first node in each each segment
+                # (when not compressed).
+                #
+                # When fix_initial is True, the first node in the phase is then removed from the
+                # design variable node indices. (So when compressed, there is no design variable).
+                #
+                # In the case of solve_segments == 'backward', the design var nodes are the last
+                # node in the phase (when compressed) or the last node in each segment (when not
+                # compressed).
+                #
+                # When fix_final is True, the last node in the phase is then removed from the design
+                # variable node indices.
+                #
+                print(num_state_input_nodes)
+                desvar_node_idxs = np.asarray(self.state_idx_map[name]['indep'])
 
-            if options['fix_initial']:
-                if options['initial_bounds'] is not None:
-                    raise ValueError('Cannot specify \'fix_initial=True\' and specify '
-                                     'initial_bounds for state {0}'.format(name))
-                if isinstance(options['fix_initial'], Iterable):
-                    idxs_to_fix = np.where(np.asarray(options['fix_initial']))[0]
-                    for idx_to_fix in reversed(sorted(idxs_to_fix)):
-                        del desvar_indices[idx_to_fix]
-                else:
-                    del desvar_indices[:size]
+                # This matrix will contain 1's for every index that is to be a design variable,
+                # otherwise the value will be zero
+                state_input_shape = (num_state_input_nodes,) + options['shape']
+                idx_mask = np.zeros(state_input_shape, dtype=int)
+                idx_mask[desvar_node_idxs, ...] = 1
 
-            elif options['connected_initial'] and not options['solve_segments']:
-                del desvar_indices[:size]
+                if options['fix_initial']:
+                    if options['initial_bounds'] is not None:
+                        raise ValueError('Cannot specify \'fix_initial=True\' and specify '
[…]
