### 2020-12-04 OpenMDAO/dymos#482 @dpsanders
Update reference to 
### Summary of Issue

Update reference to Hwang + Martins paper:
https://dl.acm.org/doi/abs/10.1145/3182393

### Issue Type

- [ ] Bug
- [ ] Enhancement
- [X] Docs
- [ ] Miscellaneous

Part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/2809)
