### 2020-12-04 #2809 @dpsanders
👋 @goerz and @robfalck. Having looked carefully at the prior paper  https://arc.aiaa.org/doi/abs/10.2514/6.2019-0976, I agree that the previous paper was about the algorithms and a separate software publication in JOSS is acceptable.

I do agree with @goerz that it is an omission that needs to be rectified that that paper is not mentioned in the JOSS submission. @robfalck: Please be explicit about how the papers are different.

Given that you think of this paper as "document[ing] the APIs and implementation of the software and how to use it", perhaps you can go into a bit more detail about that, and how it differs from other packages. Thanks!
