### 2020-12-05 #2809 @goerz
@dpsanders Thanks for weighing in!

I'm pretty much done with verifying the software package, and it's looking pretty good to me. The manuscript itself is ok, but I feel that at the moment it's still targeted at a very narrow audience, and I'm going to recommend some fixes to make it accessible to non-specialists. I'll write up my recommendations on the manuscript sometime later this week. After this small revision, I fully expect to be able to check off all items on the checklist and endorse the paper for publication.

@thowell I *may* have accidentally selected or unselected a couple of the checkmarks in your section :-/
