### 2020-12-05 OpenMDAO/dymos#471 @goerz
Thanks for the update! I was able to install and run `testflow`. It does run the testsuite, but warns about some missing (optional) packages `mpi4py` `petsc4py`. Most importantly, though, it can't find `pyOptsparse`, which causes numerous test failures. I would recommend adding `pyoptsparse` to the the 'test' extra-requirments in `setup.py` (as well as any other packages that should be installed for running the tests, to the extent that they are missing).

I would recommend uploading `pyoptsparse` to PyPI, see https://github.com/mdolab/pyoptsparse/issues/191, although in I think in principle it is also possible to put a Github address or link to a tarball in `setup.py`.

In the newly added "Running Tests" section of the documentation, you may want to list the other testing dependencies and that `testflo` and these other dependencies can also be installed with `pip install .[test]` in a `dymos` checkout directory.
