### 2020-12-05 OpenMDAO/dymos#471 @JustinSGray
We can add pyoptsparse to the test-extra requirements, but that won't fundamentally fix all the failing tests. We run some of our tests against a commercial optimizer SNOPT. These test are critical for our internal use, since we rely heavily on SNOPT.
You can certainly run Dymos without SNOPT, and we've made sure to also test many freely available optimizers. I don't want to remove the SNOPT tests but we can make sure to skip them rather than fail the tests when it is not present.

I noted your PR suggest on the pyoptsparse repo. We do collaborate with the pyoptsparse developers, but it is not our project. I think PyPI release of that package is outside the scope for this publication. Also, I will note that pyoptsparse itself was[ recently accepted to JOSS](https://joss.theoj.org/papers/10.21105/joss.02564) without having published on PyPI.
Regardless, we can absolutely make sure the git repo is listed as a target for the `test` requirements.
