### 2020-12-05 OpenMDAO/dymos#471 @goerz
Yeah, I can successfully run the test suite, and I wouldn’t put a “no failing tests” criterion on anything, so this is not going to be a blocker for JOSS.

I would recommend keeping this on your radar, though.
