+++ dymos/options.py
@@ -2,5 +2,5 @@ from openmdao.api import OptionsDictionary
 
 options = OptionsDictionary()
 
-options.declare('include_check_partials', default=False, types=bool,
+options.declare('include_check_partials', default=True, types=bool,
                 desc='If True, include dymos components when checking partials.')
