### 2020-12-07 2da3d46
timeseries plotting now skips variable paths that are not present in a given phase
