### 2020-12-09 de7cf3e
updated reference to martins and hwang 2018 in documentation
