+++ mkdocs/references.bib
@@ -69,11 +69,23 @@ year = {2018}
 }
 
 @article {hwang2018b,
-	title = {A computational architecture for coupling heterogeneous numerical models and computing coupled derivatives},
-	journal = {ACM TOMS},
-	year = {2018},
-	note = {<p>(In press)</p>},
-	author = {John T. Hwang and Joaquim R. R. A. Martins}
+author = {Hwang, John T. and Martins, Joaquim R.R.A.},
+title = {A Computational Architecture for Coupling Heterogeneous Numerical Models and Computing Coupled Derivatives},
+year = {2018},
+issue_date = {August 2018},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {44},
+number = {4},
+issn = {0098-3500},
+url = {https://doi.org/10.1145/3182393},
+doi = {10.1145/3182393},
+abstract = {One of the challenges in computational modeling is coupling models to solve multidisciplinary problems. Flow-based computational frameworks alleviate part of the challenge through a modular approach, where data flows from component to component. However, existing flow-based frameworks are inefficient when coupled derivatives are needed for optimization. To address this, we develop the modular analysis and unified derivatives (MAUD) architecture. MAUD formulates the multidisciplinary model as a nonlinear system of equations, which leads to a linear equation that unifies all methods for computing derivatives. This enables flow-based frameworks that use the MAUD architecture to provide a common interface for the chain rule, adjoint method, coupled adjoint method, and hybrid methods; MAUD automatically uses the appropriate method for the problem. A hierarchical, matrix-free approach enables modern solution techniques such as Newton--Krylov solvers to be used within this monolithic formulation without computational overhead. Two demonstration problems are solved using a Python implementation of MAUD: a nanosatellite optimization with more than 2 million unknowns and 25,000 design variables, and an aircraft optimization involving over 6,000 design variables and 23,000 constraints. MAUD is now implemented in the open source framework OpenMDAO, which has been used to solve aircraft, satellite, wind turbine, and turbofan engine design problems.},
+journal = {ACM Trans. Math. Softw.},
+month = jun,
+articleno = {37},
+numpages = {39},
+keywords = {engineering design, adjoint methods, high-performance computing, Parallel computing, complex systems, Python, multiphysics simulation, PDE-constrained optimization, optimization, multidisciplinary design optimization, derivative computation}
 }
 
 @article{Martins2013,
