### 2020-12-09 f884b50
Added Scitech 2019 paper reference to the JOSS paper
