+++ mkdocs/references.bib
@@ -37,9 +37,7 @@ year = {2016}
 abstract = {The method of collocation and nonlinear programming has been used recently to solve a number of optimal control problems. In this method polynomials are commonly used to represent the state variable time histories over subintervals of the total time of interest. These polynomials correspond to a family of modified-Gaussian quadrature rules known as the Gauss-Lobatto rules. Presently, relatively low-order rules from the Gauss-Lobatto family, such as the trapezoid and Simpson`s rule, are used to construct collocation solution schemes. In this work higher-order Gauss-Lobatto quadrature rules are formulated using collocation point selection based on a particular family of Jacobi polynomials. The advantage of using a quadrature rule of higher order is that the approximation using the higher degree polynomial may be more accurate, due to finite precision arithmetic, than a formulation based on a lower degree polynomial. In addition, the number of subintervals and, therefore, the number of nonlinear programming parameters needed to solve a problem accurately may be significantly reduced from that required if the conventional trapezoidal or Simpson`s quadrature schemes are used. An optimal trajectory maximizing final energy for a low-thrust spacecraft is used to demonstrate the benefits of using the higher-order schemes. (28 References).},
 author = {Herman, A L and Conway, B A},
 doi = {10.2514/3.21662},
-file = {:Users/rfalck/Projects/pointer.git/references/Herman Conway 1996.pdf:pdf},
 isbn = {0731-5090},
-issn = {0731-5090},
 journal = {{\{}AIAA{\}} Journal of Guidance, Control, and Dynamics},
 number = {3},
 pages = {522--529},
