### 2020-12-10 e9d742d
starting to fix some testing errors now that things are fundamentally working
