### 2020-12-10 16e2e2d
now correctly negating vector fix_initial/fix_final specification when determining indices of state_inputs that are design variables.
