+++ dymos/transcriptions/pseudospectral/components/test/test_collocation_defect_sol_opt.py
@@ -26,6 +26,59 @@ class TestCollocationCompSolOpt(unittest.TestCase):
 
     def tearDown(self):
         dm.options['include_check_partials'] = False
+        
+    def _make_state_idx_map(self, state_name, options, grid_data, state_idx_map):
+            """
+            Provides error checking for solve_segments and establishes necessary data structures.
+
+            Parameters
+            ----------
+            state_name : str
+                The name of the state being configured.
+            options : StateOptionsDictionary
+                The StateOptionsDictionary for the state being configured.
+            grid_data : GridData
+                The GridData object which holds the grid information for this test.
+            state_idx_map : dict
+                The state_idx_map to be populated by this method.  This map is equivalent to the
+                one generated by PseudospectralBase._configure_solve_segments.
+            """
+            if state_idx_map is None:
+                state_idx_map = {}
+            state_idx_map[state_name] = {'solver': None, 'indep': None}
+
+            state_input_idxs = grid_data.subset_node_indices['state_input']
+            num_state_input_nodes = grid_data.subset_num_nodes['state_input']
+            compressed = options.get('compressed', False)
+
+            # Sanity-checks for solve segments
+            # If solve_segments is used at all, we cannot fix the state at both ends of the phase.
+            if options['solve_segments']:
+                # Forward propagation
+                if options['solve_segments'] in {True, 'forward'}:
+                    if compressed:
+                        state_idx_map[state_name]['solver'] = np.arange(1, num_state_input_nodes, dtype=int)
+                        state_idx_map[state_name]['indep'] = np.zeros((1,), dtype=int)
+                    else:
+                        left_idxs = grid_data.subset_node_indices['segment_ends'][0::2]
+                        state_idx_map[state_name]['solver'] = [i for i in range(num_state_input_nodes) if state_input_idxs[i] not in left_idxs]
+                        state_idx_map[state_name]['indep'] = [i for i in range(num_state_input_nodes) if state_input_idxs[i] in left_idxs]
+
+                # Backward propagation
+                elif options['solve_segments'] in {'backward'}:
+                    if compressed:
+                        # The optimizer controls the last state input node, all others are solver-controlled
+                        state_idx_map[state_name]['indep'] = np.array([num_state_input_nodes - 1], dtype=int)
+                        state_idx_map[state_name]['solver'] = np.arange(num_state_input_nodes - 1, dtype=int)
+                    else:
+                        # The optimizer controls the last state input node in each segment, all others are solver_controlled
+                        right_idxs = grid_data.subset_node_indices['segment_ends'][1::2]
+                        state_idx_map[state_name]['indep'] = [i for i in range(num_state_input_nodes) if state_input_idxs[i] in right_idxs]
+                        state_idx_map[state_name]['solver'] = [i for i in range(num_state_input_nodes) if state_input_idxs[i] not in right_idxs]
+            else:
+                # No solver used to solve these nodes.  All state input nodes are the indep nodes.
+                state_idx_map[state_name]['solver'] = []
+                state_idx_map[state_name]['indep'] = np.arange(len(state_input_idxs), dtype=int)
 
     def make_prob(self, transcription, n_segs, order, compressed):
 
@@ -75,6 +128,11 @@ class TestCollocationCompSolOpt(unittest.TestCase):
         p.model.connect('f_computed:v', 'defect_comp.f_computed:v')
         p.model.connect('dt_dstau', 'defect_comp.dt_dstau')
         p.model.connect('defect_comp.defects:v', 'state_indep.defects:v')
+
+        self.state_idx_map = {}
+        for state_name, options in state_options.items():
+            self._make_state_idx_map(state_name, options, gd, self.state_idx_map)
+        p.model.state_indep.configure_io(self.state_idx_map)
 
         p.setup(force_alloc_complex=True)
 
