+++ dymos/transcriptions/pseudospectral/components/state_independents.py
@@ -27,11 +27,22 @@ class StateIndependentsComp(om.ImplicitComponent):
             'state_options', types=dict,
             desc='Dictionary of state names/options for the phase')
 
-    def configure_io(self, state_idx_map):
+    def configure_io(self, state_idx_map=None):
         """
         I/O creation is delayed until configure so that we can determine the shape and units for
         the states.
+
+        Parameters
+        ----------
+        state_idx_map : dict
+            The state_idx_map generated by the transcription during its configuration.  This dict
+            is keyed by state, which in turn holds a dict of the solve indices in the input state
+            vector (keyed as 'solved') and those indices managed by the optimizer (keyed as 'indep').
+            In some unittest, this map is not provided at the time of setup and is instead called
+            later.
         """
+        if state_idx_map is None:
+            return
         self.state_idx_map = state_idx_map
         state_options = self.options['state_options']
         grid_data = self.options['grid_data']
@@ -63,7 +74,7 @@ class StateIndependentsComp(om.ImplicitComponent):
                 input_name = f'initial_states:{state_name}',
                 self.add_input(name=input_name, shape=(1, ) + shape, units=units)
 
-            # compute an output contraint value since the optimizer needs it
+            # compute an output constraint value since the optimizer needs it
             if solved:
                 self.add_input(
                     name=var_names['defect'],
@@ -134,16 +145,16 @@ class StateIndependentsComp(om.ImplicitComponent):
         state_options = self.options['state_options']
 
         for state_name, options in state_options.items():
-
             state_var_name = f'states:{state_name}'
 
             solve_idx = self.state_idx_map[state_name]['solver']
             indep_idx = self.state_idx_map[state_name]['indep']
 
             var_names = self.var_names[state_name]
-            defect = inputs[var_names['defect']]
 
-            residuals[state_var_name][solve_idx, ...] = defect
+            if options['solve_segments']:
+                defect = inputs[var_names['defect']]
+                residuals[state_var_name][solve_idx, ...] = defect
 
             # really is: <idep_val> - \outputs[state_name][indep_idx] but OpenMDAO
             # implementation details mean we just set it to 0
