+++ dymos/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.16.2-dev'
+__version__ = '0.17.0'
 
 from .phase import Phase
 from .transcriptions import GaussLobatto, Radau, RungeKutta
