### 2020-12-11 0c60c3f
down to two failing tests when connect=True is used
