+++ dymos/transcriptions/pseudospectral/components/test/test_collocation_defect_solver.py
@@ -4,19 +4,22 @@ import numpy as np
 from numpy.testing import assert_almost_equal
 
 import openmdao.api as om
+from openmdao.utils.testing_utils import use_tempdirs
 
 from dymos.transcriptions.grid_data import GridData
-from dymos.transcriptions.pseudospectral.components.collocation_comp import CollocationComp
 from dymos.transcriptions.pseudospectral.components.state_independents import StateIndependentsComp
 
 # Modify class so we can run it standalone.
 import dymos as dm
-from dymos.utils.misc import CompWrapperConfig
-CollocationComp = CompWrapperConfig(CollocationComp)
-StateIndependentsComp = CompWrapperConfig(StateIndependentsComp)
+from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
+from dymos.examples.brachistochrone.test.ex_brachistochrone import brachistochrone_min_time as brach
 
 
+@use_tempdirs
 class TestCollocationBalanceIndex(unittest.TestCase):
+    """
+    Test that the indices used in the StateIndependentsComp are as expected.
+    """
 
     def setUp(self):
         dm.options['include_check_partials'] = True
@@ -41,164 +44,325 @@ class TestCollocationBalanceIndex(unittest.TestCase):
         subsys = StateIndependentsComp(grid_data=gd, state_options=state_options)
         p.model.add_subsystem('defect_comp', subsys=subsys)
 
+        self.state_idx_map = {}
+        for state_name, options in state_options.items():
+            self._make_state_idx_map(state_name, options, gd, self.state_idx_map)
+        subsys.configure_io(self.state_idx_map)
+
         p.setup()
         p.final_setup()
 
         return p
 
     def test_3_lgl(self):
+        """
+        Test one 3rd order LGL segment indices
 
-        p = self.make_prob(transcription='gauss-lobatto', n_segs=3, order=3, compressed=False)
-        defect_comp = p.model.defect_comp
-        self.assertSetEqual(set(defect_comp.state_idx_map['x']['solver']), set([1, 3, 5]))
-        self.assertSetEqual(set(defect_comp.state_idx_map['x']['indep']), set([0, 2, 4]))
+        All nodes are:   [0, 1, 2]
+        Input nodes:     [^     ^]
+        Solver nodes:    [      ^]
+        Indep nodes:     [^      ]
+        Of input nodes, solver nodes for fix_initial are: {1}
+        The indep node is just the first one: {0}
+        """
+        p = brach(transcription='gauss-lobatto', num_segments=1, transcription_order=3,
+                  compressed=True, solve_segments=True, run_driver=True)
 
-        self.assertSetEqual(set(defect_comp.state_idx_map['v']['solver']), set([0, 1, 3]))
-        self.assertSetEqual(set(defect_comp.state_idx_map['v']['indep']), set([2, 4, 5]))
+        state_indeps_comp = p.model.traj0.phases.phase0.indep_states
+
+        self.assertSetEqual(set(state_indeps_comp.state_idx_map['x']['solver']), {1})
+        self.assertSetEqual(set(state_indeps_comp.state_idx_map['x']['indep']), {0})
 
     def test_5_lgl(self):
+        """
+        Test one 5th order LGL segment indices
 
-        p = self.make_prob(transcription='gauss-lobatto', n_segs=2, order=5, compressed=False)
-        defect_comp = p.model.defect_comp
+        All nodes are:   [0, 1, 2, 3, 4]
+        Input nodes:     [^     ^     ^]
+        Solver nodes:    [      ^     ^]
[…]
