### 2020-12-11 0b6c648
included new tests for the 0.17.0 whats new section
