+++ dymos/test/test_upgrade_guide.py
@@ -480,3 +480,173 @@ class TestUpgrade_0_16_0(unittest.TestCase):
                             ('CL', None), ('CD', None), ('q', 'N/m**2'), ('f_lift', 'lbf'),
                             ('f_drag', 'N'), ('thrust', 'lbf')]:
             self.assertEqual(op_dict[f'traj.phase0.timeseries.{name}'], units)
+
+@use_tempdirs
+class TestUpgrade_0_17_0(unittest.TestCase):
+
+    def test_tags(self):
+        """
+        # upgrade_doc: begin tag_rate_source
+        self.add_output('xdot', val=np.zeros(nn), desc='velocity component in x', units='m/s')
+        self.add_output('ydot', val=np.zeros(nn), desc='velocity component in y', units='m/s')
+        self.add_output('vdot', val=np.zeros(nn), desc='acceleration magnitude', units='m/s**2')
+        # upgrade_doc: end tag_rate_source
+        # upgrade_doc: begin declare_rate_source
+        phase.add_state('x', rate_source='xdot', fix_initial=True, fix_final=True)
+        phase.add_state('y', rate_source='ydot', fix_initial=True, fix_final=True)
+        phase.add_state('v', rate_source='vdot', fix_initial=True, fix_final=False)
+        # upgrade_doc: end declare_rate_source
+        """
+        import numpy as np
+        import openmdao.api as om
+
+        class BrachistochroneODE(om.ExplicitComponent):
+
+            def initialize(self):
+                self.options.declare('num_nodes', types=int)
+                self.options.declare('static_gravity', types=(bool,), default=False,
+                                     desc='If True, treat gravity as a static (scalar) input, rather than '
+                                          'having different values at each node.')
+
+            def setup(self):
+                nn = self.options['num_nodes']
+                g_default_val = 9.80665 if self.options['static_gravity'] else 9.80665 * np.ones(nn)
+
+                # Inputs
+                self.add_input('v', val=np.zeros(nn), desc='velocity', units='m/s')
+
+                self.add_input('g', val=g_default_val, desc='grav. acceleration', units='m/s/s')
+
+                self.add_input('theta', val=np.ones(nn), desc='angle of wire', units='rad')
+
+                # upgrade_doc: begin tag_rate_source
+                self.add_output('xdot', val=np.zeros(nn), desc='velocity component in x', units='m/s',
+                                tags=['state_rate_source:x', 'state_units:m'])
+
+                self.add_output('ydot', val=np.zeros(nn), desc='velocity component in y', units='m/s',
+                                tags=['state_rate_source:y', 'state_units:m'])
+
+                self.add_output('vdot', val=np.zeros(nn), desc='acceleration magnitude', units='m/s**2',
+                                tags=['state_rate_source:v', 'state_units:m/s'])
+                # upgrade_doc: end tag_rate_source
+
+                self.add_output('check', val=np.zeros(nn), desc='check solution: v/sin(theta) = constant',
+                                units='m/s')
+
+                # Setup partials
+                arange = np.arange(self.options['num_nodes'])
+                self.declare_partials(of='vdot', wrt='theta', rows=arange, cols=arange)
+
+                self.declare_partials(of='xdot', wrt='v', rows=arange, cols=arange)
+                self.declare_partials(of='xdot', wrt='theta', rows=arange, cols=arange)
+
+                self.declare_partials(of='ydot', wrt='v', rows=arange, cols=arange)
+                self.declare_partials(of='ydot', wrt='theta', rows=arange, cols=arange)
+
+                self.declare_partials(of='check', wrt='v', rows=arange, cols=arange)
[…]
