### 2020-12-11 6ff1bfc
Fix for #488, and updated the basic brachistochrone example.
