+++ dymos/examples/brachistochrone/doc/test_doc_brachistochrone.py
@@ -80,14 +80,11 @@ class TestBrachistochroneForDocs(unittest.TestCase):
         #
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
-        phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                        fix_initial=True, fix_final=True, solve_segments=False)
+        phase.add_state('x', fix_initial=True, fix_final=True)
 
-        phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                        fix_initial=True, fix_final=True, solve_segments=False)
+        phase.add_state('y', fix_initial=True, fix_final=True)
 
-        phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                        fix_initial=True, fix_final=False, solve_segments=False)
+        phase.add_state('v', fix_initial=True, fix_final=False)
 
         phase.add_control('theta', continuity=True, rate_continuity=True,
                           units='deg', lower=0.01, upper=179.9)
@@ -165,17 +162,11 @@ class TestBrachistochroneForDocs(unittest.TestCase):
         #
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
-        phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                        units=BrachistochroneODE.states['x']['units'],
-                        fix_initial=True, fix_final=True, solve_segments=False)
+        phase.add_state('x', fix_initial=True, fix_final=True)
 
-        phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                        units=BrachistochroneODE.states['y']['units'],
-                        fix_initial=True, fix_final=True, solve_segments=False)
+        phase.add_state('y', fix_initial=True, fix_final=True)
 
-        phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                        units=BrachistochroneODE.states['v']['units'],
-                        fix_initial=True, fix_final=False, solve_segments=False)
+        phase.add_state('v', fix_initial=True, fix_final=False)
 
         phase.add_control('theta', continuity=True, rate_continuity=True,
                           units='deg', lower=0.01, upper=179.9)
@@ -255,17 +246,11 @@ class TestBrachistochroneForDocs(unittest.TestCase):
         #
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
-        phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                        units=BrachistochroneODE.states['x']['units'],
-                        fix_initial=True, fix_final=False)
+        phase.add_state('x', fix_initial=True, fix_final=False)
 
-        phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                        units=BrachistochroneODE.states['y']['units'],
-                        fix_initial=True, fix_final=False)
+        phase.add_state('y', fix_initial=True, fix_final=False)
 
-        phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                        units=BrachistochroneODE.states['v']['units'],
-                        fix_initial=True, fix_final=False)
+        phase.add_state('v', fix_initial=True, fix_final=False)
 
         phase.add_control('theta', continuity=True, rate_continuity=True,
                           units='deg', lower=0.01, upper=179.9)
