### 2020-12-11 bc5d1dd
fixed a missing parameter definition in add_linkage_constraint
