### 2020-12-11 24fc9d4
removed outdated class variables in the example BrachistochroneODE now that tagging is possible.
