+++ dymos/utils/test/brachistochrone_for_command_line.py
@@ -37,15 +37,9 @@ def brachistochrone_min_time(transcription='gauss-lobatto', num_segments=8, tran
 
     phase.set_time_options(fix_initial=True, duration_bounds=(.5, 10))
 
-    phase.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                    units=BrachistochroneODE.states['x']['units'],
-                    fix_initial=True, fix_final=False, solve_segments=False)
-    phase.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                    units=BrachistochroneODE.states['y']['units'],
-                    fix_initial=True, fix_final=False, solve_segments=False)
-    phase.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                    units=BrachistochroneODE.states['v']['units'],
-                    fix_initial=True, fix_final=False, solve_segments=False)
+    phase.add_state('x', fix_initial=True, fix_final=False)
+    phase.add_state('y', fix_initial=True, fix_final=False)
+    phase.add_state('v', fix_initial=True, fix_final=False)
 
     phase.add_control('theta', continuity=True, rate_continuity=True,
                       units='deg', lower=0.01, upper=179.9)
