+++ dymos/examples/brachistochrone/doc/test_doc_brachistochrone_tandem_phases.py
@@ -80,17 +80,11 @@ class TestBrachistochroneTandemPhases(unittest.TestCase):
 
         phase0.set_time_options(fix_initial=True, duration_bounds=(.5, 10))
 
-        phase0.add_state('x', rate_source=BrachistochroneODE.states['x']['rate_source'],
-                         units=BrachistochroneODE.states['x']['units'],
-                         fix_initial=True, fix_final=False, solve_segments=False)
+        phase0.add_state('x', fix_initial=True, fix_final=False)
 
-        phase0.add_state('y', rate_source=BrachistochroneODE.states['y']['rate_source'],
-                         units=BrachistochroneODE.states['y']['units'],
-                         fix_initial=True, fix_final=False, solve_segments=False)
+        phase0.add_state('y', fix_initial=True, fix_final=False)
 
-        phase0.add_state('v', rate_source=BrachistochroneODE.states['v']['rate_source'],
-                         units=BrachistochroneODE.states['v']['units'],
-                         fix_initial=True, fix_final=False, solve_segments=False)
+        phase0.add_state('v', fix_initial=True, fix_final=False)
 
         phase0.add_control('theta', continuity=True, rate_continuity=True,
                            units='deg', lower=0.01, upper=179.9)

[…]
