+++ dymos/utils/doc_utils.py
@@ -1,5 +1,4 @@
 import inspect
-import os
 import pathlib
 import sys
 
