### 2020-12-12 f250fda
Added method signature to API documentation.  Moved release date of 0.17.0.
