### 2020-12-14 d604290
pseudospectral base now picks the correct indices for the state design variables when connected_initial is True
