### 2020-12-14 d5c04d4
Added release notes and bumped version to 0.17.0 (#487)
* release notes added for 0.17.0

* 0.16.2-dev → 0.17.0

* added 0.17.0 features to the whats new section of docs

* included new tests for the 0.17.0 whats new section

* Fix for #488, and updated the basic brachistochrone example.

* fixed a missing parameter definition in add_linkage_constraint

* removed outdated class variables in the example BrachistochroneODE now that tagging is possible.

* Added method signature to API documentation.  Moved release date of 0.17.0.
