+++ dymos/trajectory/trajectory.py
@@ -550,7 +550,7 @@ class Trajectory(om.Group):
                     shapes[i], units[i] = get_source_metadata(phases[i]._get_subsystem(rhs_source),
                                                               vars[i], user_units=units[i],
                                                               user_shape=_unspecified)
-                except ValueError as e:
+                except ValueError:
                     raise ValueError(f'{info_str}: Unable to find variable \'{vars[i]}\' in '
                                      f'phase \'{phases[i].pathname}\' or its ODE.')
 
@@ -737,6 +737,7 @@ class Trajectory(om.Group):
         phase_a : str
             The first phase in the linkage constraint.
         phase_b : str
+            The second phase in the linkage constraint.
         var_a : str
             The linked variable from the first phase in the linkage constraint.
         var_b : str
