+++ dymos/examples/brachistochrone/test/test_path_constraints.py
@@ -19,12 +19,9 @@ class TestBrachistochronePathConstraints(unittest.TestCase):
 
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
-        phase.add_state('x', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['x']['rate_source'])
-        phase.add_state('y', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['y']['rate_source'])
-        phase.add_state('v', fix_initial=True,
-                        rate_source=BrachistochroneODE.states['v']['rate_source'])
+        phase.add_state('x', fix_initial=True, fix_final=True)
+        phase.add_state('y', fix_initial=True, fix_final=True)
+        phase.add_state('v', fix_initial=True)
 
         phase.add_control('theta', units='deg',
                           rate_continuity=False, lower=0.01, upper=179.9)
@@ -71,12 +68,9 @@ class TestBrachistochronePathConstraints(unittest.TestCase):
 
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
-        phase.add_state('x', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['x']['rate_source'])
-        phase.add_state('y', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['y']['rate_source'])
-        phase.add_state('v', fix_initial=True,
-                        rate_source=BrachistochroneODE.states['v']['rate_source'])
+        phase.add_state('x', fix_initial=True, fix_final=True)
+        phase.add_state('y', fix_initial=True, fix_final=True)
+        phase.add_state('v', fix_initial=True)
 
         phase.add_control('theta', units='deg',
                           rate_continuity=False, lower=0.01, upper=179.9)
@@ -124,12 +118,9 @@ class TestBrachistochronePathConstraints(unittest.TestCase):
 
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
-        phase.add_state('x', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['x']['rate_source'])
-        phase.add_state('y', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['y']['rate_source'])
-        phase.add_state('v', fix_initial=True,
-                        rate_source=BrachistochroneODE.states['v']['rate_source'])
+        phase.add_state('x', fix_initial=True, fix_final=True)
+        phase.add_state('y', fix_initial=True, fix_final=True)
+        phase.add_state('v', fix_initial=True)
 
         phase.add_control('theta', units='deg',
                           rate_continuity=False, lower=0.01, upper=179.9)
@@ -176,12 +167,9 @@ class TestBrachistochronePathConstraints(unittest.TestCase):
 
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
 
-        phase.add_state('x', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['x']['rate_source'])
-        phase.add_state('y', fix_initial=True, fix_final=True,
-                        rate_source=BrachistochroneODE.states['y']['rate_source'])
-        phase.add_state('v', fix_initial=True,
-                        rate_source=BrachistochroneODE.states['v']['rate_source'])
+        phase.add_state('x', fix_initial=True, fix_final=True)
+        phase.add_state('y', fix_initial=True, fix_final=True)
+        phase.add_state('v', fix_initial=True)
 
         phase.add_control('theta', units='deg',
                           rate_continuity=False, lower=0.01, upper=179.9)
