### 2020-12-14 b57594f
cleaning up solve_segments examples to use forward or backward instead of True or False.
