+++ dymos/examples/battery_multibranch/test/test_multibranch_trajectory_rk_ivp.py
@@ -36,8 +36,8 @@ class TestBatteryRKIVP(unittest.TestCase):
         transcription = dm.Radau(num_segments=5, order=5, compressed=True)
         phase1 = dm.Phase(ode_class=BatteryODE, transcription=transcription)
         phase1.set_time_options(fix_initial=False, fix_duration=True)
-        phase1.add_state('state_of_charge', fix_initial=False, fix_final=False, solve_segments=True,
-                         targets=['SOC'], rate_source='dXdt:SOC')
+        phase1.add_state('state_of_charge', fix_initial=False, fix_final=False,
+                         solve_segments='forward', targets=['SOC'], rate_source='dXdt:SOC')
         phase1.add_timeseries_output('battery.V_oc', output_name='V_oc', units='V')
         phase1.add_timeseries_output('battery.V_pack', output_name='V_pack', units='V')
         phase1.add_timeseries_output('pwr_balance.I_Li', output_name='I_Li', units='A')
@@ -48,8 +48,8 @@ class TestBatteryRKIVP(unittest.TestCase):
         phase1_bfail = dm.Phase(ode_class=BatteryODE, ode_init_kwargs={'num_battery': 2},
                                 transcription=transcription)
         phase1_bfail.set_time_options(fix_initial=False, fix_duration=True)
-        phase1_bfail.add_state('state_of_charge', fix_initial=False, fix_final=False, solve_segments=True,
-                               targets=['SOC'], rate_source='dXdt:SOC')
+        phase1_bfail.add_state('state_of_charge', fix_initial=False, fix_final=False,
+                               solve_segments='forward', targets=['SOC'], rate_source='dXdt:SOC')
         phase1_bfail.add_timeseries_output('battery.V_oc', output_name='V_oc', units='V')
         phase1_bfail.add_timeseries_output('battery.V_pack', output_name='V_pack', units='V')
         phase1_bfail.add_timeseries_output('pwr_balance.I_Li', output_name='I_Li', units='A')
@@ -60,8 +60,8 @@ class TestBatteryRKIVP(unittest.TestCase):
         phase1_mfail = dm.Phase(ode_class=BatteryODE, ode_init_kwargs={'num_motor': 2},
                                 transcription=transcription)
         phase1_mfail.set_time_options(fix_initial=False, fix_duration=True)
-        phase1_mfail.add_state('state_of_charge', fix_initial=False, fix_final=False, solve_segments=True,
-                               targets=['SOC'], rate_source='dXdt:SOC')
+        phase1_mfail.add_state('state_of_charge', fix_initial=False, fix_final=False,
+                               solve_segments='forward', targets=['SOC'], rate_source='dXdt:SOC')
         phase1_mfail.add_timeseries_output('battery.V_oc', output_name='V_oc')
         phase1_mfail.add_timeseries_output('battery.V_pack', output_name='V_pack')
         phase1_mfail.add_timeseries_output('pwr_balance.I_Li', output_name='I_Li')
@@ -180,8 +180,7 @@ class TestBatteryRKIVP(unittest.TestCase):
         phase1 = dm.Phase(ode_class=BatteryODE, transcription=transcription)
         phase1.set_time_options(fix_initial=False, fix_duration=True)
         phase1.add_state('state_of_charge', fix_initial=False, fix_final=False,
-                         solve_segments=True,
-                         targets=['SOC'], rate_source='dXdt:SOC')
+                         solve_segments='forward', targets=['SOC'], rate_source='dXdt:SOC')
         phase1.add_timeseries_output('battery.V_oc', output_name='V_oc')
         phase1.add_timeseries_output('battery.V_pack', output_name='V_pack')
         phase1.add_timeseries_output('pwr_balance.I_Li', output_name='I_Li')
[…]
