+++ dymos/transcriptions/pseudospectral/pseudospectral_base.py
@@ -3,6 +3,7 @@ from collections.abc import Iterable
 import numpy as np
 
 import openmdao.api as om
+from openmdao.utils.general_utils import warn_deprecation
 from ..transcription_base import TranscriptionBase
 from ..common import TimeComp, PseudospectralTimeseriesOutputComp
 from .components import StateIndependentsComp, StateInterpComp, CollocationComp
@@ -237,6 +238,14 @@ class PseudospectralBase(TranscriptionBase):
         # Transcription solve_segments overrides state solve_segments if its not set
         if options['solve_segments'] is None:
             options['solve_segments'] = self.options['solve_segments']
+
+        # Flag deprecated solve_segments options:
+        if options['solve_segments'] in {True, False}:
+            ss = options['solve_segments']
+            warn_deprecation(f'State {state_name} in phase {phase.name} has option '
+                             f'\'solve_segments={ss}\'.  Option \'True\' now gives forward '
+                             f'propagation. In Dymos 1.0 and later, only options \'forward\' and'
+                             f'\'backward\' will be valid.')
 
         # Sanity-checks for solve segments
         # If solve_segments is used at all, we cannot fix the state at both ends of the phase.
