+++ setup.py
@@ -28,7 +28,7 @@ optional_dependencies['all'] = sorted([
 
 
 setup(name='dymos',
-    version='0.17.0',
+    version='0.17.1-dev',
     description='Open-Source Optimization of Dynamic Multidisciplinary Systems',
     long_description='''
 Dymos is a framework for the simulation and optimization of dynamical systems within the OpenMDAO Multidisciplinary Analysis and Optimization environment.
