### 2020-12-14 3a3cebb
fixed some errant SNOPT usages
