+++ dymos/examples/brachistochrone/doc/test_doc_brachistochrone_upstream_state.py
@@ -88,8 +88,6 @@ class TestBrachistochroneUpstreamState(unittest.TestCase):
 
         # Set the driver.
         p.driver = om.ScipyOptimizeDriver()
-        p.driver = om.pyOptSparseDriver(optimizer='SNOPT')
-        p.driver.opt_settings['iSumm'] = 6
 
         # Allow OpenMDAO to automatically determine our sparsity pattern.
         # Doing so can significant speed up the execution of Dymos.
