### 2020-12-14 1023069
cleanup of deprecation warning when solve_segments=True
