### 2020-12-15 e8d2743
solve_segments=False was incorrectly triggering a deprecation warning
