### 2020-12-15 49555df
changed print_level for IPOPT since logs are getting too long for CI
