### 2020-12-16 2f49db2
Fixes solve_segments to allow it to be used when neither fix_initial nor fix_final are True (#490)
* moved solve_segments validation to _configure_solve_segments in PseudospectralBase from StateIndependentsComp.  Valid values for solve_segments are now True, False, None, "forward", and "backward".

* view_connections_bug

* reworking things to allow solve_segments="backward"

* checkpoint

* StateIndependentsComp update

* starting to fix some testing errors now that things are fundamentally working

* now correctly negating vector fix_initial/fix_final specification when determining indices o
[…]
