+++ dymos/examples/vanderpol/vanderpol_dymos.py
@@ -19,7 +19,7 @@ def vanderpol(transcription='gauss-lobatto', num_segments=8, transcription_order
         if optimizer == 'SNOPT':
             p.driver.opt_settings['iSumm'] = 6  # show detailed SNOPT output
         elif optimizer == 'IPOPT':
-            p.driver.opt_settings['print_level'] = 5
+            p.driver.opt_settings['print_level'] = 4
     p.driver.declare_coloring()
 
     # define a Trajectory object and add to model
