+++ dymos/trajectory/test/test_trajectory_input_parameters.py
@@ -181,7 +181,7 @@ def two_burn_orbit_raise_problem(transcription='gauss-lobatto', optimizer='SLSQP
     elif optimizer == 'IPOPT':
         p.driver.opt_settings['hessian_approximation'] = 'limited-memory'
         p.driver.opt_settings['nlp_scaling_method'] = 'gradient-based'
-        p.driver.opt_settings['print_level'] = 5
+        p.driver.opt_settings['print_level'] = 4
         p.driver.opt_settings['linear_solver'] = 'mumps'
         p.driver.opt_settings['mu_strategy'] = 'adaptive'
         # p.driver.opt_settings['derivative_test'] = 'first-order'
