### 2020-12-16 f23f30a
Fixed issue with vectorized scaling
