### 2020-12-16 cb78249
fixed issue with trying to divide by array valued variable
