### 2020-12-16 4fdc178
removed snopt from committed tests
