### 2020-12-17 #2809 @thowell
I've basically finished my review. Two items:

- The example in the paper doesn't seem to highlight the novelties of the package. The paper could be improved with a slightly more substantial example that clearly demonstrates where existing packages are lacking/fail. For example, plan motion for a system with 2 phases and a parameter that needs to be optimized.

- I checked the performance box because no claims are made.  However, I was curious about how fast the package is compared to other trajectory optimization offerings. It looks like there is automatic sparsity detection, does this make the package substantially faster than existing tools? Does the more general utility in the package result in slow solve times? The paper and documentation could be improved with some discussion of performance.
