### 2020-12-18 189ba68
Added vectorized scaling brachistochrone problem unit tests
