### 2020-12-18 8f5ec8a
Vectorized scaling (#491)
* moved solve_segments validation to _configure_solve_segments in PseudospectralBase from StateIndependentsComp.  Valid values for solve_segments are now True, False, None, "forward", and "backward".

* view_connections_bug

* reworking things to allow solve_segments="backward"

* checkpoint

* StateIndependentsComp update

* starting to fix some testing errors now that things are fundamentally working

* now correctly negating vector fix_initial/fix_final specification when determining indices of state_inputs that are design variables.

* continuing to get tests to p
[…]
