+++ dymos/transcriptions/pseudospectral/components/collocation_comp.py
@@ -83,14 +83,21 @@ class CollocationComp(om.ExplicitComponent):
             if 'defect_ref' in options and options['defect_ref'] is not None:
                 defect_ref = options['defect_ref']
             elif 'defect_scaler' in options and options['defect_scaler'] is not None:
-                defect_ref = 1.0 / options['defect_scaler']
+                defect_ref = np.divide(1.0, options['defect_scaler'])
             else:
                 if 'ref' in options and options['ref'] is not None:
                     defect_ref = options['ref']
                 elif 'scaler' in options and options['scaler'] is not None:
-                    defect_ref = 1.0 / options['scaler']
+                    defect_ref = np.divide(1.0, options['scaler'])
                 else:
                     defect_ref = 1.0
+
+            if not np.isscalar(defect_ref):
+                defect_ref = np.asarray(defect_ref)
+                if defect_ref.shape == shape:
+                    defect_ref = np.tile(defect_ref.flatten(), num_col_nodes)
+                else:
+                    raise ValueError('array-valued scaler/ref must length equal to state-size')
 
             if not options['solve_segments']:
                 self.add_constraint(name=var_names['defect'],
