### 2020-12-18 49a5517
First step of adding bokeh-based timeseries plots is working, but legend needs work.
