+++ dymos/examples/finite_burn_orbit_raise/test/test_ex_two_burn_orbit_raise_bokeh_plots.py
@@ -1,0 +1,309 @@
+import unittest
+
+import numpy as np
+
+import openmdao.api as om
+from openmdao.utils.assert_utils import assert_near_equal
+from openmdao.utils.general_utils import set_pyoptsparse_opt
+_, optimizer = set_pyoptsparse_opt('IPOPT', fallback=True)
+
+import dymos as dm
+from dymos.examples.finite_burn_orbit_raise.finite_burn_eom import FiniteBurnODE
+from openmdao.utils.testing_utils import use_tempdirs
+
+dm.options['plots'] = 'bokeh'
+
+
+def make_traj(transcription='gauss-lobatto', transcription_order=3, compressed=False,
+              connected=False):
+
+    t = {'gauss-lobatto': dm.GaussLobatto(num_segments=5, order=transcription_order, compressed=compressed),
+         'radau': dm.Radau(num_segments=20, order=transcription_order, compressed=compressed),
+         'runge-kutta': dm.RungeKutta(num_segments=5, compressed=compressed)}
+
+    traj = dm.Trajectory()
+
+    traj.add_parameter('c', opt=False, val=1.5, units='DU/TU',
+                       targets={'burn1': ['c'], 'burn2': ['c']})
+
+    # First Phase (burn)
+
+    burn1 = dm.Phase(ode_class=FiniteBurnODE, transcription=t[transcription])
+
+    burn1 = traj.add_phase('burn1', burn1)
+
+    burn1.set_time_options(fix_initial=True, duration_bounds=(.5, 10), units='TU')
+    burn1.add_state('r', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                    rate_source='r_dot', units='DU')
+    burn1.add_state('theta', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                    rate_source='theta_dot', units='rad')
+    burn1.add_state('vr', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                    rate_source='vr_dot', units='DU/TU')
+    burn1.add_state('vt', fix_initial=True, fix_final=False, defect_scaler=100.0,
+                    rate_source='vt_dot', units='DU/TU')
+    burn1.add_state('accel', fix_initial=True, fix_final=False,
+                    rate_source='at_dot', units='DU/TU**2')
+    burn1.add_state('deltav', fix_initial=True, fix_final=False,
+                    rate_source='deltav_dot', units='DU/TU')
+    burn1.add_control('u1', rate_continuity=True, rate2_continuity=True, units='deg', scaler=0.01,
+                      rate_continuity_scaler=0.001, rate2_continuity_scaler=0.001,
+                      lower=-30, upper=30)
+    # Second Phase (Coast)
+    coast = dm.Phase(ode_class=FiniteBurnODE, transcription=t[transcription])
+
+    coast.set_time_options(initial_bounds=(0.5, 20), duration_bounds=(.5, 50), duration_ref=50, units='TU')
+    coast.add_state('r', fix_initial=False, fix_final=False, defect_scaler=100.0,
+                    rate_source='r_dot', units='DU')
+    coast.add_state('theta', fix_initial=False, fix_final=False, defect_scaler=100.0,
+                    rate_source='theta_dot', units='rad')
+    coast.add_state('vr', fix_initial=False, fix_final=False, defect_scaler=100.0,
+                    rate_source='vr_dot', units='DU/TU')
+    coast.add_state('vt', fix_initial=False, fix_final=False, defect_scaler=100.0,
+                    rate_source='vt_dot', units='DU/TU')
+    coast.add_state('accel', fix_initial=True, fix_final=True,
+                    rate_source='at_dot', units='DU/TU**2')
[…]
