+++ dymos/visualization/timeseries_plots.py
@@ -133,11 +133,11 @@ def _mpl_timeseries_plots(varnames, time_units, var_units, phase_names, phases_n
 
 
 def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases_node_path,
-                            last_solution_case, last_simulation_case, plot_dir_path, num_cols=4,
-                            page_width=1080):
+                            last_solution_case, last_simulation_case, plot_dir_path, num_cols=2,
+                            page_width=1080, bg_fill_color='#282828'):
     from bokeh.io import output_file, show
-    from bokeh.layouts import gridplot, column
-    from bokeh.models import Legend, Circle, Line, Square, ColumnDataSource, Label
+    from bokeh.layouts import gridplot, column, row, grid, layout
+    from bokeh.models import Legend, LegendItem
     from bokeh.plotting import figure
     import bokeh.palettes as bp
 
@@ -148,6 +148,7 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
     # Prune the edges from the color map
     cmap = bp.turbo(len(phase_names) + 2)[1:-1]
     figures = []
+    colors = {}
 
     # Get the minimum and maximum times in any phase, so when we plot a variable that only exists
     # in a few phases, it is plotted against the entire time range.
@@ -163,7 +164,7 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
             time_name = f'{phase_name}.timeseries.time'
         min_time = min(min_time, np.min(last_solution_case.outputs[time_name]))
         max_time = max(max_time, np.max(last_solution_case.outputs[time_name]))
-    #     color = cmap[iphase]
+        colors[phase_name] = cmap[iphase]
     #     legend_fig = figure(title='', x_range=(0, 50), y_range=(-25, 25))
     #     legend_fig.axis.visible = False
     #     legend_fig.xgrid.visible = False
@@ -185,8 +186,9 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
 
         # add labels, title, and legend
         padding = 0.05 * (max_time - min_time)
-        fig = figure(title=title, background_fill_color="#282828",
-                     x_range=(min_time - padding, max_time + padding))
+        fig = figure(title=title, background_fill_color=bg_fill_color,
+                     x_range=(min_time - padding, max_time + padding), plot_width=180,
+                     plot_height=180)
         fig.xaxis.axis_label = time_label
         fig.yaxis.axis_label = var_label
         fig.xgrid.grid_line_color = '#666666'
@@ -210,8 +212,6 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
 
             var_val = last_solution_case.outputs[var_name_full]
             time_val = last_solution_case.outputs[time_name]
-            mean_time = np.mean(time_val)
-            mean_val = np.mean(var_val)
 
             # Plot the data
             # color = cm.colors[iphase % 20]
@@ -219,7 +219,7 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
             for idxs, i in np.ndenumerate(np.zeros(var_val.shape[1:])):
                 var_val_i = var_val[:, idxs].ravel()
                 sol_plots[phase_name] = fig.circle(time_val.ravel(), var_val_i, size=5,
-                                                   color=sol_color)
+                                                   color=sol_color, name='sol:' + phase_name)
 
             # get simulation values, if plotting simulation
             if last_simulation_case:
@@ -231,7 +231,8 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
[…]
