+++ dymos/visualization/timeseries_plots.py
@@ -134,14 +134,12 @@ def _mpl_timeseries_plots(varnames, time_units, var_units, phase_names, phases_n
 
 def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases_node_path,
                             last_solution_case, last_simulation_case, plot_dir_path, num_cols=2,
-                            page_width=1080, bg_fill_color='#282828'):
+                            bg_fill_color='#282828', grid_line_color='#666666'):
     from bokeh.io import output_file, show
     from bokeh.layouts import gridplot, column, row, grid, layout
     from bokeh.models import Legend, LegendItem
     from bokeh.plotting import figure
     import bokeh.palettes as bp
-
-    # Create a figure of annotations that will serve as the legend
 
     output_file(os.path.join(plot_dir_path, 'plots.html')) # f'{var_name.replace(":", "_")}.png')
 
@@ -149,13 +147,13 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
     cmap = bp.turbo(len(phase_names) + 2)[1:-1]
     figures = []
     colors = {}
+    sol_plots = {}
+    sim_plots = {}
 
     # Get the minimum and maximum times in any phase, so when we plot a variable that only exists
     # in a few phases, it is plotted against the entire time range.
     min_time = 1.0E21
     max_time = -1.0E21
-
-    legend_figs = []
 
     for iphase, phase_name in enumerate(phase_names):
         if phases_node_path:
@@ -165,20 +163,8 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
         min_time = min(min_time, np.min(last_solution_case.outputs[time_name]))
         max_time = max(max_time, np.max(last_solution_case.outputs[time_name]))
         colors[phase_name] = cmap[iphase]
-    #     legend_fig = figure(title='', x_range=(0, 50), y_range=(-25, 25))
-    #     legend_fig.axis.visible = False
-    #     legend_fig.xgrid.visible = False
-    #     legend_fig.ygrid.visible = False
-    #     legend_fig.square([10], [0], size=20, color=color)
-    #     legend_figs.append(legend_fig)
-    #
-    # legend_grid = gridplot(legend_figs, ncols=5, plot_width=100, plot_height=100)
-    sol_plots = {}
-    sim_plots = {}
 
     for ivar, var_name in enumerate(varnames):
-        # start a new plot
-
         # Get the labels
         time_label = f'time ({time_units[var_name]})'
         var_label = f'{var_name} ({var_units[var_name]})'
@@ -191,8 +177,8 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
                      plot_height=180)
         fig.xaxis.axis_label = time_label
         fig.yaxis.axis_label = var_label
-        fig.xgrid.grid_line_color = '#666666'
-        fig.ygrid.grid_line_color = '#666666'
+        fig.xgrid.grid_line_color = grid_line_color
+        fig.ygrid.grid_line_color = grid_line_color
 
         # Plot each phase
         for iphase, phase_name in enumerate(phase_names):
@@ -213,9 +199,6 @@ def _bokeh_timeseries_plots(varnames, time_units, var_units, phase_names, phases
             var_val = last_solution_case.outputs[var_name_full]
             time_val = last_solution_case.outputs[time_name]
 
-            # Plot the data
-            # color = cm.colors[iphase % 20]
-
             for idxs, i in np.ndenumerate(np.zeros(var_val.shape[1:])):
                 var_val_i = var_val[:, idxs].ravel()
                 sol_plots[phase_name] = fig.circle(time_val.ravel(), var_val_i, size=5,
[…]
