### 2020-12-19 52888f6
automatic noteboko plots attempt
