### 2020-12-21 cc7b106
Added vectorized brachistochrone problem to test_run_problem
