### 2020-12-21 56c6128
replaced abs with fabs to avoid issue with vectorized states
