### 2020-12-21 680b6dd
Modified refinement outputs and test problem to be faster
