### 2020-12-23 #2809 @goerz
I have checked the software package and its online documentation, and they appear ready for publication. The following are my comments on the manuscript itself. Again, sorry for the delay in writing this up...

The manuscript seems like it should be OK for someone within the narrow field of classical optimal control, e.g., users of some of the existing packages cited in the manuscript. However, I think the Summary could be improved to expand the scope towards a more "diverse, non-specialist audience". I myself seem like a good "test" audience, in fact: My background is in *quantum* optimal control theory, where I routinely optimize processes to drive a complex-valued initial state-vector toward some desired target vector, subject to constraints. There is a well-established set of methods and software tools to solve these control problems, without much overlap to the classical control community. However, at least in principle, it would seem that there *could* be more overlap between the communities, as I could rewrite my complex state vectors to real-valued vectors by separating real and imaginary parts. At least to me personally it is an open question whether and to what extent there are possible improvements or insights to be gained from classical control theory software. I would read a paper like the Dymos JOSS manuscript with this question in my; How would I have to reformulate my quantum control problems so that it would fit into Dymos' (or any similar package's) language.

First, I would recommend that the manuscript uses some (minimal) mathematical notation to be more specific without requiring the reader to be familiar with domain-specific jargon. The mathematical definitions in https://openmdao.github.io/dymos/getting_started/optimal_control.html#the-overall-optimization-problem are something I can understand, and I would recommend putting them in the manuscript, at least in some shortened form. This would then also allow to specify Dymos' unique "co-design" feature:

> Dymos [...] is unique in that it can also support broader cases where trajectory optimization is just one component in a larger physical model.

I don't understand what that means: How does the "larger physical model" not fit into the equations describing the control objectives and system dynamics? Are these not numerically accessible? Can you give a specific easy to understand example (like from the cited aircraft design paper): What is the "pure" optimal control problem here, and what is the larger physical model?

Is the presence of the DAE's (second paragraph) what you mean by the "larger physical model"? Again, a specific example or mathematical definition might help (what's an algebraic equation again?) Or, is the "co-design" just the presence of different phases (second section), with enforced boundary conditions between the phases?

> One application of this approach is the method of differential inclusions

Dymos does "differential inclusion" right? (just using a "modular" approach, instead of "monolithic algorithmic differentiation") It's not 100% clear if the "Instead" that starts the third paragraph refers to the entire second paragraph ("differential inclusion") or just the "monolithic" part. If my reading is correct, then the paragraph break should before the sentence "Some optimal control libraries tackle ..."

You might want to rewrite this part slightly, in any case:

> While effective, the monolithic nature is both less efficient(Kenway et al., 2019; Mader et al., 2008) and less flexible.

Less than what? Less than Dymos' approach? I would probably turn this around: After introducing "differential inclusions" in paragraph two, introduce what Dymos does in paragraph three, and end with something like "The modular approach is both more efficient and more flexible than tackling the numerical challenge with a monolithic algorithmic differentiation ...". Has the "modular approach" been introduced before (citation?). What are the citations from Kenway and Mader comparing?

> The available time discretization schemes in Dymos are based on pseudospectral methods, using two common direct collocation transcriptions: the high-order Gauss-Lobatto transcription (Herman & Conway, 1996) and the Radau pseudospectral method (Garg et al., 2009). Both implicit and explicit forms of these transcriptions are supported. The explicit forms can be used to build single or multiple-shooting style problem formulations. The implicit forms can be used to construct two point boundary value problems. Transcriptions are built to be totally independent of the ODE implementation, and nearly transparent to the user. This means that switching transcriptions requires very minor code changes - typically a single line in the run-script.

This paragraph has *a lot* of domain-specific jargon: collocation, transcriptions, two-point boundary problem, shooting style problems (is that what I would call "gradient-free"?). This is fine for a reader who is familiar with all of these, but it might be worth the effort to add a couple of sentences with high-level explanations.

As for the example, I have to agree with @thowell. It's a relatively straightforward optimization problem, and illustrates the basic usage of Dymos nicely, but it doesn't really show Dymos' unique features. Specifically, it doesn't give me an example for the fundamental "co-design" feature (my initial question)

### Small Typos

> in order to solve ODE

in order to solve *the* ODE(s)

> fearures

features (please do run a spell-checker!)

### Summary

I think the manuscript could use substantial work if the authors want to reach a wide audience (which JOSS would seem to encourage), while for a narrow audience, it seems mostly fine in its current form. I will leave it up to the authors to decide to what extent they want to revise based on my comments above. After revision, I would endorse the paper for publication.

I have left the related boxes in the checklist open for now, but will be happy to check them on revision.

Happy Holidays!
