### 2020-12-29 5445eec
Vector states refinement (#492)
* Added vectorized brachistochrone problem to test_run_problem

* replaced abs with fabs to avoid issue with vectorized states

* Modified refinement outputs and test problem to be faster

Co-authored-by: Kaushik Ponnapalli <kaushik.s.ponnapalli@nasa.gov>
