### 2021-01-04 d4b18db
more math formatting
