+++ joss/paper.md
@@ -37,10 +37,12 @@ This flexibility allows Dymos to efficiently solve optimal control problems cons
 Dymos can also help solve more general optimization problems where dynamics are only one part in a larger system level model with additional --- potentially computationally expensive --- calculations that come before and after the dynamic calculations.  
 These broader problems are commonly referred to co-design, controls-co-design, and multidisciplinary design optimization.
 Dymos provides specific APIs and features that make it possible to integrate traditional optimal-control models into a co-design context, while still supporting analytic derivatives that are necessary for computational efficiency in these complex use cases. 
+An example of a co-design problem that was solved with Dymos is the coupled trajectory-thermal design of an electric vertical takeoff and landing aircraft where the thermal management and propulsion systems were designed simultaneously with the flight trajectories to ensure no components overheated[@Hariton2020a].
 
 
 # Difference between optimal-control and co-design
-In the most general sense, optimal-control problems are numerically valid ways of transcribing a dynamic control problem into a nonlinear programming problem like this: 
+In the most general sense, both optimal-control and co-design problems are both are numerically valid ways of optimizations problems involving ordinary differential equations (ODE) or differential algebraic equations (DAE).  
+A general problem formulation will look like this:  
 
 \begin{align*} 
 \mathrm{Minimize}& \qquad \mathrm{J} = f_{obj}(\bar{x},t,\bar{u},\bar{d}) \\
@@ -51,26 +53,42 @@ In the most general sense, optimal-control problems are numerically valid ways o
 \mathrm{State , Variables:}& \qquad \bar{x}{lb} \leq \bar{x} \leq \bar{x}{ub} \\
 \mathrm{Dynamic , Controls:}& \qquad \bar{u}{lb} \leq \bar{u} \leq \bar{u}{ub} \\ 
 \mathrm{Design , Parameters:}& \qquad \bar{d}{lb} \leq \bar{d} \leq \bar{d}{ub} \\ 
-\mathrm{Initial , Boundary , Constraints:}& \qquad \bar{g}{0,lb} \leq g{0}(\bar{x}0,t_0,\bar{u}0, \bar{d}) \leq \bar{g}{0,ub} \\
+\mathrm{Initial , Boundary , Constraints:}& \qquad \bar{g}{0,lb} \leq g_{0}(\bar{x}0,t_0,\bar{u}0, \bar{d}) \leq \bar{g}{0,ub} \\
 \mathrm{Final , Boundary , Constraints:}& \qquad \bar{g}{f,lb} \leq g_{f}(\bar{x}f,t_f,\bar{u}f, \bar{d}) \leq \bar{g}{f,ub} \\ 
 <!-- \mathrm{Path , Constraints:}& \qquad \bar{p}{f,lb} \leq p_{f}(\bar{x},t,\bar{u},\bar{d}) \leq \bar{p}_{f,ub} 
  -->
  \end{align*}
 
-
-Optimization problems that combine physical design and dynamic control are often referred to as 
-
-We can use an electric aircraft example to clarify the difference between traditional optimal control and broader co-design problems.
-One of the primary challenges for electric aircraft is keeping their battery and electric motors cool. 
-Given a pre-designed aircraft, a traditional optimal control problem would seek to fly a fixed distance most efficiently, without overheating anything, by adjusting the schedule of throttle setting and flight path angle. These are purely time-varying controls akin to what the pilot would change in flight, and hence don't change the physical design of the aircraft at all. 
-A co-design version of this problem would allow the aircraft design itself to change (e.g. the wing span, radiator size, propeller diameter), along with the trajectory, in order to find a better overall approach. 
-It's possible that a slightly larger radiator would provide more cooling, allowing the aircraft to fly at a faster but more efficient speed and hence use less power to go the same distance. 
[…]
