+++ joss/paper.md
@@ -82,7 +82,7 @@ This kind of iterative sequential optimization look like this:
 Dymos can support sequential co-design, but its unique value is that it also enables a more tightly coupled co-design process with a single top level optimizer handling both parts of the problem simultaneously. 
 Coupled co-design is particularly challenging because it requires propagating derivative information from the static analysis to the dynamic analysis in an efficient way. 
 
-![optimal control diagram](images/coupled_co_design.png){width=100%}
+![optimal control diagram](images/coupled_co_design.png){width=75%}
 
 
 # ODE versus DAE
@@ -95,14 +95,15 @@ Since the derivatives are needed to perform optimization, DAEs are more challeng
 
 One relatively common use case for DAEs is differential inclusions, in which the state time-history is posed as a dynamic control and the traditional control schedule needed to achieve that time-history is found using a nonlinear solver within the dynamic model [@Seywald1994].
 For some problems differential inclusion provides a more natural and numerical beneficial design space for the optimizer to traverse,
-but the nonlinear solver causes difficulties when computing derivatives, 
-especially when finite-differences are used to compute derivatives for the optimizer.
-Some optimal control libraries tackle this numerical challenge with a monolithic algorithmic differentiation[@griewank2003mathematical] approach.
-While effective, the monolithic nature is both less efficient[@mader2008adjoint; @kenway2019effective] and less flexible.
+but the nonlinear solver poses numerical challenges for computing derivatives for the optimizer.
+A simple approach to this is to just finite-difference across the nonlinear solver, but this is has been shown to be expensive and numerically unstable[@gray2014derivatives]. 
+Another option, taken by some optimal control libraries, is to apply monolithic algorithmic differentiation[@griewank2003mathematical] across the nonlinear solver.
+While it does provide accurate derivatives, the monolithic approach is expensive and uses a lot of memory[@mader2008adjoint; @kenway2019effective]. 
+The most efficient approach is to use a pair of analytic derivative approachs called the direct and adjoint methods, which were unified into a single unified derivative equation (UDE) by Hwang and Martins[@hwang2018b]. 
 
-Instead, Dymos adopts the analytic derivative approach[@hwang2018b] which uses a linear solver to compute total derivatives needed by the optimizer using only partial derivatives of the residual equations in the DAE.
+Dymos adopts the UDE approach which uses a linear solver to compute total derivatives needed by the optimizer using only partial derivatives of the residual equations in the DAE.
 This approach offers two key advantages. 
-First partial derivatives of the DAE residual equations are much less computationally expensive to compute. 
+First partial derivatives of the DAE residual equations are much less computationally challenging to compute. 
 Second, using the OpenMDAO underpinnings of Dymos users can construct their DAE in a modular fashion and combine various methods of computing the partial derivatives via finite-difference, complex-step [@Martins2003CS], algorithmic differentiation, or hand differentiation as needed. 
 
 
@@ -140,33 +141,33 @@ On more challenging optimal-control problems, higher quality optimizers are impo
 
 ## Statement of Need
 
-When dealing with the design of complex systems, there are two approaches: collaborative optimization or coupled co-design [@allison2004complex].
+When dealing with the design of complex systems that include transient behavior, co-design becomes critical[@garciasans2019].
+Broadly there are two approaches: sequential co-design or coupled co-design [@allison2004complex].
 The best choice depends on the degree of interaction, or coupling, between various sub-systems. 
-In the case where you have a system where performance is heavily influenced by its transient behavior, then the dynamic behavior becomes a critical subsystem. 
-If the coupling is strong in this case, a co-design (a.k.a. controls-co-design; a.k.a. multidisciplinary design optimization) approach is necessary to achieve the best performance [@garciasans2019].
[…]
