### 2021-01-04 ed22358
more math typos
