+++ joss/paper.bib
@@ -1,3 +1,15 @@
+@article{Martins2013,
+  author = {Martins, Joaquim R. R. A. and Lambe, Andrew B.},
+  doi = {10.2514/1.J051895},
+  journal = {AIAA Journal},
+  month = sep,
+  number = {9},
+  pages = {2049--2075},
+  title = {Multidisciplinary Design Optimization: A Survey of Architectures},
+  volume = {51},
+  year = {2013}
+}
+
 @inproceedings{gray2014derivatives,
 Author = {Justin S. Gray and Tristan A. Hearn and Kenneth T. Moore and John Hwang and Joaquim Martins and Andrew Ning},
 Booktitle = "{15th AIAA/ISSMO Multidisciplinary Analysis and Optimization Conference}",
