+++ joss/paper.md
@@ -41,13 +41,27 @@ An example of a co-design problem that was solved with Dymos is the coupled traj
 
 
 # Difference between optimal-control and co-design
-In the most general sense, both optimal-control and co-design problems are both are numerically valid ways of optimizing dynamic systems. 
+In the most general sense, both optimal-control and co-design problems deal with dynamic systems. 
 There a vector of time varying state variables ($\bar{x}$) who's behavior is affected by time ($t$), a vector of dynamic controls ($\bar{u)}$), and a vector of static design parameters ($\bar{d}$). 
 The evolution of the states over time is governed by an ordinary differential equation (ODE) or differential algebraic equation (DAE):
 \begin{align}
   \dot{\bar{x}} = f_{ode}(\bar{x},t,\bar{u},\bar{d})
 \end{align}
 
+To optimize a dynamic system we also need to account for the objective function ($J$)
+\begin{align}
+  \mathrm{J} = f_{obj}(\bar{x},t,\bar{u},\bar{d})
+\end{align}
+In addition, there are constraints that typically need to be enforced. 
+\begin{align}
+  \mathrm{Time:}& \qquad {t}_{lb} \leq t \leq {t}_{ub} \\
+  \mathrm{State \, Variables:}& \qquad \bar{x}_{lb} \leq \bar{x} \leq \bar{x}_{ub} \\
+  \mathrm{Dynamic \, Controls:}& \qquad \bar{u}_{lb} \leq \bar{u} \leq \bar{u}_{ub} \\
+  \mathrm{Design \, Parameters:}& \qquad \bar{d}_{lb} \leq \bar{d} \leq \bar{d}_{ub} \\
+  \mathrm{Initial \, Boundary \, Constraints:}& \qquad \bar{g}_{0,lb} \leq g_{0}(\bar{x}_0,t_0,\bar{u}_0, \bar{d}) \leq \bar{g}_{0,ub} \\
+  \mathrm{Final \, Boundary \, Constraints:}& \qquad \bar{g}_{f,lb} \leq g_{f}(\bar{x}_f,t_f,\bar{u}_f, \bar{d}) \leq \bar{g}_{f,ub} \\
+  \mathrm{Path \, Constraints:}& \qquad \bar{p}_{f,lb} \leq p_{f}(\bar{x},t,\bar{u},\bar{d}) \leq \bar{p}_{f,ub} \\
+\end{align}
 
 A general problem formulation will look like this:  
 
