+++ joss/images/opt_control.py
@@ -1,18 +1,18 @@
 from pyxdsm.XDSM import XDSM, OPT, SUBOPT, SOLVER, DOE, IFUNC, FUNC, GROUP, IGROUP, METAMODEL
 
 
-x = XDSM()
+x = XDSM(use_sfmath=False)
 
 x.add_system('OPT', OPT, r"\text{Optimizer}")
 x.add_system('ODE', GROUP, r"\text{ODE or DAE}")
 
-x.connect('ODE', 'OPT', ["J", "g_0", "g_f"],  label_width=3)
+x.connect('ODE', 'OPT', ["J", r"\bar{g}_0", r"\bar{g}_f", r"\bar{p}"],  label_width=4)
 x.connect('OPT', 'ODE', ["t", r"\bar{x}", r"\bar{u}", r"\bar{d}"],  label_width=4)
 
 x.write('opt_control')
 
 
-x = XDSM()
+x = XDSM(use_sfmath=False)
 
 x.add_system('OPT_static', OPT, r"\text{Static Optimizer}")
 x.add_system('static', GROUP, r"\text{Static System Model}")
@@ -20,7 +20,7 @@ x.add_system('static', GROUP, r"\text{Static System Model}")
 x.add_system('OPT_dynamic', OPT, r"\text{Dynamic Optimizer}")
 x.add_system('dynamic', GROUP, r"\text{ODE or DAE}")
 
-x.connect('dynamic', 'OPT_dynamic', [r"J_\text{dynamic}", "g_0", "g_f"],  label_width=3)
+x.connect('dynamic', 'OPT_dynamic', [r"J_\text{dynamic}", r"\bar{g}_0", r"\bar{g}_f", r"\bar{p}"],  label_width=4)
 x.connect('OPT_dynamic', 'dynamic', ["t", r"\bar{x}", r"\bar{u}"],  label_width=3)
 
 x.connect('static', 'OPT_static', [r"J_\text{static}", r"g_\text{static}"],  label_width=3)
@@ -32,13 +32,13 @@ x.connect('OPT_dynamic', 'OPT_static', '')
 x.write('sequential_co_design')
 
 
-x = XDSM()
+x = XDSM(use_sfmath=False)
 
 x.add_system('OPT', OPT, r"\text{Optimizer}")
 x.add_system('static', GROUP, r"\text{Static System Model}")
 x.add_system('dynamic', GROUP, r"\text{ODE or DAE}")
 
-x.connect('dynamic', 'OPT', [r"J", "g_0", "g_f"],  label_width=3)
+x.connect('dynamic', 'OPT', [r"J", r"\bar{g}_0", r"\bar{g}_f", r"\bar{p}"],  label_width=4)
 x.connect('OPT', 'dynamic', ["t", r"\bar{x}", r"\bar{u}"],  label_width=3)
 
 x.connect('static', 'OPT', [r"g_\text{static}"],  label_width=3)
