### 2021-01-06 cce6d7d
balanced field length example is getting a reasonable answer but not converging well.
