+++ dymos/examples/balanced_field/doc/test_doc_balanced_field_length.py
@@ -29,10 +29,10 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
         p.driver = om.pyOptSparseDriver()
         p.driver.options['optimizer'] = 'SNOPT'
         p.driver.opt_settings['iSumm'] = 6
-        p.driver.opt_settings['Verify level'] = 3
-        p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-6
-        p.driver.opt_settings['Major optimality tolerance'] = 1.0E-6
-        p.driver.opt_settings['Major iterations limit'] = 50
+        # p.driver.opt_settings['Verify level'] = 3
+        p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-5
+        p.driver.opt_settings['Major optimality tolerance'] = 1.0E-5
+        p.driver.opt_settings['Major iterations limit'] = 500
         p.driver.opt_settings['Minor iterations limit'] = 100000
         p.driver.declare_coloring()
 
@@ -135,7 +135,7 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
         p3.add_boundary_constraint('v', loc='final', equals=0, ref=100, linear=True)
 
         # Minimize range at the end of the phase
-        # p3.add_objective('r', loc='final', ref=100.0)
+        p3.add_objective('r', loc='final', ref=1.0)
 
 
         # Fourth Phase - Rotate for single engine takeoff
@@ -168,7 +168,7 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
 
         p4.add_timeseries_output('*')
 
-        p4.add_boundary_constraint('F_r', loc='final', equals=0, ref=1000)
+        p4.add_boundary_constraint('F_r', loc='final', equals=0, ref=100000)
         # p4.add_boundary_constraint('alpha', loc='final', equals=10, units='deg')
 
         # p4.add_path_constraint('alpha_rate', lower=0, upper=3, units='deg/s')
@@ -186,7 +186,7 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
         #
         # Set the options on the optimization variables
         #
-        p5.set_time_options(fix_initial=False, duration_bounds=(0.1, 100), duration_ref=1.0)
+        p5.set_time_options(fix_initial=False, duration_bounds=(1, 100), duration_ref=1.0)
 
         p5.add_state('r', fix_initial=False, lower=0, ref=1000.0, defect_ref=1000.0,
                      rate_source='r_dot')
@@ -203,14 +203,17 @@ class TestBalancedFieldLengthForDocs(unittest.TestCase):
         p5.add_parameter('T', val=27000, opt=False, units='lbf')
         p5.add_parameter('m', val=174200, opt=False, units='lbm')
 
-        p5.add_polynomial_control('alpha', order=1, opt=True, units='deg', lower=0, upper=15, ref=10)
+        p5.add_control('alpha', opt=True, units='deg', lower=-10, upper=15, ref=10)
 
         p5.add_timeseries_output('*')
 
         p5.add_boundary_constraint('h', loc='final', equals=35, ref=35)
+        p5.add_boundary_constraint('gam', loc='final', equals=5, ref=5, units='deg')
+        p5.add_path_constraint('gam', lower=0, upper=5, ref=5, units='deg')
+
         p5.add_boundary_constraint('v_over_v_stall', loc='final', equals=1.2, ref=1.2)
 
-        p5.add_objective('r', loc='final', ref=1000.0)
+        # p5.add_objective('time', loc='final', ref=100.0)
 
         p.model.linear_solver = om.DirectSolver()
         
[…]
