### 2021-01-06 4d76305
balanced field length example is getting a reasonable answer but not converging well.
