+++ joss/paper.md
@@ -62,7 +62,7 @@ In addition, there are constraints that typically need to be enforced.
   \mathrm{Design \, Parameters:}& \qquad \bar{d}_{lb} \leq \bar{d} \leq \bar{d}_{ub} \\
   \mathrm{Initial \, Boundary \, Constraints:}& \qquad \bar{g}_{0,lb} \leq g_{0}(\bar{x}_0,t_0,\bar{u}_0, \bar{d}) \leq \bar{g}_{0,ub} \\
   \mathrm{Final \, Boundary \, Constraints:}& \qquad \bar{g}_{f,lb} \leq g_{f}(\bar{x}_f,t_f,\bar{u}_f, \bar{d}) \leq \bar{g}_{f,ub} \\
-  \mathrm{Path \, Constraints:}& \qquad \bar{p}_{f,lb} \leq p_{f}(\bar{x},t,\bar{u},\bar{d}) \leq \bar{p}_{f,ub} \\
+  \mathrm{Path \, Constraints:}& \qquad \bar{p}_{lb} \leq p(\bar{x},t,\bar{u},\bar{d}) \leq \bar{p}_{ub} \\
 \end{align*}
 
 
@@ -157,7 +157,7 @@ Here, an explicit phases is one where the full time history is computed starting
 From the optimizers perspective it will set values for the design parameters ($\bar{d}$) and the controls ($\bar{u}$) and can expect to be given a physically valid time history as the output.
 Wrapping an optimizer around an explicit phase gives what is traditionally called a "shooting method" in the optimal-control world.  
 In contrast, implicit phases don't provide valid time histories on their own. 
-Instead, they add the state vector ($\bar{x}$) as an additional design variable to the optimizer and add an associated set of additional defect constraints that must be driven to 0 to enforce physics at the discrete time points of the time integration. 
+Instead, they add the entire time-history of the state vector ($\bar{x}$) as an additional design variable to the optimizer and add an associated set of defect constraints that must be driven to 0 to enforce physics at the discrete points of the time integration is evaluated at. 
 The net effect is that the full time history is only known once the optimization is fully converged. 
 In the context of the multidisciplinary design optimization field, explicit phases are similar to the multidisciplinary design feasible (MDF) optimization architecture and implicit phases are similar to the simultaneous analysis and design (SAND) optimization architecture[@Martins2013]. 
 
@@ -214,11 +214,13 @@ Dymos has been used to demonstrate the coupling of flight dynamics and subsystem
 NASA's X-57 "Maxwell" is using Dymos for mission planning to maximize data collection while abiding the limits of battery storage capacity and subsystem temperatures [@Schnulo2018a; @Schnulo2019a].
 Other authors have used Dymos to perform studies of aircraft acoustics [@Ingraham2020a] and the the design of supersonic aircraft with thermal fuel management systems [@Jasa2018a].
 
-## Optimal-control example use case for Dymos
+## Optimal-control example: Brachistochone
 
 As a simple use-case of Dymos, consider the classic brachistochrone optimal control problem.
-In this problem, we seek the shape of a frictionless wire strung between two points of different heights such that a bead sliding along the wire moves from the first point to the second point in minimum time.
-To find the solution in Dymos, we first define the ordinary differential equations that govern the motion of the bead.
+There is a bead sliding along a frictionless wire strung between two points of different heights, 
+and we seek the shape of the wire such that the bead travels from start to finish in the shortest time. 
+The time-varying control is the angle of the wire at each point in time and there are no design parameters, 
[…]
