### 2021-01-06 538758e
added code to perform boundary constraint introspection, still need path constraints
