+++ joss/paper.md
@@ -335,14 +335,14 @@ fig.savefig('brachistochone_yx.png', bbox_inches='tight')
 
 The build in plotting utility in Dymos will plot all relevant quantities vs time. 
 
-![Brachistochrone Solution: y state time history](brachistochrone_states_y.png)
+![Brachistochrone Solution: y state time history](brachistochrone_states_y.png){width=70%}
 
-![Brachistochrone Solution: x state time history](brachistochrone_states_x.png)
+![Brachistochrone Solution: x state time history](brachistochrone_states_x.png){width=70%}
 
 
 The more traditional way to view the brachistochrone solution is to view the actual shape of the wire (i.e. y vs x)
 
-![Brachistochrone Solution: y as a function of x](brachistochone_yx.png)
+![Brachistochrone Solution: y as a function of x](brachistochone_yx.png){width=70%}
 
 
 ## Coupled co-design example: Designing a cannonball
@@ -588,13 +588,13 @@ if __name__ == "__main__":
 The built in plotting in Dymos will give time histories of all the time varying quantities. 
 For example, these are the time histories for the range and height: 
 
-<!-- ![Cannonball Solution: height vs time](cannonball_states_h.png) -->
+![Cannonball Solution: height vs time](cannonball_states_h.png){width=70%}
 
-<!-- ![Cannonball Solution: range vs time](cannonball_states_r.png) -->
+![Cannonball Solution: range vs time](cannonball_states_r.png){width=70%}
 
 A more natural way to view the solution is to consider height vs range: 
 
-<!-- ![Cannonball Solution: height vs time](cannonball_hr.png) -->
+![Cannonball Solution: height vs time](cannonball_hr.png){width=70%}
 
 The parabolic trajectory is slightly skewed due to the effect of air resistance slowing down the cannonball so it is moving slower during the descent than the ascent. 
 
