### 2021-01-06 69b35d3
added option for bokeh plots.  matplotlib plots no longer change backend permanently.  scarring in place for use within notebooks.
