### 2021-01-06 6a112bf
add bokeh to the travis.yml.  Skip the bokeh test if not available.
