+++ .travis.yml
@@ -94,6 +94,7 @@ install:
     python -m pip install testflo==1.3.6;
     python -m pip install pyyaml;
     python -m pip install coveralls;
+    python -m pip install bokeh;
 
     python -m pip install mkdocs
     python -m pip install mkdocs-material
