### 2021-01-07 49b5c4a
Fixes issue where show_plots argument to run_problem is permanently changing matplotlib backend. (#497)
* First step of adding bokeh-based timeseries plots is working, but legend needs work.

* acceptable legend

* added option for bokeh plots.  matplotlib plots no longer change backend permanently.  scarring in place for use within notebooks.

* add bokeh to the travis.yml.  Skip the bokeh test if not available.

* coverage improvements
