### 2021-01-07 9aeb060
update to account for get_var_meta being renamed to _get_var_meta -> prefer get_io_metadata
