### 2021-01-07 0a062aa
Tweaks for JOSS Revision (#498)
* edits

* auto build action

* adjusting path

* getting workflow in right place

* new workflow yaml

* math formatting

* more math formatting

* math

* added some figures

* spacing

* ODE vs DAE

* tweaked statement of need

* forgot a ref

* math typos

* more math typos

* tweaked xdsm

* the dymos perspective

* formatting check

* more formatting

* updating the math section

* more math

* foo

* foo

* updated xdsm figures

* optimizer and dymos-way sections

* remove eqn numbers

* reformatting brach example

* trying font size changes

* trying fon
[…]
