+++ joss/paper.md
@@ -24,102 +24,206 @@ affiliations:
     index: 2
 date: 26 June 2020
 bibliography: paper.bib
+header-includes:
+    - \usepackage{setspace}
 ---
 
 # Summary
 
-Dymos is a library for optimizing control schedules for dynamic systems --- sometimes referred to as trajectory optimization.
+Dymos is a library for optimizing control schedules for dynamic systems --- sometimes referred to as  optimal control or trajectory optimization.
 There are a number of other optimal control libraries that tackle similar kinds of problems, such as OTIS4 [@Paris2006], GPOPS-II [@Patterson2014GPOPSII],and CASADI [@Andersson2018].
-Dymos can operate in a similar manner to these codes where it is focused purely on the optimal control problem, 
-but it is unique in that it can also support broader cases where trajectory optimization is just one component in a larger physical model. 
-This allows Dymos to play a critical role in solving problems where the physical design of a system is optimized simultaneously with control schedule governing its transient performance. 
-This kind of problem is often referred to as co-design, controls-co-design, or multidisciplinary design optimization.
-Clear examples of Dymos used in this context are demonstrated in the coupled trajectory-thermal design of electric aircraft [@Hariton2020a].
+These tools all rely on gradient based optimization to solve optimal control problems, though their methods of computing the gradients vary. 
+Dymos is built on top of the OpenMDAO framework[@Gray2019a] and supports its modular derivative system which allows users to mix-and-match from finite-differencing, complex-step, hand-differentiated, and algorithmic differentiation. 
+This flexibility allows Dymos to efficiently solve optimal control problems constructed with both ordinary differential equations (ODE) and differential algebraic equations (DAE). 
+
+Dymos can also help solve more general optimization problems where dynamics are only one part in a larger system level model with additional --- potentially computationally expensive --- calculations that come before and after the dynamic calculations.  
+These broader problems are commonly referred to co-design, controls-co-design, and multidisciplinary design optimization.
+Dymos provides specific APIs and features that make it possible to integrate traditional optimal-control models into a co-design context, while still supporting analytic derivatives that are necessary for computational efficiency in these complex use cases. 
+An example of a co-design problem that was solved with Dymos is the coupled trajectory-thermal design of an electric vertical takeoff and landing aircraft where the thermal management and propulsion systems were designed simultaneously with the flight trajectories to ensure no components overheated[@Hariton2020a].
+
+
+# Difference between optimal-control and co-design
+Optimal-control and co-design problems deal with dynamic systems. 
+There a vector of time varying state variables ($\bar{x}$) who's behavior is affected by time ($t$), a vector of dynamic controls ($\bar{u)}$), and a vector of static design parameters ($\bar{d}$). 
+The evolution of the states over time is governed by an ordinary differential equation (ODE) or differential algebraic equation (DAE):
+\begin{align*}
+  \dot{\bar{x}} = f_{ode}(\bar{x},t,\bar{u},\bar{d})
+\end{align*}
+
+To optimize a dynamic system we also need to account for the objective function ($J$)
+\begin{align*}
+  \mathrm{J} = f_{obj}(\bar{x},t,\bar{u},\bar{d})
+\end{align*}
+In addition, there are constraints that typically need to be enforced. 
+\begin{align*}
+  \mathrm{Time:}& \qquad {t}_{lb} \leq t \leq {t}_{ub} \\
[…]
