+++ joss/paper.bib
@@ -1,3 +1,27 @@
+@article{Martins2013,
+  author = {Martins, Joaquim R. R. A. and Lambe, Andrew B.},
+  doi = {10.2514/1.J051895},
+  journal = {AIAA Journal},
+  month = sep,
+  number = {9},
+  pages = {2049--2075},
+  title = {Multidisciplinary Design Optimization: A Survey of Architectures},
+  volume = {51},
+  year = {2013}
+}
+
+@inproceedings{gray2014derivatives,
+Author = {Justin S. Gray and Tristan A. Hearn and Kenneth T. Moore and John Hwang and Joaquim Martins and Andrew Ning},
+Booktitle = "{15th AIAA/ISSMO Multidisciplinary Analysis and Optimization Conference}",
+Doi = {doi:10.2514/6.2014-2042},
+Month = {2014/07/08},
+Publisher = {American Institute of Aeronautics and Astronautics},
+Title = "{Automatic Evaluation of Multidisciplinary Derivatives Using a Graph-Based Problem Formulation in OpenMDAO}",
+Url = {http://dx.doi.org/10.2514/6.2014-2042},
+Year = {2014},
+pdf = {http://openmdao.org/pubs/Gray_Hearn_Moore_EtAl-_2014_-Multidisciplinary_Derivatives.pdf}
+}
+
 @article{Gray2019a,
     Author = {Justin S. Gray and John T. Hwang
               and Joaquim R. R. A. Martins
