### 2021-01-07 #2809 @robfalck
@goerz

Thank you for your feedback.  My coauthors and I agree with your point about broadening the audience and have added more detail there.  We added a discussion and diagrams covering more use-cases for more multidisciplinary/co-design applications.  While you're right that optimal control often involves modeling the static system model as part of the ODE, in our experience there are applications where it's better to separate the static design and optimal control portions of the optimization.  We added our simple cannonball design example to the paper to demonstrate the separation of the static and dynamic design elements.  It's somewhat difficult to capture a real-world case in a short format such as this paper, since more realistic applications generally involve many more models and hundreds or thousands of lines of code.

We've also included the mathematical notation describing the optimal control problems that Dymos can solve.

Regarding approaching quantum applications, that is interesting and we do have other OpenMDAO-based tools for circuit design that implement complex variables as separated real variables.

We hope our other clarifications address some of your points.

@thowell

Hopefully the simple cannonball design example, while contrived, is a bit more demonstrative of the sort of multidisciplinary/co-design capability we're focusing on.

Regarding performance, there is nothing in this paper but we do provide some comparison against OTIS in our paper on the methods behind Dymos that we previously discussed (http://openmdao.org/pubs/falck_dymos_2019_scitech.pdf). OTIS is a Boeing/NASA-developed legacy optimal control tool.  OTIS is written in Fortran and takes advantage of sparsity, but uses finite differencing to obtain derivatives for the optimizer.  We compared against the supersonic minimum time-to-climb problem.  While OTIS is faster for small numbers of "nodes" in the trajectory (the number of points at which the continuous solution is discretized), Dymos scales better as the number of nodes is increased.  It's not surprising that Fortran would be faster than Python on small problems, but the more efficient derivative calculation methods used in Dymos overcome that as the problem size grows larger.

Thank you both for your reviews.
