### 2021-01-08 b0d6727
fixes a bug where Trajectory.simulate crashes in some circumstances when trajectory (input) parameters are used.
