+++ dymos/trajectory/test/test_trajectory_input_parameters.py
@@ -188,7 +188,7 @@ def two_burn_orbit_raise_problem(transcription='gauss-lobatto', optimizer='SLSQP
 
     traj = make_traj(transcription=transcription, transcription_order=transcription_order,
                      compressed=compressed, connected=connected, param_mode=param_mode)
-    p.model.add_subsystem('traj', subsys=traj)
+    p.model.add_subsystem('orbit_transfer', subsys=traj)
 
     # Finish Problem Setup
 
@@ -200,87 +200,75 @@ def two_burn_orbit_raise_problem(transcription='gauss-lobatto', optimizer='SLSQP
     p.setup(check=True)
 
     # Set Initial Guesses
-    p.set_val('traj.parameters:c', value=1.5, units='DU/TU')
+    p.set_val('orbit_transfer.parameters:c', value=1.5, units='DU/TU')
 
-    burn1 = p.model.traj.phases.burn1
-    burn2 = p.model.traj.phases.burn2
-    coast = p.model.traj.phases.coast
+    burn1 = p.model.orbit_transfer.phases.burn1
+    burn2 = p.model.orbit_transfer.phases.burn2
+    coast = p.model.orbit_transfer.phases.coast
 
-    if burn1 in p.model.traj.phases._subsystems_myproc:
-        p.set_val('traj.burn1.t_initial', value=0.0)
-        p.set_val('traj.burn1.t_duration', value=2.25)
-        p.set_val('traj.burn1.states:r', value=burn1.interpolate(ys=[1, 1.5],
-                                                                 nodes='state_input'))
-        p.set_val('traj.burn1.states:theta', value=burn1.interpolate(ys=[0, 1.7],
-                  nodes='state_input'))
-        p.set_val('traj.burn1.states:vr', value=burn1.interpolate(ys=[0, 0],
-                                                                  nodes='state_input'))
-        p.set_val('traj.burn1.states:vt', value=burn1.interpolate(ys=[1, 1],
-                                                                  nodes='state_input'))
-        p.set_val('traj.burn1.states:accel', value=burn1.interpolate(ys=[0.1, 0],
-                  nodes='state_input'))
-        p.set_val('traj.burn1.states:deltav', value=burn1.interpolate(ys=[0, 0.1],
-                  nodes='state_input'))
-        p.set_val('traj.burn1.controls:u1',
-                  value=burn1.interpolate(ys=[-3.5, 13.0], nodes='control_input'))
+    if burn1 in p.model.orbit_transfer.phases._subsystems_myproc:
+        p.set_val('orbit_transfer.burn1.t_initial', value=0.0)
+        p.set_val('orbit_transfer.burn1.t_duration', value=2.25)
+        p.set_val('orbit_transfer.burn1.states:r', value=burn1.interpolate(ys=[1, 1.5], nodes='state_input'))
+        p.set_val('orbit_transfer.burn1.states:theta', value=burn1.interpolate(ys=[0, 1.7], nodes='state_input'))
+        p.set_val('orbit_transfer.burn1.states:vr', value=burn1.interpolate(ys=[0, 0], nodes='state_input'))
+        p.set_val('orbit_transfer.burn1.states:vt', value=burn1.interpolate(ys=[1, 1], nodes='state_input'))
+        p.set_val('orbit_transfer.burn1.states:accel', value=burn1.interpolate(ys=[0.1, 0], nodes='state_input'))
+        p.set_val('orbit_transfer.burn1.states:deltav', value=burn1.interpolate(ys=[0, 0.1], nodes='state_input'))
+        p.set_val('orbit_transfer.burn1.controls:u1', value=burn1.interpolate(ys=[-3.5, 13.0], nodes='control_input'))
 
[…]
