### 2021-01-08 03c0a6e
fixes a bug where Trajectory.simulate crashes in some circumstances when trajectory (input) parameters are used. (#500)
