### 2021-01-12 620e5e3
added detection of unsatisfiable linkages
