### 2021-01-12 ab60130
fix for the case where time bounds are (None, None)
