### 2021-01-13 151f1ff
added detection of unsatisfiable linkages (#502)
* added detection of unsatisfiable linkages

* fix for the case where time bounds are (None, None)

* logic fix

* errant test name
