### 2021-01-13 c87379b
adds a deprecation warning when using the command line interface
