+++ dymos/examples/double_integrator/test/test_double_integrator.py
@@ -243,7 +243,7 @@ class TestDoubleIntegratorExample(unittest.TestCase):
         phase.add_control('u', units='m/s**2', scaler=0.01, continuity=False, rate_continuity=False,
                           rate2_continuity=False, shape=(1, ), lower=-1.0, upper=1.0)
 
-        phase.add_boundary_constraint(name='v', loc='final', equals=0, units='m/s')
+        phase.add_boundary_constraint(name='v', loc='final', equals=0)
 
         # Maximize distance travelled in one second.
         phase.add_objective('x', loc='final', scaler=-1)
