+++ dymos/examples/brachistochrone/test/test_brachistochrone_rk4.py
@@ -266,6 +266,71 @@ class TestBrachistochroneRK4Example(unittest.TestCase):
         assert_near_equal(exp_out.get_val('phase0.timeseries.states:y')[-1, 0], 5,
                           tolerance=1.0E-3)
 
+    def test_brachistochrone_forward_shooting_path_constrained_control_unit_convert(self):
+        # Tests that we can add path or boundary constraints with a declared unit that differs from
+        # the model unit, and conversion occurs.
+        import openmdao.api as om
+        from openmdao.utils.assert_utils import assert_near_equal
+        import dymos as dm
+        from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
+
+        p = om.Problem(model=om.Group())
+        p.driver = om.ScipyOptimizeDriver()
+        p.driver.declare_coloring()
+
+        phase = dm.Phase(ode_class=BrachistochroneODE,
+                         transcription=dm.RungeKutta(num_segments=20))
+
+        p.model.add_subsystem('phase0', phase)
+
+        phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(0.5, 2.0))
+
+        phase.add_state('x', fix_initial=True, fix_final=False)
+
+        phase.add_state('y', fix_initial=True, fix_final=False)
+
+        phase.add_state('v', fix_initial=True, fix_final=False)
+
+        phase.add_control('theta', continuity=True, rate_continuity=True,
+                          units='deg', lower=0.01, upper=179.9)
+
+        phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
+
+        # Final state values can't be controlled with simple bounds in ExplicitPhase,
+        # so use nonlinear boundary constraints instead.
+        phase.add_boundary_constraint('x', loc='final', equals=1000, units='cm')
+        phase.add_boundary_constraint('y', loc='final', equals=500, units='cm')
+        phase.add_path_constraint('theta', lower=0.01*0.01745329, upper=110*0.01745329, units='rad')
+
+        # Minimize time at the end of the phase
+        phase.add_objective('time_phase', loc='final', scaler=1)
+
+        p.model.linear_solver = om.DirectSolver()
+
+        p.setup(check=True)
+
+        p['phase0.t_initial'] = 0.0
+        p['phase0.t_duration'] = 2.0
+
+        p['phase0.states:x'] = 0
+        p['phase0.states:y'] = 10
+        p['phase0.states:v'] = 0
+        p['phase0.controls:theta'] = phase.interpolate(ys=[5, 100.5], nodes='control_input')
+
+        # Solve for the optimal trajectory
+        p.run_driver()
+
+        # Test the results
+        assert_near_equal(p['phase0.time'][-1], 1.8016, tolerance=1.0E-3)
+
+        # Generate the explicitly simulated trajectory
+        exp_out = phase.simulate()
+
+        assert_near_equal(exp_out.get_val('phase0.timeseries.states:x')[-1, 0], 10,
+                          tolerance=1.0E-3)
+        assert_near_equal(exp_out.get_val('phase0.timeseries.states:y')[-1, 0], 5,
+                          tolerance=1.0E-3)
+
     def test_brachistochrone_forward_shooting_path_constrained_control_rate(self):
         import openmdao.api as om
         from openmdao.utils.assert_utils import assert_near_equal
