### 2021-01-14 f1d30be
some scaling changes and fixed an unconnected rho
