+++ dymos/examples/balanced_field/ground_roll_ode.py
@@ -17,7 +17,8 @@ class GroundRollODE(om.Group):
 
         self.add_subsystem(name='atmos',
                            subsys=USatm1976Comp(num_nodes=nn),
-                           promotes_inputs=['h'])
+                           promotes_inputs=['h'],
+                           promotes_outputs=['rho'])
 
         self.add_subsystem(name='k_comp',
                            subsys=KComp(num_nodes=nn),
