### 2021-01-14 e46323d
Added check for whether dynamic parameters when setting values for older versions of openmdao
