+++ dymos/examples/balanced_field/doc/test_doc_balanced_field_length.py
@@ -8,352 +8,352 @@ plt.style.use('ggplot')
 from dymos.utils.doc_utils import save_for_docs
 
 
-class TestBalancedFieldLengthForDocs(unittest.TestCase):
+# class TestBalancedFieldLengthForDocs(unittest.TestCase):
+#
+#     # @save_for_docs
+#     def test_balanced_field_length_for_docs(self):
+import matplotlib.pyplot as plt
 
-    # @save_for_docs
-    def test_balanced_field_length_for_docs(self):
-        import matplotlib.pyplot as plt
+import openmdao.api as om
+from openmdao.utils.assert_utils import assert_near_equal
 
-        import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
+import dymos as dm
+from dymos.examples.balanced_field.ground_roll_ode import GroundRollODE
+from dymos.examples.balanced_field.takeoff_ode import TakeoffODE
 
-        import dymos as dm
-        from dymos.examples.balanced_field.ground_roll_ode import GroundRollODE
-        from dymos.examples.balanced_field.takeoff_ode import TakeoffODE
+#
+# Instantiate the problem and configure the optimization driver
+#
+p = om.Problem(model=om.Group())
 
-        #
-        # Instantiate the problem and configure the optimization driver
-        #
-        p = om.Problem(model=om.Group())
+p.driver = om.pyOptSparseDriver()
+p.driver.options['optimizer'] = 'SNOPT'
+p.driver.opt_settings['iSumm'] = 6
+# p.driver.opt_settings['Verify level'] = 3
+p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-5
+p.driver.opt_settings['Major optimality tolerance'] = 1.0E-5
+p.driver.opt_settings['Major iterations limit'] = 500
+p.driver.opt_settings['Minor iterations limit'] = 100000
 
-        p.driver = om.pyOptSparseDriver()
-        p.driver.options['optimizer'] = 'SNOPT'
-        p.driver.opt_settings['iSumm'] = 6
-        # p.driver.opt_settings['Verify level'] = 3
-        p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-5
-        p.driver.opt_settings['Major optimality tolerance'] = 1.0E-5
-        p.driver.opt_settings['Major iterations limit'] = 500
-        p.driver.opt_settings['Minor iterations limit'] = 100000
 
+# p.driver.options['optimizer'] = 'IPOPT'
+# p.driver.opt_settings['print_level'] = 5
+# p.driver.opt_settings['nlp_scaling_method'] = 'gradient-based'
 
-        # p.driver.options['optimizer'] = 'IPOPT'
-        # p.driver.opt_settings['print_level'] = 5
-        # p.driver.opt_settings['nlp_scaling_method'] = 'gradient-based'
+p.driver.declare_coloring(tol=1.0E-12)
 
-        p.driver.declare_coloring(tol=1.0E-12)
+#
+# Instantiate the trajectory and phase
+#
+traj = dm.Trajectory()
 
-        #
-        # Instantiate the trajectory and phase
-        #
-        traj = dm.Trajectory()
+p.model.add_subsystem('traj', traj)
 
-        p.model.add_subsystem('traj', traj)
+#
+# First Phase - Brake Release to V1
+# Operating with two engines
+# We don't know what V1 is a priori, we're going to use this model to determine it.
+#
 
-        #
-        # First Phase - Brake Release to V1
-        # Operating with two engines
[…]
