### 2021-01-15 7ac8b3c
constant CL version of the balanced field problem
