### 2021-01-15 09de114
fixes for restarting trajectories
