### 2021-01-15 b638f51
working on load_case fix but have a weird error regarding set_input_defaults
