### 2021-01-15 b75b0fc
working on load_case fix but have a weird error regarding set_input_defaults
