+++ dymos/trajectory/trajectory.py
@@ -393,7 +393,7 @@ class Trajectory(om.Group):
                 elif isinstance(targets[phase_name], str) and \
                         targets[phase_name] in phs.parameter_options:
                     # Connect to an input parameter with a different name in this phase
-                    tgt = '{0}.parameters:{1}'.format(phase_name, targets[phase_name])
+                    tgt = f'{phase_name}.parameters:{targets[phase_name]}'
                     tgt_shapes[phs.name] = phs.parameter_options[targets[phase_name]]['shape']
                     tgt_units[phs.name] = phs.parameter_options[targets[phase_name]]['units']
                 elif isinstance(targets[phase_name], Sequence) and \
