### 2021-01-16 b05f3ab
Added test cases to test for using load_case on multi-phase trajectories. Updated load_case to handle situations where parameters or times are inputs instead of outputs. Added assert_cases_equal to test if two cases contain the same variables with the same values.
