+++ dymos/load_case.py
@@ -81,71 +81,61 @@ def load_case(problem, previous_solution):
     traj_paths = find_trajectories(problem.model)
     phase_paths = find_phases(problem.model)
 
-    if not phase_paths:
+    if not phase_paths and not traj_paths:
         return
 
-    prev_inputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in previous_solution['inputs']}
-    prev_outputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in previous_solution['outputs']}
+    prev_vars = {}
+    prev_vars.update({v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in previous_solution['inputs']})
+    prev_vars.update({v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in previous_solution['outputs']})
 
     problem.final_setup()  # make sure list_inputs and list_outputs can work
-
-    traj_io = {'inputs': problem.model.list_inputs(out_stream=None, units=True, prom_name=True),
-               'outputs': problem.model.list_outputs(out_stream=None, units=True, prom_name=True)}
-
-    traj_inputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in traj_io['inputs']}
-    traj_outputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in traj_io['outputs']}
 
     phase_io = {'inputs': problem.model.list_inputs(out_stream=None, units=True, prom_name=True),
                 'outputs': problem.model.list_outputs(out_stream=None, units=True, prom_name=True)}
 
-    phase_inputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in phase_io['inputs']}
-    phase_outputs = {v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in phase_io['outputs']}
+    phase_vars = {}
+    phase_vars.update({v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in phase_io['inputs']})
+    phase_vars.update({v['prom_name']: {'value': v['value'], 'units': v['units']} for k, v in phase_io['outputs']})
 
     for traj_abs_path, traj in traj_paths.items():
         traj_name = traj_abs_path.split('.')[-1]
-        for parameter_name, options in traj.parameter_options.items():
-            prev_output_matches = [s for s in prev_outputs if s.endswith(f'{traj_name}.parameters:{parameter_name}')]
-            prev_input_matches = [s for s in prev_inputs if s.endswith(f'{traj_name}.parameters:{parameter_name}')]
-            if prev_output_matches:
+        for param_name, options in traj.parameter_options.items():
+            prev_match = [s for s in prev_vars if s.endswith(f'{traj_name}.parameters:{param_name}')]
+            if prev_match:
                 # In previous outputs
-                prev_param_path = prev_output_matches[0]
-                prev_param_val = prev_outputs[prev_param_path]['value']
-                prev_param_units = prev_outputs[prev_param_path]['units']
-            elif prev_input_matches:
-                # In previous outputs
-                prev_param_path = prev_input_matches[0]
-                prev_param_val = prev_inputs[prev_param_path]['value']
-                prev_param_units = prev_inputs[prev_param_path]['units']
+                prev_data = prev_vars[prev_match[0]]
+                prev_val = prev_data['value']
+                prev_units = prev_data['units']
             else:
-                raise Warning(f'Unable to find a value for {traj_name}.parameters:{parameter_name} in the restart file.')
-            problem.set_val(f'{traj.pathname}.parameters:{parameter_name}',
-                            prev_param_val[0, ...],
-                            units=prev_param_units)
[…]

[…]
