+++ dymos/examples/balanced_field/doc/test_doc_balanced_field_length_constant_cl.py
@@ -1,364 +1,0 @@
-import unittest
-
-import matplotlib
-import matplotlib.pyplot as plt
-# plt.switch_backend('TkAgg')
-plt.style.use('ggplot')
-
-from dymos.utils.doc_utils import save_for_docs
-
-
-# class TestBalancedFieldLengthForDocs(unittest.TestCase):
-#
-#     # @save_for_docs
-#     def test_balanced_field_length_for_docs(self):
-import matplotlib.pyplot as plt
-
-import openmdao.api as om
-from openmdao.utils.assert_utils import assert_near_equal
-
-import dymos as dm
-from dymos.examples.balanced_field.ground_roll_ode import GroundRollODE
-from dymos.examples.balanced_field.takeoff_ode import TakeoffODE
-
-#
-# Instantiate the problem and configure the optimization driver
-#
-p = om.Problem(model=om.Group())
-
-p.driver = om.pyOptSparseDriver()
-# p.driver.options['optimizer'] = 'SNOPT'
-# p.driver.opt_settings['iSumm'] = 6
-# # p.driver.opt_settings['Verify level'] = 3
-# # p.driver.opt_settings['Major step limit'] = 0.1
-# p.driver.opt_settings['Major feasibility tolerance'] = 1.0E-5
-# p.driver.opt_settings['Major optimality tolerance'] = 1.0E-5
-# p.driver.opt_settings['Major iterations limit'] = 500
-# p.driver.opt_settings['Minor iterations limit'] = 100000
-# p.driver.opt_settings['Linesearch tolerance'] = 0.5
-
-
-# p.driver.options['optimizer'] = 'IPOPT'
-# p.driver.opt_settings['print_level'] = 5
-# p.driver.opt_settings['nlp_scaling_method'] = 'gradient-based'
-
-p.driver.options['optimizer'] = 'SLSQP'
-
-p.driver.declare_coloring()
-
-#
-# Instantiate the trajectory and phase
-#
-traj = dm.Trajectory()
-
-p.model.add_subsystem('traj', traj)
-
-#
-# First Phase - Brake Release to V1
-# Operating with two engines
-# We don't know what V1 is a priori, we're going to use this model to determine it.
-#
-
-p1 = dm.Phase(ode_class=GroundRollODE, transcription=dm.Radau(num_segments=3))
-
-traj.add_phase('brake_release_to_v1', p1)
-
-#
-# Set the options on the optimization variables
-#
-p1.set_time_options(fix_initial=True, duration_bounds=(1, 1000), duration_ref=10.0)
-
-p1.add_state('r', fix_initial=True, lower=0, ref=1000.0, defect_ref=1000.0,
-             rate_source='r_dot')
-
-p1.add_state('v', fix_initial=True, lower=0.0, ref=100.0, defect_ref=100.0, rate_source='v_dot')
-
-p1.add_parameter('h', opt=False, units='m')
-p1.add_parameter('T', val=27000 * 2, opt=False, units='lbf')
-p1.add_parameter('m', val=174200, opt=False, units='lbm')
-p1.add_parameter('alpha', val=0.0, opt=False, units='deg')
-p1.add_parameter('mu_r', val=0.03, opt=False, units=None)
-
-p1.add_timeseries_output('*')
-
-# Second Phase - V1 to Vr
-# Operating with one engine
-# Vr is taken to be 1.2 * the stall speed (v_stall)
-#
-
-p2 = dm.Phase(ode_class=GroundRollODE,
-                 transcription=dm.Radau(num_segments=3))
-
-traj.add_phase('v1_to_vr', p2)
-
-#
[…]
