### 2021-01-16 1640fae
Added introspection for boundary and path constraints (#506)
* added code to perform boundary constraint introspection, still need path constraints


Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
Co-authored-by: Rob Falck <rfalck@nasa.gov>
