+++ dymos/transcriptions/transcription_base.py
@@ -3,7 +3,6 @@ from collections.abc import Sequence
 import numpy as np
 
 import openmdao.api as om
-from openmdao.core.constants import _UNDEFINED
 
 from .common import BoundaryConstraintComp, ControlGroup, PolynomialControlGroup, PathConstraintComp
 from ..phase.options import StateOptionsDictionary
@@ -451,21 +450,11 @@ class TranscriptionBase(object):
         for var, options in bc_dict.items():
             con_name = options['constraint_name']
 
-            # Constraint options are a copy of options with constraint_name key removed.
-            con_options = options.copy()
-            con_options.pop('constraint_name')
-
-            src, shape, units, linear = self._get_boundary_constraint_src(var, loc, phase)
-
-            con_units = options.get('units', None)
-
-            shape = options['shape'] if shape is None else shape
-            if shape is None:
-                shape = (1,)
+            _, shape, units, linear = self._get_boundary_constraint_src(var, loc, phase)
 
             if options['indices'] is not None:
-                # Indices are provided, make sure lower/upper/equals are compatible.
-                con_shape = (len(options['indices']),)
+                # Sliced shape.
+                con_shape = (len(options['indices']), )
                 # Indices provided, make sure lower/upper/equals have shape of the indices.
                 if options['lower'] and not np.isscalar(options['lower']) and \
                         np.asarray(options['lower']).shape != con_shape:
@@ -485,7 +474,7 @@ class TranscriptionBase(object):
                                      'compatible the provided indices. Provide them as a '
                                      'flat array with the same size as indices.'.format(var))
 
-            elif 'lower' in options or 'upper' in options or 'equals' in options:
+            else:
                 # Indices not provided, make sure lower/upper/equals have shape of source.
                 if 'lower' in options and options['lower'] is not None and \
                         not np.isscalar(options['lower']) and np.asarray(options['lower']).shape != shape:
@@ -507,18 +496,18 @@ class TranscriptionBase(object):
                                      'compatible with its shape, and no indices were '
                                      'provided. Expected a shape of {1} but given shape '
                                      'is {2}'.format(var, shape, np.asarray(options['equals']).shape))
-                con_shape = (np.prod(shape),)
 
-            size = np.prod(shape)
-            con_options['shape'] = shape if shape is not None else con_shape
+            # Constraint options are a copy of options with constraint_name key removed.
+            con_options = options.copy()
+            con_options.pop('constraint_name')
+
+            # By now, all possible constraint target shapes should have been introspected.
+            con_options['shape'] = options['shape'] = shape
+
+            # If user overrides the introspected unit, then change the unit on the add_constraint call.
+            con_units = options['units']
             con_options['units'] = units if con_units is None else con_units
             con_options['linear'] = linear
-
-            # Build the correct src_indices regardless of shape
-            if loc == 'initial':
-                src_idxs = np.arange(size, dtype=int).reshape(shape)
-            else:
-                src_idxs = np.arange(-size, 0, dtype=int).reshape(shape)
 
             bc_comp._add_constraint(con_name, **con_options)
 
@@ -529,10 +518,6 @@ class TranscriptionBase(object):
             con_name = options['constraint_name']
 
             src, shape, units, linear = self._get_boundary_constraint_src(var, loc, phase)
-
-            shape = options['shape'] if shape is None else shape
-            if shape is None:
-                shape = (1,)
 
             size = np.prod(shape)
 
@@ -550,7 +535,7 @@ class TranscriptionBase(object):
 
             else:
                 phase.connect(src,
-                              '{0}_boundary_constraints.{0}_value_in:{1}'.format(loc, con_name),
[…]
