+++ dymos/examples/min_time_climb/doc/test_doc_min_time_climb.py
@@ -83,12 +83,12 @@ class TestMinTimeClimbForDocs(unittest.TestCase):
         #
         # Setup the boundary and path constraints
         #
-        phase.add_boundary_constraint('h', loc='final', equals=20000, scaler=1.0E-3, units='m')
-        phase.add_boundary_constraint('aero.mach', loc='final', equals=1.0, shape=(1,))
-        phase.add_boundary_constraint('gam', loc='final', equals=0.0, units='rad')
+        phase.add_boundary_constraint('h', loc='final', equals=20000, scaler=1.0E-3)
+        phase.add_boundary_constraint('aero.mach', loc='final', equals=1.0)
+        phase.add_boundary_constraint('gam', loc='final', equals=0.0)
 
         phase.add_path_constraint(name='h', lower=100.0, upper=20000, ref=20000)
-        phase.add_path_constraint(name='aero.mach', lower=0.1, upper=1.8, shape=(1,))
+        phase.add_path_constraint(name='aero.mach', lower=0.1, upper=1.8)
 
         # Minimize time at the end of the phase
         phase.add_objective('time', loc='final', ref=1.0)
