### 2021-01-16 e8a2d81
added test to verify that parameters are now saved to the solution file.
