+++ dymos/test/test_run_problem.py
@@ -456,5 +456,93 @@ class TestRunProblemPlotting(unittest.TestCase):
             self.assertTrue(os.path.exists(f'plots/{varname.replace(":","_")}.png'))
 
 
+@use_tempdirs
+class TestSimulateArrayParam(unittest.TestCase):
+
+    def test_simulate_array_param(self):
+        #
+        # Initialize the Problem and the optimization driver
+        #
+        p = om.Problem(model=om.Group())
+        p.driver = om.ScipyOptimizeDriver()
+        p.driver.declare_coloring()
+
+        #
+        # Create a trajectory and add a phase to it
+        #
+        traj = p.model.add_subsystem('traj', dm.Trajectory())
+
+        phase = traj.add_phase('phase0',
+                               dm.Phase(ode_class=BrachistochroneODE,
+                                        transcription=dm.GaussLobatto(num_segments=10)))
+
+        #
+        # Set the variables
+        #
+        phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
+
+        phase.add_state('x', fix_initial=True, fix_final=True, rate_source='xdot')
+
+        phase.add_state('y', fix_initial=True, fix_final=True, rate_source='ydot')
+
+        phase.add_state('v', fix_initial=True, fix_final=False, rate_source='vdot')
+
+        phase.add_control('theta', continuity=True, rate_continuity=True,
+                          units='deg', lower=0.01, upper=179.9)
+
+        phase.add_parameter('g', units='m/s**2', val=9.80665)
+        phase.add_parameter('array', units=None, shape=(10,), dynamic=False)
+
+        # dummy array of data
+        indeps = p.model.add_subsystem('indeps', om.IndepVarComp(), promotes=['*'])
+        indeps.add_output('array', np.linspace(1, 10, 10), units=None)
+        # add dummy array as a parameter and connect it
+        p.model.connect('array', 'traj.phase0.parameters:array')
+
+        #
+        # Minimize time at the end of the phase
+        #
+        phase.add_objective('time', loc='final', scaler=10)
+
+        p.model.linear_solver = om.DirectSolver()
+
+        #
+        # Setup the Problem
+        #
+        p.setup()
+
+        #
+        # Set the initial values
+        #
+        p['traj.phase0.t_initial'] = 0.0
+        p['traj.phase0.t_duration'] = 2.0
+
+        p['traj.phase0.states:x'] = phase.interpolate(ys=[0, 10], nodes='state_input')
+        p['traj.phase0.states:y'] = phase.interpolate(ys=[10, 5], nodes='state_input')
+        p['traj.phase0.states:v'] = phase.interpolate(ys=[0, 9.9], nodes='state_input')
+        p['traj.phase0.controls:theta'] = phase.interpolate(ys=[5, 100.5],
+                                                            nodes='control_input')
+
+        #
+        # Solve for the optimal trajectory
+        #
+        dm.run_problem(p, simulate=True)
+
+        # Test the results
+        sol_results = om.CaseReader('dymos_solution.db').get_case('final')
+        sim_results = om.CaseReader('dymos_solution.db').get_case('final')
+
+        sol = sol_results.get_val('traj.phase0.timeseries.parameters:array')
+        sim = sim_results.get_val('traj.phase0.timeseries.parameters:array')
+
+        assert_near_equal(sol - sim, np.zeros_like(sol))
+
+        # Test that the parameter is available in the solution and simulation files
+        sol = sol_results.get_val('traj.phase0.parameters:array')
+        sim = sim_results.get_val('traj.phase0.parameters:array')
+
+        assert_near_equal(sol - sim, np.zeros_like(sol))
+
+
 if __name__ == '__main__':  # pragma: no cover
     unittest.main()
