+++ dymos/test/test_testing_utils.py
@@ -1,0 +1,232 @@
+import os
+import unittest
+
+import numpy as np
+import openmdao.api as om
+from openmdao.utils.testing_utils import use_tempdirs
+import dymos as dm
+
+from dymos.utils.testing_utils import assert_cases_equal, assert_timeseries_near_equal
+
+
+@use_tempdirs
+class TestAssertCasesEqual(unittest.TestCase):
+
+    def tearDown(self):
+        for file in ('p1.db', 'p2.db'):
+            try:
+                os.remove(file)
+                print('removed', file)
+            except:
+                print(f'no file named {file}')
+
+    def test_different_variables(self):
+
+        p1 = om.Problem()
+        ivc = p1.model.add_subsystem('ivc', om.IndepVarComp(), promotes_outputs=['*'])
+        ivc.add_output('a', val=0.0)
+        ivc.add_output('b', val=1.0)
+        ivc.add_output('c', val=2.0)
+        ivc.add_output('x', val=3.0)
+        p1.add_recorder(om.SqliteRecorder('p1.db'))
+        p1.setup()
+
+        p2 = om.Problem()
+        ivc = p2.model.add_subsystem('ivc', om.IndepVarComp(), promotes_outputs=['*'])
+        ivc.add_output('x', val=3.0)
+        ivc.add_output('y', val=4.0)
+        ivc.add_output('z', val=5.0)
+        p2.add_recorder(om.SqliteRecorder('p2.db'))
+        p2.setup()
+
+        p1.run_model()
+        p2.run_model()
+
+        p1.record('final')
+        p1.cleanup()
+
+        p2.record('final')
+        p2.cleanup()
+
+        c1 = om.CaseReader('p1.db').get_case('final')
+        c2 = om.CaseReader('p2.db').get_case('final')
+
+        with self.assertRaises(AssertionError) as e:
+            assert_cases_equal(c1, c2)
+
+        expected = "\nrequire_same_vars=True but cases contain different variables.\nVariables in " \
+                   "case1 but not in case2: ['a', 'b', 'c']\nVariables in case2 but not in " \
+                   "case1: ['y', 'z']"
+
+        self.assertEqual(str(e.exception), expected)
+
+    def test_allow_different_variables(self):
+        p1 = om.Problem()
+        ivc = p1.model.add_subsystem('ivc', om.IndepVarComp(), promotes_outputs=['*'])
+        ivc.add_output('a', val=0.0)
+        ivc.add_output('b', val=1.0)
+        ivc.add_output('c', val=2.0)
+        ivc.add_output('x', val=3.0)
+        p1.add_recorder(om.SqliteRecorder('p1.db'))
+        p1.setup()
+
+        p2 = om.Problem()
+        ivc = p2.model.add_subsystem('ivc', om.IndepVarComp(), promotes_outputs=['*'])
+        ivc.add_output('x', val=3.0)
+        ivc.add_output('y', val=4.0)
+        ivc.add_output('z', val=5.0)
+        p2.add_recorder(om.SqliteRecorder('p2.db'))
+        p2.setup()
+
+        p1.run_model()
+        p2.run_model()
+
+        p1.record('final')
+        p1.cleanup()
+
+        p2.record('final')
+        p2.cleanup()
+
+        c1 = om.CaseReader('p1.db').get_case('final')
+        c2 = om.CaseReader('p2.db').get_case('final')
+
+        assert_cases_equal(c1, c2, require_same_vars=False)
+
+    def test_different_shapes(self):
+
+        p1 = om.Problem()
[…]
