### 2021-01-18 bebede4
removed unused MotorStaticGearboxPower from battery multibranch test
