### 2021-01-19 f4a7a30
more tests for integrating state rates.  assert_timeseries_near_equal now interpolates the more dense timeseries and checks the less dense one against that.
