+++ dymos/examples/brachistochrone/test/test_state_rate_introspection.py
@@ -1,0 +1,496 @@
+import unittest
+
+import numpy as np
+import openmdao.api as om
+import dymos as dm
+
+from openmdao.utils.assert_utils import assert_near_equal
+from dymos.utils.testing_utils import assert_timeseries_near_equal
+
+from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
+
+
+class TestBrachistochroneQuickStart(unittest.TestCase):
+
+    def test_integrate_control(self):
+
+        #
+        # Define the OpenMDAO problem
+        #
+        p = om.Problem(model=om.Group())
+
+        #
+        # Define a Trajectory object
+        #
+        traj = dm.Trajectory()
+
+        p.model.add_subsystem('traj', subsys=traj)
+
+        #
+        # Define a Dymos Phase object with GaussLobatto Transcription
+        #
+        phase = dm.Phase(ode_class=BrachistochroneODE,
+                         transcription=dm.GaussLobatto(num_segments=10, order=3))
+
+        traj.add_phase(name='phase0', phase=phase)
+
+        #
+        # Set the time options
+        # Time has no targets in our ODE.
+        # We fix the initial time so that the it is not a design variable in the optimization.
+        # The duration of the phase is allowed to be optimized, but is bounded on [0.5, 10].
+        #
+        phase.set_time_options(fix_initial=True, duration_bounds=(1.0, 10.0), units='s')
+
+        #
+        # Set the time options
+        # Initial values of positions and velocity are all fixed.
+        # The final value of position are fixed, but the final velocity is a free variable.
+        # The equations of motion are not functions of position, so 'x' and 'y' have no targets.
+        # The rate source points to the output in the ODE which provides the time derivative of the
+        # given state.
+        phase.add_state('x', fix_initial=True, fix_final=True)
+        phase.add_state('y', fix_initial=True, fix_final=True)
+        phase.add_state('v', fix_initial=True)
+
+        phase.add_state('int_theta', fix_initial=False, rate_source='theta_rate', targets=['theta'])
+
+        # Define theta as a control.
+        phase.add_control(name='theta_rate', units='rad/s', shape=(1,), targets=None)
+
+        # Minimize final time.
+        phase.add_objective('time', loc='final')
+
+        # Set the driver.
+        p.driver = om.ScipyOptimizeDriver()
+
+        # Allow OpenMDAO to automatically determine our sparsity pattern.
+        # Doing so can significant speed up the execution of Dymos.
+        p.driver.declare_coloring()
+
+        # Setup the problem
+        p.setup(check=True)
+
+        # Now that the OpenMDAO problem is setup, we can set the values of the states.
+
+        p.set_val('traj.phase0.t_initial', 0.0, units='s')
+        p.set_val('traj.phase0.t_duration', 5.0, units='s')
+
+        p.set_val('traj.phase0.states:x',
+                  phase.interpolate(ys=[0, 10], nodes='state_input'),
+                  units='m')
+
+        p.set_val('traj.phase0.states:y',
+                  phase.interpolate(ys=[10, 5], nodes='state_input'),
+                  units='m')
+
+        p.set_val('traj.phase0.states:v',
+                  phase.interpolate(ys=[0, 5], nodes='state_input'),
+                  units='m/s')
+
+        p.set_val('traj.phase0.states:int_theta',
+                  phase.interpolate(ys=[0.1, 45], nodes='state_input'),
+                  units='deg')
+
+        # Additional states to test rate sources
[…]
