+++ dymos/examples/brachistochrone/test/test_state_rate_introspection.py
@@ -5,14 +5,16 @@ import openmdao.api as om
 import dymos as dm
 
 from openmdao.utils.assert_utils import assert_near_equal
+from openmdao.utils.testing_utils import use_tempdirs
 from dymos.utils.testing_utils import assert_timeseries_near_equal
 
 from dymos.examples.brachistochrone.brachistochrone_ode import BrachistochroneODE
 
 
-class TestBrachistochroneQuickStart(unittest.TestCase):
+@use_tempdirs
+class TestIntegrateControl(unittest.TestCase):
 
-    def test_integrate_control(self):
+    def _test_integrate_control(self, transcription):
 
         #
         # Define the OpenMDAO problem
@@ -30,7 +32,7 @@ class TestBrachistochroneQuickStart(unittest.TestCase):
         # Define a Dymos Phase object with GaussLobatto Transcription
         #
         phase = dm.Phase(ode_class=BrachistochroneODE,
-                         transcription=dm.GaussLobatto(num_segments=10, order=3))
+                         transcription=transcription(num_segments=10, order=3))
 
         traj.add_phase(name='phase0', phase=phase)
 
@@ -92,7 +94,6 @@ class TestBrachistochroneQuickStart(unittest.TestCase):
                   phase.interpolate(ys=[0.1, 45], nodes='state_input'),
                   units='deg')
 
-        # Additional states to test rate sources
         p.set_val('traj.phase0.controls:theta_rate',
                   phase.interpolate(ys=[10, 10], nodes='control_input'),
                   units='deg/s')
@@ -127,8 +128,7 @@ class TestBrachistochroneQuickStart(unittest.TestCase):
         assert_timeseries_near_equal(t_sol, int_theta_sol, t_sim, int_theta_sim, tolerance=1.0E-3)
         assert_timeseries_near_equal(t_sol, theta_rate_sol, t_sim, theta_rate_sim, tolerance=1.0E-3)
 
-
-    def test_integrate_control_rate(self):
+    def _test_integrate_control_rate(self, transcription):
 
         #
         # Define the OpenMDAO problem
@@ -146,7 +146,7 @@ class TestBrachistochroneQuickStart(unittest.TestCase):
         # Define a Dymos Phase object with GaussLobatto Transcription
         #
         phase = dm.Phase(ode_class=BrachistochroneODE,
-                         transcription=dm.GaussLobatto(num_segments=10, order=5))
+                         transcription=transcription(num_segments=10, order=3))
 
         traj.add_phase(name='phase0', phase=phase)
 
@@ -173,6 +173,11 @@ class TestBrachistochroneQuickStart(unittest.TestCase):
 
         # Define theta as a control.
         phase.add_control(name='theta', units='rad', shape=(1,), targets=None)
+
+        # Force the initial value of the theta polynomial control to equal the initial value of the theta state.
+        traj.add_linkage_constraint(phase_a='phase0', phase_b='phase0',
+                                    var_a='theta', var_b='int_theta',
+                                    loc_a='initial', loc_b='initial')
 
         # Minimize final time.
         phase.add_objective('time', loc='final')
@@ -208,7 +213,6 @@ class TestBrachistochroneQuickStart(unittest.TestCase):
                   phase.interpolate(ys=[0.1, 45], nodes='state_input'),
                   units='deg')
 
-        # Additional states to test rate sources
         p.set_val('traj.phase0.controls:theta',
                   phase.interpolate(ys=[0, 100], nodes='control_input'),
                   units='deg')
@@ -242,11 +246,10 @@ class TestBrachistochroneQuickStart(unittest.TestCase):
[…]
