### 2021-01-20 521e061
removed unnecessary parameter traj_param option.  addressed some of kens suggestions in load_case.
