### 2021-01-20 ef4d9f3
Fixed a bug that prevented restart from working with multiphase trajectories, and now include static parameters in the solution file. (#510)
* fixes for restarting trajectories

* working on load_case fix but have a weird error regarding set_input_defaults

* Added test cases to test for using load_case on multi-phase trajectories.
Updated load_case to handle situations where parameters or times are inputs instead of outputs.
Added assert_cases_equal to test if two cases contain the same variables with the same values.

* added test to verify that parameters are now saved to the solution file.
[…]
