+++ dymos/utils/testing_utils.py
@@ -1,5 +1,4 @@
 import numpy as np
-from numpy.testing import assert_array_almost_equal
 from scipy.interpolate import interp1d
 import openmdao.utils.assert_utils as _om_assert_utils
 
@@ -12,6 +11,80 @@ def assert_check_partials(data, atol=1.0E-6, rtol=1.0E-6):
     assert len(data) >= 1, "No check partials data found.  Is " \
                            "dymos.options['include_check_partials'] set to True?"
     _om_assert_utils.assert_check_partials(data, atol, rtol)
+
+
+def assert_cases_equal(case1, case2, tol=1.0E-12, require_same_vars=True):
+    """
+    Raise AssertionError if the data in two OpenMDAO Cases is different.
+
+    Parameters
+    ----------
+    case1 : om.Case
+        The first OpenMDAO Case for comparison.
+    case2 : om.Case
+        The second OpenMDAO Case for comparison.
+    tol : float
+        The absolute value of the allowable difference in values between two variables.
+    require_same_vars : bool
+        If True, require that the two files contain the same set of variables.
+
+    Raises
+    ------
+    AssertionError
+        Raised in the following cases:  If require_same_vars is True, then AssertionError is raised
+        if the two cases contain different variables.  Otherwise, this error is raised if case1
+        and case2 contain the same variable but the variable has a different size/shape in the two
+        cases, or if the variables have the same shape but different values (as given by tol).
+    """
+    case1_vars = {t[1]['prom_name']: t[1] for t in
+                  case1.list_inputs(values=True, units=True, prom_name=True, out_stream=None)}
+    case1_vars.update({t[1]['prom_name']: t[1] for t in
+                       case1.list_outputs(values=True, units=True, prom_name=True, out_stream=None)})
+
+    case2_vars = {t[1]['prom_name']: t[1] for t in
+                  case2.list_inputs(values=True, units=True, prom_name=True, out_stream=None)}
+    case2_vars.update({t[1]['prom_name']: t[1] for t in
+                       case2.list_outputs(values=True, units=True, prom_name=True, out_stream=None)})
+
+    # Warn if a and b don't contain the same sets of variables
+    diff_err_msg = ''
+    if require_same_vars:
+        case1_minus_case2 = set(case1_vars.keys()) - set(case2_vars.keys())
+        case2_minus_case1 = set(case2_vars.keys()) - set(case1_vars.keys())
+        if case1_minus_case2 or case2_minus_case1:
+            diff_err_msg = '\nrequire_same_vars=True but cases contain different variables.'
+        if case1_minus_case2:
+            diff_err_msg += f'\nVariables in case1 but not in case2: {sorted(case1_minus_case2)}'
+        if case2_minus_case1:
+            diff_err_msg += f'\nVariables in case2 but not in case1: {sorted(case2_minus_case1)}'
+
+    shape_errors = set()
+    val_errors = set()
+    shape_err_msg = '\nThe following variables have different shapes/sizes:'
+    val_err_msg = '\nThe following variables contain different values:\nvar: error'
+
+    for var in sorted(set(case1_vars.keys()).intersection(case2_vars.keys())):
+        a = case1_vars[var]['value']
+        b = case2_vars[var]['value']
+        if a.shape != b.shape:
+            shape_errors.add(var)
+            shape_err_msg += f'\n{var} has shape {a.shape} in case1 but shape {b.shape} in case2'
+            continue
+        err = np.abs(a - b)
+        if np.any(err > tol):
+            val_errors.add(var)
[…]
