+++ dymos/run_problem.py
@@ -112,14 +112,20 @@ def run_problem(problem, refine_method='hp', refine_iteration_limit=0, run_drive
     plot_dir : str
         Path to directory for plot files.
     """
+    if restart is not None:
+        case = om.CaseReader(restart).get_case('final')
+
     if solution_record_file not in [rec._filepath for rec in iter(problem._rec_mgr)]:
         recorder = om.SqliteRecorder(solution_record_file)
         problem.add_recorder(recorder)
+        # record_inputs is needed to capture potential input parameters that aren't connected
+        problem.recording_options['record_inputs'] = True
+        # record_outputs is need to capture the timeseries outputs
+        problem.recording_options['record_outputs'] = True
 
     problem.final_setup()  # make sure command line option hook has a chance to run
 
     if restart is not None:
-        case = om.CaseReader(restart).get_case('final')
         load_case(problem, case)
 
     if run_driver:
