### 2021-01-20 1d8b076
added tests for various state rate sources.  fixed a few transcription issues for some rate source types.
