+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -608,7 +608,7 @@ class SolveIVP(TranscriptionBase):
         elif phase.control_options is not None and var in phase.control_options:
             rate_path = f'control_values:{var}'
         elif phase.polynomial_control_options is not None and var in phase.polynomial_control_options:
-            rate_path = f'polynomial_controls:{var}'
+            rate_path = f'polynomial_control_values:{var}'
         elif phase.parameter_options is not None and var in phase.parameter_options:
             rate_path = f'parameters:{var}'
         elif var.endswith('_rate') and phase.control_options is not None and \
