+++ dymos/transcriptions/transcription_base.py
@@ -215,7 +215,7 @@ class TranscriptionBase(object):
             control_name = rate_src[:-6]
             control = phase.control_options[control_name]
             rate_src_units = get_rate_units(control['units'], time_units, deriv=2)
-            rate_src_shape = phase.control_options[control_name]['shape']
+            rate_src_shape = control['shape']
         elif rate_src_type == 'polynomial_control_rate':
             control_name = rate_src[:-5]
             control = phase.polynomial_control_options[control_name]
