### 2021-01-20 388a849
Add coverage for various rate source types (#513)
* added tests for various state rate sources.  fixed a few transcription issues for some rate source types.

* minor cleanup
