+++ dymos/utils/testing_utils.py
@@ -1,6 +1,7 @@
 import numpy as np
 from scipy.interpolate import interp1d
 import openmdao.utils.assert_utils as _om_assert_utils
+from openmdao.utils.general_utils import warn_deprecation
 
 
 def assert_check_partials(data, atol=1.0E-6, rtol=1.0E-6):
@@ -87,7 +88,7 @@ def assert_cases_equal(case1, case2, tol=1.0E-12, require_same_vars=True):
         raise AssertionError(err_msg)
 
 
-def assert_timeseries_near_equal(t1, x1, t2, x2, tolerance=1.0E-6, num_points=20):
+def assert_timeseries_near_equal(t1, x1, t2, x2, tolerance=1.0E-6, num_points=None):
     """
     Assert that two timeseries of data are approximately equal.
     This is done by fitting a 1D interpolant to each index of each timeseries, and then comparing
@@ -114,6 +115,9 @@ def assert_timeseries_near_equal(t1, x1, t2, x2, tolerance=1.0E-6, num_points=20
         When one or more elements of the interpolated timeseries are not within the
         desired tolerance.
     """
+    if num_points is not None:
+        warn_deprecation('Argument num_points is deprecated and will be removed in dymos 1.0.0')
+
     shape1 = x1.shape[1:]
     shape2 = x2.shape[1:]
 
@@ -139,12 +143,22 @@ def assert_timeseries_near_equal(t1, x1, t2, x2, tolerance=1.0E-6, num_points=20
     a2 = np.reshape(x2, newshape=(nn2, size))
     t2_unique, idxs2 = np.unique(t2.ravel(), return_index=True)
 
-    interp1 = interp1d(x=t1_unique, y=a1[idxs1, ...], kind='slinear', axis=0)
-    interp2 = interp1d(x=t2_unique, y=a2[idxs2, ...], kind='slinear', axis=0)
+    if nn1 > nn2:
+        # The first timeseries is more dense
+        t_unique = t1_unique
+        x_to_interp = a1[idxs1, ...]
+        t_check = t2.ravel()
+        x_check = x2
+    else:
+        # The second timeseries is more dense
+        t_unique = t2_unique
+        x_to_interp = a2[idxs2, ...]
+        t_check = t1.ravel()
+        x_check = x1
 
-    t_interp = np.linspace(t1[0], t1[-1], num_points)
+    interp = interp1d(x=t_unique, y=x_to_interp, kind='slinear', axis=0)
+    num_points = np.prod(t_check.shape)
 
-    y1 = np.reshape(interp1(t_interp), newshape=(num_points,) + shape1)
-    y2 = np.reshape(interp2(t_interp), newshape=(num_points,) + shape2)
+    y_interp = np.reshape(interp(t_check), newshape=(num_points,) + shape1)
 
-    _om_assert_utils.assert_near_equal(y1, y2, tolerance=tolerance)
+    _om_assert_utils.assert_near_equal(y_interp, x_check, tolerance=tolerance)
