+++ release_notes.md
@@ -1,4 +1,34 @@
 ********************************
+# Release Notes for Dymos 0.18.0
+
+January 20, 2021
+
+This release of Dymos brings a few improvements and bug fixes.
+
+## Backwards Incompatible API Changes & Deprecations
+
+* The command line interface is deprecated due to inconsistent behavior. [#505](https://github.com/OpenMDAO/dymos/pull/505)
+
+## Enhancements
+
+* Changed CoerceDesvar to handle vector values for ref/ref0/scaler/adder. Carried over changes to for defect ref/defect scaler. [#464](https://github.com/OpenMDAO/dymos/pull/464)
+* Changed solve_segments to allow it to be used when neither fix_initial nor fix_final are True. [#490](https://github.com/OpenMDAO/dymos/pull/490)
+* Added detection of unsatisfiable linkage constraints (where quantities on both sides cannot be changed by the optimizer). [#502](https://github.com/OpenMDAO/dymos/pull/502)
+* Added introspection for boundary and path constraints. [#506](https://github.com/OpenMDAO/dymos/pull/506)
+
+## Bug Fixes
+
+* Fixed a bug with using grid refinement if one or more of the states are vector-valued. [#492](https://github.com/OpenMDAO/dymos/pull/492)
+* Fixed issue where show_plots argument to run_problem is permanently changing matplotlib backend. [#497](https://github.com/OpenMDAO/dymos/pull/497)
+* Fixed a bug where trajectory.simulate() would crash if trajectory were not named the common 'traj' name. [#500](https://github.com/OpenMDAO/dymos/pull/500)
+* Fixed a bug that prevented restart from working with multiphase trajectories, and now include static parameters in the solution file. [#510](https://github.com/OpenMDAO/dymos/pull/510)
+* Fixed a bug associated with polynomial control rate sources, and added coverage of more cases. [#513](https://github.com/OpenMDAO/dymos/pull/513)
+
+## Miscellaneous
+
+
+
+****************************************************************
 # Release Notes for Dymos 0.17.0
 
 December 14, 2020
