### 2021-01-20 23e19d9
Allow None as a traj param target for a phase.
