### 2021-01-20 e1d554e
Fix for exception when simulating a trajectory with a parameter for which None was declared as the target for one of its phases. (#515)
* Allow None as a traj param target for a phase.

Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
