### 2021-01-20 02bc02d
added last issue to the release notes
