+++ dymos/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.18.0'
+__version__ = '0.18.1-dev'
 
 from .phase import Phase
 from .transcriptions import GaussLobatto, Radau, RungeKutta
