### 2021-01-21 9ffd23d
timeseries outputs will now ignore outputs that are not sized with the first dimension being num_nodes, and raise a warning when doing so.  Docs updated to reflect this.
