### 2021-01-21 c5aab93
Docstring linting: the beginning
