### 2021-01-21 51b9249
fixed static timeseries output for runge kutta
