### 2021-01-22 9cd169f
Added is_static_ode_output method to TranscriptionBase to reduce code reuse in derived Transcription classes.
