+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -523,13 +523,9 @@ class GaussLobatto(PseudospectralBase):
                         continue
 
                     # If the full shape does not start with num_nodes, skip this variable.
-                    ode_outputs = {opts['prom_name']: opts for (k, opts) in
-                                   phase.rhs_disc.get_io_metadata(iotypes=('output',), get_remote=True).items()}
-                    ode_shape = ode_outputs[v]['shape']
-                    if ode_shape[0] != self.grid_data.subset_num_nodes['state_disc']:
+                    if self.is_static_ode_output(v, phase):
                         warnings.warn(f'Cannot add ODE output {v} to the timeseries output. It is '
-                                      f'sized such that its first dimension != num_nodes. '
-                                      f'The shape is {ode_shape}.')
+                                      f'sized such that its first dimension != num_nodes.')
                         continue
 
                     try:
