+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -535,14 +535,9 @@ class SolveIVP(TranscriptionBase):
 
                 # Skip the timeseries output if it does not appear to be shaped as a dynamic variable
                 # If the full shape does not start with num_nodes, skip this variable.
-                ode_outputs = {opts['prom_name']: opts for (k, opts) in
-                               phase.ode.get_io_metadata(iotypes=('output',),
-                                                         get_remote=True).items()}
-                ode_shape = ode_outputs[v]['shape']
-                if ode_shape[0] != phase._get_subsystem('ode').options['num_nodes']:
+                if self.is_static_ode_output(v, phase):
                     warnings.warn(f'Cannot add ODE output {v} to the timeseries output. It is '
-                                  f'sized such that its first dimension != num_nodes. '
-                                  f'The shape is {ode_shape}.')
+                                  f'sized such that its first dimension != num_nodes.')
                     continue
 
                 shape, units = get_source_metadata(phase.ode, src=v, user_shape=options['shape'],
