### 2021-01-22 3724021
errant return from is_static_ode_output
