### 2021-01-22 151f4ea
Fixed a bug where timeseries outputs of non-dynamic ODE outputs would cause an exception. (#521)
* timeseries outputs will now ignore outputs that are not sized with the first dimension being num_nodes, and raise a warning when doing so.  Docs updated to reflect this.

* Added is_static_ode_output method to TranscriptionBase to reduce code reuse in derived Transcription classes.
