+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -1,4 +1,5 @@
 from fnmatch import filter
+import warnings
 
 import numpy as np
 
@@ -530,6 +531,13 @@ class SolveIVP(TranscriptionBase):
 
                 # Ignore any variables that we've already added (states, times, controls, etc)
                 if var_type != 'ode':
+                    continue
+
+                # Skip the timeseries output if it does not appear to be shaped as a dynamic variable
+                # If the full shape does not start with num_nodes, skip this variable.
+                if self.is_static_ode_output(v, phase):
+                    warnings.warn(f'Cannot add ODE output {v} to the timeseries output. It is '
+                                  f'sized such that its first dimension != num_nodes.')
                     continue
 
                 shape, units = get_source_metadata(phase.ode, src=v, user_shape=options['shape'],
