+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -1,3 +1,5 @@
+import warnings
+
 import numpy as np
 import openmdao.api as om
 
@@ -518,6 +520,12 @@ class GaussLobatto(PseudospectralBase):
 
                     # Ignore any variables that we've already added (states, times, controls, etc)
                     if var_type != 'ode':
+                        continue
+
+                    # If the full shape does not start with num_nodes, skip this variable.
+                    if self.is_static_ode_output(v, phase):
+                        warnings.warn(f'Cannot add ODE output {v} to the timeseries output. It is '
+                                      f'sized such that its first dimension != num_nodes.')
                         continue
 
                     try:
