+++ dymos/phase/phase.py
@@ -45,7 +45,7 @@ class Phase(om.Group):
         A phase instance from which the initialized phase should copy its data.
     transcription : <TranscriptionBase>
         The transcription to be utilized within the Phase.
-    ode_class
+    ode_class : classs of type System
         An OpenMDAO system class serving as the ODE for the phase.
     ode_init_kwargs : dict
         Keyword arguments used to initialize ode_class.
@@ -95,6 +95,9 @@ class Phase(om.Group):
         super(Phase, self).__init__(**_kwargs)
 
     def initialize(self):
+        """
+        Declare instantiation options for the phase.
+        """
         self.options.declare('ode_class', default=None,
                              desc='System defining the ODE')
         self.options.declare('ode_init_kwargs', types=dict, default={},
@@ -153,7 +156,7 @@ class Phase(om.Group):
         ref0 : float or ndarray or None (None)
             The zero-reference value of the state at the nodes of the phase.
         ref : float or ndarray or None (None)
-            The unit-reference value of the state at the nodes of the phase
+            The unit-reference value of the state at the nodes of the phase.
         defect_scaler : float or ndarray
             The scaler of the state defect at the collocation nodes of the phase.
         defect_ref : float or ndarray
@@ -419,7 +422,6 @@ class Phase(om.Group):
         Notes
         -----
         rate and rate2 continuity are not enforced for input controls.
-
         """
         if name not in self.control_options:
             self.check_parameter(name)
@@ -517,7 +519,6 @@ class Phase(om.Group):
         Notes
         -----
         rate and rate2 continuity are not enforced for input controls.
-
         """
         if units is not _unspecified:
             self.control_options[name]['units'] = units
@@ -616,11 +617,11 @@ class Phase(om.Group):
         order : int
             The order of the interpolating polynomial used to represent the control valeu in
             phase tau space.
+        desc : str
+            A description of the polynomial control.
         val : float or ndarray
             Default value of the control at all nodes.  If val scalar and the control
             is dynamic it will be broadcast.
-        desc : str
-            A description of the polynomial control.
         units : str or None or 0
             Units in which the control variable is defined.  If 0, use the units declared
             for the parameter in the ODE.
@@ -645,7 +646,7 @@ class Phase(om.Group):
         ref0 : float or ndarray
             The zero-reference value of the control at the nodes of the phase.
         ref : float or ndarray
-            The unit-reference value of the control at the nodes of the phase
+            The unit-reference value of the control at the nodes of the phase.
         targets : Sequence of str or None
             Targets in the ODE to which this polynomial control is connected.
         rate_targets : None or str
@@ -1081,7 +1082,7 @@ class Phase(om.Group):
         targets : Sequence of str or None
             Targets in the ODE to which this parameter is connected.
         desc : str
-            A description of the input parameter
+            A description of the input parameter.
         shape : Sequence of str or None
             The shape of the input parameter.
         dynamic : bool
@@ -1139,12 +1140,12 @@ class Phase(om.Group):
 
         Parameters
         ----------
-        name : string
+        name : str
             Name of the variable to constrain.  If name is not a state, control, or 'time',
             then this is assumed to be the path of the variable to be constrained in the ODE.
-        loc : string
-            The location of the boundary constraint ('initial' or 'final')
-        constraint_name : string or None
+        loc : str
+            The location of the boundary constraint ('initial' or 'final').
+        constraint_name : str or None
             The name of the variable as provided to the boundary constraint comp.  By
             default this is the last element in `name` when split by dots.  The user may
             override the constraint name if splitting the path causes name collisions.
@@ -1161,21 +1162,21 @@ class Phase(om.Group):
             flattening.  For instance, when constraining element [0, 1] of a variable of shape
             [2, 2], indices would be [3].
[…]
