### 2021-01-22 97be2e4
clean up implementation of wildcard units when adding multiple timeseries at once.
