+++ dymos/transcriptions/solve_ivp/solve_ivp.py
@@ -505,7 +505,7 @@ class SolveIVP(TranscriptionBase):
         for var, options in phase._timeseries['timeseries']['outputs'].items():
             output_name = options['output_name']
             units = options.get('units', None)
-            timeseries_units = options.get('timeseries_units', None)
+            wildcard_units = options.get('wildcard_units', None)
 
             if '*' in var:  # match outputs from the ODE
                 ode_outputs = {opts['prom_name']: opts for (k, opts) in
@@ -518,9 +518,9 @@ class SolveIVP(TranscriptionBase):
                 if '*' in var:
                     output_name = v.split('.')[-1]
                     units = ode_outputs[v]['units']
-                    # check for timeseries_units override of ODE units
-                    if v in timeseries_units:
-                        units = timeseries_units[v]
+                    # check for wildcard_units override of ODE units
+                    if v in wildcard_units:
+                        units = wildcard_units[v]
 
                 # Determine the path to the variable which we will be constraining
                 # This is more complicated for path constraints since, for instance,
