### 2021-01-22 8f1916b
Clean up implementation of wildcard units when adding multiple timeseries at once. (#523)
* clean up implementation of wildcard units when adding multiple timeseries at once.

Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
