+++ dymos/transcriptions/pseudospectral/gauss_lobatto.py
@@ -15,7 +15,7 @@ from fnmatch import filter
 
 class GaussLobatto(PseudospectralBase):
     """
-    High-order Gauss Lobatto Transcription
+    High-order Gauss Lobatto Transcription.
 
     References
     ----------
@@ -28,16 +28,24 @@ class GaussLobatto(PseudospectralBase):
         self._rhs_source = 'rhs_disc'
 
     def init_grid(self):
+        """
+        Setup the GridData object for the Transcription.
+        """
         self.grid_data = GridData(num_segments=self.options['num_segments'],
                                   transcription='gauss-lobatto',
                                   transcription_order=self.options['order'],
                                   segment_ends=self.options['segment_ends'],
                                   compressed=self.options['compressed'])
 
-    def setup_time(self, phase):
-        super(GaussLobatto, self).setup_time(phase)
-
     def configure_time(self, phase):
+        """
+        Configure the inputs/outputs on the time component.
+
+        Parameters
+        ----------
+        phase : dymos.Phase
+            The phase object to which this transcription instance applies.
+        """
         super(GaussLobatto, self).configure_time(phase)
         options = phase.time_options
 
@@ -62,6 +70,14 @@ class GaussLobatto(PseudospectralBase):
                               src_indices=disc_src_idxs)
 
     def configure_controls(self, phase):
+        """
+        Configure the inputs/outputs for the controls.
+
+        Parameters
+        ----------
+        phase : dymos.Phase
+            The phase object to which this transcription instance applies.
+        """
         super(GaussLobatto, self).configure_controls(phase)
 
         grid_data = self.grid_data
@@ -112,6 +128,14 @@ class GaussLobatto(PseudospectralBase):
                               src_indices=col_src_idxs, flat_src_indices=True)
 
     def configure_polynomial_controls(self, phase):
+        """
+        Configure the inputs/outputs for the polynomial controls.
+
+        Parameters
+        ----------
+        phase : dymos.Phase
+            The phase object to which this transcription instance applies.
+        """
         super(GaussLobatto, self).configure_polynomial_controls(phase)
         grid_data = self.grid_data
 
@@ -158,6 +182,14 @@ class GaussLobatto(PseudospectralBase):
                               src_indices=col_src_idxs, flat_src_indices=True)
 
     def setup_ode(self, phase):
+        """
+        Setup the ode for this transcription.
+
+        Parameters
+        ----------
+        phase : dymos.phase
+            The phase object to which this transcription instance applies.
+        """
         grid_data = self.grid_data
         ode_class = phase.options['ode_class']
 
@@ -177,6 +209,14 @@ class GaussLobatto(PseudospectralBase):
         phase.add_subsystem('interleave_comp', GaussLobattoInterleaveComp(grid_data=self.grid_data))
 
     def configure_ode(self, phase):
+        """
+        Create connections to the introspected states.
+
+        Parameters
+        ----------
+        phase : dymos.phase
+            The phase object to which this transcription instance applies.
+        """
         super(GaussLobatto, self).configure_ode(phase)
 
         map_input_indices_to_disc = self.grid_data.input_maps['state_input_to_disc']
@@ -235,7 +275,14 @@ class GaussLobatto(PseudospectralBase):
         self.configure_interleave_comp(phase)
 
     def configure_interleave_comp(self, phase):
+        """
+        Create connections to the interleave_comp.
 
+        Parameters
+        ----------
+        phase : dymos.phase
+            The phase object to which this transcription instance applies.
+        """
         map_input_indices_to_disc = self.grid_data.input_maps['state_input_to_disc']
 
         time_units = phase.time_options['units']
[…]
