+++ dymos/transcriptions/grid_data.py
@@ -26,7 +26,8 @@ def gauss_lobatto_subsets_and_nodes(n, seg_idx, compressed=False):
 
     Returns
     -------
-    subsets : A dictionary with the following keys:
+    dict
+        A dictionary with the following keys:
         'disc' gives the indices of the state discretization nodes (deprecated)
         'state_disc' gives the indices of the state discretization nodes
         'state_input' gives the indices of the state input nodes
@@ -34,7 +35,7 @@ def gauss_lobatto_subsets_and_nodes(n, seg_idx, compressed=False):
         'control_input' gives the indices of the control input nodes
         'segment_ends' gives the indices of the nodes at the start (even) and end (odd) of a segment
         'col' gives the indices of the collocation nodes
-        'all' gives all node indices
+        'all' gives all node indices.
 
     Notes
     -----
@@ -63,8 +64,7 @@ def gauss_lobatto_subsets_and_nodes(n, seg_idx, compressed=False):
 
 def radau_pseudospectral_subsets_and_nodes(n, seg_idx, compressed=False):
     """
-    Returns the subset dictionary corresponding to the Radau Pseudospectral
-    transcription.
+    Returns the subset dictionary corresponding to the Radau Pseudospectral transcription.
 
     Parameters
     ----------
@@ -77,7 +77,8 @@ def radau_pseudospectral_subsets_and_nodes(n, seg_idx, compressed=False):
 
     Returns
     -------
-    subsets : A dictionary with the following keys:
+    dict
+        A dictionary with the following keys:
         'disc' gives the indices of the state discretization nodes (deprecated)
         'state_disc' gives the indices of the state discretization nodes
         'state_input' gives the indices of the state input nodes
@@ -85,7 +86,7 @@ def radau_pseudospectral_subsets_and_nodes(n, seg_idx, compressed=False):
         'control_input' gives the indices of the control input nodes
         'segment_ends' gives the indices of the nodes at the start (even) and end (odd) of a segment
         'col' gives the indices of the collocation nodes
-        'all' gives all node indices
+        'all' gives all node indices.
 
     Notes
     -----
@@ -116,7 +117,7 @@ def rk_subsets_and_nodes(method, seg_idx, compressed=False):
     Parameters
     ----------
     method : str
-        One of the available RK methods in dymos.utils.rk_methods
+        One of the available RK methods in dymos.utils.rk_methods.
     seg_idx : int
         The index of this segment within its phase.
     compressed : bool
@@ -124,22 +125,21 @@ def rk_subsets_and_nodes(method, seg_idx, compressed=False):
 
     Returns
     -------
-    subsets : dict of {str: np.ndarray}
+    dict of {str: np.ndarray}
         'state_disc' gives the indices of the state discretization nodes
         'state_input' gives the indices of the state input nodes
         'control_disc' gives the indices of the control discretization nodes
         'control_input' gives the indices of the control input nodes
         'segment_ends' gives the indices of the nodes at the start (even) and end (odd) of a segment
         'step' gives the indices of the nodes at step boundaries
-        'all' gives all node indices
-    nodes : np.ndarray
-        The location of all nodes on the interval -1, 1.
+        'all' gives all node indices.
+    ndarray
+        The location of all nodes on the interval [-1, 1].
 
     Notes
     -----
     (subsets, nodes)
     Subset 'state_input' is the same as subset 'state_disc'.
-
     """
     # transform c onto [-1, 1]
     nodes = 2.0 * np.asarray(rk_methods[method]['c']) - 1.0
@@ -172,26 +172,25 @@ def explicit_subsets_and_nodes(n, seg_idx, compressed=False, shooting='single'):
     compressed : bool
         True if the subset requested is for a phase with compressed transcription.
     shooting : str
[…]
