+++ dymos/examples/racecar/tracks.py
@@ -1,0 +1,47 @@
+import numpy as np
+from Track import Track
+
+#Track descriptor information:
+#0 - straight, 1 - corner
+#0 - left, 1 - right
+#[type,length/sweep,radius,direction]
+
+segments = np.array([[0,150,0,-1],[1,np.pi/2,50,0],[0,100,0,-1],[1,np.pi/2,90,0],[0,300,0,-1],[1,np.pi/2,50,0],[0,100,0,-1],[1,np.pi/2,90,0],[0,155,0,-1]])
+ovaltrack = Track(segments)
+
+segments = np.array([[0,550,0,-1],[1,2.7,12,1],[0,300,0,-1],[1,0.2,160,1],[0,340,0,-1],[1,0.3,120,0],[0,40,0,-1],[1,0.8,120,1],[0,40,0,-1],[1,0.4,120,0],[0,260,0,-1],[1,0.4,160,1],[0,750,0,-1],[1,1.2,30,1],[0,30,0,-1],[1,1.1,40,0],[0,80,0,-1],[1,1.4,55,1],[0,310,0,-1],[1,3.14,55,1],[0,140,0,-1],[1,1.45,50,0],[0,440,0,-1],[1,1.4,70,0],[0,60,0,-1],[1,1.2,130,0],[0,340,0,-1],[1,1.45,66,1],[0,60,0,-1],[1,1.35,66,0],[0,200,0,-1],[1,1.57,50,1],[0,140,0,-1],[1,1.6,140,1],[0,200,0,-1],[1,0.7,150,1],[0,270,0,-1],[1,0.5,160,0],[0,220,0,-1],[1,1.1,140,0],[0,481.920,0,-1],[1,1.90,10,1],[0,23.9815,0,-1],[1,1.94,17,0]])
+Spa = Track(segments)
+
+segments = np.array([[0,927,0,-1],[1,1.48,40,1],[0,32,0,-1],[1,1.04,50,0],[0,87,0,-1],[1,1.10,104,1],[0,30,0,-1],[1,1.60,150,1],[0,263,0,-1],[1,1.48,44,1],[0,10,0,-1],[1,1.56,100,1],[0,177,0,-1],[1,2.54,42,0],[0,130,0,-1],[1,0.55,150,0],[0,115,0,-1],[1,1.48,52,0],[0,45,0,-1],[1,0.4,100,1],[0,181,0,-1],[1,1.72,85,1],[0,505,0,-1],[1,2.60,34,0],[0,100,0,-1],[1,0.8,30,0],[0,10,0,-1],[1,3.20,60,1],[0,135,0,-1],[1,1.40,30,1],[0,100,0,-1],[1,1.65,25,0],[0,10,0,-1],[1,1.50,33,1],[0,74.84,0,-1],[1,1.503,90,1],[0,142.0,0,-1]])
+Barcelona = Track(segments)
+
[…]
