### 2021-01-22 #2809 @robfalck
It's very odd, we've been using the github action to compile the paper and it's been fine (including rebuilding it a few times today).  Are there ever issues with whedon?
