+++ dymos/examples/balanced_field/ground_roll_eom_2d.py
@@ -1,0 +1,149 @@
+import numpy as np
+
+import openmdao.api as om
+
+
+class GroundRollEOM2D(om.ExplicitComponent):
+    """
+    Computes the acceleration of an aircraft on a runway, per Raymer Eq. 17.97 _[1]
+    Computes the position and velocity equations of motion using a modification of the 2D flight
+    path parameterization of states per equations 4.42 - 4.46 of _[1].  Flight path angle
+    and altitude are static quantities during the ground roll and are not integrated as states.
+
+    References
+    ----------
+    .. [1] Raymer, Daniel. Aircraft design: a conceptual approach. American Institute of
+    Aeronautics and Astronautics, Inc., 2012.
+    """
+    def initialize(self):
+        self.options.declare('num_nodes', types=int)
+        self.options.declare('g', types=(float, int), default=9.80665, desc='gravitational acceleration in m/s**2)')
+
+    def setup(self):
+        g = 9.80665
+        nn = self.options['num_nodes']
+
+        self.add_input(name='m',
+                       val=79016 * np.ones(nn),  # MTOW of 174200 lb
+                       desc='aircraft mass',
+                       units='kg')
+
+        self.add_input(name='v',
+                       val=np.ones(nn),
+                       desc='aircraft velocity magnitude',
+                       units='m/s')
+
+        self.add_input(name='T',
+                       val=120101.98 * np.ones(nn),  # 27000 lbf
+                       desc='thrust',
+                       units='N')
+
+        self.add_input(name='alpha',
+                       val=np.ones(nn),
+                       desc='angle of attack',
+                       units='rad')
+
+        self.add_input(name='L',
+                       val=np.ones(nn),
+                       desc='lift force',
+                       units='N')
+
+        self.add_input(name='D',
+                       val=np.ones(nn),
+                       desc='drag force',
+                       units='N')
+
+        self.add_input(name='mu_r',
+                       val=0.05 * np.ones(nn),
+                       desc='runway friction coefficient',
+                       units=None)
+
+        self.add_output(name='v_dot',
+                        val=np.ones(nn),
+                        desc='rate of change of velocity magnitude',
+                        units='m/s**2')
+
+        self.add_output(name='r_dot',
+                        val=np.ones(nn),
+                        desc='rate of change of range',
+                        units='m/s')
+
+        self.add_output(name='F_r',
+                        val=np.ones(nn),
+                        desc='runway normal force',
+                        units='N')
+
+        self.add_output(name='W',
+                        val=np.ones(nn),
+                        desc='aircraft weight',
+                        units='N')
+
+        ar = np.arange(nn)
+
+        self.declare_partials('v_dot', 'T', rows=ar, cols=ar)
+        self.declare_partials('v_dot', 'D', rows=ar, cols=ar)
+        self.declare_partials('v_dot', 'm', rows=ar, cols=ar)
+        self.declare_partials('v_dot', 'alpha', rows=ar, cols=ar)
+        self.declare_partials('v_dot', 'mu_r', rows=ar, cols=ar)
+
+        self.declare_partials('r_dot', 'v', rows=ar, cols=ar, val=1.0)
+
+        self.declare_partials('F_r', 'm', rows=ar, cols=ar, val=g)
+        self.declare_partials('F_r', 'L', rows=ar, cols=ar)
[…]

[…]
