### 2021-01-25 #2809 @goerz
I'll read through it by the end of the week. At first glance, it looks very much improved. I immediately saw some typos, though (like, 3rd paragraph, 1st sentence: missing "as"), so you might want to proof-read again.
