+++ dymos/phase/options.py
@@ -9,8 +9,12 @@ from ..utils.misc import _unspecified
 class ControlOptionsDictionary(om.OptionsDictionary):
     """
     An OptionsDictionary specific to controls.
-    """
 
+    Parameters
+    ----------
+    read_only : bool
+        If True, setting (via __setitem__ or update) is not permitted.
+    """
     def __init__(self, read_only=False):
         super(ControlOptionsDictionary, self).__init__(read_only)
 
@@ -118,8 +122,12 @@ class ControlOptionsDictionary(om.OptionsDictionary):
 class PolynomialControlOptionsDictionary(om.OptionsDictionary):
     """
     An OptionsDictionary specific to controls.
-    """
 
+    Parameters
+    ----------
+    read_only : bool
+        If True, setting (via __setitem__ or update) is not permitted.
+    """
     def __init__(self, read_only=False):
         super(PolynomialControlOptionsDictionary, self).__init__(read_only)
 
@@ -225,6 +233,11 @@ def check_valid_shape(name, value):
 class ParameterOptionsDictionary(om.OptionsDictionary):
     """
     An OptionsDictionary specific to parameters.
+
+    Parameters
+    ----------
+    read_only : bool
+        If True, setting (via __setitem__ or update) is not permitted.
     """
     def __init__(self, read_only=False):
         super(ParameterOptionsDictionary, self).__init__(read_only)
@@ -292,8 +305,12 @@ class ParameterOptionsDictionary(om.OptionsDictionary):
 class TrajParameterOptionsDictionary(ParameterOptionsDictionary):
     """
     An OptionsDictionary specific to trajectory design parameters.
-    """
 
+    Parameters
+    ----------
+    read_only : bool
+        If True, setting (via __setitem__ or update) is not permitted.
+    """
     def __init__(self, read_only=False):
         super(TrajParameterOptionsDictionary, self).__init__(read_only)
 
@@ -317,8 +334,12 @@ class TrajParameterOptionsDictionary(ParameterOptionsDictionary):
 class StateOptionsDictionary(om.OptionsDictionary):
     """
     An OptionsDictionary specific to states.
-    """
 
+    Parameters
+    ----------
+    read_only : bool
+        If True, setting (via __setitem__ or update) is not permitted.
+    """
     def __init__(self, read_only=False):
         super(StateOptionsDictionary, self).__init__(read_only)
 
@@ -444,8 +465,12 @@ class StateOptionsDictionary(om.OptionsDictionary):
 class TimeOptionsDictionary(om.OptionsDictionary):
     """
     An OptionsDictionary for time options.
-    """
 
+    Parameters
+    ----------
+    read_only : bool
+        If True, setting (via __setitem__ or update) is not permitted.
+    """
     def __init__(self, read_only=False):
         super(TimeOptionsDictionary, self).__init__(read_only)
 
@@ -525,8 +550,12 @@ class TimeOptionsDictionary(om.OptionsDictionary):
 class GridRefinementOptionsDictionary(om.OptionsDictionary):
     """
     An OptionsDictionary for grid refinement options in a Phase.
-    """
 
+    Parameters
+    ----------
+    read_only : bool
+        If True, setting (via __setitem__ or update) is not permitted.
+    """
     def __init__(self, read_only=False):
         super(GridRefinementOptionsDictionary, self).__init__(read_only)
 
