+++ dymos/transcriptions/runge_kutta/components/runge_kutta_state_predict_comp.py
@@ -10,8 +10,12 @@ from ....options import options as dymos_options
 class RungeKuttaStatePredictComp(om.ExplicitComponent):
     """
     Class definition for the RungeKuttaStatePredictComp.
-    """
 
+    Parameters
+    ----------
+    **kwargs : dict
+        Dictionary of optional arguments.
+    """
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self._no_check_partials = not dymos_options['include_check_partials']

[…]
