### 2021-01-25 ab6ac96
numpydocs 1.1 a test dependency
