### 2021-01-25 332707f
Added a test that checks all docstrings vs the NumpyDoc standard. (#526)
* clean up implementation of wildcard units when adding multiple timeseries at once.

* docstring linting is ready

* numpydocs 1.1 a test dependency

Co-authored-by: Kenneth Moore <KenMooreZ1@gmail.com>
