+++ dymos/phase/phase.py
@@ -41,15 +41,10 @@ class Phase(om.Group):
 
     Parameters
     ----------
-    from_phase: Phase or None
+    from_phase : <Phase> or None
         A phase instance from which the initialized phase should copy its data.
-    transcription: TranscriptionBase
-        The transcription to be utilized within the Phase.
-    ode_class
-        An OpenMDAO system class serving as the ODE for the phase.
-    ode_init_kwargs: dict
-        Keyword arguments used to initialize ode_class.
-
+    **kwargs : dict
+        Dictionary of optional phase arguments.
     """
 
     def __init__(self, from_phase=None, **kwargs):
@@ -96,6 +91,9 @@ class Phase(om.Group):
         super(Phase, self).__init__(**_kwargs)
 
     def initialize(self):
+        """
+        Declare instantiation options for the phase.
+        """
         self.options.declare('ode_class', default=None,
                              desc='System defining the ODE')
         self.options.declare('ode_init_kwargs', types=dict, default={},
@@ -154,7 +152,7 @@ class Phase(om.Group):
         ref0 : float or ndarray or None (None)
             The zero-reference value of the state at the nodes of the phase.
         ref : float or ndarray or None (None)
-            The unit-reference value of the state at the nodes of the phase
+            The unit-reference value of the state at the nodes of the phase.
         defect_scaler : float or ndarray
             The scaler of the state defect at the collocation nodes of the phase.
         defect_ref : float or ndarray
@@ -228,7 +226,7 @@ class Phase(om.Group):
         ref0 : float or ndarray or None (None)
             The zero-reference value of the state at the nodes of the phase.
         ref : float or ndarray or None (None)
-            The unit-reference value of the state at the nodes of the phase
+            The unit-reference value of the state at the nodes of the phase.
         defect_scaler : float or ndarray
             The scaler of the state defect at the collocation nodes of the phase.
         defect_ref : float or ndarray
@@ -420,7 +418,6 @@ class Phase(om.Group):
         Notes
         -----
         rate and rate2 continuity are not enforced for input controls.
-
         """
         if name not in self.control_options:
             self.check_parameter(name)
@@ -518,7 +515,6 @@ class Phase(om.Group):
         Notes
         -----
         rate and rate2 continuity are not enforced for input controls.
-
         """
         if units is not _unspecified:
             self.control_options[name]['units'] = units
@@ -617,18 +613,18 @@ class Phase(om.Group):
         order : int
             The order of the interpolating polynomial used to represent the control valeu in
             phase tau space.
+        desc : str
+            A description of the polynomial control.
         val : float or ndarray
             Default value of the control at all nodes.  If val scalar and the control
             is dynamic it will be broadcast.
-        desc : str
-            A description of the polynomial control.
         units : str or None or 0
             Units in which the control variable is defined.  If 0, use the units declared
             for the parameter in the ODE.
         opt : bool
             If True (default) the value(s) of this control will be design variables in
             the optimization problem, in the path 'phase_name.indep_controls.controls:control_name'.
-            If False, the values of this control will exist as input controls:{name}
+            If False, the values of this control will exist as input controls:{name}.
         fix_initial : bool
             If True, the given initial value of the polynomial control is not a design variable and
             will not be changed during the optimization.
@@ -646,7 +642,7 @@ class Phase(om.Group):
         ref0 : float or ndarray
             The zero-reference value of the control at the nodes of the phase.
         ref : float or ndarray
-            The unit-reference value of the control at the nodes of the phase
+            The unit-reference value of the control at the nodes of the phase.
         targets : Sequence of str or None
             Targets in the ODE to which this polynomial control is connected.
         rate_targets : None or str
[…]
