### 2021-01-26 3a11fd1
The phase instantiation can now take ode_class as a callable that returns an OpenMDAO system, where the callable has the arguments (num_nodes=<int>, **ode_init_kwargs). Added some documentation. Added tests for invalid callable ODEs with appropriate error messages. Added test for copying an existing ODE instance via deepcopy, though this is discouraged in the documentation.
