### 2021-01-26 a78d7e4
added test to use setup ode as the ode class for a subsequent phase
