### 2021-02-01 #2809 @goerz
I've re-read the manuscript in its current, updated version. I'm happy to see that the authors substantially revised the text, resulting in a much more accessible publication. I have no fundamental objections to seeing the manuscript published in its current form.

However, the writing is sloppy, and further editing will be required before publication. A thorough proof-reading should have caught most of these issues. I expect authors to proof-read their submissions before sending them out for review!

A few of the issues:

* Line 29: extra parenthesis after the vector u. Actually, almost all the parentheses around symbols in the text are superfluous; simply write "... affected by time t, a vector of dynamic controls u, ...". Leave out "(J)" before the equation for J. Put a period after the equation (equations should have punctuation according to the text they are embedded in).
* Line 34: "etc" together with "such as" is redundant
* Line 41: "fits" should be "fit"
* Line 47: "going first performing" ...?
* Fig 2/Fig3: do the small empty boxes at the edges of any of the arrows signify anything? These should probably be left out.
* Line 78 and the rest of the manuscript: Isn't "time-history" redundant? Maybe just "history" or "time-evolution"?
* Line 81: "numerical" should be "numerically"
* Line 83: I don't think "finite-difference" is a verb
* Paragraph starting in line 115: This is sloppy writing. Rephrase! If RK is only used "in some cases", then it's not really "widely used", right?
* Line 143: "the" should be "they"
* Line 156: "shipt" should be "ship" (how did your spellchecker not catch this?)
* Line 159: the word "library" should be before the citation.
* Line 165: Addressing the reader as "you" is stylistically inconsistent with the more imporsonal / passive voice style of the rest of the paper
* Line 214: "effect" should be "effects"

I probably missed some additional typos and/or grammatical errors. I'm not going to read through this again, so please make sure to fix them.
