### 2021-02-05 #2809 @robfalck
These have been addressed (in addition to a few others).
