### 2021-02-23 #2809 @kthyng
(@dpsanders not trying to step on any toes but trying to help handling editors out by taking over when it looks possible)
