### 2021-03-29 #2809 @dpsanders
@JustinSGray: Various references are missing the journal / conference, e.g. Aretskin, Falck. Please
