### 2021-03-30 #2809 @JustinSGray
@dpsanders thanks for your patience.

I found the problem. A number of the references had the wrong bib type and that confused things. All DOIs added, all references have the journal name or conference name now.

We've released version 1.0.0 of Dymos to coincide with this publication.: https://github.com/OpenMDAO/dymos/archive/refs/tags/1.0.0.zip

The tagged 1.0.0 release obviously doesn't have these very latest updates to the paper, but it has all the actual software code.
If necessary, we could make another incremental release with the this version of the paper. Please let us know if that is needed.
