### 2021-03-30 #2809 @dpsanders
Thanks @JustinSGray, it's all looking good, thanks.
Just one nit:

- Herman et al: `{Gauss-Lobatto}` (to get the capitalisation right)

I don't think it's necessary to release a new version.

Please let me know the archive address at zenodo.org or similar. Thanks!
