### 2021-03-30 #2809 @JustinSGray
Ok, the capitalization issue is fixed.

Zenodo upload:
https://zenodo.org/record/4646412#.YGJ_XUhKjUI

Here is the doi for version 1.0.0: 10.5281/zenodo.4646412
